@@ -1,7 +1,7 @@
-module github.com/uapl/ebay-buy-browse-go
+module github.com/GIT_USER_ID/GIT_REPO_ID
 
-go 1.13
+go 1.18
 
 require (
-	golang.org/x/oauth2 v0.0.0-20210218202405-ba52d332ba99
+	golang.org/x/oauth2 v0.0.0-20210323180902-22b0adad7558
 )
