@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -12,83 +12,94 @@ package buybrowse
 
 import (
 	"bytes"
-	_context "context"
-	_ioutil "io/ioutil"
-	_nethttp "net/http"
-	_neturl "net/url"
+	"context"
+	"io"
+	"net/http"
+	"net/url"
 	"strings"
 )
 
-// Linger please
-var (
-	_ _context.Context
-)
 
-// ItemApiService ItemApi service
-type ItemApiService service
+// ItemAPIService ItemAPI service
+type ItemAPIService service
 
 type ApiCheckCompatibilityRequest struct {
-	ctx                  _context.Context
-	ApiService           *ItemApiService
-	itemId               string
-	xEBAYCMARKETPLACEID  *string
+	ctx context.Context
+	ApiService *ItemAPIService
+	itemId string
+	contentType *string
+	xEBAYCMARKETPLACEID *string
+	acceptLanguage *string
 	compatibilityPayload *CompatibilityPayload
 }
 
-// The ID of the eBay marketplace you want to use. &lt;b&gt; Note: &lt;/b&gt; This value is case sensitive.&lt;br /&gt;&lt;br /&gt;For example: &lt;br /&gt;&amp;nbsp;&amp;nbsp;&lt;code&gt;X-EBAY-C-MARKETPLACE-ID &#x3D; EBAY_US&lt;/code&gt;  &lt;br /&gt;&lt;br /&gt; For a list of supported sites see, &lt;a href&#x3D;\&quot;/api-docs/buy/browse/overview.html#API\&quot;&gt;API Restrictions&lt;/a&gt;.
-func (r *ApiCheckCompatibilityRequest) XEBAYCMARKETPLACEID(xEBAYCMARKETPLACEID string) *ApiCheckCompatibilityRequest {
+// This header indicates the format of the request body provided by the client.&lt;br&gt;&lt;br&gt;Its value should be set to &lt;code&gt;application/json&lt;/code&gt;.&lt;br&gt;&lt;br&gt;For more information, refer to &lt;a href&#x3D;\&quot;/api-docs/static/rest-request-components.html#HTTP\&quot; target&#x3D;\&quot;_blank \&quot;&gt;HTTP request headers&lt;/a&gt; in the &lt;a href&#x3D;\&quot;/api-docs/static/ebay-rest-landing.html\&quot; target&#x3D;\&quot;_blank\&quot;&gt;Using eBay RESTful APIs&lt;/a&gt; guide.
+func (r ApiCheckCompatibilityRequest) ContentType(contentType string) ApiCheckCompatibilityRequest {
+	r.contentType = &contentType
+	return r
+}
+
+// This header identifies the seller&#39;s eBay marketplace. It is required for all marketplaces outside of the US.&lt;br&gt;&lt;br&gt;&lt;span class&#x3D;\&quot;tablenote\&quot;&gt;&lt;b&gt;Note:&lt;/b&gt; If a marketplace ID value is not provided, the default value of &lt;code&gt;EBAY_US&lt;/code&gt; is used.&lt;/span&gt;&lt;br&gt;See &lt;a href&#x3D;\&quot;/api-docs/buy/browse/types/ba:MarketplaceIdEnum\&quot; target&#x3D;\&quot;_blank\&quot;&gt;MarketplaceIdEnum&lt;/a&gt; for supported values.
+func (r ApiCheckCompatibilityRequest) XEBAYCMARKETPLACEID(xEBAYCMARKETPLACEID string) ApiCheckCompatibilityRequest {
 	r.xEBAYCMARKETPLACEID = &xEBAYCMARKETPLACEID
 	return r
 }
-func (r *ApiCheckCompatibilityRequest) CompatibilityPayload(compatibilityPayload CompatibilityPayload) *ApiCheckCompatibilityRequest {
+
+// This header is used to indicate the natural language and locale preferred by the user for the response.&lt;br&gt;&lt;br&gt;This header is required when targeting a specific locale of a marketplace that supports multiple locales. For example:&lt;ul&gt;&lt;li&gt;When targeting the French locale of the Belgium marketplace, it is required to pass in &lt;code&gt;fr-BE&lt;/code&gt; to specify this. If this locale is not specified, the language will default to Dutch.&lt;/li&gt;&lt;li&gt;When targeting the French locale of the Canadian marketplace, it is required to pass in &lt;code&gt;fr-CA&lt;/code&gt; to specify this. If this locale is not specified, the language will default to English.&lt;/li&gt;&lt;/ul&gt;
+func (r ApiCheckCompatibilityRequest) AcceptLanguage(acceptLanguage string) ApiCheckCompatibilityRequest {
+	r.acceptLanguage = &acceptLanguage
+	return r
+}
+
+func (r ApiCheckCompatibilityRequest) CompatibilityPayload(compatibilityPayload CompatibilityPayload) ApiCheckCompatibilityRequest {
 	r.compatibilityPayload = &compatibilityPayload
 	return r
 }
 
-func (r *ApiCheckCompatibilityRequest) Execute() (CompatibilityResponse, *_nethttp.Response, error) {
+func (r ApiCheckCompatibilityRequest) Execute() (*CompatibilityResponse, *http.Response, error) {
 	return r.ApiService.CheckCompatibilityExecute(r)
 }
 
 /*
 CheckCompatibility Method for CheckCompatibility
 
-This method checks if a product is compatible with the specified item. You can use this method to check the compatibility of cars, trucks, and motorcycles with a specific part listed on eBay. <br /><br />For example, to check the compatibility of a part, you pass in the item ID of the part as a URI parameter and specify all the attributes used to define a specific car in the <b> compatibilityProperties</b> container. If the call is successful, the response will be <b> COMPATIBLE</b>, <b> NOT_COMPATIBLE</b>, or <b> UNDETERMINED</b>. See <a href="/api-docs/buy/browse/resources/item/methods/checkCompatibility#response.compatibilityStatus">compatibilityStatus</a> for details.   <br /><br /> <span class="tablenote"><b> Note: </b> The only products supported are cars, trucks, and motorcycles. </span><p>  To find the attributes and values for a specific marketplace, you can use the compatibility methods in the <a href="/api-docs/commerce/taxonomy/resources/methods">Taxonomy API</a>. You can use this data to create menus to help buyers specify the product, such as their car.</p> <p> For more details and a list of the required attributes for the US marketplace that describe motor vehicles, see <a href="/api-docs/buy/static/api-browse.html#Check">Check compatibility</a> in the Buy Integration Guide</a>.</p>   <p>For an example, see the <a href="/api-docs/buy/browse/resources/item/methods/checkCompatibility#h2-samples">Samples</a> section. </p>    <h3>URLs for this method</h3>  <p><b> Production URL: </b> <code>https://api.ebay.com/buy/browse/v1/item/{item_id}/check_compatibility</code> </p>         <p><span class="tablenote"><b> Note: </b> This method is supported only on Production. </span></p>         <h3><b> Restrictions </b></h3> <p>For a list of supported sites and other restrictions, see <a href="/api-docs/buy/browse/overview.html#API">API Restrictions</a>.</p>
+This method checks if a product is compatible with the specified item. You can use this method to check the compatibility of cars, trucks, and motorcycles with a specific part listed on eBay.<br><br>For example, to check the compatibility of a part, you pass in the <b>item_id</b> of the part as a URI parameter and specify all the attributes used to define a specific car within the <b>compatibilityProperties</b> container. If the call is successful, the response will be <code>COMPATIBLE</code>, <code>NOT_COMPATIBLE</code>, or <code>UNDETERMINED</code>. Refer to <a href="/api-docs/buy/browse/resources/item/methods/checkCompatibility#response.compatibilityStatus">compatibilityStatus</a> for details.   <br><br><span class="tablenote"><b>Note:</b> The only products supported are cars, trucks, and motorcycles.</span><br>To find the attributes and values for a specific marketplace, you can use the compatibility methods in the <a href="/api-docs/commerce/taxonomy/resources/methods" target="_blank">Taxonomy API</a>. You can use this data to create menus to help buyers specify the product, such as their car.<br><br>For more information and a list of required attributes for the US marketplace that describe motor vehicles, refer to <a href="/api-docs/buy/static/api-browse.html#Check" target="_blank">Check compatibility</a> in the <i>Buying Integration Guide</i>.<br><br>For an example, refer to the <a href="/api-docs/buy/browse/resources/item/methods/checkCompatibility#h2-samples">Samples</a> section.<br><br><span class="tablenote"><b>Note:</b> This method is supported in Sandbox but <i>only</i> when passing in the specified <code>item_id</code> and compatibility name-value pairs listed in <a href="/api-docs/buy/browse/resources/item/methods/checkCompatibility#s0-1-22-6-7-7-6-SandboxSample-1">Sample 2: Sandbox Sample</a>.</span><h3>Restrictions</h3>For a list of supported sites and other restrictions, refer to <a href="/api-docs/buy/browse/overview.html#API">API Restrictions</a>.
 
- @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- @param itemId The eBay RESTful identifier of an item (such as a part you want to check). This ID is returned by the <b> Browse</b> and <b> Feed</b> API methods.  <br /><br /> <b> RESTful Item ID Format: </b><code>v1</code>|<code><i>#</i></code>|<code><i>#</i></code> <br />For example: <code>v1|2**********2|0</code> or <code>v1|1**********2|4**********2</code> <br /><br />For more information about item ID for RESTful APIs, see the <a href=\"/api-docs/buy/static/api-browse.html#Legacy\">Legacy API compatibility</a> section of the <i>Buy APIs Overview</i>.
+ @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+ @param itemId This path parameter specifies the unique RESTful identifier of an item (such as the park you want to check).<br><br><b>RESTful Item ID Format: </b><code>v1</code>|<code><i>#</i></code>|<code><i>#</i></code><br><br>For a single SKU listing, pass in the item ID: <pre>v1|2**********2|0</pre>For a multi-SKU listing, pass in the identifier of the variation:<pre>v1|1**********2|4**********2</pre><br>For more information about item IDs for RESTful APIs, refer to <a href=\"/api-docs/buy/static/api-browse.html#Legacy\" target=\"_blank\">Item ID legacy API compatibility overview</a> in the <a href=\"/api-docs/buy/static/buying-ig-landing.html\" target=\"_blank\">Buying Integration Guide</a>.
  @return ApiCheckCompatibilityRequest
 */
-func (a *ItemApiService) CheckCompatibility(ctx _context.Context, itemId string) ApiCheckCompatibilityRequest {
+func (a *ItemAPIService) CheckCompatibility(ctx context.Context, itemId string) ApiCheckCompatibilityRequest {
 	return ApiCheckCompatibilityRequest{
 		ApiService: a,
-		ctx:        ctx,
-		itemId:     itemId,
+		ctx: ctx,
+		itemId: itemId,
 	}
 }
 
 // Execute executes the request
 //  @return CompatibilityResponse
-func (a *ItemApiService) CheckCompatibilityExecute(r *ApiCheckCompatibilityRequest) (CompatibilityResponse, *_nethttp.Response, error) {
+func (a *ItemAPIService) CheckCompatibilityExecute(r ApiCheckCompatibilityRequest) (*CompatibilityResponse, *http.Response, error) {
 	var (
-		localVarHTTPMethod  = _nethttp.MethodPost
-		localVarPostBody    interface{}
-		formFiles           []formFile
-		localVarReturnValue CompatibilityResponse
+		localVarHTTPMethod   = http.MethodPost
+		localVarPostBody     interface{}
+		formFiles            []formFile
+		localVarReturnValue  *CompatibilityResponse
 	)
 
-	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "ItemApiService.CheckCompatibility")
+	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "ItemAPIService.CheckCompatibility")
 	if err != nil {
-		return localVarReturnValue, nil, GenericOpenAPIError{error: err.Error()}
+		return localVarReturnValue, nil, &GenericOpenAPIError{error: err.Error()}
 	}
 
 	localVarPath := localBasePath + "/item/{item_id}/check_compatibility"
-	localVarPath = strings.Replace(localVarPath, "{"+"item_id"+"}", _neturl.PathEscape(parameterToString(r.itemId, "")), -1)
+	localVarPath = strings.Replace(localVarPath, "{"+"item_id"+"}", url.PathEscape(parameterValueToString(r.itemId, "itemId")), -1)
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
-	if r.xEBAYCMARKETPLACEID == nil {
-		return localVarReturnValue, nil, reportError("xEBAYCMARKETPLACEID is required and must be specified")
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
+	if r.contentType == nil {
+		return localVarReturnValue, nil, reportError("contentType is required and must be specified")
 	}
 
 	// to determine the Content-Type header
@@ -108,7 +119,13 @@ func (a *ItemApiService) CheckCompatibilityExecute(r *ApiCheckCompatibilityReque
 	if localVarHTTPHeaderAccept != "" {
 		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
 	}
-	localVarHeaderParams["X-EBAY-C-MARKETPLACE-ID"] = parameterToString(*r.xEBAYCMARKETPLACEID, "")
+	if r.xEBAYCMARKETPLACEID != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "X-EBAY-C-MARKETPLACE-ID", r.xEBAYCMARKETPLACEID, "simple", "")
+	}
+	parameterAddToHeaderOrQuery(localVarHeaderParams, "Content-Type", r.contentType, "simple", "")
+	if r.acceptLanguage != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "Accept-Language", r.acceptLanguage, "simple", "")
+	}
 	// body params
 	localVarPostBody = r.compatibilityPayload
 	req, err := a.client.prepareRequest(r.ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, formFiles)
@@ -121,15 +138,15 @@ func (a *ItemApiService) CheckCompatibilityExecute(r *ApiCheckCompatibilityReque
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
-	localVarHTTPResponse.Body = _ioutil.NopCloser(bytes.NewBuffer(localVarBody))
+	localVarHTTPResponse.Body = io.NopCloser(bytes.NewBuffer(localVarBody))
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
 	if localVarHTTPResponse.StatusCode >= 300 {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: localVarHTTPResponse.Status,
 		}
@@ -138,7 +155,7 @@ func (a *ItemApiService) CheckCompatibilityExecute(r *ApiCheckCompatibilityReque
 
 	err = a.client.decode(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
 	if err != nil {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: err.Error(),
 		}
@@ -149,63 +166,84 @@ func (a *ItemApiService) CheckCompatibilityExecute(r *ApiCheckCompatibilityReque
 }
 
 type ApiGetItemRequest struct {
-	ctx         _context.Context
-	ApiService  *ItemApiService
-	itemId      string
+	ctx context.Context
+	ApiService *ItemAPIService
+	itemId string
 	fieldgroups *string
+	xEBAYCENDUSERCTX *string
+	xEBAYCMARKETPLACEID *string
+	acceptLanguage *string
 }
 
-// This parameter lets you control what is returned in the response. If you do not set this field, the method returns all the details of the item.   &lt;br /&gt; &lt;br /&gt; &lt;b&gt; Valid Values: &lt;/b&gt; &lt;ul&gt;  &lt;li&gt; &lt;b&gt; PRODUCT&lt;/b&gt; - This adds the &lt;code&gt;additionalImages&lt;/code&gt;, &lt;code&gt;additionalProductIdentities&lt;/code&gt;, &lt;code&gt;aspectGroups&lt;/code&gt;, &lt;code&gt;description&lt;/code&gt;, &lt;code&gt;gtins&lt;/code&gt;, &lt;code&gt;image&lt;/code&gt;, and &lt;code&gt;title&lt;/code&gt; product fields to the response, which describe the product associated with the item. See &lt;a href&#x3D;\&quot;/api-docs/buy/browse/resources/item/methods/getItem#response.product\&quot;&gt;Product&lt;/a&gt; for more information about these fields.&lt;/li&gt;          &lt;li&gt;&lt;b&gt; COMPACT&lt;/b&gt; -  This returns only the following fields, which let you quickly check if the availability or price of the item has changed, if the item has been revised by the seller, or if an item&#39;s top-rated plus status has changed for items you have stored.  &lt;ul&gt; &lt;li&gt; &lt;b&gt; itemId&lt;/b&gt; - The identifier of the item.&lt;/li&gt; &lt;li&gt;&lt;b&gt;bidCount&lt;/b&gt; - This integer value indicates the total number of bids that have been placed against an auction item.&lt;/li&gt; &lt;li&gt;&lt;b&gt;currentBidPrice&lt;/b&gt; - This container shows the current highest bid for an auction item. This container will only be returned for auction items.&lt;/li&gt;  &lt;li&gt;&lt;b&gt;eligibleForInlineCheckout&lt;/b&gt; - This parameter returns items based on whether or not the items can be purchased using the Buy &lt;a href&#x3D;\&quot;/api-docs/buy/order/resources/methods\&quot;&gt;Order API&lt;/a&gt;. &lt;ul&gt; &lt;li&gt;If the value of this field is &lt;code&gt;true&lt;/code&gt;, this indicates that the item can be purchased using the &lt;b&gt; Order API&lt;/b&gt;. &lt;/li&gt;  &lt;li&gt;If the value of this field is &lt;code&gt;false&lt;/code&gt;, this indicates that the item cannot be purchased using the &lt;b&gt; Order API&lt;/b&gt; and must be purchased on the eBay site.&lt;/li&gt; &lt;/ul&gt; &lt;li&gt;&lt;b&gt; estimatedAvailabilities&lt;/b&gt; -  Returns the item availability information, which is based on the item&#39;s quantity. &lt;b&gt; Note:&lt;/b&gt; Changes in quantity are not tracked by &lt;b&gt;sellerItemRevision&lt;/b&gt;.&lt;/li&gt; &lt;li&gt;&lt;b&gt;itemAffiliateWebURL&lt;/b&gt; - The URL of the View Item page of the item, which includes the affiliate tracking ID. This field is only returned if the eBay partner enables affiliate tracking for the item by including the &lt;code&gt;X-EBAY-C-ENDUSERCTX&lt;/code&gt; request header in the method.&lt;/li&gt; &lt;li&gt;&lt;b&gt;itemEndDate&lt;/b&gt; - This is the scheduled end time of the listing.&lt;/li&gt; &lt;li&gt;&lt;b&gt;ItemWebURL&lt;/b&gt; - The URL of the View Item page of the item. This enables you to include a \&quot;Report Item on eBay\&quot; link that takes the buyer to the View Item page on eBay. From there they can report any issues regarding this item to eBay.&lt;/li&gt; &lt;li&gt;&lt;b&gt;legacyItemId&lt;/b&gt; - The unique identifier of the eBay listing that contains the item. This is the traditional/legacy ID that is often seen in the URL of the listing View Item page.&lt;/li&gt; &lt;li&gt;&lt;b&gt;minimumPriceToBid&lt;/b&gt; - This container shows the minimum bid amount that would be accepted as a qualifying bid in an auction listing. This container will only be returned for auction items.&lt;/li&gt; &lt;li&gt;&lt;b&gt;price&lt;/b&gt; - This is tracked by the revision ID but is returned here to enable you to quickly verify the price of the item.&lt;/li&gt; &lt;li&gt;&lt;b&gt;priorityListing&lt;/b&gt; - This field is returned as &lt;code&gt;true&lt;/code&gt; if the listing is part of a Promoted Listing campaign. Promoted Listings are available to Above Standard and Top Rated sellers with recent sales activity.&lt;/li&gt; &lt;li&gt;&lt;b&gt;reservePriceMet&lt;/b&gt; - This field indicates whether or not an auction&#39;s reserve price (if set by the seller) has been met yet. This field will only be returned for auction items.&lt;/li&gt; &lt;li&gt;&lt;b&gt; sellerItemRevision&lt;/b&gt; - An identifier generated/incremented when a seller revises the item. The following are the two types of item revisions:   &lt;ul&gt;  &lt;li&gt;&lt;b&gt; Seller changes&lt;/b&gt;, such as changing the title&lt;/li&gt;  &lt;li&gt;  &lt;b&gt; eBay system changes&lt;/b&gt;, such as changing the quantity when an item is purchased.&lt;/li&gt;  &lt;/ul&gt; This ID is changed &lt;em&gt;only&lt;/em&gt; when the seller makes a change to the item. This means you cannot use this value to determine if the quantity has changed. To check if the quantity has changed, use &lt;b&gt; estimatedAvailabilities.&lt;/b&gt;&lt;/li&gt; &lt;li&gt;&lt;b&gt;taxes&lt;/b&gt; - A container for the tax information for the item, such as the tax jurisdiction, the tax percentage, and the tax type.&lt;/li&gt; &lt;li&gt;&lt;b&gt; topRatedBuyingExperience&lt;/b&gt; - A boolean value indicating if this item is a top-rated plus item. A change in the item&#39;s top rated plus standing is not tracked by the revision ID. See &lt;a href&#x3D;\&quot;/api-docs/buy/browse/resources/item/methods/getItem#response.topRatedBuyingExperience\&quot;&gt;topRatedBuyingExperience&lt;/a&gt; for more information.&lt;/li&gt; &lt;li&gt;&lt;b&gt;uniqueBidderCount&lt;/b&gt; - This integer value indicates the number of different eBay users who have placed one or more bids on an auction item. This field is only applicable to auction items.&lt;/li&gt;&lt;/ul&gt;    &lt;b&gt; For Example&lt;/b&gt; &lt;br /&gt; &lt;br /&gt;To check if a stored item&#39;s information is current, do following.  &lt;ol&gt;    &lt;li&gt;Pass in the item ID and set &lt;b&gt; fieldgroups&lt;/b&gt; to COMPACT. &lt;br /&gt; &lt;br /&gt;&lt;code&gt;item/v1|4**********8|0?fieldgroups&#x3D;COMPACT&lt;/code&gt; &lt;/li&gt;     &lt;li&gt;Do one of the following:    &lt;ul&gt;     &lt;li&gt;If the &lt;b&gt; sellerItemRevision&lt;/b&gt; field is returned and you &lt;em&gt;haven&#39;t&lt;/em&gt; stored a revision number for this item, record the number and pass in the item ID in the &lt;b&gt; getItem&lt;/b&gt; method to get the latest information.&lt;/li&gt;   &lt;li&gt;If the revision number is different from the value you have stored, update the value and pass in the item ID in the &lt;b&gt; getItem&lt;/b&gt; method to get the latest information.&lt;/li&gt;     &lt;li&gt;If the &lt;b&gt; sellerItemRevision&lt;/b&gt; field is &lt;em&gt;not&lt;/em&gt; returned or has not changed, where needed, update the item information with the information returned in the response.&lt;/li&gt;  &lt;/ul&gt;  &lt;/li&gt; &lt;/ol&gt;&lt;/li&gt; &lt;/ul&gt;  &lt;/ul&gt;    &lt;p&gt;&lt;b&gt; Maximum value: &lt;/b&gt; 1 &lt;br /&gt;If more than one values is specified, the first value will be used.
-func (r *ApiGetItemRequest) Fieldgroups(fieldgroups string) *ApiGetItemRequest {
+// This parameter controls what is returned in the response. If this field is not set, the method returns all the details of the item.&lt;br&gt;&lt;br&gt;&lt;span class&#x3D;\&quot;tablenote\&quot;&gt;&lt;b&gt;Note:&lt;/b&gt; Multiple &lt;b&gt;fieldgroups&lt;/b&gt; can be set and applied simultaneously. However, &lt;b&gt;COMPACT&lt;/b&gt; &lt;b&gt;must&lt;/b&gt; be used alone. Otherwise, an error will occur.&lt;/span&gt;&lt;br&gt;&lt;b&gt;Valid Values:&lt;/b&gt;&lt;ul&gt;&lt;li&gt;&lt;b&gt;PRODUCT&lt;/b&gt;&lt;br&gt;This field group adds the &lt;a href&#x3D;\&quot;/api-docs/buy/browse/resources/item/methods/getItem#response.product\&quot;&gt;product&lt;/a&gt; container to the response.&lt;/li&gt;&lt;li&gt;&lt;b&gt;COMPACT&lt;/b&gt;&lt;/br&gt;This field group returns only the following fields which let you quickly check if the availability or price of the item has changed, if the item has been revised by the seller, or if an item&#39;s top-rated plus status has changed for items you have stored.&lt;ul&gt;&lt;li&gt;&lt;b&gt;itemId&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;bidCount&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;currentBidPrice&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;eligibleForInlineCheckout&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;estimatedAvailabilities&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;gtin&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;itemAffiliateWebURL&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;itemCreationDate&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;ItemWebURL&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;legacyItemId&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;minimumPriceToBid&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;price&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;priorityListing&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;reservePriceMet&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;sellerItemRevision&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;shippingOptions&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;taxes&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;topRatedBuyingExperience&lt;/b&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;uniqueBidderCount&lt;/b&gt;&lt;/li&gt;&lt;/ul&gt;For Example:&lt;br&gt;To determine if a stored item&#39;s information is current, perform the following:&lt;ol&gt;&lt;li&gt;Pass in the item ID and set &lt;code&gt;fieldgroups&lt;/code&gt; to &lt;code&gt;COMPACT&lt;/code&gt;.&lt;pre&gt;item/v1|4**********8|0?fieldgroups&#x3D;COMPACT&lt;/pre&gt;&lt;/li&gt;&lt;li&gt;Do one of the following:&lt;ul&gt;&lt;li&gt;If &lt;b&gt;sellerItemRevision&lt;/b&gt; is returned and a revision number for this item &lt;i&gt;has not&lt;/i&gt; been stored, record the number and pass in the item ID in the &lt;b&gt;getItem&lt;/b&gt; method to retrieve the most recent information.&lt;/li&gt;&lt;li&gt;If the revision number is different from the value you have stored, update the value and pass in the item ID in the &lt;b&gt;getItem&lt;/b&gt; method to retrieve the most recent information.&lt;/li&gt;&lt;li&gt;If &lt;code&gt;sellerItemRevision&lt;/code&gt; is &lt;i&gt;not&lt;/i&gt; returned or has not changed, where needed, update the item information with the information returned in the response.&lt;/li&gt;&lt;/ul&gt;&lt;/li&gt;&lt;/ol&gt;&lt;/li&gt;&lt;li&gt;&lt;b&gt;ADDITIONAL_SELLER_DETAILS&lt;/b&gt;&lt;br&gt;This adds the &lt;a href&#x3D;\&quot;/api-docs/buy/browse/resources/item/methods/getItem#response.seller.userId\&quot;&gt;userId&lt;/a&gt; field to the response.&lt;/li&gt;&lt;/ul&gt;
+func (r ApiGetItemRequest) Fieldgroups(fieldgroups string) ApiGetItemRequest {
 	r.fieldgroups = &fieldgroups
 	return r
 }
 
-func (r *ApiGetItemRequest) Execute() (Item, *_nethttp.Response, error) {
+// This header is required to support revenue sharing for eBay Partner Network and to improve the accuracy of shipping and delivery time estimations.&lt;br&gt;&lt;br&gt;For additional information, refer to &lt;a href&#x3D;\&quot;/api-docs/buy/static/api-browse.html#Headers\&quot; target&#x3D;\&quot;_blank \&quot;&gt;Use request headers&lt;/a&gt; in the &lt;a href&#x3D;\&quot;/api-docs/buy/static/buying-ig-landing.html\&quot; target&#x3D;\&quot;_blank\&quot;&gt;Buying Integration Guide&lt;/a&gt;.
+func (r ApiGetItemRequest) XEBAYCENDUSERCTX(xEBAYCENDUSERCTX string) ApiGetItemRequest {
+	r.xEBAYCENDUSERCTX = &xEBAYCENDUSERCTX
+	return r
+}
+
+// This header identifies the seller&#39;s eBay marketplace. It is required for all marketplaces outside of the US.&lt;br&gt;&lt;br&gt;&lt;span class&#x3D;\&quot;tablenote\&quot;&gt;&lt;b&gt;Note:&lt;/b&gt; If a marketplace ID value is not provided, the default value of &lt;code&gt;EBAY_US&lt;/code&gt; is used.&lt;/span&gt;&lt;br&gt;See &lt;a href&#x3D;\&quot;/api-docs/buy/browse/types/ba:MarketplaceIdEnum\&quot; target&#x3D;\&quot;_blank\&quot;&gt;MarketplaceIdEnum&lt;/a&gt; for a list of supported marketplaces.
+func (r ApiGetItemRequest) XEBAYCMARKETPLACEID(xEBAYCMARKETPLACEID string) ApiGetItemRequest {
+	r.xEBAYCMARKETPLACEID = &xEBAYCMARKETPLACEID
+	return r
+}
+
+// This header is used to indicate the natural language and locale preferred by the user for the response.&lt;br&gt;&lt;br&gt;This header is required when targeting a specific locale of a marketplace that supports multiple locales. For example:&lt;ul&gt;&lt;li&gt;When targeting the French locale of the Belgium marketplace, it is required to pass in &lt;code&gt;fr-BE&lt;/code&gt; to specify this. If this locale is not specified, the language will default to Dutch.&lt;/li&gt;&lt;li&gt;When targeting the French locale of the Canadian marketplace, it is required to pass in &lt;code&gt;fr-CA&lt;/code&gt; to specify this. If this locale is not specified, the language will default to English.&lt;/li&gt;&lt;/ul&gt;
+func (r ApiGetItemRequest) AcceptLanguage(acceptLanguage string) ApiGetItemRequest {
+	r.acceptLanguage = &acceptLanguage
+	return r
+}
+
+func (r ApiGetItemRequest) Execute() (*Item, *http.Response, error) {
 	return r.ApiService.GetItemExecute(r)
 }
 
 /*
 GetItem Method for GetItem
 
-<p>This method retrieves the details of a specific item, such as description, price, category, all item aspects, condition, return policies, seller feedback and score, shipping options, shipping costs, estimated delivery, and other information the buyer needs to make a purchasing decision.</p><p>The Buy APIs are designed to let you create an eBay shopping experience in your app or website. This means you will need to know when something, such as the availability, quantity, etc., has changed in any eBay item you are offering. You can do this easily by setting the <b> fieldgroups</b> URI parameter. This parameter lets you control what is returned in the response.</p>        <p>Setting <b> fieldgroups</b> to <code>COMPACT</code> reduces the response to only those fields that you need in order to check if any item detail has changed.  Setting <b> fieldgroups</b> to <code>PRODUCT</code>, adds additional fields to the default response that return information about the product of the item. You can use either <code>COMPACT</code> or <code>PRODUCT</code> but not both. For more information, see <a href="/api-docs/buy/browse/resources/item/methods/getItem#uri.fieldgroups">fieldgroups</a>.</p>      <h3>URLs for this method</h3>           <p><ul>            <li><b> Production URL: </b> <code>https://api.ebay.com/buy/browse/v1/item/{item_id}</code></li>            <li><b> Sandbox URL:  </b><code>https://api.sandbox.ebay.com/buy/browse/v1/item/{item_id}</code></li>           </ul>    </p>                   <h3><b> Request headers</b></h3> This method uses the  <b>X-EBAY-C-ENDUSERCTX</b> request header to support revenue sharing for eBay Partner Networks and to improve the accuracy of shipping and delivery time estimations.  For details see, <a href="/api-docs/buy/static/api-browse.html#Headers">Request headers</a> in the Buying Integration Guide.     <h3><b> Restrictions </b></h3> <p>For a list of supported sites and other restrictions, see <a href="/api-docs/buy/browse/overview.html#API">API Restrictions</a>.</p> <span class="tablenote"><b>eBay Partner Network: </b> In order to be commissioned for your sales, you must use the URL returned in the <code>itemAffiliateWebUrl</code> field to forward your buyer to the ebay.com site. </span>
+This method retrieves the details of a specific item, such as description, price, category, all item aspects, condition, return policies, seller feedback and score, shipping options, shipping costs, estimated delivery, and other information the buyer needs to make a purchasing decision.<br><br>The Buy APIs are designed to let you create an eBay shopping experience in your app or website. This means you will need to know when something, such as the availability, quantity, etc., has changed in any eBay item you are offering. This is easily achieved by setting the <b>fieldgroups</b> URI parameter to one of the following values:<ul><li><code>COMPACT</code><br>Reduces the response to only those fields necessary in order to determine if any item detail has changed. This field group <b>must</b> be used alone.</li><li><code>PRODUCT</code><br>Adds fields to the default response that return information about the product/item. This field group may also be used in conjunction with <code>ADDITIONAL_SELLER_DETAILS</code>.</li><li><code>ADDITIONAL_SELLER_DETAILS</code><br>Adds an additional field to the response that returns the seller's user ID. This field group may also be used in conjunction with <code>PRODUCT</code>.</li></ul>For additional information, refer to <a href="/api-docs/buy/browse/resources/item/methods/getItem#uri.fieldgroups">fieldgroups</a>.<h3>Restrictions</h3>For a list of supported sites and other restrictions, refer to <a href="/api-docs/buy/browse/overview.html#API">API Restrictions</a>.<br><br><span class="tablenote"><b>eBay Partner Network:</b> In order to be commissioned for your sales, you must use the URL returned in the <code>itemAffiliateWebUrl</code> field to forward your buyer to the ebay.com site.</span>
 
- @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
- @param itemId The eBay RESTful identifier of an item. This ID is returned by the <b> Browse</b> and <b> Feed</b> API methods.  <br /><br /> <b> RESTful Item ID Format: </b><code>v1</code>|<code><i>#</i></code>|<code><i>#</i></code> <br />For example: <code>v1|2**********2|0</code> or <code>v1|1**********2|4**********2</code> <br /><br />For more information about item ID for RESTful APIs, see the <a href=\"/api-docs/buy/static/api-browse.html#Legacy\">Legacy API compatibility</a> section of the <i>Buy APIs Overview</i>.
+ @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+ @param itemId This path parameter specifies the unique RESTful identifier of the item being retrieved. <br><br><b>RESTful Item ID Format: </b><code>v1</code>|<code><i>#</i></code>|<code><i>#</i></code><br><br>For a single SKU listing, pass in the item ID: <pre>v1|2**********2|0</pre>For a multi-SKU listing, pass in the identifier of the variation:<pre>v1|1**********2|4**********2</pre><br>For more information about item IDs for RESTful APIs, refer to <a href=\"/api-docs/buy/static/api-browse.html#Legacy\" target=\"_blank\">Item ID legacy API compatibility overview</a> in the <a href=\"/api-docs/buy/static/buying-ig-landing.html\" target=\"_blank\">Buying Integration Guide</a>.
  @return ApiGetItemRequest
 */
-func (a *ItemApiService) GetItem(ctx _context.Context, itemId string) *ApiGetItemRequest {
-	return &ApiGetItemRequest{
+func (a *ItemAPIService) GetItem(ctx context.Context, itemId string) ApiGetItemRequest {
+	return ApiGetItemRequest{
 		ApiService: a,
-		ctx:        ctx,
-		itemId:     itemId,
+		ctx: ctx,
+		itemId: itemId,
 	}
 }
 
 // Execute executes the request
 //  @return Item
-func (a *ItemApiService) GetItemExecute(r *ApiGetItemRequest) (Item, *_nethttp.Response, error) {
+func (a *ItemAPIService) GetItemExecute(r ApiGetItemRequest) (*Item, *http.Response, error) {
 	var (
-		localVarHTTPMethod  = _nethttp.MethodGet
-		localVarPostBody    interface{}
-		formFiles           []formFile
-		localVarReturnValue Item
+		localVarHTTPMethod   = http.MethodGet
+		localVarPostBody     interface{}
+		formFiles            []formFile
+		localVarReturnValue  *Item
 	)
 
-	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "ItemApiService.GetItem")
+	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "ItemAPIService.GetItem")
 	if err != nil {
-		return localVarReturnValue, nil, GenericOpenAPIError{error: err.Error()}
+		return localVarReturnValue, nil, &GenericOpenAPIError{error: err.Error()}
 	}
 
 	localVarPath := localBasePath + "/item/{item_id}"
-	localVarPath = strings.Replace(localVarPath, "{"+"item_id"+"}", _neturl.PathEscape(parameterToString(r.itemId, "")), -1)
+	localVarPath = strings.Replace(localVarPath, "{"+"item_id"+"}", url.PathEscape(parameterValueToString(r.itemId, "itemId")), -1)
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
 	if r.fieldgroups != nil {
-		localVarQueryParams.Add("fieldgroups", parameterToString(*r.fieldgroups, ""))
+		parameterAddToHeaderOrQuery(localVarQueryParams, "fieldgroups", r.fieldgroups, "form", "")
 	}
 	// to determine the Content-Type header
 	localVarHTTPContentTypes := []string{}
@@ -224,6 +262,15 @@ func (a *ItemApiService) GetItemExecute(r *ApiGetItemRequest) (Item, *_nethttp.R
 	if localVarHTTPHeaderAccept != "" {
 		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
 	}
+	if r.xEBAYCENDUSERCTX != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "X-EBAY-C-ENDUSERCTX", r.xEBAYCENDUSERCTX, "simple", "")
+	}
+	if r.xEBAYCMARKETPLACEID != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "X-EBAY-C-MARKETPLACE-ID", r.xEBAYCMARKETPLACEID, "simple", "")
+	}
+	if r.acceptLanguage != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "Accept-Language", r.acceptLanguage, "simple", "")
+	}
 	req, err := a.client.prepareRequest(r.ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, formFiles)
 	if err != nil {
 		return localVarReturnValue, nil, err
@@ -234,15 +281,15 @@ func (a *ItemApiService) GetItemExecute(r *ApiGetItemRequest) (Item, *_nethttp.R
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
-	localVarHTTPResponse.Body = _ioutil.NopCloser(bytes.NewBuffer(localVarBody))
+	localVarHTTPResponse.Body = io.NopCloser(bytes.NewBuffer(localVarBody))
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
 	if localVarHTTPResponse.StatusCode >= 300 {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: localVarHTTPResponse.Status,
 		}
@@ -251,7 +298,7 @@ func (a *ItemApiService) GetItemExecute(r *ApiGetItemRequest) (Item, *_nethttp.R
 
 	err = a.client.decode(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
 	if err != nil {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: err.Error(),
 		}
@@ -262,90 +309,111 @@ func (a *ItemApiService) GetItemExecute(r *ApiGetItemRequest) (Item, *_nethttp.R
 }
 
 type ApiGetItemByLegacyIdRequest struct {
-	ctx                _context.Context
-	ApiService         *ItemApiService
-	legacyItemId       *string
-	fieldgroups        *string
-	legacyVariationId  *string
+	ctx context.Context
+	ApiService *ItemAPIService
+	legacyItemId *string
+	fieldgroups *string
+	legacyVariationId *string
 	legacyVariationSku *string
+	xEBAYCENDUSERCTX *string
+	xEBAYCMARKETPLACEID *string
+	acceptLanguage *string
 }
 
-// Specifies either: &lt;ul&gt; &lt;li&gt;The legacy item ID of an item that is &lt;em&gt;not&lt;/em&gt; part of a group. &lt;/li&gt; &lt;li&gt;The legacy item ID of a group, which is the ID of the \&quot;parent\&quot; of the group of items. &lt;br /&gt; &lt;br /&gt;&lt;span class&#x3D;\&quot;tablenote\&quot;&gt; &lt;b&gt; Note: &lt;/b&gt; If you pass in a group ID, you must also use the &lt;b&gt; legacy_variation_id&lt;/b&gt; field and pass in the legacy ID of the specific item variation (child ID).&lt;/span&gt;&lt;/li&gt;&lt;/ul&gt;  Legacy ids are returned by APIs, such as the &lt;a href&#x3D;\&quot;https://developer.ebay.com/devzone/finding/callref/index.html\&quot; target&#x3D;\&quot;_blank\&quot;&gt;Finding API&lt;/a&gt;.  &lt;br /&gt;&lt;br /&gt;The following is an example of using the value of the &lt;b&gt; ItemID&lt;/b&gt; field for a specific item from Finding to get the RESTful &lt;b&gt; itemId&lt;/b&gt; value. &lt;br /&gt; &lt;br /&gt;&amp;nbsp;&amp;nbsp;&amp;nbsp;&lt;code&gt; browse/v1/item/get_item_by_legacy_id?legacy_item_id&#x3D;1**********9  &lt;/code&gt;&lt;br /&gt;&lt;br /&gt;&lt;b&gt; Maximum: &lt;/b&gt; 1
-func (r *ApiGetItemByLegacyIdRequest) LegacyItemId(legacyItemId string) *ApiGetItemByLegacyIdRequest {
+// This query parameter is the unique identifier that specifies the item being retrieved.&lt;br&gt;&lt;br&gt;&lt;span class&#x3D;\&quot;tablenote\&quot;&gt;&lt;b&gt; Note:&lt;/b&gt; When passing in the ID for a multi-variation listing, you must also use the &lt;code&gt;legacy_variation_id&lt;/code&gt; field and pass in the ID of the specific item variation. If not, all variation within the multi-SKU listing will be retrieved.&lt;/span&gt;&lt;/li&gt;&lt;/ul&gt;&lt;br&gt;The following is an example of using the value of the &lt;b&gt;ItemID&lt;/b&gt; field for a specific item to get the RESTful &lt;code&gt;itemId&lt;/code&gt; value.&lt;pre&gt;&lt;code&gt;browse/v1/item/get_item_by_legacy_id?legacy_item_id&#x3D;1**********9&lt;/code&gt;&lt;/pre&gt;
+func (r ApiGetItemByLegacyIdRequest) LegacyItemId(legacyItemId string) ApiGetItemByLegacyIdRequest {
 	r.legacyItemId = &legacyItemId
 	return r
 }
 
-// This field lets you control what is returned in the response. If you do not set this field, the method returns all the details of the item. &lt;b&gt; Note&lt;/b&gt;: In this method, the only value supported is &lt;code&gt;PRODUCT&lt;/code&gt;. &lt;p&gt;&lt;b&gt; Valid Values: &lt;/b&gt;&lt;br /&gt;&lt;br /&gt; &lt;b&gt; PRODUCT&lt;/b&gt; - This adds the &lt;code&gt;additionalImages&lt;/code&gt;, &lt;code&gt;additionalProductIdentities&lt;/code&gt;, &lt;code&gt;aspectGroups&lt;/code&gt;, &lt;code&gt;description&lt;/code&gt;, &lt;code&gt;gtins&lt;/code&gt;, &lt;code&gt;image&lt;/code&gt;, and &lt;code&gt;title&lt;/code&gt; fields to the response, which describe the item&#39;s product.  See &lt;a href&#x3D;\&quot;/api-docs/buy/browse/resources/item/methods/getItemByLegacyItem#response.product\&quot;&gt;Product&lt;/a&gt; for more information about these fields. &lt;br /&gt;&lt;br /&gt;Code so that your app gracefully handles any future changes to this list.
-func (r *ApiGetItemByLegacyIdRequest) Fieldgroups(fieldgroups string) *ApiGetItemByLegacyIdRequest {
+// This field controls what is returned in the response. If this field is not set, the method returns all details about the item. Multiple &lt;code&gt;fieldgroups&lt;/code&gt; can be set.&lt;br&gt;&lt;br&gt;&lt;b&gt;Valid Values:&lt;/b&gt;&lt;ul&gt;&lt;li&gt; &lt;b&gt;PRODUCT&lt;/b&gt;&lt;br&gt;This field group adds the &lt;a href&#x3D;\&quot;/api-docs/buy/browse/resources/item/methods/getItemByLegacyId#response.product\&quot;&gt;product&lt;/a&gt; container to the response. &lt;/li&gt;&lt;li&gt;&lt;b&gt;ADDITIONAL_SELLER_DETAILS&lt;/b&gt;&lt;br&gt;This field group adds the &lt;a href&#x3D;\&quot;/api-docs/buy/browse/resources/item/methods/getItemByLegacyId#response.seller.userId\&quot;&gt;userId&lt;/a&gt; field to the response.&lt;/li&gt;&lt;/ul&gt;
+func (r ApiGetItemByLegacyIdRequest) Fieldgroups(fieldgroups string) ApiGetItemByLegacyIdRequest {
 	r.fieldgroups = &fieldgroups
 	return r
 }
 
-// Specifies the legacy item ID of a specific item in an item group, such as the red shirt size L. &lt;br /&gt;&lt;br /&gt;Legacy ids are returned by APIs, such as the &lt;a href&#x3D;\&quot;https://developer.ebay.com/devzone/finding/callref/index.html\&quot; target&#x3D;\&quot;_blank\&quot;&gt;Finding API&lt;/a&gt;.     &lt;br /&gt;&lt;br /&gt;&lt;b&gt; Maximum: &lt;/b&gt; 1 &lt;br /&gt;&lt;b&gt; Requirement: &lt;/b&gt; You must &lt;b&gt; always&lt;/b&gt; pass in the &lt;b&gt; legacy_item_id &lt;/b&gt; with the &lt;b&gt; legacy_variation_id&lt;/b&gt;
-func (r *ApiGetItemByLegacyIdRequest) LegacyVariationId(legacyVariationId string) *ApiGetItemByLegacyIdRequest {
+// This query parameter specifies the legacy item ID of a specific item in a multi-variation listing, such as that for the &lt;i&gt;red shirt size L&lt;/i&gt; item.&lt;br&gt;&lt;br&gt;&lt;div class&#x3D;\&quot;msgbox_important\&quot;&gt;&lt;p class&#x3D;\&quot;msgbox_importantInDiv\&quot; data-mc-autonum&#x3D;\&quot;&amp;lt;b&amp;gt;&amp;lt;span style&#x3D;&amp;quot;color: #dd1e31;&amp;quot; class&#x3D;&amp;quot;mcFormatColor&amp;quot;&amp;gt;Important! &amp;lt;/span&amp;gt;&amp;lt;/b&amp;gt;\&quot;&gt;&lt;span class&#x3D;\&quot;autonumber\&quot;&gt;&lt;span&gt;&lt;b&gt;&lt;span style&#x3D;\&quot;color: #dd1e31;\&quot; class&#x3D;\&quot;mcFormatColor\&quot;&gt;Important!&lt;/span&gt;&lt;/b&gt;&lt;/span&gt;&lt;/span&gt; A &lt;code&gt;legacy_item_id&lt;/code&gt; value must always be passed in when specifying a &lt;code&gt;legacy_variation_id&lt;/code&gt; value.&lt;/p&gt;&lt;/div&gt;
+func (r ApiGetItemByLegacyIdRequest) LegacyVariationId(legacyVariationId string) ApiGetItemByLegacyIdRequest {
 	r.legacyVariationId = &legacyVariationId
 	return r
 }
 
-// Specifics the legacy SKU of the item. SKU are item ids created by the seller. &lt;br /&gt;&lt;br /&gt;Legacy SKUs are returned by eBay the  &lt;a href&#x3D;\&quot;https://developer.ebay.com/Devzone/shopping/docs/CallRef/index.html\&quot; target&#x3D;\&quot;_blank\&quot;&gt;Shopping API&lt;/a&gt;. &lt;br /&gt;&lt;br /&gt;The following is an example of using the value of the &lt;b&gt; ItemID&lt;/b&gt; and &lt;b&gt; SKU&lt;/b&gt; fields to get the RESTful &lt;b&gt; itemId&lt;/b&gt; value. &lt;br /&gt; &lt;br /&gt;&amp;nbsp;&amp;nbsp;&amp;nbsp;&lt;code&gt; browse/v1/item/get_item_by_legacy_id?legacy_item_id&#x3D;1**********9&amp;amp;legacy_variation_sku&#x3D;V**********M&lt;/code&gt;&lt;br /&gt;&lt;br /&gt;&lt;b&gt; Maximum: &lt;/b&gt; 1 &lt;br /&gt;&lt;b&gt; Requirement: &lt;/b&gt; You must &lt;b&gt; always&lt;/b&gt; pass in the &lt;b&gt; legacy_item_id &lt;/b&gt; with the &lt;b&gt; legacy_variation_sku&lt;/b&gt;
-func (r *ApiGetItemByLegacyIdRequest) LegacyVariationSku(legacyVariationSku string) *ApiGetItemByLegacyIdRequest {
+// This query parameter specifies the legacy SKU of an item. SKUs are the unique identifiers of an item created by the seller.&lt;br&gt;&lt;br&gt;The following is an example of using the value of the &lt;code&gt;ItemID&lt;/code&gt; and &lt;code&gt;SKU&lt;/code&gt; fields to get the RESTful &lt;code&gt;itemId&lt;/code&gt; value.&lt;pre&gt;&lt;code&gt;browse/v1/item/get_item_by_legacy_id?legacy_item_id&#x3D;1**********9&amp;amp;legacy_variation_sku&#x3D;V**********M&lt;/code&gt;&lt;/pre&gt;&lt;div class&#x3D;\&quot;msgbox_important\&quot;&gt;&lt;p class&#x3D;\&quot;msgbox_importantInDiv\&quot; data-mc-autonum&#x3D;\&quot;&amp;lt;b&amp;gt;&amp;lt;span style&#x3D;&amp;quot;color: #dd1e31;&amp;quot; class&#x3D;&amp;quot;mcFormatColor&amp;quot;&amp;gt;Important! &amp;lt;/span&amp;gt;&amp;lt;/b&amp;gt;\&quot;&gt;&lt;span class&#x3D;\&quot;autonumber\&quot;&gt;&lt;span&gt;&lt;b&gt;&lt;span style&#x3D;\&quot;color: #dd1e31;\&quot; class&#x3D;\&quot;mcFormatColor\&quot;&gt;Important!&lt;/span&gt;&lt;/b&gt;&lt;/span&gt;&lt;/span&gt; A &lt;code&gt;legacy_item_id&lt;/code&gt; value must always be passed in when specifying a &lt;code&gt;legacy_variation_sku&lt;/code&gt; value.&lt;/p&gt;&lt;/div&gt;
+func (r ApiGetItemByLegacyIdRequest) LegacyVariationSku(legacyVariationSku string) ApiGetItemByLegacyIdRequest {
 	r.legacyVariationSku = &legacyVariationSku
 	return r
 }
 
-func (r *ApiGetItemByLegacyIdRequest) Execute() (Item, *_nethttp.Response, error) {
+// This header is required to support revenue sharing for eBay Partner Network and to improve the accuracy of shipping and delivery time estimations.&lt;br&gt;&lt;br&gt;For additional information, refer to &lt;a href&#x3D;\&quot;/api-docs/buy/static/api-browse.html#Headers\&quot; target&#x3D;\&quot;_blank \&quot;&gt;Use request headers&lt;/a&gt; in the &lt;a href&#x3D;\&quot;/api-docs/buy/static/buying-ig-landing.html\&quot; target&#x3D;\&quot;_blank\&quot;&gt;Buying Integration Guide&lt;/a&gt;.
+func (r ApiGetItemByLegacyIdRequest) XEBAYCENDUSERCTX(xEBAYCENDUSERCTX string) ApiGetItemByLegacyIdRequest {
+	r.xEBAYCENDUSERCTX = &xEBAYCENDUSERCTX
+	return r
+}
+
+// This header identifies the seller&#39;s eBay marketplace. It is required for all marketplaces outside of the US.&lt;br&gt;&lt;br&gt;&lt;span class&#x3D;\&quot;tablenote\&quot;&gt;&lt;b&gt;Note:&lt;/b&gt; If a marketplace ID value is not provided, the default value of &lt;code&gt;EBAY_US&lt;/code&gt; is used.&lt;/span&gt;&lt;br&gt;See &lt;a href&#x3D;\&quot;/api-docs/buy/browse/types/ba:MarketplaceIdEnum\&quot; target&#x3D;\&quot;_blank\&quot;&gt;MarketplaceIdEnum&lt;/a&gt; for a list of supported marketplaces.
+func (r ApiGetItemByLegacyIdRequest) XEBAYCMARKETPLACEID(xEBAYCMARKETPLACEID string) ApiGetItemByLegacyIdRequest {
+	r.xEBAYCMARKETPLACEID = &xEBAYCMARKETPLACEID
+	return r
+}
+
+// This header is used to indicate the natural language and locale preferred by the user for the response.&lt;br&gt;&lt;br&gt;This header is required when targeting a specific locale of a marketplace that supports multiple locales. For example:&lt;ul&gt;&lt;li&gt;When targeting the French locale of the Belgium marketplace, it is required to pass in &lt;code&gt;fr-BE&lt;/code&gt; to specify this. If this locale is not specified, the language will default to Dutch.&lt;/li&gt;&lt;li&gt;When targeting the French locale of the Canadian marketplace, it is required to pass in &lt;code&gt;fr-CA&lt;/code&gt; to specify this. If this locale is not specified, the language will default to English.&lt;/li&gt;&lt;/ul&gt;
+func (r ApiGetItemByLegacyIdRequest) AcceptLanguage(acceptLanguage string) ApiGetItemByLegacyIdRequest {
+	r.acceptLanguage = &acceptLanguage
+	return r
+}
+
+func (r ApiGetItemByLegacyIdRequest) Execute() (*Item, *http.Response, error) {
 	return r.ApiService.GetItemByLegacyIdExecute(r)
 }
 
 /*
 GetItemByLegacyId Method for GetItemByLegacyId
 
-  <p>This method is a bridge between the eBay legacy APIs, such as  <b> Shopping</b>, and <b> Finding</b> and the eBay Buy APIs. There are differences between how legacy APIs and RESTful APIs return the identifier of an "item" and what the item ID represents. This method lets you use the legacy item ids retrieve the details of a specific item, such as description, price, and other information the buyer needs to make a purchasing decision. It also returns the RESTful item ID, which you can use with all the Buy API  methods.</p>  <p>For more information about how to use legacy ids with the Buy APIs, see <a href="/api-docs/buy/static/api-browse.html#Legacy">Legacy API compatibility</a> in the Buying Integration guide.</p>  <p>This method returns the item details and requires you to pass in either the item ID of a non-variation item or the item ids of both the parent and child of an item group. An item group is an item that has various aspect differences, such as color, size, storage capacity, etc.</p> <p>When an item group is created, one of the item variations, such as the red shirt size L, is chosen as the "parent". All the other items in the group are the children, such as the blue shirt size L, red shirt size M, etc.</p>    <p>The <b> fieldgroups</b> URI parameter lets you control what is returned in the response. Setting <b> fieldgroups</b> to <code>PRODUCT</code>, adds additional fields to the default response that return information about the product of the item. For more information, see <a href="/api-docs/buy/browse/resources/item/methods/getItemByLegacyItem#uri.fieldgroups">fieldgroups</a>.</p>       <h3>URLs for this method</h3>           <p><ul>            <li><b> Production URL: </b> <code>https://api.ebay.com/buy/browse/v1/item/get_item_by_legacy_id?</code></li>            <li><b> Sandbox URL:  </b><code>https://api.sandbox.ebay.com/buy/browse/v1/item/get_item_by_legacy_id?</code></li>           </ul>    </p>              <h3><b> Request headers</b></h3> This method uses the  <b>X-EBAY-C-ENDUSERCTX</b> request header to support revenue sharing for eBay Partner Networks and to improve the accuracy of shipping and delivery time estimations.   For details see, <a href="/api-docs/buy/static/api-browse.html#Headers">Request headers</a> in the Buying Integration Guide.    <h3><b> Restrictions </b></h3> <p>For a list of supported sites and other restrictions, see <a href="/api-docs/buy/browse/overview.html#API">API Restrictions</a>.</p> <span class="tablenote"><b>eBay Partner Network: </b> In order to be commissioned for your sales, you must use the URL returned in the <code>itemAffiliateWebUrl</code> field to forward your buyer to the ebay.com site. </span>
+This method is a bridge between the eBay legacy APIs, such as <b>Shopping</b> and <b>Finding</b>, and the eBay Buy APIs. There are differences between how legacy APIs and RESTful APIs return the identifier of an "item" and what the item ID represents. This method lets you use the legacy item ids retrieve the details of a specific item, such as description, price, and other information the buyer needs to make a purchasing decision. It also returns the RESTful <b>item_id</b>, which you can use with all the Buy API  methods.<br><br>For additional information about how to use legacy ids with the Buy APIs, refer to <a href="/api-docs/buy/static/api-browse.html#Legacy" target="_blank">Item ID legacy API compatibility overview</a> in the Buying Integration guide.</br><br>This method returns the item details and requires you to pass in either the <b>item_id</b> of a non-variation item or the <b>item_id</b> values for both the parent and child of an item group.<br><br><span class="tablenote"><b>Note:</b> An item group is an item that has various aspect differences, such as color, size, storage capacity, etc.</span></br>When an item group is created, one of the item variations, such as the red shirt size L, is chosen as the "parent". All other items in the group are the children, such as the blue shirt size L, red shirt size M, etc.<br><br>The <b>fieldgroups</b> URI parameter lets you control what is returned in the response:<ul><li>Setting <b>fieldgroups</b> to <code>PRODUCT</code> adds additional fields to the default response that return information about the product of the item.</li><li>Setting the <b>fieldgroups</b> to <code>ADDITIONAL_SELLER_DETAILS</code> adds an additional field to the response that returns the seller's user ID.</li></ul>These <b>fieldgroups</b> can be used independently or at the same time. For additional information, refer to <a href="/api-docs/buy/browse/resources/item/methods/getItemByLegacyId#uri.fieldgroups">fieldgroups</a>.<h3>Restrictions</h3>For a list of supported sites and other restrictions, refer to <a href="/api-docs/buy/browse/overview.html#API">API Restrictions</a>.<br><br><span class="tablenote"><b>eBay Partner Network:</b> In order to be commissioned for your sales, you must use the URL returned in the <code>itemAffiliateWebUrl</code> field to forward your buyer to the ebay.com site.</span>
 
- @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+ @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
  @return ApiGetItemByLegacyIdRequest
 */
-func (a *ItemApiService) GetItemByLegacyId(ctx _context.Context) *ApiGetItemByLegacyIdRequest {
-	return &ApiGetItemByLegacyIdRequest{
+func (a *ItemAPIService) GetItemByLegacyId(ctx context.Context) ApiGetItemByLegacyIdRequest {
+	return ApiGetItemByLegacyIdRequest{
 		ApiService: a,
-		ctx:        ctx,
+		ctx: ctx,
 	}
 }
 
 // Execute executes the request
 //  @return Item
-func (a *ItemApiService) GetItemByLegacyIdExecute(r *ApiGetItemByLegacyIdRequest) (Item, *_nethttp.Response, error) {
+func (a *ItemAPIService) GetItemByLegacyIdExecute(r ApiGetItemByLegacyIdRequest) (*Item, *http.Response, error) {
 	var (
-		localVarHTTPMethod  = _nethttp.MethodGet
-		localVarPostBody    interface{}
-		formFiles           []formFile
-		localVarReturnValue Item
+		localVarHTTPMethod   = http.MethodGet
+		localVarPostBody     interface{}
+		formFiles            []formFile
+		localVarReturnValue  *Item
 	)
 
-	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "ItemApiService.GetItemByLegacyId")
+	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "ItemAPIService.GetItemByLegacyId")
 	if err != nil {
-		return localVarReturnValue, nil, GenericOpenAPIError{error: err.Error()}
+		return localVarReturnValue, nil, &GenericOpenAPIError{error: err.Error()}
 	}
 
 	localVarPath := localBasePath + "/item/get_item_by_legacy_id"
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 	if r.legacyItemId == nil {
 		return localVarReturnValue, nil, reportError("legacyItemId is required and must be specified")
 	}
 
 	if r.fieldgroups != nil {
-		localVarQueryParams.Add("fieldgroups", parameterToString(*r.fieldgroups, ""))
+		parameterAddToHeaderOrQuery(localVarQueryParams, "fieldgroups", r.fieldgroups, "form", "")
 	}
-	localVarQueryParams.Add("legacy_item_id", parameterToString(*r.legacyItemId, ""))
+	parameterAddToHeaderOrQuery(localVarQueryParams, "legacy_item_id", r.legacyItemId, "form", "")
 	if r.legacyVariationId != nil {
-		localVarQueryParams.Add("legacy_variation_id", parameterToString(*r.legacyVariationId, ""))
+		parameterAddToHeaderOrQuery(localVarQueryParams, "legacy_variation_id", r.legacyVariationId, "form", "")
 	}
 	if r.legacyVariationSku != nil {
-		localVarQueryParams.Add("legacy_variation_sku", parameterToString(*r.legacyVariationSku, ""))
+		parameterAddToHeaderOrQuery(localVarQueryParams, "legacy_variation_sku", r.legacyVariationSku, "form", "")
 	}
 	// to determine the Content-Type header
 	localVarHTTPContentTypes := []string{}
@@ -364,6 +432,15 @@ func (a *ItemApiService) GetItemByLegacyIdExecute(r *ApiGetItemByLegacyIdRequest
 	if localVarHTTPHeaderAccept != "" {
 		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
 	}
+	if r.xEBAYCENDUSERCTX != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "X-EBAY-C-ENDUSERCTX", r.xEBAYCENDUSERCTX, "simple", "")
+	}
+	if r.xEBAYCMARKETPLACEID != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "X-EBAY-C-MARKETPLACE-ID", r.xEBAYCMARKETPLACEID, "simple", "")
+	}
+	if r.acceptLanguage != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "Accept-Language", r.acceptLanguage, "simple", "")
+	}
 	req, err := a.client.prepareRequest(r.ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, formFiles)
 	if err != nil {
 		return localVarReturnValue, nil, err
@@ -374,15 +451,15 @@ func (a *ItemApiService) GetItemByLegacyIdExecute(r *ApiGetItemByLegacyIdRequest
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
-	localVarHTTPResponse.Body = _ioutil.NopCloser(bytes.NewBuffer(localVarBody))
+	localVarHTTPResponse.Body = io.NopCloser(bytes.NewBuffer(localVarBody))
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
 	if localVarHTTPResponse.StatusCode >= 300 {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: localVarHTTPResponse.Status,
 		}
@@ -391,7 +468,7 @@ func (a *ItemApiService) GetItemByLegacyIdExecute(r *ApiGetItemByLegacyIdRequest
 
 	err = a.client.decode(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
 	if err != nil {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: err.Error(),
 		}
@@ -402,69 +479,90 @@ func (a *ItemApiService) GetItemByLegacyIdExecute(r *ApiGetItemByLegacyIdRequest
 }
 
 type ApiGetItemsRequest struct {
-	ctx          _context.Context
-	ApiService   *ItemApiService
-	itemIds      *string
+	ctx context.Context
+	ApiService *ItemAPIService
+	itemIds *string
 	itemGroupIds *string
+	xEBAYCENDUSERCTX *string
+	xEBAYCMARKETPLACEID *string
+	acceptLanguage *string
 }
 
-// A list of item IDs. Item IDs are the eBay RESTful identifier of items. &lt;br&gt;&lt;br&gt;&lt;b&gt; RESTful Item ID Format: &lt;/b&gt;&lt;code&gt;v1&lt;/code&gt;|&lt;code&gt;&lt;i&gt;#&lt;/i&gt;&lt;/code&gt;|&lt;code&gt;&lt;i&gt;#&lt;/i&gt;&lt;/code&gt;&lt;br&gt;For example: &lt;code&gt;v1|2**********2|0&lt;/code&gt; or &lt;code&gt;v1|1**********2|4**********2&lt;/code&gt; &lt;br&gt;&lt;br&gt;In any given request, either item_ids or item_group_ids can be retrieved. Attempting to retrieve both will result in an error. &lt;br&gt;&lt;br&gt; In a request, multiple item_ids can be passed as comma separated values.&lt;br&gt;&lt;br&gt;&lt;b&gt; Maximum allowed itemIDs: &lt;/b&gt; 20 &lt;br&gt;&lt;br&gt;For more information about item IDs for RESTful APIs, see the &lt;a href&#x3D;\&quot;/api-docs/buy/static/api-browse.html#Legacy\&quot;&gt;Legacy API compatibility&lt;/a&gt; section of the &lt;i&gt;Buy APIs Overview&lt;/i&gt;.
-func (r *ApiGetItemsRequest) ItemIds(itemIds string) *ApiGetItemsRequest {
+// A comma separated list of the unique identifiers of the items to retrieve (maximum 20).&lt;br&gt;&lt;br&gt;&lt;span class&#x3D;\&quot;tablenote\&quot;&gt;&lt;b&gt;Note:&lt;/b&gt; In any given request, either &lt;code&gt;item_ids&lt;/code&gt; &lt;b&gt;or&lt;/b&gt; &lt;code&gt;item_group_ids&lt;/code&gt; can be retrieved. Attempting to retrieve both will result in an error.&lt;/span&gt;&lt;br&gt;&lt;b&gt;RESTful Item ID Format: &lt;/b&gt;&lt;code&gt;v1&lt;/code&gt;|&lt;code&gt;&lt;i&gt;#&lt;/i&gt;&lt;/code&gt;|&lt;code&gt;&lt;i&gt;#&lt;/i&gt;&lt;/code&gt;&lt;br&gt;&lt;br&gt;For a single SKU listing, pass in the item ID: &lt;pre&gt;v1|2**********2|0&lt;/pre&gt;For a multi-SKU listing, pass in the identifier of the variation:&lt;pre&gt;v1|1**********2|4**********2&lt;/pre&gt;&lt;br&gt;For more information about item IDs for RESTful APIs, refer to &lt;a href&#x3D;\&quot;/api-docs/buy/static/api-browse.html#Legacy\&quot; target&#x3D;\&quot;_blank\&quot;&gt;Item ID legacy API compatibility overview&lt;/a&gt; in the &lt;a href&#x3D;\&quot;/api-docs/buy/static/buying-ig-landing.html\&quot; target&#x3D;\&quot;_blank\&quot;&gt;Buying Integration Guide&lt;/a&gt;.
+func (r ApiGetItemsRequest) ItemIds(itemIds string) ApiGetItemsRequest {
 	r.itemIds = &itemIds
 	return r
 }
 
-// A list of item group IDs. Item group IDs are the eBay RESTful identifier of item groups. &lt;br&gt;&lt;br&gt;&lt;b&gt; RESTful Group Item ID Format: &lt;/b&gt;&lt;code&gt;############&lt;/code&gt;&lt;br&gt;For example: &lt;code&gt;3**********9&lt;/code&gt;&lt;br&gt;&lt;br&gt;In any given request, either item_ids or item_group_ids can be retrieved. Attempting to retrieve both will result in an error.&lt;br&gt;&lt;br&gt;In a request, multiple item_group_ids can be passed as comma separated values.&lt;br&gt;&lt;br&gt;&lt;b&gt; Maximum allowed itemGroupIDs: &lt;/b&gt; 10 &lt;br&gt;&lt;br&gt;
-func (r *ApiGetItemsRequest) ItemGroupIds(itemGroupIds string) *ApiGetItemsRequest {
+// A comma separated list of the unique identifiers of the item groups being retrieved (maximum 10).&lt;br&gt;&lt;br&gt;&lt;span class&#x3D;\&quot;tablenote\&quot;&gt;&lt;b&gt;Note:&lt;/b&gt; In any given request, either &lt;code&gt;item_ids&lt;/code&gt; &lt;b&gt;or&lt;/b&gt; &lt;code&gt;item_group_ids&lt;/code&gt; can be retrieved. Attempting to retrieve both will result in an error.&lt;/span&gt;&lt;br&gt;&lt;b&gt;RESTful Group Item ID Format:&lt;/b&gt; &lt;code&gt;############&lt;/code&gt;&lt;br&gt;&lt;br&gt;For example:&lt;pre&gt;3**********9&lt;/pre&gt;
+func (r ApiGetItemsRequest) ItemGroupIds(itemGroupIds string) ApiGetItemsRequest {
 	r.itemGroupIds = &itemGroupIds
 	return r
 }
 
-func (r *ApiGetItemsRequest) Execute() (Items, *_nethttp.Response, error) {
+// This header is required to support revenue sharing for eBay Partner Network and to improve the accuracy of shipping and delivery time estimations.&lt;br&gt;&lt;br&gt;For additional information, refer to &lt;a href&#x3D;\&quot;/api-docs/buy/static/api-browse.html#Headers\&quot; target&#x3D;\&quot;_blank \&quot;&gt;Use request headers&lt;/a&gt; in the &lt;a href&#x3D;\&quot;/api-docs/buy/static/buying-ig-landing.html\&quot; target&#x3D;\&quot;_blank\&quot;&gt;Buying Integration Guide&lt;/a&gt;.
+func (r ApiGetItemsRequest) XEBAYCENDUSERCTX(xEBAYCENDUSERCTX string) ApiGetItemsRequest {
+	r.xEBAYCENDUSERCTX = &xEBAYCENDUSERCTX
+	return r
+}
+
+// This header identifies the seller&#39;s eBay marketplace. It is required for all marketplaces outside of the US.&lt;br&gt;&lt;br&gt;&lt;span class&#x3D;\&quot;tablenote\&quot;&gt;&lt;b&gt;Note:&lt;/b&gt; If a marketplace ID value is not provided, the default value of &lt;code&gt;EBAY_US&lt;/code&gt; is used.&lt;/span&gt;&lt;br&gt;See &lt;a href&#x3D;\&quot;/api-docs/buy/browse/types/ba:MarketplaceIdEnum\&quot; target&#x3D;\&quot;_blank\&quot;&gt;MarketplaceIdEnum&lt;/a&gt; for a list of supported marketplaces.
+func (r ApiGetItemsRequest) XEBAYCMARKETPLACEID(xEBAYCMARKETPLACEID string) ApiGetItemsRequest {
+	r.xEBAYCMARKETPLACEID = &xEBAYCMARKETPLACEID
+	return r
+}
+
+// This header is used to indicate the natural language and locale preferred by the user for the response.&lt;br&gt;&lt;br&gt;This header is required when targeting a specific locale of a marketplace that supports multiple locales. For example:&lt;ul&gt;&lt;li&gt;When targeting the French locale of the Belgium marketplace, it is required to pass in &lt;code&gt;fr-BE&lt;/code&gt; to specify this. If this locale is not specified, the language will default to Dutch.&lt;/li&gt;&lt;li&gt;When targeting the French locale of the Canadian marketplace, it is required to pass in &lt;code&gt;fr-CA&lt;/code&gt; to specify this. If this locale is not specified, the language will default to English.&lt;/li&gt;&lt;/ul&gt;
+func (r ApiGetItemsRequest) AcceptLanguage(acceptLanguage string) ApiGetItemsRequest {
+	r.acceptLanguage = &acceptLanguage
+	return r
+}
+
+func (r ApiGetItemsRequest) Execute() (*Items, *http.Response, error) {
 	return r.ApiService.GetItemsExecute(r)
 }
 
 /*
 GetItems Method for GetItems
 
-This method retrieves the details of specific items that the buyer needs to make a purchasing decision.  <br><br><span class="tablenote"> <b>Note:</b> This is a <a href="https://developer.ebay.com/api-docs/static/versioning.html#limited" target="_blank"> <img src="/cms/img/docs/partners-api.svg" class="legend-icon partners-icon" title="Limited Release"  alt="Limited Release" />(Limited Release)</a> available only to select Partners. <br><br>For this method, only the following fields are returned: <code>bidCount</code>, <code>currentBidPrice</code>, <code>eligibleForInlineCheckout</code>, <code>enabledForGuestCheckout</code>, <code>estimatedAvailabilities</code>, <code>itemAffiliateWebUrl</code>, <code>itemId</code>, <code>itemWebUrl</code>, <code>legacyItemId</code>, <code>minimumPriceToBid</code>, <code>price</code>, <code>priorityListing</code>, <code>reservePriceMet</code>, <code>sellerItemRevision</code>, <code>taxes</code>, <code>topRatedBuyingExperience</code>, and <code>uniqueBidderCount</code>.</span> <h3>URLs for this method</h3>           <p><ul>            <li><b> Production URL: </b> <code>https://api.ebay.com/buy/browse/v1/item?</code></li>            <li><b> Sandbox URL:  </b><code>https://api.sandbox.ebay.com/buy/browse/v1/item?</code></li>           </ul>    </p>            <h3><b> Request headers</b></h3> This method uses the  <b>X-EBAY-C-ENDUSERCTX</b> request header to support revenue sharing for eBay Partner Networks and to improve the accuracy of shipping and delivery time estimations.   For details see, <a href="/api-docs/buy/static/api-browse.html#Headers">Request headers</a> in the Buying Integration Guide.   <h3><b> Restrictions </b></h3> <p>For a list of supported sites and other restrictions, see <a href="/api-docs/buy/browse/overview.html#API">API Restrictions</a>.</p> <span class="tablenote"><b>eBay Partner Network:</b> In order to be commissioned for your sales, you must use the URL returned in the itemAffiliateWebUrl field to forward your buyer to the ebay.com site.
+This method retrieves the details about specific items that buyers need to make a purchasing decision.<br><br><span class="tablenote"><b>Note:</b> This is a <a href="/api-docs/static/versioning.html#limited " target="_blank"><img src="/cms/img/docs/partners-api.svg" class="legend-icon partners-icon" title="Limited Release"  alt="Limited Release" />(Limited Release)</a> available only to select Partners.<br><br>For this method, only the following fields are returned: <code>bidCount</code>, <code>currentBidPrice</code>, <code>eligibleForInlineCheckout</code>, <code>enabledForGuestCheckout</code>, <code>estimatedAvailabilities</code>, <code>gtin</code>, <code>itemAffiliateWebUrl</code>, <code>itemCreationDate</code>, <code>itemId</code>, <code>itemWebUrl</code>, <code>legacyItemId</code>, <code>minimumPriceToBid</code>, <code>price</code>, <code>priorityListing</code>, <code>reservePriceMet</code>, <code>sellerItemRevision</code>, <code>taxes</code>, <code>topRatedBuyingExperience</code>, and <code>uniqueBidderCount</code>.<br><br>The array <code>shippingOptions</code>, which comprises multiple fields, is also returned.</span><h3>Restrictions</h3>For a list of supported sites and other restrictions, refer to <a href="/api-docs/buy/browse/overview.html#API">API Restrictions</a>.<br><br><span class="tablenote"><b>eBay Partner Network:</b> In order to be commissioned for your sales, you must use the URL returned in the itemAffiliateWebUrl field to forward your buyer to the ebay.com site.
 
- @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+ @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
  @return ApiGetItemsRequest
 */
-func (a *ItemApiService) GetItems(ctx _context.Context) *ApiGetItemsRequest {
-	return &ApiGetItemsRequest{
+func (a *ItemAPIService) GetItems(ctx context.Context) ApiGetItemsRequest {
+	return ApiGetItemsRequest{
 		ApiService: a,
-		ctx:        ctx,
+		ctx: ctx,
 	}
 }
 
 // Execute executes the request
 //  @return Items
-func (a *ItemApiService) GetItemsExecute(r *ApiGetItemsRequest) (Items, *_nethttp.Response, error) {
+func (a *ItemAPIService) GetItemsExecute(r ApiGetItemsRequest) (*Items, *http.Response, error) {
 	var (
-		localVarHTTPMethod  = _nethttp.MethodGet
-		localVarPostBody    interface{}
-		formFiles           []formFile
-		localVarReturnValue Items
+		localVarHTTPMethod   = http.MethodGet
+		localVarPostBody     interface{}
+		formFiles            []formFile
+		localVarReturnValue  *Items
 	)
 
-	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "ItemApiService.GetItems")
+	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "ItemAPIService.GetItems")
 	if err != nil {
-		return localVarReturnValue, nil, GenericOpenAPIError{error: err.Error()}
+		return localVarReturnValue, nil, &GenericOpenAPIError{error: err.Error()}
 	}
 
 	localVarPath := localBasePath + "/item/"
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 
 	if r.itemIds != nil {
-		localVarQueryParams.Add("item_ids", parameterToString(*r.itemIds, ""))
+		parameterAddToHeaderOrQuery(localVarQueryParams, "item_ids", r.itemIds, "form", "")
 	}
 	if r.itemGroupIds != nil {
-		localVarQueryParams.Add("item_group_ids", parameterToString(*r.itemGroupIds, ""))
+		parameterAddToHeaderOrQuery(localVarQueryParams, "item_group_ids", r.itemGroupIds, "form", "")
 	}
 	// to determine the Content-Type header
 	localVarHTTPContentTypes := []string{}
@@ -483,6 +581,15 @@ func (a *ItemApiService) GetItemsExecute(r *ApiGetItemsRequest) (Items, *_nethtt
 	if localVarHTTPHeaderAccept != "" {
 		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
 	}
+	if r.xEBAYCENDUSERCTX != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "X-EBAY-C-ENDUSERCTX", r.xEBAYCENDUSERCTX, "simple", "")
+	}
+	if r.xEBAYCMARKETPLACEID != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "X-EBAY-C-MARKETPLACE-ID", r.xEBAYCMARKETPLACEID, "simple", "")
+	}
+	if r.acceptLanguage != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "Accept-Language", r.acceptLanguage, "simple", "")
+	}
 	req, err := a.client.prepareRequest(r.ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, formFiles)
 	if err != nil {
 		return localVarReturnValue, nil, err
@@ -493,15 +600,15 @@ func (a *ItemApiService) GetItemsExecute(r *ApiGetItemsRequest) (Items, *_nethtt
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
-	localVarHTTPResponse.Body = _ioutil.NopCloser(bytes.NewBuffer(localVarBody))
+	localVarHTTPResponse.Body = io.NopCloser(bytes.NewBuffer(localVarBody))
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
 	if localVarHTTPResponse.StatusCode >= 300 {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: localVarHTTPResponse.Status,
 		}
@@ -510,7 +617,7 @@ func (a *ItemApiService) GetItemsExecute(r *ApiGetItemsRequest) (Items, *_nethtt
 
 	err = a.client.decode(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
 	if err != nil {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: err.Error(),
 		}
@@ -521,61 +628,92 @@ func (a *ItemApiService) GetItemsExecute(r *ApiGetItemsRequest) (Items, *_nethtt
 }
 
 type ApiGetItemsByItemGroupRequest struct {
-	ctx         _context.Context
-	ApiService  *ItemApiService
+	ctx context.Context
+	ApiService *ItemAPIService
 	itemGroupId *string
+	fieldgroups *string
+	xEBAYCENDUSERCTX *string
+	xEBAYCMARKETPLACEID *string
+	acceptLanguage *string
 }
 
-// Identifier of the item group to return.  An item group is an item that has various aspect differences, such as color, size, storage capacity, etc. &lt;/p&gt; &lt;p&gt;This ID is returned in the &lt;b&gt; itemGroupHref&lt;/b&gt; field of the &lt;a href&#x3D;\&quot;/api-docs/buy/browse/resources/item_summary/methods/search\&quot;&gt;search&lt;/a&gt; and &lt;a href&#x3D;\&quot;/api-docs/buy/browse/resources/item/methods/getItem\&quot;&gt;getItem&lt;/a&gt; methods. &lt;br /&gt;&lt;br /&gt;&lt;b&gt; For Example: &lt;/b&gt;&lt;code&gt; https://api.ebay.com/buy/browse/v1/item/get_items_by_item_group?item_group_id&#x3D;3**********6&lt;/code&gt;
-func (r *ApiGetItemsByItemGroupRequest) ItemGroupId(itemGroupId string) *ApiGetItemsByItemGroupRequest {
+// This query parameter specifies the unique identifier of an item group for which information is to be returned. An item group is an item that has various aspect differences, such as color, size, storage capacity, etc.&lt;br&gt;&lt;br&gt;This ID is returned in the &lt;b&gt;itemGroupHref&lt;/b&gt; field of the &lt;a href&#x3D;\&quot;/api-docs/buy/browse/resources/item_summary/methods/search\&quot;&gt;search&lt;/a&gt; and &lt;a href&#x3D;\&quot;/api-docs/buy/browse/resources/item/methods/getItem\&quot;&gt;getItem&lt;/a&gt; methods.&lt;br&gt;&lt;br&gt;&lt;b&gt;For Example:&lt;/b&gt;&lt;pre&gt;https://api.ebay.com/buy/browse/v1/item/get_items_by_item_group?item_group_id&#x3D;3**********6&lt;/pre&gt;
+func (r ApiGetItemsByItemGroupRequest) ItemGroupId(itemGroupId string) ApiGetItemsByItemGroupRequest {
 	r.itemGroupId = &itemGroupId
 	return r
 }
 
-func (r *ApiGetItemsByItemGroupRequest) Execute() (ItemGroup, *_nethttp.Response, error) {
+// This field controls what is returned in the response. If this field is not set, the method returns all details about the item.&lt;br&gt;&lt;br&gt;&lt;b&gt;Valid Values:&lt;/b&gt;&lt;br&gt;&lt;br&gt;&lt;b&gt;ADDITIONAL_SELLER_DETAILS&lt;/b&gt; - This field group adds the &lt;a href&#x3D;\&quot;/api-docs/buy/browse/resources/item/methods/getItemsByItemGroup#response.items.seller.userId\&quot;&gt;userId&lt;/a&gt; field to the response.
+func (r ApiGetItemsByItemGroupRequest) Fieldgroups(fieldgroups string) ApiGetItemsByItemGroupRequest {
+	r.fieldgroups = &fieldgroups
+	return r
+}
+
+// This header is required to support revenue sharing for eBay Partner Network and to improve the accuracy of shipping and delivery time estimations.&lt;br&gt;&lt;br&gt;For additional information, refer to &lt;a href&#x3D;\&quot;/api-docs/buy/static/api-browse.html#Headers\&quot; target&#x3D;\&quot;_blank \&quot;&gt;Use request headers&lt;/a&gt; in the &lt;a href&#x3D;\&quot;/api-docs/buy/static/buying-ig-landing.html\&quot; target&#x3D;\&quot;_blank\&quot;&gt;Buying Integration Guide&lt;/a&gt;.
+func (r ApiGetItemsByItemGroupRequest) XEBAYCENDUSERCTX(xEBAYCENDUSERCTX string) ApiGetItemsByItemGroupRequest {
+	r.xEBAYCENDUSERCTX = &xEBAYCENDUSERCTX
+	return r
+}
+
+// This header identifies the seller&#39;s eBay marketplace. It is required for all marketplaces outside of the US.&lt;br&gt;&lt;br&gt;&lt;span class&#x3D;\&quot;tablenote\&quot;&gt;&lt;b&gt;Note:&lt;/b&gt; If a marketplace ID value is not provided, the default value of &lt;code&gt;EBAY_US&lt;/code&gt; is used.&lt;/span&gt;&lt;br&gt;See &lt;a href&#x3D;\&quot;/api-docs/buy/browse/types/ba:MarketplaceIdEnum\&quot; target&#x3D;\&quot;_blank\&quot;&gt;MarketplaceIdEnum&lt;/a&gt; for a list of supported marketplaces.
+func (r ApiGetItemsByItemGroupRequest) XEBAYCMARKETPLACEID(xEBAYCMARKETPLACEID string) ApiGetItemsByItemGroupRequest {
+	r.xEBAYCMARKETPLACEID = &xEBAYCMARKETPLACEID
+	return r
+}
+
+// This header is used to indicate the natural language and locale preferred by the user for the response.&lt;br&gt;&lt;br&gt;This header is required when targeting a specific locale of a marketplace that supports multiple locales. For example:&lt;ul&gt;&lt;li&gt;When targeting the French locale of the Belgium marketplace, it is required to pass in &lt;code&gt;fr-BE&lt;/code&gt; to specify this. If this locale is not specified, the language will default to Dutch.&lt;/li&gt;&lt;li&gt;When targeting the French locale of the Canadian marketplace, it is required to pass in &lt;code&gt;fr-CA&lt;/code&gt; to specify this. If this locale is not specified, the language will default to English.&lt;/li&gt;&lt;/ul&gt;
+func (r ApiGetItemsByItemGroupRequest) AcceptLanguage(acceptLanguage string) ApiGetItemsByItemGroupRequest {
+	r.acceptLanguage = &acceptLanguage
+	return r
+}
+
+func (r ApiGetItemsByItemGroupRequest) Execute() (*ItemGroup, *http.Response, error) {
 	return r.ApiService.GetItemsByItemGroupExecute(r)
 }
 
 /*
 GetItemsByItemGroup Method for GetItemsByItemGroup
 
- <p>This method retrieves the details of the individual items in an item group. An item group is an item that has various aspect differences, such as color, size, storage capacity, etc. </p>  <p>You pass in the item group ID as a URI parameter. You use this method to show the item details of items with multiple aspects, such as color, size, storage capacity, etc.  </p>  <p>This method returns two main containers;  <b> items</b> and <b> commonDescriptions</b>. The <b> items</b> container has an array of  containers with the details of each item in the group. The <b> commonDescriptions</b> container has an array of containers for a description and the item ids of all the items that have this exact description. Because items within an item group often have the same description, this decreases the size of the response. </p>         <h3>URLs for this method</h3>           <p><ul>            <li><b> Production URL: </b> <code>https://api.ebay.com/buy/browse/v1/item/get_items_by_item_group?</code></li>            <li><b> Sandbox URL:  </b><code>https://api.sandbox.ebay.com/buy/browse/v1/item/get_items_by_item_group?</code></li>           </ul>    </p>            <h3><b> Request headers</b></h3> This method uses the  <b>X-EBAY-C-ENDUSERCTX</b> request header to support revenue sharing for eBay Partner Networks and to improve the accuracy of shipping and delivery time estimations.   For details see, <a href="/api-docs/buy/static/api-browse.html#Headers">Request headers</a> in the Buying Integration Guide.   <h3><b> Restrictions </b></h3> <p>For a list of supported sites and other restrictions, see <a href="/api-docs/buy/browse/overview.html#API">API Restrictions</a>.</p> <span class="tablenote"><b>eBay Partner Network: </b> In order to be commissioned for your sales, you must use the URL returned in the <code>itemAffiliateWebUrl</code> field to forward your buyer to the ebay.com site. </span>
+This method retrieves details about individual items in an item group. An item group is an item that has various aspect differences, such as color, size, storage capacity, etc.<br><br>You pass in the <b>item_group_id</b> as a URI parameter.<br><br>This method returns two main containers:<ul><li><b>items</b><br>This container has an array of containers with the details about each item in the group.</li><li><b>commonDescriptions</b><br>This container has an array of containers for a description and the <code>item_ids</code> for all items that have this exact description. Because items within an item group often have the same description, this decreases the size of the response.</li></ul>Setting the <b>fieldgroup</b> to <code>ADDITIONAL_SELLER_DETAILS</code> adds an additional field to the response that returns the seller's user ID. For more information, refer to <a href="/api-docs/buy/browse/resources/item/methods/getItemsByItemGroup#uri.fieldgroups">fieldgroups</a>.<h3>Restrictions</h3>For a list of supported sites and other restrictions, refer to <a href="/api-docs/buy/browse/overview.html#API">API Restrictions</a>.<br><br><span class="tablenote"><b>eBay Partner Network: </b> In order to be commissioned for your sales, you must use the URL returned in the <code>itemAffiliateWebUrl</code> field to forward your buyer to the ebay.com site.</span>
 
- @param ctx _context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
+ @param ctx context.Context - for authentication, logging, cancellation, deadlines, tracing, etc. Passed from http.Request or context.Background().
  @return ApiGetItemsByItemGroupRequest
 */
-func (a *ItemApiService) GetItemsByItemGroup(ctx _context.Context) *ApiGetItemsByItemGroupRequest {
-	return &ApiGetItemsByItemGroupRequest{
+func (a *ItemAPIService) GetItemsByItemGroup(ctx context.Context) ApiGetItemsByItemGroupRequest {
+	return ApiGetItemsByItemGroupRequest{
 		ApiService: a,
-		ctx:        ctx,
+		ctx: ctx,
 	}
 }
 
 // Execute executes the request
 //  @return ItemGroup
-func (a *ItemApiService) GetItemsByItemGroupExecute(r *ApiGetItemsByItemGroupRequest) (ItemGroup, *_nethttp.Response, error) {
+func (a *ItemAPIService) GetItemsByItemGroupExecute(r ApiGetItemsByItemGroupRequest) (*ItemGroup, *http.Response, error) {
 	var (
-		localVarHTTPMethod  = _nethttp.MethodGet
-		localVarPostBody    interface{}
-		formFiles           []formFile
-		localVarReturnValue ItemGroup
+		localVarHTTPMethod   = http.MethodGet
+		localVarPostBody     interface{}
+		formFiles            []formFile
+		localVarReturnValue  *ItemGroup
 	)
 
-	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "ItemApiService.GetItemsByItemGroup")
+	localBasePath, err := a.client.cfg.ServerURLWithContext(r.ctx, "ItemAPIService.GetItemsByItemGroup")
 	if err != nil {
-		return localVarReturnValue, nil, GenericOpenAPIError{error: err.Error()}
+		return localVarReturnValue, nil, &GenericOpenAPIError{error: err.Error()}
 	}
 
 	localVarPath := localBasePath + "/item/get_items_by_item_group"
 
 	localVarHeaderParams := make(map[string]string)
-	localVarQueryParams := _neturl.Values{}
-	localVarFormParams := _neturl.Values{}
+	localVarQueryParams := url.Values{}
+	localVarFormParams := url.Values{}
 	if r.itemGroupId == nil {
 		return localVarReturnValue, nil, reportError("itemGroupId is required and must be specified")
 	}
 
-	localVarQueryParams.Add("item_group_id", parameterToString(*r.itemGroupId, ""))
+	if r.fieldgroups != nil {
+		parameterAddToHeaderOrQuery(localVarQueryParams, "fieldgroups", r.fieldgroups, "form", "")
+	}
+	parameterAddToHeaderOrQuery(localVarQueryParams, "item_group_id", r.itemGroupId, "form", "")
 	// to determine the Content-Type header
 	localVarHTTPContentTypes := []string{}
 
@@ -593,6 +731,15 @@ func (a *ItemApiService) GetItemsByItemGroupExecute(r *ApiGetItemsByItemGroupReq
 	if localVarHTTPHeaderAccept != "" {
 		localVarHeaderParams["Accept"] = localVarHTTPHeaderAccept
 	}
+	if r.xEBAYCENDUSERCTX != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "X-EBAY-C-ENDUSERCTX", r.xEBAYCENDUSERCTX, "simple", "")
+	}
+	if r.xEBAYCMARKETPLACEID != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "X-EBAY-C-MARKETPLACE-ID", r.xEBAYCMARKETPLACEID, "simple", "")
+	}
+	if r.acceptLanguage != nil {
+		parameterAddToHeaderOrQuery(localVarHeaderParams, "Accept-Language", r.acceptLanguage, "simple", "")
+	}
 	req, err := a.client.prepareRequest(r.ctx, localVarPath, localVarHTTPMethod, localVarPostBody, localVarHeaderParams, localVarQueryParams, localVarFormParams, formFiles)
 	if err != nil {
 		return localVarReturnValue, nil, err
@@ -603,15 +750,15 @@ func (a *ItemApiService) GetItemsByItemGroupExecute(r *ApiGetItemsByItemGroupReq
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
-	localVarBody, err := _ioutil.ReadAll(localVarHTTPResponse.Body)
+	localVarBody, err := io.ReadAll(localVarHTTPResponse.Body)
 	localVarHTTPResponse.Body.Close()
-	localVarHTTPResponse.Body = _ioutil.NopCloser(bytes.NewBuffer(localVarBody))
+	localVarHTTPResponse.Body = io.NopCloser(bytes.NewBuffer(localVarBody))
 	if err != nil {
 		return localVarReturnValue, localVarHTTPResponse, err
 	}
 
 	if localVarHTTPResponse.StatusCode >= 300 {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: localVarHTTPResponse.Status,
 		}
@@ -620,7 +767,7 @@ func (a *ItemApiService) GetItemsByItemGroupExecute(r *ApiGetItemsByItemGroupReq
 
 	err = a.client.decode(&localVarReturnValue, localVarBody, localVarHTTPResponse.Header.Get("Content-Type"))
 	if err != nil {
-		newErr := GenericOpenAPIError{
+		newErr := &GenericOpenAPIError{
 			body:  localVarBody,
 			error: err.Error(),
 		}
