@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,12 +14,15 @@ import (
 	"encoding/json"
 )
 
+// checks if the Aspect type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &Aspect{}
+
 // Aspect The type that defines the fields for the name/value pairs for the aspects of the product. For example: BRAND/Apple
 type Aspect struct {
 	// The text representing the name of the aspect for the name/value pair, such as Brand.
 	LocalizedName *string `json:"localizedName,omitempty"`
 	// The text representing the value of the aspect for the name/value pair, such as Apple.
-	LocalizedValues *[]string `json:"localizedValues,omitempty"`
+	LocalizedValues []string `json:"localizedValues,omitempty"`
 }
 
 // NewAspect instantiates a new Aspect object
@@ -41,7 +44,7 @@ func NewAspectWithDefaults() *Aspect {
 
 // GetLocalizedName returns the LocalizedName field value if set, zero value otherwise.
 func (o *Aspect) GetLocalizedName() string {
-	if o == nil || o.LocalizedName == nil {
+	if o == nil || IsNil(o.LocalizedName) {
 		var ret string
 		return ret
 	}
@@ -51,7 +54,7 @@ func (o *Aspect) GetLocalizedName() string {
 // GetLocalizedNameOk returns a tuple with the LocalizedName field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Aspect) GetLocalizedNameOk() (*string, bool) {
-	if o == nil || o.LocalizedName == nil {
+	if o == nil || IsNil(o.LocalizedName) {
 		return nil, false
 	}
 	return o.LocalizedName, true
@@ -59,7 +62,7 @@ func (o *Aspect) GetLocalizedNameOk() (*string, bool) {
 
 // HasLocalizedName returns a boolean if a field has been set.
 func (o *Aspect) HasLocalizedName() bool {
-	if o != nil && o.LocalizedName != nil {
+	if o != nil && !IsNil(o.LocalizedName) {
 		return true
 	}
 
@@ -73,17 +76,17 @@ func (o *Aspect) SetLocalizedName(v string) {
 
 // GetLocalizedValues returns the LocalizedValues field value if set, zero value otherwise.
 func (o *Aspect) GetLocalizedValues() []string {
-	if o == nil || o.LocalizedValues == nil {
+	if o == nil || IsNil(o.LocalizedValues) {
 		var ret []string
 		return ret
 	}
-	return *o.LocalizedValues
+	return o.LocalizedValues
 }
 
 // GetLocalizedValuesOk returns a tuple with the LocalizedValues field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Aspect) GetLocalizedValuesOk() (*[]string, bool) {
-	if o == nil || o.LocalizedValues == nil {
+func (o *Aspect) GetLocalizedValuesOk() ([]string, bool) {
+	if o == nil || IsNil(o.LocalizedValues) {
 		return nil, false
 	}
 	return o.LocalizedValues, true
@@ -91,7 +94,7 @@ func (o *Aspect) GetLocalizedValuesOk() (*[]string, bool) {
 
 // HasLocalizedValues returns a boolean if a field has been set.
 func (o *Aspect) HasLocalizedValues() bool {
-	if o != nil && o.LocalizedValues != nil {
+	if o != nil && !IsNil(o.LocalizedValues) {
 		return true
 	}
 
@@ -100,18 +103,26 @@ func (o *Aspect) HasLocalizedValues() bool {
 
 // SetLocalizedValues gets a reference to the given []string and assigns it to the LocalizedValues field.
 func (o *Aspect) SetLocalizedValues(v []string) {
-	o.LocalizedValues = &v
+	o.LocalizedValues = v
 }
 
 func (o Aspect) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.LocalizedName != nil {
-		toSerialize["localizedName"] = o.LocalizedName
-	}
-	if o.LocalizedValues != nil {
-		toSerialize["localizedValues"] = o.LocalizedValues
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o Aspect) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.LocalizedName) {
+		toSerialize["localizedName"] = o.LocalizedName
+	}
+	if !IsNil(o.LocalizedValues) {
+		toSerialize["localizedValues"] = o.LocalizedValues
+	}
+	return toSerialize, nil
 }
 
 type NullableAspect struct {
