@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,11 +14,14 @@ import (
 	"encoding/json"
 )
 
+// checks if the AspectDistribution type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &AspectDistribution{}
+
 // AspectDistribution The type that define the fields for the aspect information. Aspects are the variations of an item, such as color, size, etc.
 type AspectDistribution struct {
-	// An array of containers for the various values of the aspect and the match count and a HATEOAS reference (<b> refinementHref</b>) for this aspect.
-	AspectValueDistributions *[]AspectValueDistribution `json:"aspectValueDistributions,omitempty"`
-	// The name of an aspect, such as Brand, Color, etc.
+	// An array of containers for the various values of the aspect and the match count, and a HATEOAS reference (<code>refinementHref</code>) for this aspect.
+	AspectValueDistributions []AspectValueDistribution `json:"aspectValueDistributions,omitempty"`
+	// The name of an aspect, such as <i>Brand</i>, <i>Color</i>, etc.
 	LocalizedAspectName *string `json:"localizedAspectName,omitempty"`
 }
 
@@ -41,17 +44,17 @@ func NewAspectDistributionWithDefaults() *AspectDistribution {
 
 // GetAspectValueDistributions returns the AspectValueDistributions field value if set, zero value otherwise.
 func (o *AspectDistribution) GetAspectValueDistributions() []AspectValueDistribution {
-	if o == nil || o.AspectValueDistributions == nil {
+	if o == nil || IsNil(o.AspectValueDistributions) {
 		var ret []AspectValueDistribution
 		return ret
 	}
-	return *o.AspectValueDistributions
+	return o.AspectValueDistributions
 }
 
 // GetAspectValueDistributionsOk returns a tuple with the AspectValueDistributions field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *AspectDistribution) GetAspectValueDistributionsOk() (*[]AspectValueDistribution, bool) {
-	if o == nil || o.AspectValueDistributions == nil {
+func (o *AspectDistribution) GetAspectValueDistributionsOk() ([]AspectValueDistribution, bool) {
+	if o == nil || IsNil(o.AspectValueDistributions) {
 		return nil, false
 	}
 	return o.AspectValueDistributions, true
@@ -59,7 +62,7 @@ func (o *AspectDistribution) GetAspectValueDistributionsOk() (*[]AspectValueDist
 
 // HasAspectValueDistributions returns a boolean if a field has been set.
 func (o *AspectDistribution) HasAspectValueDistributions() bool {
-	if o != nil && o.AspectValueDistributions != nil {
+	if o != nil && !IsNil(o.AspectValueDistributions) {
 		return true
 	}
 
@@ -68,12 +71,12 @@ func (o *AspectDistribution) HasAspectValueDistributions() bool {
 
 // SetAspectValueDistributions gets a reference to the given []AspectValueDistribution and assigns it to the AspectValueDistributions field.
 func (o *AspectDistribution) SetAspectValueDistributions(v []AspectValueDistribution) {
-	o.AspectValueDistributions = &v
+	o.AspectValueDistributions = v
 }
 
 // GetLocalizedAspectName returns the LocalizedAspectName field value if set, zero value otherwise.
 func (o *AspectDistribution) GetLocalizedAspectName() string {
-	if o == nil || o.LocalizedAspectName == nil {
+	if o == nil || IsNil(o.LocalizedAspectName) {
 		var ret string
 		return ret
 	}
@@ -83,7 +86,7 @@ func (o *AspectDistribution) GetLocalizedAspectName() string {
 // GetLocalizedAspectNameOk returns a tuple with the LocalizedAspectName field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AspectDistribution) GetLocalizedAspectNameOk() (*string, bool) {
-	if o == nil || o.LocalizedAspectName == nil {
+	if o == nil || IsNil(o.LocalizedAspectName) {
 		return nil, false
 	}
 	return o.LocalizedAspectName, true
@@ -91,7 +94,7 @@ func (o *AspectDistribution) GetLocalizedAspectNameOk() (*string, bool) {
 
 // HasLocalizedAspectName returns a boolean if a field has been set.
 func (o *AspectDistribution) HasLocalizedAspectName() bool {
-	if o != nil && o.LocalizedAspectName != nil {
+	if o != nil && !IsNil(o.LocalizedAspectName) {
 		return true
 	}
 
@@ -104,14 +107,22 @@ func (o *AspectDistribution) SetLocalizedAspectName(v string) {
 }
 
 func (o AspectDistribution) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.AspectValueDistributions != nil {
-		toSerialize["aspectValueDistributions"] = o.AspectValueDistributions
-	}
-	if o.LocalizedAspectName != nil {
-		toSerialize["localizedAspectName"] = o.LocalizedAspectName
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o AspectDistribution) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.AspectValueDistributions) {
+		toSerialize["aspectValueDistributions"] = o.AspectValueDistributions
+	}
+	if !IsNil(o.LocalizedAspectName) {
+		toSerialize["localizedAspectName"] = o.LocalizedAspectName
+	}
+	return toSerialize, nil
 }
 
 type NullableAspectDistribution struct {
