@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,11 +14,14 @@ import (
 	"encoding/json"
 )
 
+// checks if the AspectGroup type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &AspectGroup{}
+
 // AspectGroup struct for AspectGroup
 type AspectGroup struct {
 	// An array of the name/value pairs for the aspects of the product. For example: BRAND/Apple
-	Aspects *[]Aspect `json:"aspects,omitempty"`
-	// The name of a group of aspects. <br /><br />In the following example, <b> Product Identifiers</b> and <b> Process</b> are product aspect group names. Under the group name are the product aspect name/value pairs. <p><b> Product Identifiers</b> <br />&nbsp;&nbsp;&nbsp;Brand/Apple <br />&nbsp;&nbsp;&nbsp;Product Family/iMac</p> <p><b> Processor</b><br />&nbsp;&nbsp;&nbsp;Processor Type/Intel <br />&nbsp;&nbsp;&nbsp;Processor Speed/3.10</p> 
+	Aspects []Aspect `json:"aspects,omitempty"`
+	// The name of a group of aspects.<br><br>In the following example, <b>Product Identifiers</b> and <b>Process</b> are product aspect group names. Under the group name are the product aspect name/value pairs.<br><br><b>Product Identifiers</b><br>&nbsp;&nbsp;&nbsp;Brand/Apple<br>&nbsp;&nbsp;&nbsp;Product Family/iMac<br><br><b>Processor</b><br>&nbsp;&nbsp;&nbsp;Processor Type/Intel<br>&nbsp;&nbsp;&nbsp;Processor Speed/3.10
 	LocalizedGroupName *string `json:"localizedGroupName,omitempty"`
 }
 
@@ -41,17 +44,17 @@ func NewAspectGroupWithDefaults() *AspectGroup {
 
 // GetAspects returns the Aspects field value if set, zero value otherwise.
 func (o *AspectGroup) GetAspects() []Aspect {
-	if o == nil || o.Aspects == nil {
+	if o == nil || IsNil(o.Aspects) {
 		var ret []Aspect
 		return ret
 	}
-	return *o.Aspects
+	return o.Aspects
 }
 
 // GetAspectsOk returns a tuple with the Aspects field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *AspectGroup) GetAspectsOk() (*[]Aspect, bool) {
-	if o == nil || o.Aspects == nil {
+func (o *AspectGroup) GetAspectsOk() ([]Aspect, bool) {
+	if o == nil || IsNil(o.Aspects) {
 		return nil, false
 	}
 	return o.Aspects, true
@@ -59,7 +62,7 @@ func (o *AspectGroup) GetAspectsOk() (*[]Aspect, bool) {
 
 // HasAspects returns a boolean if a field has been set.
 func (o *AspectGroup) HasAspects() bool {
-	if o != nil && o.Aspects != nil {
+	if o != nil && !IsNil(o.Aspects) {
 		return true
 	}
 
@@ -68,12 +71,12 @@ func (o *AspectGroup) HasAspects() bool {
 
 // SetAspects gets a reference to the given []Aspect and assigns it to the Aspects field.
 func (o *AspectGroup) SetAspects(v []Aspect) {
-	o.Aspects = &v
+	o.Aspects = v
 }
 
 // GetLocalizedGroupName returns the LocalizedGroupName field value if set, zero value otherwise.
 func (o *AspectGroup) GetLocalizedGroupName() string {
-	if o == nil || o.LocalizedGroupName == nil {
+	if o == nil || IsNil(o.LocalizedGroupName) {
 		var ret string
 		return ret
 	}
@@ -83,7 +86,7 @@ func (o *AspectGroup) GetLocalizedGroupName() string {
 // GetLocalizedGroupNameOk returns a tuple with the LocalizedGroupName field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AspectGroup) GetLocalizedGroupNameOk() (*string, bool) {
-	if o == nil || o.LocalizedGroupName == nil {
+	if o == nil || IsNil(o.LocalizedGroupName) {
 		return nil, false
 	}
 	return o.LocalizedGroupName, true
@@ -91,7 +94,7 @@ func (o *AspectGroup) GetLocalizedGroupNameOk() (*string, bool) {
 
 // HasLocalizedGroupName returns a boolean if a field has been set.
 func (o *AspectGroup) HasLocalizedGroupName() bool {
-	if o != nil && o.LocalizedGroupName != nil {
+	if o != nil && !IsNil(o.LocalizedGroupName) {
 		return true
 	}
 
@@ -104,14 +107,22 @@ func (o *AspectGroup) SetLocalizedGroupName(v string) {
 }
 
 func (o AspectGroup) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.Aspects != nil {
-		toSerialize["aspects"] = o.Aspects
-	}
-	if o.LocalizedGroupName != nil {
-		toSerialize["localizedGroupName"] = o.LocalizedGroupName
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o AspectGroup) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.Aspects) {
+		toSerialize["aspects"] = o.Aspects
+	}
+	if !IsNil(o.LocalizedGroupName) {
+		toSerialize["localizedGroupName"] = o.LocalizedGroupName
+	}
+	return toSerialize, nil
 }
 
 type NullableAspectGroup struct {
