@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,7 +14,10 @@ import (
 	"encoding/json"
 )
 
-// AspectValueDistribution The container that defines the fields for the conditions refinements. This container is returned when <b> fieldgroups</b> is set to <code>ASPECT_REFINEMENTS</code> or <code>FULL</code> in the request.
+// checks if the AspectValueDistribution type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &AspectValueDistribution{}
+
+// AspectValueDistribution The container that defines the fields for the conditions refinements. This container is returned when <code>fieldgroups</code> is set to <code>ASPECT_REFINEMENTS</code> or <code>FULL</code> in the request.
 type AspectValueDistribution struct {
 	// The value of an aspect. For example, Red is a value for the aspect Color.
 	LocalizedAspectValue *string `json:"localizedAspectValue,omitempty"`
@@ -43,7 +46,7 @@ func NewAspectValueDistributionWithDefaults() *AspectValueDistribution {
 
 // GetLocalizedAspectValue returns the LocalizedAspectValue field value if set, zero value otherwise.
 func (o *AspectValueDistribution) GetLocalizedAspectValue() string {
-	if o == nil || o.LocalizedAspectValue == nil {
+	if o == nil || IsNil(o.LocalizedAspectValue) {
 		var ret string
 		return ret
 	}
@@ -53,7 +56,7 @@ func (o *AspectValueDistribution) GetLocalizedAspectValue() string {
 // GetLocalizedAspectValueOk returns a tuple with the LocalizedAspectValue field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AspectValueDistribution) GetLocalizedAspectValueOk() (*string, bool) {
-	if o == nil || o.LocalizedAspectValue == nil {
+	if o == nil || IsNil(o.LocalizedAspectValue) {
 		return nil, false
 	}
 	return o.LocalizedAspectValue, true
@@ -61,7 +64,7 @@ func (o *AspectValueDistribution) GetLocalizedAspectValueOk() (*string, bool) {
 
 // HasLocalizedAspectValue returns a boolean if a field has been set.
 func (o *AspectValueDistribution) HasLocalizedAspectValue() bool {
-	if o != nil && o.LocalizedAspectValue != nil {
+	if o != nil && !IsNil(o.LocalizedAspectValue) {
 		return true
 	}
 
@@ -75,7 +78,7 @@ func (o *AspectValueDistribution) SetLocalizedAspectValue(v string) {
 
 // GetMatchCount returns the MatchCount field value if set, zero value otherwise.
 func (o *AspectValueDistribution) GetMatchCount() int32 {
-	if o == nil || o.MatchCount == nil {
+	if o == nil || IsNil(o.MatchCount) {
 		var ret int32
 		return ret
 	}
@@ -85,7 +88,7 @@ func (o *AspectValueDistribution) GetMatchCount() int32 {
 // GetMatchCountOk returns a tuple with the MatchCount field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AspectValueDistribution) GetMatchCountOk() (*int32, bool) {
-	if o == nil || o.MatchCount == nil {
+	if o == nil || IsNil(o.MatchCount) {
 		return nil, false
 	}
 	return o.MatchCount, true
@@ -93,7 +96,7 @@ func (o *AspectValueDistribution) GetMatchCountOk() (*int32, bool) {
 
 // HasMatchCount returns a boolean if a field has been set.
 func (o *AspectValueDistribution) HasMatchCount() bool {
-	if o != nil && o.MatchCount != nil {
+	if o != nil && !IsNil(o.MatchCount) {
 		return true
 	}
 
@@ -107,7 +110,7 @@ func (o *AspectValueDistribution) SetMatchCount(v int32) {
 
 // GetRefinementHref returns the RefinementHref field value if set, zero value otherwise.
 func (o *AspectValueDistribution) GetRefinementHref() string {
-	if o == nil || o.RefinementHref == nil {
+	if o == nil || IsNil(o.RefinementHref) {
 		var ret string
 		return ret
 	}
@@ -117,7 +120,7 @@ func (o *AspectValueDistribution) GetRefinementHref() string {
 // GetRefinementHrefOk returns a tuple with the RefinementHref field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AspectValueDistribution) GetRefinementHrefOk() (*string, bool) {
-	if o == nil || o.RefinementHref == nil {
+	if o == nil || IsNil(o.RefinementHref) {
 		return nil, false
 	}
 	return o.RefinementHref, true
@@ -125,7 +128,7 @@ func (o *AspectValueDistribution) GetRefinementHrefOk() (*string, bool) {
 
 // HasRefinementHref returns a boolean if a field has been set.
 func (o *AspectValueDistribution) HasRefinementHref() bool {
-	if o != nil && o.RefinementHref != nil {
+	if o != nil && !IsNil(o.RefinementHref) {
 		return true
 	}
 
@@ -138,17 +141,25 @@ func (o *AspectValueDistribution) SetRefinementHref(v string) {
 }
 
 func (o AspectValueDistribution) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.LocalizedAspectValue != nil {
-		toSerialize["localizedAspectValue"] = o.LocalizedAspectValue
-	}
-	if o.MatchCount != nil {
-		toSerialize["matchCount"] = o.MatchCount
-	}
-	if o.RefinementHref != nil {
-		toSerialize["refinementHref"] = o.RefinementHref
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o AspectValueDistribution) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.LocalizedAspectValue) {
+		toSerialize["localizedAspectValue"] = o.LocalizedAspectValue
+	}
+	if !IsNil(o.MatchCount) {
+		toSerialize["matchCount"] = o.MatchCount
+	}
+	if !IsNil(o.RefinementHref) {
+		toSerialize["refinementHref"] = o.RefinementHref
+	}
+	return toSerialize, nil
 }
 
 type NullableAspectValueDistribution struct {
