@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,17 +14,20 @@ import (
 	"encoding/json"
 )
 
+// checks if the AvailableCoupon type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &AvailableCoupon{}
+
 // AvailableCoupon struct for AvailableCoupon
 type AvailableCoupon struct {
 	Constraint *CouponConstraint `json:"constraint,omitempty"`
 	DiscountAmount *Amount `json:"discountAmount,omitempty"`
 	// The type of discount that the coupon applies. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:CouponDiscountType'>eBay API documentation</a>
 	DiscountType *string `json:"discountType,omitempty"`
-	// A description of the coupon.<br /><br /><span class=\"tablenote\"><b> Note: </b> The value returned in the <b>termsWebUrl</b> field should appear for all experiences when displaying coupons. The value in the <b>availableCoupons.message</b> field must also be included, if returned in the API response.</span>
+	// A description of the coupon.<br><br><span class=\"tablenote\"><b>Note:</b> The value returned in the <code>termsWebUrl</code> field should appear for all experiences when displaying coupons. The value in the <code>availableCoupons.message</code> field must also be included if returned in the API response.</span>
 	Message *string `json:"message,omitempty"`
 	// The coupon code.
 	RedemptionCode *string `json:"redemptionCode,omitempty"`
-	// The URL to the coupon terms of use.<br /><br /><span class=\"tablenote\"><b> Note: </b> The value returned in the <b>termsWebUrl</b> field should appear for all experiences when displaying coupons. The value in the <b>availableCoupons.message</b> field must also be included, if returned in the API response.</span>
+	// The URL to the coupon terms of use.<br><br><span class=\"tablenote\"><b>Note:</b> The value returned in the <code>termsWebUrl</code> field should appear for all experiences when displaying coupons. The value in the <code>availableCoupons.message</code> field must also be included if returned in the API response.</span>
 	TermsWebUrl *string `json:"termsWebUrl,omitempty"`
 }
 
@@ -47,7 +50,7 @@ func NewAvailableCouponWithDefaults() *AvailableCoupon {
 
 // GetConstraint returns the Constraint field value if set, zero value otherwise.
 func (o *AvailableCoupon) GetConstraint() CouponConstraint {
-	if o == nil || o.Constraint == nil {
+	if o == nil || IsNil(o.Constraint) {
 		var ret CouponConstraint
 		return ret
 	}
@@ -57,7 +60,7 @@ func (o *AvailableCoupon) GetConstraint() CouponConstraint {
 // GetConstraintOk returns a tuple with the Constraint field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AvailableCoupon) GetConstraintOk() (*CouponConstraint, bool) {
-	if o == nil || o.Constraint == nil {
+	if o == nil || IsNil(o.Constraint) {
 		return nil, false
 	}
 	return o.Constraint, true
@@ -65,7 +68,7 @@ func (o *AvailableCoupon) GetConstraintOk() (*CouponConstraint, bool) {
 
 // HasConstraint returns a boolean if a field has been set.
 func (o *AvailableCoupon) HasConstraint() bool {
-	if o != nil && o.Constraint != nil {
+	if o != nil && !IsNil(o.Constraint) {
 		return true
 	}
 
@@ -79,7 +82,7 @@ func (o *AvailableCoupon) SetConstraint(v CouponConstraint) {
 
 // GetDiscountAmount returns the DiscountAmount field value if set, zero value otherwise.
 func (o *AvailableCoupon) GetDiscountAmount() Amount {
-	if o == nil || o.DiscountAmount == nil {
+	if o == nil || IsNil(o.DiscountAmount) {
 		var ret Amount
 		return ret
 	}
@@ -89,7 +92,7 @@ func (o *AvailableCoupon) GetDiscountAmount() Amount {
 // GetDiscountAmountOk returns a tuple with the DiscountAmount field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AvailableCoupon) GetDiscountAmountOk() (*Amount, bool) {
-	if o == nil || o.DiscountAmount == nil {
+	if o == nil || IsNil(o.DiscountAmount) {
 		return nil, false
 	}
 	return o.DiscountAmount, true
@@ -97,7 +100,7 @@ func (o *AvailableCoupon) GetDiscountAmountOk() (*Amount, bool) {
 
 // HasDiscountAmount returns a boolean if a field has been set.
 func (o *AvailableCoupon) HasDiscountAmount() bool {
-	if o != nil && o.DiscountAmount != nil {
+	if o != nil && !IsNil(o.DiscountAmount) {
 		return true
 	}
 
@@ -111,7 +114,7 @@ func (o *AvailableCoupon) SetDiscountAmount(v Amount) {
 
 // GetDiscountType returns the DiscountType field value if set, zero value otherwise.
 func (o *AvailableCoupon) GetDiscountType() string {
-	if o == nil || o.DiscountType == nil {
+	if o == nil || IsNil(o.DiscountType) {
 		var ret string
 		return ret
 	}
@@ -121,7 +124,7 @@ func (o *AvailableCoupon) GetDiscountType() string {
 // GetDiscountTypeOk returns a tuple with the DiscountType field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AvailableCoupon) GetDiscountTypeOk() (*string, bool) {
-	if o == nil || o.DiscountType == nil {
+	if o == nil || IsNil(o.DiscountType) {
 		return nil, false
 	}
 	return o.DiscountType, true
@@ -129,7 +132,7 @@ func (o *AvailableCoupon) GetDiscountTypeOk() (*string, bool) {
 
 // HasDiscountType returns a boolean if a field has been set.
 func (o *AvailableCoupon) HasDiscountType() bool {
-	if o != nil && o.DiscountType != nil {
+	if o != nil && !IsNil(o.DiscountType) {
 		return true
 	}
 
@@ -143,7 +146,7 @@ func (o *AvailableCoupon) SetDiscountType(v string) {
 
 // GetMessage returns the Message field value if set, zero value otherwise.
 func (o *AvailableCoupon) GetMessage() string {
-	if o == nil || o.Message == nil {
+	if o == nil || IsNil(o.Message) {
 		var ret string
 		return ret
 	}
@@ -153,7 +156,7 @@ func (o *AvailableCoupon) GetMessage() string {
 // GetMessageOk returns a tuple with the Message field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AvailableCoupon) GetMessageOk() (*string, bool) {
-	if o == nil || o.Message == nil {
+	if o == nil || IsNil(o.Message) {
 		return nil, false
 	}
 	return o.Message, true
@@ -161,7 +164,7 @@ func (o *AvailableCoupon) GetMessageOk() (*string, bool) {
 
 // HasMessage returns a boolean if a field has been set.
 func (o *AvailableCoupon) HasMessage() bool {
-	if o != nil && o.Message != nil {
+	if o != nil && !IsNil(o.Message) {
 		return true
 	}
 
@@ -175,7 +178,7 @@ func (o *AvailableCoupon) SetMessage(v string) {
 
 // GetRedemptionCode returns the RedemptionCode field value if set, zero value otherwise.
 func (o *AvailableCoupon) GetRedemptionCode() string {
-	if o == nil || o.RedemptionCode == nil {
+	if o == nil || IsNil(o.RedemptionCode) {
 		var ret string
 		return ret
 	}
@@ -185,7 +188,7 @@ func (o *AvailableCoupon) GetRedemptionCode() string {
 // GetRedemptionCodeOk returns a tuple with the RedemptionCode field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AvailableCoupon) GetRedemptionCodeOk() (*string, bool) {
-	if o == nil || o.RedemptionCode == nil {
+	if o == nil || IsNil(o.RedemptionCode) {
 		return nil, false
 	}
 	return o.RedemptionCode, true
@@ -193,7 +196,7 @@ func (o *AvailableCoupon) GetRedemptionCodeOk() (*string, bool) {
 
 // HasRedemptionCode returns a boolean if a field has been set.
 func (o *AvailableCoupon) HasRedemptionCode() bool {
-	if o != nil && o.RedemptionCode != nil {
+	if o != nil && !IsNil(o.RedemptionCode) {
 		return true
 	}
 
@@ -207,7 +210,7 @@ func (o *AvailableCoupon) SetRedemptionCode(v string) {
 
 // GetTermsWebUrl returns the TermsWebUrl field value if set, zero value otherwise.
 func (o *AvailableCoupon) GetTermsWebUrl() string {
-	if o == nil || o.TermsWebUrl == nil {
+	if o == nil || IsNil(o.TermsWebUrl) {
 		var ret string
 		return ret
 	}
@@ -217,7 +220,7 @@ func (o *AvailableCoupon) GetTermsWebUrl() string {
 // GetTermsWebUrlOk returns a tuple with the TermsWebUrl field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *AvailableCoupon) GetTermsWebUrlOk() (*string, bool) {
-	if o == nil || o.TermsWebUrl == nil {
+	if o == nil || IsNil(o.TermsWebUrl) {
 		return nil, false
 	}
 	return o.TermsWebUrl, true
@@ -225,7 +228,7 @@ func (o *AvailableCoupon) GetTermsWebUrlOk() (*string, bool) {
 
 // HasTermsWebUrl returns a boolean if a field has been set.
 func (o *AvailableCoupon) HasTermsWebUrl() bool {
-	if o != nil && o.TermsWebUrl != nil {
+	if o != nil && !IsNil(o.TermsWebUrl) {
 		return true
 	}
 
@@ -238,26 +241,34 @@ func (o *AvailableCoupon) SetTermsWebUrl(v string) {
 }
 
 func (o AvailableCoupon) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.Constraint != nil {
-		toSerialize["constraint"] = o.Constraint
-	}
-	if o.DiscountAmount != nil {
-		toSerialize["discountAmount"] = o.DiscountAmount
-	}
-	if o.DiscountType != nil {
-		toSerialize["discountType"] = o.DiscountType
-	}
-	if o.Message != nil {
-		toSerialize["message"] = o.Message
-	}
-	if o.RedemptionCode != nil {
-		toSerialize["redemptionCode"] = o.RedemptionCode
-	}
-	if o.TermsWebUrl != nil {
-		toSerialize["termsWebUrl"] = o.TermsWebUrl
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o AvailableCoupon) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.Constraint) {
+		toSerialize["constraint"] = o.Constraint
+	}
+	if !IsNil(o.DiscountAmount) {
+		toSerialize["discountAmount"] = o.DiscountAmount
+	}
+	if !IsNil(o.DiscountType) {
+		toSerialize["discountType"] = o.DiscountType
+	}
+	if !IsNil(o.Message) {
+		toSerialize["message"] = o.Message
+	}
+	if !IsNil(o.RedemptionCode) {
+		toSerialize["redemptionCode"] = o.RedemptionCode
+	}
+	if !IsNil(o.TermsWebUrl) {
+		toSerialize["termsWebUrl"] = o.TermsWebUrl
+	}
+	return toSerialize, nil
 }
 
 type NullableAvailableCoupon struct {
