@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,9 +14,12 @@ import (
 	"encoding/json"
 )
 
-// BuyingOptionDistribution The container that defines the fields for the buying options refinements. This container is returned when <b> fieldgroups</b> is set to <code>BUYING_OPTION_REFINEMENTS</code> or <code>FULL</code> in the request.
+// checks if the BuyingOptionDistribution type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &BuyingOptionDistribution{}
+
+// BuyingOptionDistribution The container that defines the fields for the buying options refinements. This container is returned when <code>fieldgroups</code> is set to <code>BUYING_OPTION_REFINEMENTS</code> or <code>FULL</code> in the request.
 type BuyingOptionDistribution struct {
-	// The container that returns the buying option type. This will be AUCTION or FIXED_PRICE or both. For details, see <a href=\"/api-docs/buy/browse/resources/item_summary/methods/search#response.itemSummaries.buyingOptions\">buyingOptions</a>.
+	// The container that returns the buying option type. This will be AUCTION, FIXED_PRICE, CLASSIFIED_AD, or a combination of these options. For details, see <a href=\"/api-docs/buy/browse/resources/item_summary/methods/search#response.itemSummaries.buyingOptions\">buyingOptions</a>.
 	BuyingOption *string `json:"buyingOption,omitempty"`
 	// The number of items having this buying option.
 	MatchCount *int32 `json:"matchCount,omitempty"`
@@ -43,7 +46,7 @@ func NewBuyingOptionDistributionWithDefaults() *BuyingOptionDistribution {
 
 // GetBuyingOption returns the BuyingOption field value if set, zero value otherwise.
 func (o *BuyingOptionDistribution) GetBuyingOption() string {
-	if o == nil || o.BuyingOption == nil {
+	if o == nil || IsNil(o.BuyingOption) {
 		var ret string
 		return ret
 	}
@@ -53,7 +56,7 @@ func (o *BuyingOptionDistribution) GetBuyingOption() string {
 // GetBuyingOptionOk returns a tuple with the BuyingOption field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *BuyingOptionDistribution) GetBuyingOptionOk() (*string, bool) {
-	if o == nil || o.BuyingOption == nil {
+	if o == nil || IsNil(o.BuyingOption) {
 		return nil, false
 	}
 	return o.BuyingOption, true
@@ -61,7 +64,7 @@ func (o *BuyingOptionDistribution) GetBuyingOptionOk() (*string, bool) {
 
 // HasBuyingOption returns a boolean if a field has been set.
 func (o *BuyingOptionDistribution) HasBuyingOption() bool {
-	if o != nil && o.BuyingOption != nil {
+	if o != nil && !IsNil(o.BuyingOption) {
 		return true
 	}
 
@@ -75,7 +78,7 @@ func (o *BuyingOptionDistribution) SetBuyingOption(v string) {
 
 // GetMatchCount returns the MatchCount field value if set, zero value otherwise.
 func (o *BuyingOptionDistribution) GetMatchCount() int32 {
-	if o == nil || o.MatchCount == nil {
+	if o == nil || IsNil(o.MatchCount) {
 		var ret int32
 		return ret
 	}
@@ -85,7 +88,7 @@ func (o *BuyingOptionDistribution) GetMatchCount() int32 {
 // GetMatchCountOk returns a tuple with the MatchCount field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *BuyingOptionDistribution) GetMatchCountOk() (*int32, bool) {
-	if o == nil || o.MatchCount == nil {
+	if o == nil || IsNil(o.MatchCount) {
 		return nil, false
 	}
 	return o.MatchCount, true
@@ -93,7 +96,7 @@ func (o *BuyingOptionDistribution) GetMatchCountOk() (*int32, bool) {
 
 // HasMatchCount returns a boolean if a field has been set.
 func (o *BuyingOptionDistribution) HasMatchCount() bool {
-	if o != nil && o.MatchCount != nil {
+	if o != nil && !IsNil(o.MatchCount) {
 		return true
 	}
 
@@ -107,7 +110,7 @@ func (o *BuyingOptionDistribution) SetMatchCount(v int32) {
 
 // GetRefinementHref returns the RefinementHref field value if set, zero value otherwise.
 func (o *BuyingOptionDistribution) GetRefinementHref() string {
-	if o == nil || o.RefinementHref == nil {
+	if o == nil || IsNil(o.RefinementHref) {
 		var ret string
 		return ret
 	}
@@ -117,7 +120,7 @@ func (o *BuyingOptionDistribution) GetRefinementHref() string {
 // GetRefinementHrefOk returns a tuple with the RefinementHref field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *BuyingOptionDistribution) GetRefinementHrefOk() (*string, bool) {
-	if o == nil || o.RefinementHref == nil {
+	if o == nil || IsNil(o.RefinementHref) {
 		return nil, false
 	}
 	return o.RefinementHref, true
@@ -125,7 +128,7 @@ func (o *BuyingOptionDistribution) GetRefinementHrefOk() (*string, bool) {
 
 // HasRefinementHref returns a boolean if a field has been set.
 func (o *BuyingOptionDistribution) HasRefinementHref() bool {
-	if o != nil && o.RefinementHref != nil {
+	if o != nil && !IsNil(o.RefinementHref) {
 		return true
 	}
 
@@ -138,17 +141,25 @@ func (o *BuyingOptionDistribution) SetRefinementHref(v string) {
 }
 
 func (o BuyingOptionDistribution) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.BuyingOption != nil {
-		toSerialize["buyingOption"] = o.BuyingOption
-	}
-	if o.MatchCount != nil {
-		toSerialize["matchCount"] = o.MatchCount
-	}
-	if o.RefinementHref != nil {
-		toSerialize["refinementHref"] = o.RefinementHref
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o BuyingOptionDistribution) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.BuyingOption) {
+		toSerialize["buyingOption"] = o.BuyingOption
+	}
+	if !IsNil(o.MatchCount) {
+		toSerialize["matchCount"] = o.MatchCount
+	}
+	if !IsNil(o.RefinementHref) {
+		toSerialize["refinementHref"] = o.RefinementHref
+	}
+	return toSerialize, nil
 }
 
 type NullableBuyingOptionDistribution struct {
