@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,10 +14,15 @@ import (
 	"encoding/json"
 )
 
-// Category This type is used by the <b> categories</b>  container in the response of the <b> search</b>  method, and contains the primary item category ID of the item, as well as the secondary item category if the item was listed in two categories.
+// checks if the Category type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &Category{}
+
+// Category This type is used by the <code>categories</code> container in the response of the <b>search</b>  method, and contains the name and ID of the item category.
 type Category struct {
-	// The unique identifier of the primary item category of the item, as well as the secondary item category if item was listed in two categories.
+	// The unique identifier of the category.
 	CategoryId *string `json:"categoryId,omitempty"`
+	// The name of the category.
+	CategoryName *string `json:"categoryName,omitempty"`
 }
 
 // NewCategory instantiates a new Category object
@@ -39,7 +44,7 @@ func NewCategoryWithDefaults() *Category {
 
 // GetCategoryId returns the CategoryId field value if set, zero value otherwise.
 func (o *Category) GetCategoryId() string {
-	if o == nil || o.CategoryId == nil {
+	if o == nil || IsNil(o.CategoryId) {
 		var ret string
 		return ret
 	}
@@ -49,7 +54,7 @@ func (o *Category) GetCategoryId() string {
 // GetCategoryIdOk returns a tuple with the CategoryId field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Category) GetCategoryIdOk() (*string, bool) {
-	if o == nil || o.CategoryId == nil {
+	if o == nil || IsNil(o.CategoryId) {
 		return nil, false
 	}
 	return o.CategoryId, true
@@ -57,7 +62,7 @@ func (o *Category) GetCategoryIdOk() (*string, bool) {
 
 // HasCategoryId returns a boolean if a field has been set.
 func (o *Category) HasCategoryId() bool {
-	if o != nil && o.CategoryId != nil {
+	if o != nil && !IsNil(o.CategoryId) {
 		return true
 	}
 
@@ -69,12 +74,55 @@ func (o *Category) SetCategoryId(v string) {
 	o.CategoryId = &v
 }
 
+// GetCategoryName returns the CategoryName field value if set, zero value otherwise.
+func (o *Category) GetCategoryName() string {
+	if o == nil || IsNil(o.CategoryName) {
+		var ret string
+		return ret
+	}
+	return *o.CategoryName
+}
+
+// GetCategoryNameOk returns a tuple with the CategoryName field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *Category) GetCategoryNameOk() (*string, bool) {
+	if o == nil || IsNil(o.CategoryName) {
+		return nil, false
+	}
+	return o.CategoryName, true
+}
+
+// HasCategoryName returns a boolean if a field has been set.
+func (o *Category) HasCategoryName() bool {
+	if o != nil && !IsNil(o.CategoryName) {
+		return true
+	}
+
+	return false
+}
+
+// SetCategoryName gets a reference to the given string and assigns it to the CategoryName field.
+func (o *Category) SetCategoryName(v string) {
+	o.CategoryName = &v
+}
+
 func (o Category) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.CategoryId != nil {
-		toSerialize["categoryId"] = o.CategoryId
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o Category) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.CategoryId) {
+		toSerialize["categoryId"] = o.CategoryId
+	}
+	if !IsNil(o.CategoryName) {
+		toSerialize["categoryName"] = o.CategoryName
+	}
+	return toSerialize, nil
 }
 
 type NullableCategory struct {
