@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,11 +14,14 @@ import (
 	"encoding/json"
 )
 
-// CategoryDistribution The container that defines the fields for the category refinements. This container is returned when <b> fieldgroups</b> is set to <code>CATEGORY_REFINEMENTS</code> or <code>FULL</code> in the request.
+// checks if the CategoryDistribution type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &CategoryDistribution{}
+
+// CategoryDistribution The container that defines the fields for the category refinements. This container is returned when <code>fieldgroups</code> is set to <code>CATEGORY_REFINEMENTS</code> or <code>FULL</code> in the request.
 type CategoryDistribution struct {
-	// The identifier of the category.
+	// The unique identifier of the category.
 	CategoryId *string `json:"categoryId,omitempty"`
-	// The name of the category, such as Baby & Toddler Clothing.
+	// The name of the category, such as <b>Baby &amp; Toddler Clothing</b>.
 	CategoryName *string `json:"categoryName,omitempty"`
 	// The number of items in this category.
 	MatchCount *int32 `json:"matchCount,omitempty"`
@@ -45,7 +48,7 @@ func NewCategoryDistributionWithDefaults() *CategoryDistribution {
 
 // GetCategoryId returns the CategoryId field value if set, zero value otherwise.
 func (o *CategoryDistribution) GetCategoryId() string {
-	if o == nil || o.CategoryId == nil {
+	if o == nil || IsNil(o.CategoryId) {
 		var ret string
 		return ret
 	}
@@ -55,7 +58,7 @@ func (o *CategoryDistribution) GetCategoryId() string {
 // GetCategoryIdOk returns a tuple with the CategoryId field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *CategoryDistribution) GetCategoryIdOk() (*string, bool) {
-	if o == nil || o.CategoryId == nil {
+	if o == nil || IsNil(o.CategoryId) {
 		return nil, false
 	}
 	return o.CategoryId, true
@@ -63,7 +66,7 @@ func (o *CategoryDistribution) GetCategoryIdOk() (*string, bool) {
 
 // HasCategoryId returns a boolean if a field has been set.
 func (o *CategoryDistribution) HasCategoryId() bool {
-	if o != nil && o.CategoryId != nil {
+	if o != nil && !IsNil(o.CategoryId) {
 		return true
 	}
 
@@ -77,7 +80,7 @@ func (o *CategoryDistribution) SetCategoryId(v string) {
 
 // GetCategoryName returns the CategoryName field value if set, zero value otherwise.
 func (o *CategoryDistribution) GetCategoryName() string {
-	if o == nil || o.CategoryName == nil {
+	if o == nil || IsNil(o.CategoryName) {
 		var ret string
 		return ret
 	}
@@ -87,7 +90,7 @@ func (o *CategoryDistribution) GetCategoryName() string {
 // GetCategoryNameOk returns a tuple with the CategoryName field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *CategoryDistribution) GetCategoryNameOk() (*string, bool) {
-	if o == nil || o.CategoryName == nil {
+	if o == nil || IsNil(o.CategoryName) {
 		return nil, false
 	}
 	return o.CategoryName, true
@@ -95,7 +98,7 @@ func (o *CategoryDistribution) GetCategoryNameOk() (*string, bool) {
 
 // HasCategoryName returns a boolean if a field has been set.
 func (o *CategoryDistribution) HasCategoryName() bool {
-	if o != nil && o.CategoryName != nil {
+	if o != nil && !IsNil(o.CategoryName) {
 		return true
 	}
 
@@ -109,7 +112,7 @@ func (o *CategoryDistribution) SetCategoryName(v string) {
 
 // GetMatchCount returns the MatchCount field value if set, zero value otherwise.
 func (o *CategoryDistribution) GetMatchCount() int32 {
-	if o == nil || o.MatchCount == nil {
+	if o == nil || IsNil(o.MatchCount) {
 		var ret int32
 		return ret
 	}
@@ -119,7 +122,7 @@ func (o *CategoryDistribution) GetMatchCount() int32 {
 // GetMatchCountOk returns a tuple with the MatchCount field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *CategoryDistribution) GetMatchCountOk() (*int32, bool) {
-	if o == nil || o.MatchCount == nil {
+	if o == nil || IsNil(o.MatchCount) {
 		return nil, false
 	}
 	return o.MatchCount, true
@@ -127,7 +130,7 @@ func (o *CategoryDistribution) GetMatchCountOk() (*int32, bool) {
 
 // HasMatchCount returns a boolean if a field has been set.
 func (o *CategoryDistribution) HasMatchCount() bool {
-	if o != nil && o.MatchCount != nil {
+	if o != nil && !IsNil(o.MatchCount) {
 		return true
 	}
 
@@ -141,7 +144,7 @@ func (o *CategoryDistribution) SetMatchCount(v int32) {
 
 // GetRefinementHref returns the RefinementHref field value if set, zero value otherwise.
 func (o *CategoryDistribution) GetRefinementHref() string {
-	if o == nil || o.RefinementHref == nil {
+	if o == nil || IsNil(o.RefinementHref) {
 		var ret string
 		return ret
 	}
@@ -151,7 +154,7 @@ func (o *CategoryDistribution) GetRefinementHref() string {
 // GetRefinementHrefOk returns a tuple with the RefinementHref field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *CategoryDistribution) GetRefinementHrefOk() (*string, bool) {
-	if o == nil || o.RefinementHref == nil {
+	if o == nil || IsNil(o.RefinementHref) {
 		return nil, false
 	}
 	return o.RefinementHref, true
@@ -159,7 +162,7 @@ func (o *CategoryDistribution) GetRefinementHrefOk() (*string, bool) {
 
 // HasRefinementHref returns a boolean if a field has been set.
 func (o *CategoryDistribution) HasRefinementHref() bool {
-	if o != nil && o.RefinementHref != nil {
+	if o != nil && !IsNil(o.RefinementHref) {
 		return true
 	}
 
@@ -172,20 +175,28 @@ func (o *CategoryDistribution) SetRefinementHref(v string) {
 }
 
 func (o CategoryDistribution) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.CategoryId != nil {
-		toSerialize["categoryId"] = o.CategoryId
-	}
-	if o.CategoryName != nil {
-		toSerialize["categoryName"] = o.CategoryName
-	}
-	if o.MatchCount != nil {
-		toSerialize["matchCount"] = o.MatchCount
-	}
-	if o.RefinementHref != nil {
-		toSerialize["refinementHref"] = o.RefinementHref
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o CategoryDistribution) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.CategoryId) {
+		toSerialize["categoryId"] = o.CategoryId
+	}
+	if !IsNil(o.CategoryName) {
+		toSerialize["categoryName"] = o.CategoryName
+	}
+	if !IsNil(o.MatchCount) {
+		toSerialize["matchCount"] = o.MatchCount
+	}
+	if !IsNil(o.RefinementHref) {
+		toSerialize["refinementHref"] = o.RefinementHref
+	}
+	return toSerialize, nil
 }
 
 type NullableCategoryDistribution struct {
