@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,12 +14,15 @@ import (
 	"encoding/json"
 )
 
-// CommonDescriptions The type that defines the fields for the item ids that all use a common description.  Often the item variations within an item group all have the same description. Instead of repeating this description in the item details of each item, a description that is shared by at least one other item is returned in this container. If the description is unique, it is returned in the <b> items.description</b> field.
+// checks if the CommonDescriptions type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &CommonDescriptions{}
+
+// CommonDescriptions The type that defines the fields for the <code>item_id</code> values that all use a common description. Often the item variations within an item group all have the same description. Instead of repeating this description in the item details of each item, a description that is shared by at least one other item is returned in this container. If the description is unique, it is returned in the <code>items.description</code> field.
 type CommonDescriptions struct {
 	// The item description that is used by more than one of the item variations.
 	Description *string `json:"description,omitempty"`
 	// A list of item ids that have this description.
-	ItemIds *[]string `json:"itemIds,omitempty"`
+	ItemIds []string `json:"itemIds,omitempty"`
 }
 
 // NewCommonDescriptions instantiates a new CommonDescriptions object
@@ -41,7 +44,7 @@ func NewCommonDescriptionsWithDefaults() *CommonDescriptions {
 
 // GetDescription returns the Description field value if set, zero value otherwise.
 func (o *CommonDescriptions) GetDescription() string {
-	if o == nil || o.Description == nil {
+	if o == nil || IsNil(o.Description) {
 		var ret string
 		return ret
 	}
@@ -51,7 +54,7 @@ func (o *CommonDescriptions) GetDescription() string {
 // GetDescriptionOk returns a tuple with the Description field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *CommonDescriptions) GetDescriptionOk() (*string, bool) {
-	if o == nil || o.Description == nil {
+	if o == nil || IsNil(o.Description) {
 		return nil, false
 	}
 	return o.Description, true
@@ -59,7 +62,7 @@ func (o *CommonDescriptions) GetDescriptionOk() (*string, bool) {
 
 // HasDescription returns a boolean if a field has been set.
 func (o *CommonDescriptions) HasDescription() bool {
-	if o != nil && o.Description != nil {
+	if o != nil && !IsNil(o.Description) {
 		return true
 	}
 
@@ -73,17 +76,17 @@ func (o *CommonDescriptions) SetDescription(v string) {
 
 // GetItemIds returns the ItemIds field value if set, zero value otherwise.
 func (o *CommonDescriptions) GetItemIds() []string {
-	if o == nil || o.ItemIds == nil {
+	if o == nil || IsNil(o.ItemIds) {
 		var ret []string
 		return ret
 	}
-	return *o.ItemIds
+	return o.ItemIds
 }
 
 // GetItemIdsOk returns a tuple with the ItemIds field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *CommonDescriptions) GetItemIdsOk() (*[]string, bool) {
-	if o == nil || o.ItemIds == nil {
+func (o *CommonDescriptions) GetItemIdsOk() ([]string, bool) {
+	if o == nil || IsNil(o.ItemIds) {
 		return nil, false
 	}
 	return o.ItemIds, true
@@ -91,7 +94,7 @@ func (o *CommonDescriptions) GetItemIdsOk() (*[]string, bool) {
 
 // HasItemIds returns a boolean if a field has been set.
 func (o *CommonDescriptions) HasItemIds() bool {
-	if o != nil && o.ItemIds != nil {
+	if o != nil && !IsNil(o.ItemIds) {
 		return true
 	}
 
@@ -100,18 +103,26 @@ func (o *CommonDescriptions) HasItemIds() bool {
 
 // SetItemIds gets a reference to the given []string and assigns it to the ItemIds field.
 func (o *CommonDescriptions) SetItemIds(v []string) {
-	o.ItemIds = &v
+	o.ItemIds = v
 }
 
 func (o CommonDescriptions) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.Description != nil {
-		toSerialize["description"] = o.Description
-	}
-	if o.ItemIds != nil {
-		toSerialize["itemIds"] = o.ItemIds
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o CommonDescriptions) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.Description) {
+		toSerialize["description"] = o.Description
+	}
+	if !IsNil(o.ItemIds) {
+		toSerialize["itemIds"] = o.ItemIds
+	}
+	return toSerialize, nil
 }
 
 type NullableCommonDescriptions struct {
