@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,10 +14,13 @@ import (
 	"encoding/json"
 )
 
-// CompatibilityPayload An array of attribute name/value pairs used to define a specific product. For example: If you wanted to specify a specific car, one of the name/value pairs would be <br /><code>\"name\" : \"Year\", <br />\"value\" : \"2019\"</code>  <p> For a list of the attributes required for cars and trucks and motorcycles see <a href=\"/api-docs/buy/static/api-browse.html#Check\">Check compatibility</a> in the Buy Integration Guide.</p>
+// checks if the CompatibilityPayload type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &CompatibilityPayload{}
+
+// CompatibilityPayload An array of attribute name/value pairs used to define a specific product. For example: If you wanted to specify a specific car, one of the name/value pairs would be:<br><br><code>\"name\" : \"Year\",<br>\"value\" : \"2019\"</code><br><br>For a list of the attributes required for cars and trucks and motorcycles refer to <a href=\"/api-docs/buy/static/api-browse.html#Check\" target=\"_blank\">Check compatibility</a> in the <a href=\"/api-docs/buy/static/buying-ig-landing.html\" target=\"_blank\">Buying Integration Guide</a>.
 type CompatibilityPayload struct {
-	// An array of attribute name/value pairs used to define a specific product. For example: If you wanted to specify a specific car, one of the name/value pairs would be <br /><code>\"name\" : \"Year\", <br />\"value\" : \"2019\"</code>  <p> For a list of the attributes required for cars and trucks and motorcycles see <a href=\"/api-docs/buy/static/api-browse.html#Check\">Check compatibility</a> in the Buy Integration Guide.</p>
-	CompatibilityProperties *[]AttributeNameValue `json:"compatibilityProperties,omitempty"`
+	// An array of attribute name/value pairs used to define a specific product. For example: If you wanted to specify a specific car, one of the name/value pairs would be<br><br><code>\"name\" : \"Year\", <br>\"value\" : \"2019\"</code><br><br>For a list of the attributes required for cars and trucks and motorcycles see <a href=\"/api-docs/buy/static/api-browse.html#Check\">Check compatibility</a> in the Buy Integration Guide.
+	CompatibilityProperties []AttributeNameValue `json:"compatibilityProperties,omitempty"`
 }
 
 // NewCompatibilityPayload instantiates a new CompatibilityPayload object
@@ -39,17 +42,17 @@ func NewCompatibilityPayloadWithDefaults() *CompatibilityPayload {
 
 // GetCompatibilityProperties returns the CompatibilityProperties field value if set, zero value otherwise.
 func (o *CompatibilityPayload) GetCompatibilityProperties() []AttributeNameValue {
-	if o == nil || o.CompatibilityProperties == nil {
+	if o == nil || IsNil(o.CompatibilityProperties) {
 		var ret []AttributeNameValue
 		return ret
 	}
-	return *o.CompatibilityProperties
+	return o.CompatibilityProperties
 }
 
 // GetCompatibilityPropertiesOk returns a tuple with the CompatibilityProperties field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *CompatibilityPayload) GetCompatibilityPropertiesOk() (*[]AttributeNameValue, bool) {
-	if o == nil || o.CompatibilityProperties == nil {
+func (o *CompatibilityPayload) GetCompatibilityPropertiesOk() ([]AttributeNameValue, bool) {
+	if o == nil || IsNil(o.CompatibilityProperties) {
 		return nil, false
 	}
 	return o.CompatibilityProperties, true
@@ -57,7 +60,7 @@ func (o *CompatibilityPayload) GetCompatibilityPropertiesOk() (*[]AttributeNameV
 
 // HasCompatibilityProperties returns a boolean if a field has been set.
 func (o *CompatibilityPayload) HasCompatibilityProperties() bool {
-	if o != nil && o.CompatibilityProperties != nil {
+	if o != nil && !IsNil(o.CompatibilityProperties) {
 		return true
 	}
 
@@ -66,15 +69,23 @@ func (o *CompatibilityPayload) HasCompatibilityProperties() bool {
 
 // SetCompatibilityProperties gets a reference to the given []AttributeNameValue and assigns it to the CompatibilityProperties field.
 func (o *CompatibilityPayload) SetCompatibilityProperties(v []AttributeNameValue) {
-	o.CompatibilityProperties = &v
+	o.CompatibilityProperties = v
 }
 
 func (o CompatibilityPayload) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.CompatibilityProperties != nil {
-		toSerialize["compatibilityProperties"] = o.CompatibilityProperties
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o CompatibilityPayload) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.CompatibilityProperties) {
+		toSerialize["compatibilityProperties"] = o.CompatibilityProperties
+	}
+	return toSerialize, nil
 }
 
 type NullableCompatibilityPayload struct {
