@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,13 +14,16 @@ import (
 	"encoding/json"
 )
 
-// CompatibilityProperty This container returns the product attribute name/value pairs that are compatible with the keyword. These attributes are submitted in the  <b>compatibility_filter</b> request field.
+// checks if the CompatibilityProperty type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &CompatibilityProperty{}
+
+// CompatibilityProperty This container returns the product attribute name/value pairs that are compatible with the keyword. These attributes are submitted in the <code>compatibility_filter</code> request field.
 type CompatibilityProperty struct {
 	// The name of the product attribute that as been translated to the language of the site.
 	LocalizedName *string `json:"localizedName,omitempty"`
 	// The name of the product attribute, such as Make, Model, Year, etc.
 	Name *string `json:"name,omitempty"`
-	// The value for the <b> name</b> attribute, such as BMW, R1200GS, 2011, etc.
+	// The value for the <code>name</code> attribute, such as <b>BMW</b>, <b>R1200GS</b>, <b>2011</b>, etc.
 	Value *string `json:"value,omitempty"`
 }
 
@@ -43,7 +46,7 @@ func NewCompatibilityPropertyWithDefaults() *CompatibilityProperty {
 
 // GetLocalizedName returns the LocalizedName field value if set, zero value otherwise.
 func (o *CompatibilityProperty) GetLocalizedName() string {
-	if o == nil || o.LocalizedName == nil {
+	if o == nil || IsNil(o.LocalizedName) {
 		var ret string
 		return ret
 	}
@@ -53,7 +56,7 @@ func (o *CompatibilityProperty) GetLocalizedName() string {
 // GetLocalizedNameOk returns a tuple with the LocalizedName field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *CompatibilityProperty) GetLocalizedNameOk() (*string, bool) {
-	if o == nil || o.LocalizedName == nil {
+	if o == nil || IsNil(o.LocalizedName) {
 		return nil, false
 	}
 	return o.LocalizedName, true
@@ -61,7 +64,7 @@ func (o *CompatibilityProperty) GetLocalizedNameOk() (*string, bool) {
 
 // HasLocalizedName returns a boolean if a field has been set.
 func (o *CompatibilityProperty) HasLocalizedName() bool {
-	if o != nil && o.LocalizedName != nil {
+	if o != nil && !IsNil(o.LocalizedName) {
 		return true
 	}
 
@@ -75,7 +78,7 @@ func (o *CompatibilityProperty) SetLocalizedName(v string) {
 
 // GetName returns the Name field value if set, zero value otherwise.
 func (o *CompatibilityProperty) GetName() string {
-	if o == nil || o.Name == nil {
+	if o == nil || IsNil(o.Name) {
 		var ret string
 		return ret
 	}
@@ -85,7 +88,7 @@ func (o *CompatibilityProperty) GetName() string {
 // GetNameOk returns a tuple with the Name field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *CompatibilityProperty) GetNameOk() (*string, bool) {
-	if o == nil || o.Name == nil {
+	if o == nil || IsNil(o.Name) {
 		return nil, false
 	}
 	return o.Name, true
@@ -93,7 +96,7 @@ func (o *CompatibilityProperty) GetNameOk() (*string, bool) {
 
 // HasName returns a boolean if a field has been set.
 func (o *CompatibilityProperty) HasName() bool {
-	if o != nil && o.Name != nil {
+	if o != nil && !IsNil(o.Name) {
 		return true
 	}
 
@@ -107,7 +110,7 @@ func (o *CompatibilityProperty) SetName(v string) {
 
 // GetValue returns the Value field value if set, zero value otherwise.
 func (o *CompatibilityProperty) GetValue() string {
-	if o == nil || o.Value == nil {
+	if o == nil || IsNil(o.Value) {
 		var ret string
 		return ret
 	}
@@ -117,7 +120,7 @@ func (o *CompatibilityProperty) GetValue() string {
 // GetValueOk returns a tuple with the Value field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *CompatibilityProperty) GetValueOk() (*string, bool) {
-	if o == nil || o.Value == nil {
+	if o == nil || IsNil(o.Value) {
 		return nil, false
 	}
 	return o.Value, true
@@ -125,7 +128,7 @@ func (o *CompatibilityProperty) GetValueOk() (*string, bool) {
 
 // HasValue returns a boolean if a field has been set.
 func (o *CompatibilityProperty) HasValue() bool {
-	if o != nil && o.Value != nil {
+	if o != nil && !IsNil(o.Value) {
 		return true
 	}
 
@@ -138,17 +141,25 @@ func (o *CompatibilityProperty) SetValue(v string) {
 }
 
 func (o CompatibilityProperty) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.LocalizedName != nil {
-		toSerialize["localizedName"] = o.LocalizedName
-	}
-	if o.Name != nil {
-		toSerialize["name"] = o.Name
-	}
-	if o.Value != nil {
-		toSerialize["value"] = o.Value
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o CompatibilityProperty) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.LocalizedName) {
+		toSerialize["localizedName"] = o.LocalizedName
+	}
+	if !IsNil(o.Name) {
+		toSerialize["name"] = o.Name
+	}
+	if !IsNil(o.Value) {
+		toSerialize["value"] = o.Value
+	}
+	return toSerialize, nil
 }
 
 type NullableCompatibilityProperty struct {
