@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,11 +14,14 @@ import (
 	"encoding/json"
 )
 
-// ConditionDistribution The container that defines the fields for the conditions refinements. This container is returned when <b> fieldgroups</b> is set to <code>CONDITION_REFINEMENTS</code> or <code>FULL</code> in the request.
+// checks if the ConditionDistribution type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &ConditionDistribution{}
+
+// ConditionDistribution The container that defines the fields for the conditions refinements. This container is returned when <code>fieldgroups</code> is set to <code>CONDITION_REFINEMENTS</code> or <code>FULL</code> in the request.
 type ConditionDistribution struct {
-	// The text describing the condition of the item, such as New or Used. For a list of condition names, see <a href=\"https://developer.ebay.com/devzone/finding/callref/enums/conditionIdList.html\" target=\"_blank\">Item Condition IDs and Names</a>.  <br /><br />Code so that your app gracefully handles any future changes to this list.<br /><br /><span class=\"tablenote\"><b> Note: </b>As of September 1, 2021, condition ID 2500 ('Seller Refurbished') is no longer a valid item condition in the <b>Cell Phones & Smartphones</b> category (category ID 9355) for the following marketplaces: US, Canada, UK, Germany, and Australia. This refurbished item condition has been replaced by three new refurbished values, which include 'Excellent - Refurbished' (condition ID 2010), 'Very Good - Refurbished' (condition ID 2020), and 'Good - Refurbished' (condition ID 2030).</span>
+	// The text describing the condition of the item, such as <i>New</i> or <i>Used</i>. For a list of condition names, refer to <a href=\"/devzone/finding/callref/enums/conditionIdList.html \" target=\"_blank\">Item Condition IDs and Names</a>.
 	Condition *string `json:"condition,omitempty"`
-	// The identifier of the condition. For example, 1000 is the identifier for NEW.<br /><br /><span class=\"tablenote\"><b> Note: </b>As of September 1, 2021, condition ID 2500 ('Seller Refurbished') is no longer a valid item condition in the <b>Cell Phones & Smartphones</b> category (category ID 9355) for the following marketplaces: US, Canada, UK, Germany, and Australia. This refurbished item condition has been replaced by three new refurbished values, which include 'Excellent - Refurbished' (condition ID 2010), 'Very Good - Refurbished' (condition ID 2020), and 'Good - Refurbished' (condition ID 2030).</span>
+	// The identifier of the condition. For example, <code>1000</code> is the identifier for <code>NEW</code>.
 	ConditionId *string `json:"conditionId,omitempty"`
 	// The number of items having the condition.
 	MatchCount *int32 `json:"matchCount,omitempty"`
@@ -45,7 +48,7 @@ func NewConditionDistributionWithDefaults() *ConditionDistribution {
 
 // GetCondition returns the Condition field value if set, zero value otherwise.
 func (o *ConditionDistribution) GetCondition() string {
-	if o == nil || o.Condition == nil {
+	if o == nil || IsNil(o.Condition) {
 		var ret string
 		return ret
 	}
@@ -55,7 +58,7 @@ func (o *ConditionDistribution) GetCondition() string {
 // GetConditionOk returns a tuple with the Condition field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ConditionDistribution) GetConditionOk() (*string, bool) {
-	if o == nil || o.Condition == nil {
+	if o == nil || IsNil(o.Condition) {
 		return nil, false
 	}
 	return o.Condition, true
@@ -63,7 +66,7 @@ func (o *ConditionDistribution) GetConditionOk() (*string, bool) {
 
 // HasCondition returns a boolean if a field has been set.
 func (o *ConditionDistribution) HasCondition() bool {
-	if o != nil && o.Condition != nil {
+	if o != nil && !IsNil(o.Condition) {
 		return true
 	}
 
@@ -77,7 +80,7 @@ func (o *ConditionDistribution) SetCondition(v string) {
 
 // GetConditionId returns the ConditionId field value if set, zero value otherwise.
 func (o *ConditionDistribution) GetConditionId() string {
-	if o == nil || o.ConditionId == nil {
+	if o == nil || IsNil(o.ConditionId) {
 		var ret string
 		return ret
 	}
@@ -87,7 +90,7 @@ func (o *ConditionDistribution) GetConditionId() string {
 // GetConditionIdOk returns a tuple with the ConditionId field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ConditionDistribution) GetConditionIdOk() (*string, bool) {
-	if o == nil || o.ConditionId == nil {
+	if o == nil || IsNil(o.ConditionId) {
 		return nil, false
 	}
 	return o.ConditionId, true
@@ -95,7 +98,7 @@ func (o *ConditionDistribution) GetConditionIdOk() (*string, bool) {
 
 // HasConditionId returns a boolean if a field has been set.
 func (o *ConditionDistribution) HasConditionId() bool {
-	if o != nil && o.ConditionId != nil {
+	if o != nil && !IsNil(o.ConditionId) {
 		return true
 	}
 
@@ -109,7 +112,7 @@ func (o *ConditionDistribution) SetConditionId(v string) {
 
 // GetMatchCount returns the MatchCount field value if set, zero value otherwise.
 func (o *ConditionDistribution) GetMatchCount() int32 {
-	if o == nil || o.MatchCount == nil {
+	if o == nil || IsNil(o.MatchCount) {
 		var ret int32
 		return ret
 	}
@@ -119,7 +122,7 @@ func (o *ConditionDistribution) GetMatchCount() int32 {
 // GetMatchCountOk returns a tuple with the MatchCount field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ConditionDistribution) GetMatchCountOk() (*int32, bool) {
-	if o == nil || o.MatchCount == nil {
+	if o == nil || IsNil(o.MatchCount) {
 		return nil, false
 	}
 	return o.MatchCount, true
@@ -127,7 +130,7 @@ func (o *ConditionDistribution) GetMatchCountOk() (*int32, bool) {
 
 // HasMatchCount returns a boolean if a field has been set.
 func (o *ConditionDistribution) HasMatchCount() bool {
-	if o != nil && o.MatchCount != nil {
+	if o != nil && !IsNil(o.MatchCount) {
 		return true
 	}
 
@@ -141,7 +144,7 @@ func (o *ConditionDistribution) SetMatchCount(v int32) {
 
 // GetRefinementHref returns the RefinementHref field value if set, zero value otherwise.
 func (o *ConditionDistribution) GetRefinementHref() string {
-	if o == nil || o.RefinementHref == nil {
+	if o == nil || IsNil(o.RefinementHref) {
 		var ret string
 		return ret
 	}
@@ -151,7 +154,7 @@ func (o *ConditionDistribution) GetRefinementHref() string {
 // GetRefinementHrefOk returns a tuple with the RefinementHref field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ConditionDistribution) GetRefinementHrefOk() (*string, bool) {
-	if o == nil || o.RefinementHref == nil {
+	if o == nil || IsNil(o.RefinementHref) {
 		return nil, false
 	}
 	return o.RefinementHref, true
@@ -159,7 +162,7 @@ func (o *ConditionDistribution) GetRefinementHrefOk() (*string, bool) {
 
 // HasRefinementHref returns a boolean if a field has been set.
 func (o *ConditionDistribution) HasRefinementHref() bool {
-	if o != nil && o.RefinementHref != nil {
+	if o != nil && !IsNil(o.RefinementHref) {
 		return true
 	}
 
@@ -172,20 +175,28 @@ func (o *ConditionDistribution) SetRefinementHref(v string) {
 }
 
 func (o ConditionDistribution) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.Condition != nil {
-		toSerialize["condition"] = o.Condition
-	}
-	if o.ConditionId != nil {
-		toSerialize["conditionId"] = o.ConditionId
-	}
-	if o.MatchCount != nil {
-		toSerialize["matchCount"] = o.MatchCount
-	}
-	if o.RefinementHref != nil {
-		toSerialize["refinementHref"] = o.RefinementHref
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o ConditionDistribution) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.Condition) {
+		toSerialize["condition"] = o.Condition
+	}
+	if !IsNil(o.ConditionId) {
+		toSerialize["conditionId"] = o.ConditionId
+	}
+	if !IsNil(o.MatchCount) {
+		toSerialize["matchCount"] = o.MatchCount
+	}
+	if !IsNil(o.RefinementHref) {
+		toSerialize["refinementHref"] = o.RefinementHref
+	}
+	return toSerialize, nil
 }
 
 type NullableConditionDistribution struct {
