@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,15 +14,18 @@ import (
 	"encoding/json"
 )
 
+// checks if the ConvertedAmount type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &ConvertedAmount{}
+
 // ConvertedAmount This type defines the monetary value of an amount. It can provide the amount in both the currency used on the eBay site where an item is being offered and the conversion of that value into another currency, if applicable.
 type ConvertedAmount struct {
-	// The three-letter <a href=\"https://www.iso.org/iso-4217-currency-codes.html\" target=\"_blank\">ISO 4217</a> code representing the currency of the amount in the <b> convertedFromValue</b> field. This value is required or returned only if currency conversion/localization is required, and represents the pre-conversion currency. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/ba:CurrencyCodeEnum'>eBay API documentation</a>
+	// The three-letter <a href=\"https://www.iso.org/iso-4217-currency-codes.html \" target=\"_blank\">ISO 4217</a> code representing the currency of the amount in the <code>convertedFromValue</code> field. This value is required or returned only if currency conversion/localization is required, and represents the pre-conversion currency. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/ba:CurrencyCodeEnum'>eBay API documentation</a>
 	ConvertedFromCurrency *string `json:"convertedFromCurrency,omitempty"`
-	// The monetary amount before any conversion is performed, in the currency specified by the <b> convertedFromCurrency</b> field. This value is required or returned only if currency conversion/localization is required. The <b> value</b> field contains the converted amount of this value, in the currency specified by the <b> currency</b> field.
+	// The monetary amount before any conversion is performed, in the currency specified by the <code>convertedFromCurrency</code> field. This value is required or returned only if currency conversion/localization is required. The <code>value</code> field contains the converted amount of this value, in the currency specified by the <code>currency</code> field.
 	ConvertedFromValue *string `json:"convertedFromValue,omitempty"`
-	// The three-letter <a href=\"https://www.iso.org/iso-4217-currency-codes.html\" target=\"_blank\">ISO 4217</a> code representing the currency of the amount in the <b> value</b> field. If currency conversion/localization is required, this is the post-conversion currency of the amount in the <b> value</b> field.   <br /><br /><b> Default:</b> The currency of the authenticated user's country. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/ba:CurrencyCodeEnum'>eBay API documentation</a>
+	// The three-letter <a href=\"https://www.iso.org/iso-4217-currency-codes.html \" target=\"_blank\">ISO 4217</a> code representing the currency of the amount in the <code>value</code> field. If currency conversion/localization is required, this is the post-conversion currency of the amount in the <code>value</code> field.<br><br><b>Default:</b> The currency of the authenticated user's country. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/ba:CurrencyCodeEnum'>eBay API documentation</a>
 	Currency *string `json:"currency,omitempty"`
-	// The monetary amount, in the currency specified by the <b> currency</b> field. If currency conversion/localization is required, this value is the converted amount, and the <b> convertedFromValue</b> field contains the amount in the original currency. 
+	// The monetary amount in the currency specified by the <code>currency</code> field. If currency conversion/localization is required, this value is the converted amount, and the <code>convertedFromValue</code> field contains the amount in the original currency.
 	Value *string `json:"value,omitempty"`
 }
 
@@ -45,7 +48,7 @@ func NewConvertedAmountWithDefaults() *ConvertedAmount {
 
 // GetConvertedFromCurrency returns the ConvertedFromCurrency field value if set, zero value otherwise.
 func (o *ConvertedAmount) GetConvertedFromCurrency() string {
-	if o == nil || o.ConvertedFromCurrency == nil {
+	if o == nil || IsNil(o.ConvertedFromCurrency) {
 		var ret string
 		return ret
 	}
@@ -55,7 +58,7 @@ func (o *ConvertedAmount) GetConvertedFromCurrency() string {
 // GetConvertedFromCurrencyOk returns a tuple with the ConvertedFromCurrency field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ConvertedAmount) GetConvertedFromCurrencyOk() (*string, bool) {
-	if o == nil || o.ConvertedFromCurrency == nil {
+	if o == nil || IsNil(o.ConvertedFromCurrency) {
 		return nil, false
 	}
 	return o.ConvertedFromCurrency, true
@@ -63,7 +66,7 @@ func (o *ConvertedAmount) GetConvertedFromCurrencyOk() (*string, bool) {
 
 // HasConvertedFromCurrency returns a boolean if a field has been set.
 func (o *ConvertedAmount) HasConvertedFromCurrency() bool {
-	if o != nil && o.ConvertedFromCurrency != nil {
+	if o != nil && !IsNil(o.ConvertedFromCurrency) {
 		return true
 	}
 
@@ -77,7 +80,7 @@ func (o *ConvertedAmount) SetConvertedFromCurrency(v string) {
 
 // GetConvertedFromValue returns the ConvertedFromValue field value if set, zero value otherwise.
 func (o *ConvertedAmount) GetConvertedFromValue() string {
-	if o == nil || o.ConvertedFromValue == nil {
+	if o == nil || IsNil(o.ConvertedFromValue) {
 		var ret string
 		return ret
 	}
@@ -87,7 +90,7 @@ func (o *ConvertedAmount) GetConvertedFromValue() string {
 // GetConvertedFromValueOk returns a tuple with the ConvertedFromValue field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ConvertedAmount) GetConvertedFromValueOk() (*string, bool) {
-	if o == nil || o.ConvertedFromValue == nil {
+	if o == nil || IsNil(o.ConvertedFromValue) {
 		return nil, false
 	}
 	return o.ConvertedFromValue, true
@@ -95,7 +98,7 @@ func (o *ConvertedAmount) GetConvertedFromValueOk() (*string, bool) {
 
 // HasConvertedFromValue returns a boolean if a field has been set.
 func (o *ConvertedAmount) HasConvertedFromValue() bool {
-	if o != nil && o.ConvertedFromValue != nil {
+	if o != nil && !IsNil(o.ConvertedFromValue) {
 		return true
 	}
 
@@ -109,7 +112,7 @@ func (o *ConvertedAmount) SetConvertedFromValue(v string) {
 
 // GetCurrency returns the Currency field value if set, zero value otherwise.
 func (o *ConvertedAmount) GetCurrency() string {
-	if o == nil || o.Currency == nil {
+	if o == nil || IsNil(o.Currency) {
 		var ret string
 		return ret
 	}
@@ -119,7 +122,7 @@ func (o *ConvertedAmount) GetCurrency() string {
 // GetCurrencyOk returns a tuple with the Currency field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ConvertedAmount) GetCurrencyOk() (*string, bool) {
-	if o == nil || o.Currency == nil {
+	if o == nil || IsNil(o.Currency) {
 		return nil, false
 	}
 	return o.Currency, true
@@ -127,7 +130,7 @@ func (o *ConvertedAmount) GetCurrencyOk() (*string, bool) {
 
 // HasCurrency returns a boolean if a field has been set.
 func (o *ConvertedAmount) HasCurrency() bool {
-	if o != nil && o.Currency != nil {
+	if o != nil && !IsNil(o.Currency) {
 		return true
 	}
 
@@ -141,7 +144,7 @@ func (o *ConvertedAmount) SetCurrency(v string) {
 
 // GetValue returns the Value field value if set, zero value otherwise.
 func (o *ConvertedAmount) GetValue() string {
-	if o == nil || o.Value == nil {
+	if o == nil || IsNil(o.Value) {
 		var ret string
 		return ret
 	}
@@ -151,7 +154,7 @@ func (o *ConvertedAmount) GetValue() string {
 // GetValueOk returns a tuple with the Value field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ConvertedAmount) GetValueOk() (*string, bool) {
-	if o == nil || o.Value == nil {
+	if o == nil || IsNil(o.Value) {
 		return nil, false
 	}
 	return o.Value, true
@@ -159,7 +162,7 @@ func (o *ConvertedAmount) GetValueOk() (*string, bool) {
 
 // HasValue returns a boolean if a field has been set.
 func (o *ConvertedAmount) HasValue() bool {
-	if o != nil && o.Value != nil {
+	if o != nil && !IsNil(o.Value) {
 		return true
 	}
 
@@ -172,20 +175,28 @@ func (o *ConvertedAmount) SetValue(v string) {
 }
 
 func (o ConvertedAmount) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.ConvertedFromCurrency != nil {
-		toSerialize["convertedFromCurrency"] = o.ConvertedFromCurrency
-	}
-	if o.ConvertedFromValue != nil {
-		toSerialize["convertedFromValue"] = o.ConvertedFromValue
-	}
-	if o.Currency != nil {
-		toSerialize["currency"] = o.Currency
-	}
-	if o.Value != nil {
-		toSerialize["value"] = o.Value
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o ConvertedAmount) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.ConvertedFromCurrency) {
+		toSerialize["convertedFromCurrency"] = o.ConvertedFromCurrency
+	}
+	if !IsNil(o.ConvertedFromValue) {
+		toSerialize["convertedFromValue"] = o.ConvertedFromValue
+	}
+	if !IsNil(o.Currency) {
+		toSerialize["currency"] = o.Currency
+	}
+	if !IsNil(o.Value) {
+		toSerialize["value"] = o.Value
+	}
+	return toSerialize, nil
 }
 
 type NullableConvertedAmount struct {
