@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,24 +14,27 @@ import (
 	"encoding/json"
 )
 
+// checks if the Error type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &Error{}
+
 // Error The type that defines the fields that can be returned in an error.
 type Error struct {
-	// This string value indicates the error category. There are three categories of errors: request errors, application errors, and system errors. 
+	// This string value indicates the error category. There are three categories of errors: <i>request errors</i>, <i>application errors</i>, and <i>system errors</i>.
 	Category *string `json:"category,omitempty"`
 	// The name of the primary system where the error occurred. This is relevant for application errors.
 	Domain *string `json:"domain,omitempty"`
 	// A unique code that identifies the particular error or warning that occurred. Your application can use error codes as identifiers in your customized error-handling algorithms.
 	ErrorId *int32 `json:"errorId,omitempty"`
 	// An array of reference IDs that identify the specific request elements most closely associated to the error or warning, if any.
-	InputRefIds *[]string `json:"inputRefIds,omitempty"`
+	InputRefIds []string `json:"inputRefIds,omitempty"`
 	// A detailed description of the condition that caused the error or warning, and information on what to do to correct the problem.
 	LongMessage *string `json:"longMessage,omitempty"`
 	// A description of the condition that caused the error or warning.
 	Message *string `json:"message,omitempty"`
 	// An array of reference IDs that identify the specific response elements most closely associated to the error or warning, if any.
-	OutputRefIds *[]string `json:"outputRefIds,omitempty"`
+	OutputRefIds []string `json:"outputRefIds,omitempty"`
 	// An array of warning and error messages that return one or more variables contextual information about the error or warning. This is often the field or value that triggered the error or warning.
-	Parameters *[]ErrorParameter `json:"parameters,omitempty"`
+	Parameters []ErrorParameter `json:"parameters,omitempty"`
 	// The name of the subdomain in which the error or warning occurred.
 	Subdomain *string `json:"subdomain,omitempty"`
 }
@@ -55,7 +58,7 @@ func NewErrorWithDefaults() *Error {
 
 // GetCategory returns the Category field value if set, zero value otherwise.
 func (o *Error) GetCategory() string {
-	if o == nil || o.Category == nil {
+	if o == nil || IsNil(o.Category) {
 		var ret string
 		return ret
 	}
@@ -65,7 +68,7 @@ func (o *Error) GetCategory() string {
 // GetCategoryOk returns a tuple with the Category field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Error) GetCategoryOk() (*string, bool) {
-	if o == nil || o.Category == nil {
+	if o == nil || IsNil(o.Category) {
 		return nil, false
 	}
 	return o.Category, true
@@ -73,7 +76,7 @@ func (o *Error) GetCategoryOk() (*string, bool) {
 
 // HasCategory returns a boolean if a field has been set.
 func (o *Error) HasCategory() bool {
-	if o != nil && o.Category != nil {
+	if o != nil && !IsNil(o.Category) {
 		return true
 	}
 
@@ -87,7 +90,7 @@ func (o *Error) SetCategory(v string) {
 
 // GetDomain returns the Domain field value if set, zero value otherwise.
 func (o *Error) GetDomain() string {
-	if o == nil || o.Domain == nil {
+	if o == nil || IsNil(o.Domain) {
 		var ret string
 		return ret
 	}
@@ -97,7 +100,7 @@ func (o *Error) GetDomain() string {
 // GetDomainOk returns a tuple with the Domain field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Error) GetDomainOk() (*string, bool) {
-	if o == nil || o.Domain == nil {
+	if o == nil || IsNil(o.Domain) {
 		return nil, false
 	}
 	return o.Domain, true
@@ -105,7 +108,7 @@ func (o *Error) GetDomainOk() (*string, bool) {
 
 // HasDomain returns a boolean if a field has been set.
 func (o *Error) HasDomain() bool {
-	if o != nil && o.Domain != nil {
+	if o != nil && !IsNil(o.Domain) {
 		return true
 	}
 
@@ -119,7 +122,7 @@ func (o *Error) SetDomain(v string) {
 
 // GetErrorId returns the ErrorId field value if set, zero value otherwise.
 func (o *Error) GetErrorId() int32 {
-	if o == nil || o.ErrorId == nil {
+	if o == nil || IsNil(o.ErrorId) {
 		var ret int32
 		return ret
 	}
@@ -129,7 +132,7 @@ func (o *Error) GetErrorId() int32 {
 // GetErrorIdOk returns a tuple with the ErrorId field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Error) GetErrorIdOk() (*int32, bool) {
-	if o == nil || o.ErrorId == nil {
+	if o == nil || IsNil(o.ErrorId) {
 		return nil, false
 	}
 	return o.ErrorId, true
@@ -137,7 +140,7 @@ func (o *Error) GetErrorIdOk() (*int32, bool) {
 
 // HasErrorId returns a boolean if a field has been set.
 func (o *Error) HasErrorId() bool {
-	if o != nil && o.ErrorId != nil {
+	if o != nil && !IsNil(o.ErrorId) {
 		return true
 	}
 
@@ -151,17 +154,17 @@ func (o *Error) SetErrorId(v int32) {
 
 // GetInputRefIds returns the InputRefIds field value if set, zero value otherwise.
 func (o *Error) GetInputRefIds() []string {
-	if o == nil || o.InputRefIds == nil {
+	if o == nil || IsNil(o.InputRefIds) {
 		var ret []string
 		return ret
 	}
-	return *o.InputRefIds
+	return o.InputRefIds
 }
 
 // GetInputRefIdsOk returns a tuple with the InputRefIds field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Error) GetInputRefIdsOk() (*[]string, bool) {
-	if o == nil || o.InputRefIds == nil {
+func (o *Error) GetInputRefIdsOk() ([]string, bool) {
+	if o == nil || IsNil(o.InputRefIds) {
 		return nil, false
 	}
 	return o.InputRefIds, true
@@ -169,7 +172,7 @@ func (o *Error) GetInputRefIdsOk() (*[]string, bool) {
 
 // HasInputRefIds returns a boolean if a field has been set.
 func (o *Error) HasInputRefIds() bool {
-	if o != nil && o.InputRefIds != nil {
+	if o != nil && !IsNil(o.InputRefIds) {
 		return true
 	}
 
@@ -178,12 +181,12 @@ func (o *Error) HasInputRefIds() bool {
 
 // SetInputRefIds gets a reference to the given []string and assigns it to the InputRefIds field.
 func (o *Error) SetInputRefIds(v []string) {
-	o.InputRefIds = &v
+	o.InputRefIds = v
 }
 
 // GetLongMessage returns the LongMessage field value if set, zero value otherwise.
 func (o *Error) GetLongMessage() string {
-	if o == nil || o.LongMessage == nil {
+	if o == nil || IsNil(o.LongMessage) {
 		var ret string
 		return ret
 	}
@@ -193,7 +196,7 @@ func (o *Error) GetLongMessage() string {
 // GetLongMessageOk returns a tuple with the LongMessage field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Error) GetLongMessageOk() (*string, bool) {
-	if o == nil || o.LongMessage == nil {
+	if o == nil || IsNil(o.LongMessage) {
 		return nil, false
 	}
 	return o.LongMessage, true
@@ -201,7 +204,7 @@ func (o *Error) GetLongMessageOk() (*string, bool) {
 
 // HasLongMessage returns a boolean if a field has been set.
 func (o *Error) HasLongMessage() bool {
-	if o != nil && o.LongMessage != nil {
+	if o != nil && !IsNil(o.LongMessage) {
 		return true
 	}
 
@@ -215,7 +218,7 @@ func (o *Error) SetLongMessage(v string) {
 
 // GetMessage returns the Message field value if set, zero value otherwise.
 func (o *Error) GetMessage() string {
-	if o == nil || o.Message == nil {
+	if o == nil || IsNil(o.Message) {
 		var ret string
 		return ret
 	}
@@ -225,7 +228,7 @@ func (o *Error) GetMessage() string {
 // GetMessageOk returns a tuple with the Message field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Error) GetMessageOk() (*string, bool) {
-	if o == nil || o.Message == nil {
+	if o == nil || IsNil(o.Message) {
 		return nil, false
 	}
 	return o.Message, true
@@ -233,7 +236,7 @@ func (o *Error) GetMessageOk() (*string, bool) {
 
 // HasMessage returns a boolean if a field has been set.
 func (o *Error) HasMessage() bool {
-	if o != nil && o.Message != nil {
+	if o != nil && !IsNil(o.Message) {
 		return true
 	}
 
@@ -247,17 +250,17 @@ func (o *Error) SetMessage(v string) {
 
 // GetOutputRefIds returns the OutputRefIds field value if set, zero value otherwise.
 func (o *Error) GetOutputRefIds() []string {
-	if o == nil || o.OutputRefIds == nil {
+	if o == nil || IsNil(o.OutputRefIds) {
 		var ret []string
 		return ret
 	}
-	return *o.OutputRefIds
+	return o.OutputRefIds
 }
 
 // GetOutputRefIdsOk returns a tuple with the OutputRefIds field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Error) GetOutputRefIdsOk() (*[]string, bool) {
-	if o == nil || o.OutputRefIds == nil {
+func (o *Error) GetOutputRefIdsOk() ([]string, bool) {
+	if o == nil || IsNil(o.OutputRefIds) {
 		return nil, false
 	}
 	return o.OutputRefIds, true
@@ -265,7 +268,7 @@ func (o *Error) GetOutputRefIdsOk() (*[]string, bool) {
 
 // HasOutputRefIds returns a boolean if a field has been set.
 func (o *Error) HasOutputRefIds() bool {
-	if o != nil && o.OutputRefIds != nil {
+	if o != nil && !IsNil(o.OutputRefIds) {
 		return true
 	}
 
@@ -274,22 +277,22 @@ func (o *Error) HasOutputRefIds() bool {
 
 // SetOutputRefIds gets a reference to the given []string and assigns it to the OutputRefIds field.
 func (o *Error) SetOutputRefIds(v []string) {
-	o.OutputRefIds = &v
+	o.OutputRefIds = v
 }
 
 // GetParameters returns the Parameters field value if set, zero value otherwise.
 func (o *Error) GetParameters() []ErrorParameter {
-	if o == nil || o.Parameters == nil {
+	if o == nil || IsNil(o.Parameters) {
 		var ret []ErrorParameter
 		return ret
 	}
-	return *o.Parameters
+	return o.Parameters
 }
 
 // GetParametersOk returns a tuple with the Parameters field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Error) GetParametersOk() (*[]ErrorParameter, bool) {
-	if o == nil || o.Parameters == nil {
+func (o *Error) GetParametersOk() ([]ErrorParameter, bool) {
+	if o == nil || IsNil(o.Parameters) {
 		return nil, false
 	}
 	return o.Parameters, true
@@ -297,7 +300,7 @@ func (o *Error) GetParametersOk() (*[]ErrorParameter, bool) {
 
 // HasParameters returns a boolean if a field has been set.
 func (o *Error) HasParameters() bool {
-	if o != nil && o.Parameters != nil {
+	if o != nil && !IsNil(o.Parameters) {
 		return true
 	}
 
@@ -306,12 +309,12 @@ func (o *Error) HasParameters() bool {
 
 // SetParameters gets a reference to the given []ErrorParameter and assigns it to the Parameters field.
 func (o *Error) SetParameters(v []ErrorParameter) {
-	o.Parameters = &v
+	o.Parameters = v
 }
 
 // GetSubdomain returns the Subdomain field value if set, zero value otherwise.
 func (o *Error) GetSubdomain() string {
-	if o == nil || o.Subdomain == nil {
+	if o == nil || IsNil(o.Subdomain) {
 		var ret string
 		return ret
 	}
@@ -321,7 +324,7 @@ func (o *Error) GetSubdomain() string {
 // GetSubdomainOk returns a tuple with the Subdomain field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Error) GetSubdomainOk() (*string, bool) {
-	if o == nil || o.Subdomain == nil {
+	if o == nil || IsNil(o.Subdomain) {
 		return nil, false
 	}
 	return o.Subdomain, true
@@ -329,7 +332,7 @@ func (o *Error) GetSubdomainOk() (*string, bool) {
 
 // HasSubdomain returns a boolean if a field has been set.
 func (o *Error) HasSubdomain() bool {
-	if o != nil && o.Subdomain != nil {
+	if o != nil && !IsNil(o.Subdomain) {
 		return true
 	}
 
@@ -342,35 +345,43 @@ func (o *Error) SetSubdomain(v string) {
 }
 
 func (o Error) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.Category != nil {
-		toSerialize["category"] = o.Category
-	}
-	if o.Domain != nil {
-		toSerialize["domain"] = o.Domain
-	}
-	if o.ErrorId != nil {
-		toSerialize["errorId"] = o.ErrorId
-	}
-	if o.InputRefIds != nil {
-		toSerialize["inputRefIds"] = o.InputRefIds
-	}
-	if o.LongMessage != nil {
-		toSerialize["longMessage"] = o.LongMessage
-	}
-	if o.Message != nil {
-		toSerialize["message"] = o.Message
-	}
-	if o.OutputRefIds != nil {
-		toSerialize["outputRefIds"] = o.OutputRefIds
-	}
-	if o.Parameters != nil {
-		toSerialize["parameters"] = o.Parameters
-	}
-	if o.Subdomain != nil {
-		toSerialize["subdomain"] = o.Subdomain
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o Error) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.Category) {
+		toSerialize["category"] = o.Category
+	}
+	if !IsNil(o.Domain) {
+		toSerialize["domain"] = o.Domain
+	}
+	if !IsNil(o.ErrorId) {
+		toSerialize["errorId"] = o.ErrorId
+	}
+	if !IsNil(o.InputRefIds) {
+		toSerialize["inputRefIds"] = o.InputRefIds
+	}
+	if !IsNil(o.LongMessage) {
+		toSerialize["longMessage"] = o.LongMessage
+	}
+	if !IsNil(o.Message) {
+		toSerialize["message"] = o.Message
+	}
+	if !IsNil(o.OutputRefIds) {
+		toSerialize["outputRefIds"] = o.OutputRefIds
+	}
+	if !IsNil(o.Parameters) {
+		toSerialize["parameters"] = o.Parameters
+	}
+	if !IsNil(o.Subdomain) {
+		toSerialize["subdomain"] = o.Subdomain
+	}
+	return toSerialize, nil
 }
 
 type NullableError struct {
