@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,18 +14,23 @@ import (
 	"encoding/json"
 )
 
+// checks if the EstimatedAvailability type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &EstimatedAvailability{}
+
 // EstimatedAvailability The type that defines the fields for the estimated item availability information.
 type EstimatedAvailability struct {
-	// This field is return only when the seller sets their '<a href=\"#display-item-quantity\">display item quantity</a>' preference to <b> Display \"More than 10 available\" in your listing (if applicable)</b>. The value of this field will be \"10\", which is the threshold value. <br /><br />Code so that your app gracefully handles any future changes to this value.
+	// This field is return only when the seller sets their '<a href=\"#display-item-quantity\">display item quantity</a>' preference to <b> Display \"More than 10 available\" in your listing (if applicable)</b>. The value of this field will be \"10\", which is the threshold value. <br><br>Code so that your app gracefully handles any future changes to this value.
 	AvailabilityThreshold *int32 `json:"availabilityThreshold,omitempty"`
-	// <a name=\"display-item-quantity\"></a> This field is return only when the seller sets their <b> Display Item Quantity</b> preference to <b> Display \"More than 10 available\" in your listing (if applicable)</b>. The value of this field will be <code> MORE_THAN</code>. This indicates that the seller has more than the 'quantity display preference', which is 10, in stock for this item.    <br /><br /> The following are the display item quantity preferences the seller can set. <br /><ul><li> <b> Display \"More than 10 available\" in your listing (if applicable) </b><ul> <li>If the seller enables this preference, this field is returned as long as there are more than 10 of this item in inventory.</li>  <li> If the quantity is equal to 10 or drops below 10, this field is not returned and the estimated quantity of the item is returned in the <b> estimatedAvailableQuantity</b> field.</li></ul> </li> <li> <b> Display the exact quantity in your items</b> <br />If the seller enables this preference, the <b> availabilityThresholdType</b> and <b> availabilityThreshold</b> fields are not returned and the estimated quantity of the item is returned in the <b> estimatedAvailableQuantity</b> field.<br /><br /><b> Note: </b> Because the quantity of an item can change several times within a second, it is impossible to return the exact quantity. </li></ul>   <br />Code so that your app gracefully handles any future changes to these preferences. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:AvailabilityThresholdEnum'>eBay API documentation</a>
+	// <a name=\"display-item-quantity\"></a> This field is return only when the seller sets their <b> Display Item Quantity</b> preference to <b> Display \"More than 10 available\" in your listing (if applicable)</b>. The value of this field will be <code> MORE_THAN</code>. This indicates that the seller has more than the 'quantity display preference', which is 10, in stock for this item.    <br><br> The following are the display item quantity preferences the seller can set. <br><ul><li> <b> Display \"More than 10 available\" in your listing (if applicable) </b><ul> <li>If the seller enables this preference, this field is returned as long as there are more than 10 of this item in inventory.</li>  <li> If the quantity is equal to 10 or drops below 10, this field is not returned and the estimated quantity of the item is returned in the <b> estimatedAvailableQuantity</b> field.</li></ul> </li> <li> <b> Display the exact quantity in your items</b> <br>If the seller enables this preference, the <b> availabilityThresholdType</b> and <b> availabilityThreshold</b> fields are not returned and the estimated quantity of the item is returned in the <b> estimatedAvailableQuantity</b> field.<br><br><b> Note: </b> Because the quantity of an item can change several times within a second, it is impossible to return the exact quantity. </li></ul>   <br>Code so that your app gracefully handles any future changes to these preferences. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:AvailabilityThresholdEnum'>eBay API documentation</a>
 	AvailabilityThresholdType *string `json:"availabilityThresholdType,omitempty"`
-	// An array of available delivery options. <br><br><b> Valid Values: </b> SHIP_TO_HOME, SELLER_ARRANGED_LOCAL_PICKUP, IN_STORE_PICKUP, PICKUP_DROP_OFF, or DIGITAL_DELIVERY <br /><br />Code so that your app gracefully handles any future changes to this list. 
-	DeliveryOptions *[]string `json:"deliveryOptions,omitempty"`
-	// An enumeration value representing the inventory status of this item.<br /><br /><span class=\"tablenote\"><b> Note: </b>Be sure to review the <b>itemEndDate</b> field to determine whether the item is available for purchase.</span><br><br><b> Valid Values: </b> IN_STOCK, LIMITED_STOCK, or OUT_OF_STOCK <br /><br />Code so that your app gracefully handles any future changes to this list. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:AvailabilityStatusEnum'>eBay API documentation</a>
+	// An array of available delivery options. <br><br><b> Valid Values: </b> SHIP_TO_HOME, SELLER_ARRANGED_LOCAL_PICKUP, IN_STORE_PICKUP, PICKUP_DROP_OFF, or DIGITAL_DELIVERY <br><br>Code so that your app gracefully handles any future changes to this list. 
+	DeliveryOptions []string `json:"deliveryOptions,omitempty"`
+	// An enumeration value representing the inventory status of this item.<br><br><span class=\"tablenote\"><b> Note: </b>Be sure to review the <b>itemEndDate</b> field to determine whether the item is available for purchase.</span><br><br><b> Valid Values: </b> IN_STOCK, LIMITED_STOCK, or OUT_OF_STOCK <br><br>Code so that your app gracefully handles any future changes to this list. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:AvailabilityStatusEnum'>eBay API documentation</a>
 	EstimatedAvailabilityStatus *string `json:"estimatedAvailabilityStatus,omitempty"`
-	// The estimated number of this item that are available for purchase. Because the quantity of an item can change several times within a second, it is impossible to return the exact quantity. So instead of returning quantity, the estimated availability of the item is returned.
+	// The estimated number of this item that are available for purchase. Because the quantity of an item can change several times within a second, it is impossible to return the exact quantity. So instead of returning quantity, the estimated availability of the item is returned.<br><br><span class=\"tablenote\"><b>Note:</b> To see if a listing is available for purchase, review the <a href=\"/api-docs/buy/browse/resources/item/methods/getItem#response.itemEndDate\">itemEndDate</a> and <a href=\"/api-docs/buy/browse/resources/item/methods/getItem#response.estimatedAvailabilities.estimatedAvailabilityStatus\">estimatedAvailablityStatus</a> fields. If the item has an <b>EndDate</b> in the past, or the <b>estimatedAvailabilityStatus</b> is <code>OUT_OF_STOCK</code>, the item is unavailable for purchase.</span>
 	EstimatedAvailableQuantity *int32 `json:"estimatedAvailableQuantity,omitempty"`
+	// The estimated number of this item that are available for purchase. Because the quantity of an item can change several times within a second, it is impossible to return the exact quantity. So instead of returning quantity, the estimated availability of the item is returned.<br><br><span class=\"tablenote\"><b>Note:</b> To see if a listing is available for purchase, review the <a href=\"/api-docs/buy/browse/resources/item/methods/getItem#response.itemEndDate\">itemEndDate</a> and <a href=\"/api-docs/buy/browse/resources/item/methods/getItem#response.estimatedAvailabilities.estimatedAvailabilityStatus\">estimatedAvailablityStatus</a> fields. If the item has an <b>EndDate</b> in the past, or the <b>estimatedAvailabilityStatus</b> is <code>OUT_OF_STOCK</code>, the item is unavailable for purchase.</span>
+	EstimatedRemainingQuantity *int32 `json:"estimatedRemainingQuantity,omitempty"`
 	// The estimated number of this item that have been sold.
 	EstimatedSoldQuantity *int32 `json:"estimatedSoldQuantity,omitempty"`
 }
@@ -49,7 +54,7 @@ func NewEstimatedAvailabilityWithDefaults() *EstimatedAvailability {
 
 // GetAvailabilityThreshold returns the AvailabilityThreshold field value if set, zero value otherwise.
 func (o *EstimatedAvailability) GetAvailabilityThreshold() int32 {
-	if o == nil || o.AvailabilityThreshold == nil {
+	if o == nil || IsNil(o.AvailabilityThreshold) {
 		var ret int32
 		return ret
 	}
@@ -59,7 +64,7 @@ func (o *EstimatedAvailability) GetAvailabilityThreshold() int32 {
 // GetAvailabilityThresholdOk returns a tuple with the AvailabilityThreshold field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *EstimatedAvailability) GetAvailabilityThresholdOk() (*int32, bool) {
-	if o == nil || o.AvailabilityThreshold == nil {
+	if o == nil || IsNil(o.AvailabilityThreshold) {
 		return nil, false
 	}
 	return o.AvailabilityThreshold, true
@@ -67,7 +72,7 @@ func (o *EstimatedAvailability) GetAvailabilityThresholdOk() (*int32, bool) {
 
 // HasAvailabilityThreshold returns a boolean if a field has been set.
 func (o *EstimatedAvailability) HasAvailabilityThreshold() bool {
-	if o != nil && o.AvailabilityThreshold != nil {
+	if o != nil && !IsNil(o.AvailabilityThreshold) {
 		return true
 	}
 
@@ -81,7 +86,7 @@ func (o *EstimatedAvailability) SetAvailabilityThreshold(v int32) {
 
 // GetAvailabilityThresholdType returns the AvailabilityThresholdType field value if set, zero value otherwise.
 func (o *EstimatedAvailability) GetAvailabilityThresholdType() string {
-	if o == nil || o.AvailabilityThresholdType == nil {
+	if o == nil || IsNil(o.AvailabilityThresholdType) {
 		var ret string
 		return ret
 	}
@@ -91,7 +96,7 @@ func (o *EstimatedAvailability) GetAvailabilityThresholdType() string {
 // GetAvailabilityThresholdTypeOk returns a tuple with the AvailabilityThresholdType field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *EstimatedAvailability) GetAvailabilityThresholdTypeOk() (*string, bool) {
-	if o == nil || o.AvailabilityThresholdType == nil {
+	if o == nil || IsNil(o.AvailabilityThresholdType) {
 		return nil, false
 	}
 	return o.AvailabilityThresholdType, true
@@ -99,7 +104,7 @@ func (o *EstimatedAvailability) GetAvailabilityThresholdTypeOk() (*string, bool)
 
 // HasAvailabilityThresholdType returns a boolean if a field has been set.
 func (o *EstimatedAvailability) HasAvailabilityThresholdType() bool {
-	if o != nil && o.AvailabilityThresholdType != nil {
+	if o != nil && !IsNil(o.AvailabilityThresholdType) {
 		return true
 	}
 
@@ -113,17 +118,17 @@ func (o *EstimatedAvailability) SetAvailabilityThresholdType(v string) {
 
 // GetDeliveryOptions returns the DeliveryOptions field value if set, zero value otherwise.
 func (o *EstimatedAvailability) GetDeliveryOptions() []string {
-	if o == nil || o.DeliveryOptions == nil {
+	if o == nil || IsNil(o.DeliveryOptions) {
 		var ret []string
 		return ret
 	}
-	return *o.DeliveryOptions
+	return o.DeliveryOptions
 }
 
 // GetDeliveryOptionsOk returns a tuple with the DeliveryOptions field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *EstimatedAvailability) GetDeliveryOptionsOk() (*[]string, bool) {
-	if o == nil || o.DeliveryOptions == nil {
+func (o *EstimatedAvailability) GetDeliveryOptionsOk() ([]string, bool) {
+	if o == nil || IsNil(o.DeliveryOptions) {
 		return nil, false
 	}
 	return o.DeliveryOptions, true
@@ -131,7 +136,7 @@ func (o *EstimatedAvailability) GetDeliveryOptionsOk() (*[]string, bool) {
 
 // HasDeliveryOptions returns a boolean if a field has been set.
 func (o *EstimatedAvailability) HasDeliveryOptions() bool {
-	if o != nil && o.DeliveryOptions != nil {
+	if o != nil && !IsNil(o.DeliveryOptions) {
 		return true
 	}
 
@@ -140,12 +145,12 @@ func (o *EstimatedAvailability) HasDeliveryOptions() bool {
 
 // SetDeliveryOptions gets a reference to the given []string and assigns it to the DeliveryOptions field.
 func (o *EstimatedAvailability) SetDeliveryOptions(v []string) {
-	o.DeliveryOptions = &v
+	o.DeliveryOptions = v
 }
 
 // GetEstimatedAvailabilityStatus returns the EstimatedAvailabilityStatus field value if set, zero value otherwise.
 func (o *EstimatedAvailability) GetEstimatedAvailabilityStatus() string {
-	if o == nil || o.EstimatedAvailabilityStatus == nil {
+	if o == nil || IsNil(o.EstimatedAvailabilityStatus) {
 		var ret string
 		return ret
 	}
@@ -155,7 +160,7 @@ func (o *EstimatedAvailability) GetEstimatedAvailabilityStatus() string {
 // GetEstimatedAvailabilityStatusOk returns a tuple with the EstimatedAvailabilityStatus field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *EstimatedAvailability) GetEstimatedAvailabilityStatusOk() (*string, bool) {
-	if o == nil || o.EstimatedAvailabilityStatus == nil {
+	if o == nil || IsNil(o.EstimatedAvailabilityStatus) {
 		return nil, false
 	}
 	return o.EstimatedAvailabilityStatus, true
@@ -163,7 +168,7 @@ func (o *EstimatedAvailability) GetEstimatedAvailabilityStatusOk() (*string, boo
 
 // HasEstimatedAvailabilityStatus returns a boolean if a field has been set.
 func (o *EstimatedAvailability) HasEstimatedAvailabilityStatus() bool {
-	if o != nil && o.EstimatedAvailabilityStatus != nil {
+	if o != nil && !IsNil(o.EstimatedAvailabilityStatus) {
 		return true
 	}
 
@@ -177,7 +182,7 @@ func (o *EstimatedAvailability) SetEstimatedAvailabilityStatus(v string) {
 
 // GetEstimatedAvailableQuantity returns the EstimatedAvailableQuantity field value if set, zero value otherwise.
 func (o *EstimatedAvailability) GetEstimatedAvailableQuantity() int32 {
-	if o == nil || o.EstimatedAvailableQuantity == nil {
+	if o == nil || IsNil(o.EstimatedAvailableQuantity) {
 		var ret int32
 		return ret
 	}
@@ -187,7 +192,7 @@ func (o *EstimatedAvailability) GetEstimatedAvailableQuantity() int32 {
 // GetEstimatedAvailableQuantityOk returns a tuple with the EstimatedAvailableQuantity field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *EstimatedAvailability) GetEstimatedAvailableQuantityOk() (*int32, bool) {
-	if o == nil || o.EstimatedAvailableQuantity == nil {
+	if o == nil || IsNil(o.EstimatedAvailableQuantity) {
 		return nil, false
 	}
 	return o.EstimatedAvailableQuantity, true
@@ -195,7 +200,7 @@ func (o *EstimatedAvailability) GetEstimatedAvailableQuantityOk() (*int32, bool)
 
 // HasEstimatedAvailableQuantity returns a boolean if a field has been set.
 func (o *EstimatedAvailability) HasEstimatedAvailableQuantity() bool {
-	if o != nil && o.EstimatedAvailableQuantity != nil {
+	if o != nil && !IsNil(o.EstimatedAvailableQuantity) {
 		return true
 	}
 
@@ -207,9 +212,41 @@ func (o *EstimatedAvailability) SetEstimatedAvailableQuantity(v int32) {
 	o.EstimatedAvailableQuantity = &v
 }
 
+// GetEstimatedRemainingQuantity returns the EstimatedRemainingQuantity field value if set, zero value otherwise.
+func (o *EstimatedAvailability) GetEstimatedRemainingQuantity() int32 {
+	if o == nil || IsNil(o.EstimatedRemainingQuantity) {
+		var ret int32
+		return ret
+	}
+	return *o.EstimatedRemainingQuantity
+}
+
+// GetEstimatedRemainingQuantityOk returns a tuple with the EstimatedRemainingQuantity field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *EstimatedAvailability) GetEstimatedRemainingQuantityOk() (*int32, bool) {
+	if o == nil || IsNil(o.EstimatedRemainingQuantity) {
+		return nil, false
+	}
+	return o.EstimatedRemainingQuantity, true
+}
+
+// HasEstimatedRemainingQuantity returns a boolean if a field has been set.
+func (o *EstimatedAvailability) HasEstimatedRemainingQuantity() bool {
+	if o != nil && !IsNil(o.EstimatedRemainingQuantity) {
+		return true
+	}
+
+	return false
+}
+
+// SetEstimatedRemainingQuantity gets a reference to the given int32 and assigns it to the EstimatedRemainingQuantity field.
+func (o *EstimatedAvailability) SetEstimatedRemainingQuantity(v int32) {
+	o.EstimatedRemainingQuantity = &v
+}
+
 // GetEstimatedSoldQuantity returns the EstimatedSoldQuantity field value if set, zero value otherwise.
 func (o *EstimatedAvailability) GetEstimatedSoldQuantity() int32 {
-	if o == nil || o.EstimatedSoldQuantity == nil {
+	if o == nil || IsNil(o.EstimatedSoldQuantity) {
 		var ret int32
 		return ret
 	}
@@ -219,7 +256,7 @@ func (o *EstimatedAvailability) GetEstimatedSoldQuantity() int32 {
 // GetEstimatedSoldQuantityOk returns a tuple with the EstimatedSoldQuantity field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *EstimatedAvailability) GetEstimatedSoldQuantityOk() (*int32, bool) {
-	if o == nil || o.EstimatedSoldQuantity == nil {
+	if o == nil || IsNil(o.EstimatedSoldQuantity) {
 		return nil, false
 	}
 	return o.EstimatedSoldQuantity, true
@@ -227,7 +264,7 @@ func (o *EstimatedAvailability) GetEstimatedSoldQuantityOk() (*int32, bool) {
 
 // HasEstimatedSoldQuantity returns a boolean if a field has been set.
 func (o *EstimatedAvailability) HasEstimatedSoldQuantity() bool {
-	if o != nil && o.EstimatedSoldQuantity != nil {
+	if o != nil && !IsNil(o.EstimatedSoldQuantity) {
 		return true
 	}
 
@@ -240,26 +277,37 @@ func (o *EstimatedAvailability) SetEstimatedSoldQuantity(v int32) {
 }
 
 func (o EstimatedAvailability) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.AvailabilityThreshold != nil {
-		toSerialize["availabilityThreshold"] = o.AvailabilityThreshold
-	}
-	if o.AvailabilityThresholdType != nil {
-		toSerialize["availabilityThresholdType"] = o.AvailabilityThresholdType
-	}
-	if o.DeliveryOptions != nil {
-		toSerialize["deliveryOptions"] = o.DeliveryOptions
-	}
-	if o.EstimatedAvailabilityStatus != nil {
-		toSerialize["estimatedAvailabilityStatus"] = o.EstimatedAvailabilityStatus
-	}
-	if o.EstimatedAvailableQuantity != nil {
-		toSerialize["estimatedAvailableQuantity"] = o.EstimatedAvailableQuantity
-	}
-	if o.EstimatedSoldQuantity != nil {
-		toSerialize["estimatedSoldQuantity"] = o.EstimatedSoldQuantity
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o EstimatedAvailability) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.AvailabilityThreshold) {
+		toSerialize["availabilityThreshold"] = o.AvailabilityThreshold
+	}
+	if !IsNil(o.AvailabilityThresholdType) {
+		toSerialize["availabilityThresholdType"] = o.AvailabilityThresholdType
+	}
+	if !IsNil(o.DeliveryOptions) {
+		toSerialize["deliveryOptions"] = o.DeliveryOptions
+	}
+	if !IsNil(o.EstimatedAvailabilityStatus) {
+		toSerialize["estimatedAvailabilityStatus"] = o.EstimatedAvailabilityStatus
+	}
+	if !IsNil(o.EstimatedAvailableQuantity) {
+		toSerialize["estimatedAvailableQuantity"] = o.EstimatedAvailableQuantity
+	}
+	if !IsNil(o.EstimatedRemainingQuantity) {
+		toSerialize["estimatedRemainingQuantity"] = o.EstimatedRemainingQuantity
+	}
+	if !IsNil(o.EstimatedSoldQuantity) {
+		toSerialize["estimatedSoldQuantity"] = o.EstimatedSoldQuantity
+	}
+	return toSerialize, nil
 }
 
 type NullableEstimatedAvailability struct {
