@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,10 +14,13 @@ import (
 	"encoding/json"
 )
 
-// ItemGroupSummary The type that defines the fields for the details of each item in an item group. An item group is  an item that has various aspect differences, such as color, size, storage capacity, etc. When an item group is created, one of the item variations, such as the red shirt size L, is chosen as the \"parent\". All the other items in the group are the children, such as the blue shirt size L, red shirt size M, etc. <br /><br /><span class=\"tablenote\"><b> Note: </b> This container is returned only if the <b> item_id</b> in the request is an item group (parent ID of an item with variations).</span>
+// checks if the ItemGroupSummary type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &ItemGroupSummary{}
+
+// ItemGroupSummary The type that defines the fields for the details of each item in an item group. An item group is  an item that has various aspect differences, such as color, size, storage capacity, etc. When an item group is created, one of the item variations, such as the red shirt size L, is chosen as the \"parent\". All the other items in the group are the children, such as the blue shirt size L, red shirt size M, etc. <br><br><span class=\"tablenote\"><b> Note: </b> This container is returned only if the <b> item_id</b> in the request is an item group (parent ID of an item with variations).</span>
 type ItemGroupSummary struct {
 	// An array of containers with the URLs for images that are in addition to the primary image of the item group.  The primary image is returned in the <b> itemGroupImage</b> field.
-	ItemGroupAdditionalImages *[]Image `json:"itemGroupAdditionalImages,omitempty"`
+	ItemGroupAdditionalImages []Image `json:"itemGroupAdditionalImages,omitempty"`
 	// The HATEOAS reference of the parent page of the item group. An item group is an item that has various aspect differences, such as color, size, storage capacity, etc. 
 	ItemGroupHref *string `json:"itemGroupHref,omitempty"`
 	// The unique identifier for the item group. An item group is an item that has various aspect differences, such as color, size, storage capacity, etc. 
@@ -48,17 +51,17 @@ func NewItemGroupSummaryWithDefaults() *ItemGroupSummary {
 
 // GetItemGroupAdditionalImages returns the ItemGroupAdditionalImages field value if set, zero value otherwise.
 func (o *ItemGroupSummary) GetItemGroupAdditionalImages() []Image {
-	if o == nil || o.ItemGroupAdditionalImages == nil {
+	if o == nil || IsNil(o.ItemGroupAdditionalImages) {
 		var ret []Image
 		return ret
 	}
-	return *o.ItemGroupAdditionalImages
+	return o.ItemGroupAdditionalImages
 }
 
 // GetItemGroupAdditionalImagesOk returns a tuple with the ItemGroupAdditionalImages field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *ItemGroupSummary) GetItemGroupAdditionalImagesOk() (*[]Image, bool) {
-	if o == nil || o.ItemGroupAdditionalImages == nil {
+func (o *ItemGroupSummary) GetItemGroupAdditionalImagesOk() ([]Image, bool) {
+	if o == nil || IsNil(o.ItemGroupAdditionalImages) {
 		return nil, false
 	}
 	return o.ItemGroupAdditionalImages, true
@@ -66,7 +69,7 @@ func (o *ItemGroupSummary) GetItemGroupAdditionalImagesOk() (*[]Image, bool) {
 
 // HasItemGroupAdditionalImages returns a boolean if a field has been set.
 func (o *ItemGroupSummary) HasItemGroupAdditionalImages() bool {
-	if o != nil && o.ItemGroupAdditionalImages != nil {
+	if o != nil && !IsNil(o.ItemGroupAdditionalImages) {
 		return true
 	}
 
@@ -75,12 +78,12 @@ func (o *ItemGroupSummary) HasItemGroupAdditionalImages() bool {
 
 // SetItemGroupAdditionalImages gets a reference to the given []Image and assigns it to the ItemGroupAdditionalImages field.
 func (o *ItemGroupSummary) SetItemGroupAdditionalImages(v []Image) {
-	o.ItemGroupAdditionalImages = &v
+	o.ItemGroupAdditionalImages = v
 }
 
 // GetItemGroupHref returns the ItemGroupHref field value if set, zero value otherwise.
 func (o *ItemGroupSummary) GetItemGroupHref() string {
-	if o == nil || o.ItemGroupHref == nil {
+	if o == nil || IsNil(o.ItemGroupHref) {
 		var ret string
 		return ret
 	}
@@ -90,7 +93,7 @@ func (o *ItemGroupSummary) GetItemGroupHref() string {
 // GetItemGroupHrefOk returns a tuple with the ItemGroupHref field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemGroupSummary) GetItemGroupHrefOk() (*string, bool) {
-	if o == nil || o.ItemGroupHref == nil {
+	if o == nil || IsNil(o.ItemGroupHref) {
 		return nil, false
 	}
 	return o.ItemGroupHref, true
@@ -98,7 +101,7 @@ func (o *ItemGroupSummary) GetItemGroupHrefOk() (*string, bool) {
 
 // HasItemGroupHref returns a boolean if a field has been set.
 func (o *ItemGroupSummary) HasItemGroupHref() bool {
-	if o != nil && o.ItemGroupHref != nil {
+	if o != nil && !IsNil(o.ItemGroupHref) {
 		return true
 	}
 
@@ -112,7 +115,7 @@ func (o *ItemGroupSummary) SetItemGroupHref(v string) {
 
 // GetItemGroupId returns the ItemGroupId field value if set, zero value otherwise.
 func (o *ItemGroupSummary) GetItemGroupId() string {
-	if o == nil || o.ItemGroupId == nil {
+	if o == nil || IsNil(o.ItemGroupId) {
 		var ret string
 		return ret
 	}
@@ -122,7 +125,7 @@ func (o *ItemGroupSummary) GetItemGroupId() string {
 // GetItemGroupIdOk returns a tuple with the ItemGroupId field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemGroupSummary) GetItemGroupIdOk() (*string, bool) {
-	if o == nil || o.ItemGroupId == nil {
+	if o == nil || IsNil(o.ItemGroupId) {
 		return nil, false
 	}
 	return o.ItemGroupId, true
@@ -130,7 +133,7 @@ func (o *ItemGroupSummary) GetItemGroupIdOk() (*string, bool) {
 
 // HasItemGroupId returns a boolean if a field has been set.
 func (o *ItemGroupSummary) HasItemGroupId() bool {
-	if o != nil && o.ItemGroupId != nil {
+	if o != nil && !IsNil(o.ItemGroupId) {
 		return true
 	}
 
@@ -144,7 +147,7 @@ func (o *ItemGroupSummary) SetItemGroupId(v string) {
 
 // GetItemGroupImage returns the ItemGroupImage field value if set, zero value otherwise.
 func (o *ItemGroupSummary) GetItemGroupImage() Image {
-	if o == nil || o.ItemGroupImage == nil {
+	if o == nil || IsNil(o.ItemGroupImage) {
 		var ret Image
 		return ret
 	}
@@ -154,7 +157,7 @@ func (o *ItemGroupSummary) GetItemGroupImage() Image {
 // GetItemGroupImageOk returns a tuple with the ItemGroupImage field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemGroupSummary) GetItemGroupImageOk() (*Image, bool) {
-	if o == nil || o.ItemGroupImage == nil {
+	if o == nil || IsNil(o.ItemGroupImage) {
 		return nil, false
 	}
 	return o.ItemGroupImage, true
@@ -162,7 +165,7 @@ func (o *ItemGroupSummary) GetItemGroupImageOk() (*Image, bool) {
 
 // HasItemGroupImage returns a boolean if a field has been set.
 func (o *ItemGroupSummary) HasItemGroupImage() bool {
-	if o != nil && o.ItemGroupImage != nil {
+	if o != nil && !IsNil(o.ItemGroupImage) {
 		return true
 	}
 
@@ -176,7 +179,7 @@ func (o *ItemGroupSummary) SetItemGroupImage(v Image) {
 
 // GetItemGroupTitle returns the ItemGroupTitle field value if set, zero value otherwise.
 func (o *ItemGroupSummary) GetItemGroupTitle() string {
-	if o == nil || o.ItemGroupTitle == nil {
+	if o == nil || IsNil(o.ItemGroupTitle) {
 		var ret string
 		return ret
 	}
@@ -186,7 +189,7 @@ func (o *ItemGroupSummary) GetItemGroupTitle() string {
 // GetItemGroupTitleOk returns a tuple with the ItemGroupTitle field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemGroupSummary) GetItemGroupTitleOk() (*string, bool) {
-	if o == nil || o.ItemGroupTitle == nil {
+	if o == nil || IsNil(o.ItemGroupTitle) {
 		return nil, false
 	}
 	return o.ItemGroupTitle, true
@@ -194,7 +197,7 @@ func (o *ItemGroupSummary) GetItemGroupTitleOk() (*string, bool) {
 
 // HasItemGroupTitle returns a boolean if a field has been set.
 func (o *ItemGroupSummary) HasItemGroupTitle() bool {
-	if o != nil && o.ItemGroupTitle != nil {
+	if o != nil && !IsNil(o.ItemGroupTitle) {
 		return true
 	}
 
@@ -208,7 +211,7 @@ func (o *ItemGroupSummary) SetItemGroupTitle(v string) {
 
 // GetItemGroupType returns the ItemGroupType field value if set, zero value otherwise.
 func (o *ItemGroupSummary) GetItemGroupType() string {
-	if o == nil || o.ItemGroupType == nil {
+	if o == nil || IsNil(o.ItemGroupType) {
 		var ret string
 		return ret
 	}
@@ -218,7 +221,7 @@ func (o *ItemGroupSummary) GetItemGroupType() string {
 // GetItemGroupTypeOk returns a tuple with the ItemGroupType field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemGroupSummary) GetItemGroupTypeOk() (*string, bool) {
-	if o == nil || o.ItemGroupType == nil {
+	if o == nil || IsNil(o.ItemGroupType) {
 		return nil, false
 	}
 	return o.ItemGroupType, true
@@ -226,7 +229,7 @@ func (o *ItemGroupSummary) GetItemGroupTypeOk() (*string, bool) {
 
 // HasItemGroupType returns a boolean if a field has been set.
 func (o *ItemGroupSummary) HasItemGroupType() bool {
-	if o != nil && o.ItemGroupType != nil {
+	if o != nil && !IsNil(o.ItemGroupType) {
 		return true
 	}
 
@@ -239,26 +242,34 @@ func (o *ItemGroupSummary) SetItemGroupType(v string) {
 }
 
 func (o ItemGroupSummary) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.ItemGroupAdditionalImages != nil {
-		toSerialize["itemGroupAdditionalImages"] = o.ItemGroupAdditionalImages
-	}
-	if o.ItemGroupHref != nil {
-		toSerialize["itemGroupHref"] = o.ItemGroupHref
-	}
-	if o.ItemGroupId != nil {
-		toSerialize["itemGroupId"] = o.ItemGroupId
-	}
-	if o.ItemGroupImage != nil {
-		toSerialize["itemGroupImage"] = o.ItemGroupImage
-	}
-	if o.ItemGroupTitle != nil {
-		toSerialize["itemGroupTitle"] = o.ItemGroupTitle
-	}
-	if o.ItemGroupType != nil {
-		toSerialize["itemGroupType"] = o.ItemGroupType
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o ItemGroupSummary) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.ItemGroupAdditionalImages) {
+		toSerialize["itemGroupAdditionalImages"] = o.ItemGroupAdditionalImages
+	}
+	if !IsNil(o.ItemGroupHref) {
+		toSerialize["itemGroupHref"] = o.ItemGroupHref
+	}
+	if !IsNil(o.ItemGroupId) {
+		toSerialize["itemGroupId"] = o.ItemGroupId
+	}
+	if !IsNil(o.ItemGroupImage) {
+		toSerialize["itemGroupImage"] = o.ItemGroupImage
+	}
+	if !IsNil(o.ItemGroupTitle) {
+		toSerialize["itemGroupTitle"] = o.ItemGroupTitle
+	}
+	if !IsNil(o.ItemGroupType) {
+		toSerialize["itemGroupType"] = o.ItemGroupType
+	}
+	return toSerialize, nil
 }
 
 type NullableItemGroupSummary struct {
