@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,19 +14,22 @@ import (
 	"encoding/json"
 )
 
+// checks if the ItemLocationImpl type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &ItemLocationImpl{}
+
 // ItemLocationImpl The type that defines the fields for the location of an item, such as information typically used for an address, including postal code, county, state/province, street address, city, and country (2-digit ISO code).
 type ItemLocationImpl struct {
 	// The first line of the street address.
 	AddressLine1 *string `json:"addressLine1,omitempty"`
 	// The second line of the street address. This field may contain such values as an apartment or suite number.
 	AddressLine2 *string `json:"addressLine2,omitempty"`
-	// The city in which the item is located. <br /><br /><b>Restriction:</b> This field is populated in the <b> search</b> method response <i> only</i> when <b> fieldgroups</b> = <code>EXTENDED</code>.
+	// The city in which the item is located.<br><br><b>Restriction:</b> This field is populated in the <code>search</code> method response <i>only</i> when <code>fieldgroups</code> = <code>EXTENDED</code>.
 	City *string `json:"city,omitempty"`
-	// The two-letter <a href=\"https://www.iso.org/iso-3166-country-codes.html\">ISO 3166</a> standard code that indicates the country in which the item is located.  For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/ba:CountryCodeEnum'>eBay API documentation</a>
+	// The two-letter <a href=\"https://www.iso.org/iso-3166-country-codes.html \" target=\"_blank\">ISO 3166</a> standard code that indicates the country in which the item is located. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/ba:CountryCodeEnum'>eBay API documentation</a>
 	Country *string `json:"country,omitempty"`
 	// The county in which the item is located.
 	County *string `json:"county,omitempty"`
-	// The postal code (or zip code in US) where the item is located. Sellers set a postal code (or zip code in US) for items when they are listed. The postal code is used for calculating proximity searches. It is anonymized when returned in <b>itemLocation.postalCode</b> via the API.
+	// The postal code (or zip code in US) where the item is located. Sellers set a postal code for items when they are listed. The postal code is used for calculating proximity searches. It is anonymized when returned in <code>itemLocation.postalCode</code> via the API.
 	PostalCode *string `json:"postalCode,omitempty"`
 	// The state or province in which the item is located.
 	StateOrProvince *string `json:"stateOrProvince,omitempty"`
@@ -51,7 +54,7 @@ func NewItemLocationImplWithDefaults() *ItemLocationImpl {
 
 // GetAddressLine1 returns the AddressLine1 field value if set, zero value otherwise.
 func (o *ItemLocationImpl) GetAddressLine1() string {
-	if o == nil || o.AddressLine1 == nil {
+	if o == nil || IsNil(o.AddressLine1) {
 		var ret string
 		return ret
 	}
@@ -61,7 +64,7 @@ func (o *ItemLocationImpl) GetAddressLine1() string {
 // GetAddressLine1Ok returns a tuple with the AddressLine1 field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemLocationImpl) GetAddressLine1Ok() (*string, bool) {
-	if o == nil || o.AddressLine1 == nil {
+	if o == nil || IsNil(o.AddressLine1) {
 		return nil, false
 	}
 	return o.AddressLine1, true
@@ -69,7 +72,7 @@ func (o *ItemLocationImpl) GetAddressLine1Ok() (*string, bool) {
 
 // HasAddressLine1 returns a boolean if a field has been set.
 func (o *ItemLocationImpl) HasAddressLine1() bool {
-	if o != nil && o.AddressLine1 != nil {
+	if o != nil && !IsNil(o.AddressLine1) {
 		return true
 	}
 
@@ -83,7 +86,7 @@ func (o *ItemLocationImpl) SetAddressLine1(v string) {
 
 // GetAddressLine2 returns the AddressLine2 field value if set, zero value otherwise.
 func (o *ItemLocationImpl) GetAddressLine2() string {
-	if o == nil || o.AddressLine2 == nil {
+	if o == nil || IsNil(o.AddressLine2) {
 		var ret string
 		return ret
 	}
@@ -93,7 +96,7 @@ func (o *ItemLocationImpl) GetAddressLine2() string {
 // GetAddressLine2Ok returns a tuple with the AddressLine2 field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemLocationImpl) GetAddressLine2Ok() (*string, bool) {
-	if o == nil || o.AddressLine2 == nil {
+	if o == nil || IsNil(o.AddressLine2) {
 		return nil, false
 	}
 	return o.AddressLine2, true
@@ -101,7 +104,7 @@ func (o *ItemLocationImpl) GetAddressLine2Ok() (*string, bool) {
 
 // HasAddressLine2 returns a boolean if a field has been set.
 func (o *ItemLocationImpl) HasAddressLine2() bool {
-	if o != nil && o.AddressLine2 != nil {
+	if o != nil && !IsNil(o.AddressLine2) {
 		return true
 	}
 
@@ -115,7 +118,7 @@ func (o *ItemLocationImpl) SetAddressLine2(v string) {
 
 // GetCity returns the City field value if set, zero value otherwise.
 func (o *ItemLocationImpl) GetCity() string {
-	if o == nil || o.City == nil {
+	if o == nil || IsNil(o.City) {
 		var ret string
 		return ret
 	}
@@ -125,7 +128,7 @@ func (o *ItemLocationImpl) GetCity() string {
 // GetCityOk returns a tuple with the City field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemLocationImpl) GetCityOk() (*string, bool) {
-	if o == nil || o.City == nil {
+	if o == nil || IsNil(o.City) {
 		return nil, false
 	}
 	return o.City, true
@@ -133,7 +136,7 @@ func (o *ItemLocationImpl) GetCityOk() (*string, bool) {
 
 // HasCity returns a boolean if a field has been set.
 func (o *ItemLocationImpl) HasCity() bool {
-	if o != nil && o.City != nil {
+	if o != nil && !IsNil(o.City) {
 		return true
 	}
 
@@ -147,7 +150,7 @@ func (o *ItemLocationImpl) SetCity(v string) {
 
 // GetCountry returns the Country field value if set, zero value otherwise.
 func (o *ItemLocationImpl) GetCountry() string {
-	if o == nil || o.Country == nil {
+	if o == nil || IsNil(o.Country) {
 		var ret string
 		return ret
 	}
@@ -157,7 +160,7 @@ func (o *ItemLocationImpl) GetCountry() string {
 // GetCountryOk returns a tuple with the Country field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemLocationImpl) GetCountryOk() (*string, bool) {
-	if o == nil || o.Country == nil {
+	if o == nil || IsNil(o.Country) {
 		return nil, false
 	}
 	return o.Country, true
@@ -165,7 +168,7 @@ func (o *ItemLocationImpl) GetCountryOk() (*string, bool) {
 
 // HasCountry returns a boolean if a field has been set.
 func (o *ItemLocationImpl) HasCountry() bool {
-	if o != nil && o.Country != nil {
+	if o != nil && !IsNil(o.Country) {
 		return true
 	}
 
@@ -179,7 +182,7 @@ func (o *ItemLocationImpl) SetCountry(v string) {
 
 // GetCounty returns the County field value if set, zero value otherwise.
 func (o *ItemLocationImpl) GetCounty() string {
-	if o == nil || o.County == nil {
+	if o == nil || IsNil(o.County) {
 		var ret string
 		return ret
 	}
@@ -189,7 +192,7 @@ func (o *ItemLocationImpl) GetCounty() string {
 // GetCountyOk returns a tuple with the County field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemLocationImpl) GetCountyOk() (*string, bool) {
-	if o == nil || o.County == nil {
+	if o == nil || IsNil(o.County) {
 		return nil, false
 	}
 	return o.County, true
@@ -197,7 +200,7 @@ func (o *ItemLocationImpl) GetCountyOk() (*string, bool) {
 
 // HasCounty returns a boolean if a field has been set.
 func (o *ItemLocationImpl) HasCounty() bool {
-	if o != nil && o.County != nil {
+	if o != nil && !IsNil(o.County) {
 		return true
 	}
 
@@ -211,7 +214,7 @@ func (o *ItemLocationImpl) SetCounty(v string) {
 
 // GetPostalCode returns the PostalCode field value if set, zero value otherwise.
 func (o *ItemLocationImpl) GetPostalCode() string {
-	if o == nil || o.PostalCode == nil {
+	if o == nil || IsNil(o.PostalCode) {
 		var ret string
 		return ret
 	}
@@ -221,7 +224,7 @@ func (o *ItemLocationImpl) GetPostalCode() string {
 // GetPostalCodeOk returns a tuple with the PostalCode field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemLocationImpl) GetPostalCodeOk() (*string, bool) {
-	if o == nil || o.PostalCode == nil {
+	if o == nil || IsNil(o.PostalCode) {
 		return nil, false
 	}
 	return o.PostalCode, true
@@ -229,7 +232,7 @@ func (o *ItemLocationImpl) GetPostalCodeOk() (*string, bool) {
 
 // HasPostalCode returns a boolean if a field has been set.
 func (o *ItemLocationImpl) HasPostalCode() bool {
-	if o != nil && o.PostalCode != nil {
+	if o != nil && !IsNil(o.PostalCode) {
 		return true
 	}
 
@@ -243,7 +246,7 @@ func (o *ItemLocationImpl) SetPostalCode(v string) {
 
 // GetStateOrProvince returns the StateOrProvince field value if set, zero value otherwise.
 func (o *ItemLocationImpl) GetStateOrProvince() string {
-	if o == nil || o.StateOrProvince == nil {
+	if o == nil || IsNil(o.StateOrProvince) {
 		var ret string
 		return ret
 	}
@@ -253,7 +256,7 @@ func (o *ItemLocationImpl) GetStateOrProvince() string {
 // GetStateOrProvinceOk returns a tuple with the StateOrProvince field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemLocationImpl) GetStateOrProvinceOk() (*string, bool) {
-	if o == nil || o.StateOrProvince == nil {
+	if o == nil || IsNil(o.StateOrProvince) {
 		return nil, false
 	}
 	return o.StateOrProvince, true
@@ -261,7 +264,7 @@ func (o *ItemLocationImpl) GetStateOrProvinceOk() (*string, bool) {
 
 // HasStateOrProvince returns a boolean if a field has been set.
 func (o *ItemLocationImpl) HasStateOrProvince() bool {
-	if o != nil && o.StateOrProvince != nil {
+	if o != nil && !IsNil(o.StateOrProvince) {
 		return true
 	}
 
@@ -274,29 +277,37 @@ func (o *ItemLocationImpl) SetStateOrProvince(v string) {
 }
 
 func (o ItemLocationImpl) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.AddressLine1 != nil {
-		toSerialize["addressLine1"] = o.AddressLine1
-	}
-	if o.AddressLine2 != nil {
-		toSerialize["addressLine2"] = o.AddressLine2
-	}
-	if o.City != nil {
-		toSerialize["city"] = o.City
-	}
-	if o.Country != nil {
-		toSerialize["country"] = o.Country
-	}
-	if o.County != nil {
-		toSerialize["county"] = o.County
-	}
-	if o.PostalCode != nil {
-		toSerialize["postalCode"] = o.PostalCode
-	}
-	if o.StateOrProvince != nil {
-		toSerialize["stateOrProvince"] = o.StateOrProvince
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o ItemLocationImpl) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.AddressLine1) {
+		toSerialize["addressLine1"] = o.AddressLine1
+	}
+	if !IsNil(o.AddressLine2) {
+		toSerialize["addressLine2"] = o.AddressLine2
+	}
+	if !IsNil(o.City) {
+		toSerialize["city"] = o.City
+	}
+	if !IsNil(o.Country) {
+		toSerialize["country"] = o.Country
+	}
+	if !IsNil(o.County) {
+		toSerialize["county"] = o.County
+	}
+	if !IsNil(o.PostalCode) {
+		toSerialize["postalCode"] = o.PostalCode
+	}
+	if !IsNil(o.StateOrProvince) {
+		toSerialize["stateOrProvince"] = o.StateOrProvince
+	}
+	return toSerialize, nil
 }
 
 type NullableItemLocationImpl struct {
