@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,11 +14,14 @@ import (
 	"encoding/json"
 )
 
+// checks if the ItemReturnTerms type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &ItemReturnTerms{}
+
 // ItemReturnTerms The type that defines the fields for the seller's return policy.
 type ItemReturnTerms struct {
 	// This indicates if the seller has enabled the Extended Holiday Returns feature on the item. Extended Holiday Returns are only applicable during the US holiday season, and gives buyers extra time to return an item. This 'extra time' will typically extend beyond what is set through the <b> returnPeriod</b> value.
 	ExtendedHolidayReturnsOffered *bool `json:"extendedHolidayReturnsOffered,omitempty"`
-	// An enumeration value that indicates how a buyer is refunded when an item is returned. <br><br><b> Valid Values: </b> MONEY_BACK or MERCHANDISE_CREDIT  <br /><br />Code so that your app gracefully handles any future changes to this list. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:RefundMethodEnum'>eBay API documentation</a>
+	// An enumeration value that indicates how a buyer is refunded when an item is returned. <br><br><b> Valid Values: </b> MONEY_BACK or MERCHANDISE_CREDIT  <br><br>Code so that your app gracefully handles any future changes to this list. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:RefundMethodEnum'>eBay API documentation</a>
 	RefundMethod *string `json:"refundMethod,omitempty"`
 	// This string field indicates the restocking fee percentage that the seller has set on the item. Sellers have the option of setting no restocking fee for an item, or they can set the percentage to 10, 15, or 20 percent. So, if the cost of the item was $100, and the restocking percentage was 20 percent, the buyer would be charged $20 to return that item, so instead of receiving a $100 refund, they would receive $80 due to the restocking fee.
 	RestockingFeePercentage *string `json:"restockingFeePercentage,omitempty"`
@@ -52,7 +55,7 @@ func NewItemReturnTermsWithDefaults() *ItemReturnTerms {
 
 // GetExtendedHolidayReturnsOffered returns the ExtendedHolidayReturnsOffered field value if set, zero value otherwise.
 func (o *ItemReturnTerms) GetExtendedHolidayReturnsOffered() bool {
-	if o == nil || o.ExtendedHolidayReturnsOffered == nil {
+	if o == nil || IsNil(o.ExtendedHolidayReturnsOffered) {
 		var ret bool
 		return ret
 	}
@@ -62,7 +65,7 @@ func (o *ItemReturnTerms) GetExtendedHolidayReturnsOffered() bool {
 // GetExtendedHolidayReturnsOfferedOk returns a tuple with the ExtendedHolidayReturnsOffered field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemReturnTerms) GetExtendedHolidayReturnsOfferedOk() (*bool, bool) {
-	if o == nil || o.ExtendedHolidayReturnsOffered == nil {
+	if o == nil || IsNil(o.ExtendedHolidayReturnsOffered) {
 		return nil, false
 	}
 	return o.ExtendedHolidayReturnsOffered, true
@@ -70,7 +73,7 @@ func (o *ItemReturnTerms) GetExtendedHolidayReturnsOfferedOk() (*bool, bool) {
 
 // HasExtendedHolidayReturnsOffered returns a boolean if a field has been set.
 func (o *ItemReturnTerms) HasExtendedHolidayReturnsOffered() bool {
-	if o != nil && o.ExtendedHolidayReturnsOffered != nil {
+	if o != nil && !IsNil(o.ExtendedHolidayReturnsOffered) {
 		return true
 	}
 
@@ -84,7 +87,7 @@ func (o *ItemReturnTerms) SetExtendedHolidayReturnsOffered(v bool) {
 
 // GetRefundMethod returns the RefundMethod field value if set, zero value otherwise.
 func (o *ItemReturnTerms) GetRefundMethod() string {
-	if o == nil || o.RefundMethod == nil {
+	if o == nil || IsNil(o.RefundMethod) {
 		var ret string
 		return ret
 	}
@@ -94,7 +97,7 @@ func (o *ItemReturnTerms) GetRefundMethod() string {
 // GetRefundMethodOk returns a tuple with the RefundMethod field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemReturnTerms) GetRefundMethodOk() (*string, bool) {
-	if o == nil || o.RefundMethod == nil {
+	if o == nil || IsNil(o.RefundMethod) {
 		return nil, false
 	}
 	return o.RefundMethod, true
@@ -102,7 +105,7 @@ func (o *ItemReturnTerms) GetRefundMethodOk() (*string, bool) {
 
 // HasRefundMethod returns a boolean if a field has been set.
 func (o *ItemReturnTerms) HasRefundMethod() bool {
-	if o != nil && o.RefundMethod != nil {
+	if o != nil && !IsNil(o.RefundMethod) {
 		return true
 	}
 
@@ -116,7 +119,7 @@ func (o *ItemReturnTerms) SetRefundMethod(v string) {
 
 // GetRestockingFeePercentage returns the RestockingFeePercentage field value if set, zero value otherwise.
 func (o *ItemReturnTerms) GetRestockingFeePercentage() string {
-	if o == nil || o.RestockingFeePercentage == nil {
+	if o == nil || IsNil(o.RestockingFeePercentage) {
 		var ret string
 		return ret
 	}
@@ -126,7 +129,7 @@ func (o *ItemReturnTerms) GetRestockingFeePercentage() string {
 // GetRestockingFeePercentageOk returns a tuple with the RestockingFeePercentage field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemReturnTerms) GetRestockingFeePercentageOk() (*string, bool) {
-	if o == nil || o.RestockingFeePercentage == nil {
+	if o == nil || IsNil(o.RestockingFeePercentage) {
 		return nil, false
 	}
 	return o.RestockingFeePercentage, true
@@ -134,7 +137,7 @@ func (o *ItemReturnTerms) GetRestockingFeePercentageOk() (*string, bool) {
 
 // HasRestockingFeePercentage returns a boolean if a field has been set.
 func (o *ItemReturnTerms) HasRestockingFeePercentage() bool {
-	if o != nil && o.RestockingFeePercentage != nil {
+	if o != nil && !IsNil(o.RestockingFeePercentage) {
 		return true
 	}
 
@@ -148,7 +151,7 @@ func (o *ItemReturnTerms) SetRestockingFeePercentage(v string) {
 
 // GetReturnInstructions returns the ReturnInstructions field value if set, zero value otherwise.
 func (o *ItemReturnTerms) GetReturnInstructions() string {
-	if o == nil || o.ReturnInstructions == nil {
+	if o == nil || IsNil(o.ReturnInstructions) {
 		var ret string
 		return ret
 	}
@@ -158,7 +161,7 @@ func (o *ItemReturnTerms) GetReturnInstructions() string {
 // GetReturnInstructionsOk returns a tuple with the ReturnInstructions field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemReturnTerms) GetReturnInstructionsOk() (*string, bool) {
-	if o == nil || o.ReturnInstructions == nil {
+	if o == nil || IsNil(o.ReturnInstructions) {
 		return nil, false
 	}
 	return o.ReturnInstructions, true
@@ -166,7 +169,7 @@ func (o *ItemReturnTerms) GetReturnInstructionsOk() (*string, bool) {
 
 // HasReturnInstructions returns a boolean if a field has been set.
 func (o *ItemReturnTerms) HasReturnInstructions() bool {
-	if o != nil && o.ReturnInstructions != nil {
+	if o != nil && !IsNil(o.ReturnInstructions) {
 		return true
 	}
 
@@ -180,7 +183,7 @@ func (o *ItemReturnTerms) SetReturnInstructions(v string) {
 
 // GetReturnMethod returns the ReturnMethod field value if set, zero value otherwise.
 func (o *ItemReturnTerms) GetReturnMethod() string {
-	if o == nil || o.ReturnMethod == nil {
+	if o == nil || IsNil(o.ReturnMethod) {
 		var ret string
 		return ret
 	}
@@ -190,7 +193,7 @@ func (o *ItemReturnTerms) GetReturnMethod() string {
 // GetReturnMethodOk returns a tuple with the ReturnMethod field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemReturnTerms) GetReturnMethodOk() (*string, bool) {
-	if o == nil || o.ReturnMethod == nil {
+	if o == nil || IsNil(o.ReturnMethod) {
 		return nil, false
 	}
 	return o.ReturnMethod, true
@@ -198,7 +201,7 @@ func (o *ItemReturnTerms) GetReturnMethodOk() (*string, bool) {
 
 // HasReturnMethod returns a boolean if a field has been set.
 func (o *ItemReturnTerms) HasReturnMethod() bool {
-	if o != nil && o.ReturnMethod != nil {
+	if o != nil && !IsNil(o.ReturnMethod) {
 		return true
 	}
 
@@ -212,7 +215,7 @@ func (o *ItemReturnTerms) SetReturnMethod(v string) {
 
 // GetReturnPeriod returns the ReturnPeriod field value if set, zero value otherwise.
 func (o *ItemReturnTerms) GetReturnPeriod() TimeDuration {
-	if o == nil || o.ReturnPeriod == nil {
+	if o == nil || IsNil(o.ReturnPeriod) {
 		var ret TimeDuration
 		return ret
 	}
@@ -222,7 +225,7 @@ func (o *ItemReturnTerms) GetReturnPeriod() TimeDuration {
 // GetReturnPeriodOk returns a tuple with the ReturnPeriod field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemReturnTerms) GetReturnPeriodOk() (*TimeDuration, bool) {
-	if o == nil || o.ReturnPeriod == nil {
+	if o == nil || IsNil(o.ReturnPeriod) {
 		return nil, false
 	}
 	return o.ReturnPeriod, true
@@ -230,7 +233,7 @@ func (o *ItemReturnTerms) GetReturnPeriodOk() (*TimeDuration, bool) {
 
 // HasReturnPeriod returns a boolean if a field has been set.
 func (o *ItemReturnTerms) HasReturnPeriod() bool {
-	if o != nil && o.ReturnPeriod != nil {
+	if o != nil && !IsNil(o.ReturnPeriod) {
 		return true
 	}
 
@@ -244,7 +247,7 @@ func (o *ItemReturnTerms) SetReturnPeriod(v TimeDuration) {
 
 // GetReturnsAccepted returns the ReturnsAccepted field value if set, zero value otherwise.
 func (o *ItemReturnTerms) GetReturnsAccepted() bool {
-	if o == nil || o.ReturnsAccepted == nil {
+	if o == nil || IsNil(o.ReturnsAccepted) {
 		var ret bool
 		return ret
 	}
@@ -254,7 +257,7 @@ func (o *ItemReturnTerms) GetReturnsAccepted() bool {
 // GetReturnsAcceptedOk returns a tuple with the ReturnsAccepted field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemReturnTerms) GetReturnsAcceptedOk() (*bool, bool) {
-	if o == nil || o.ReturnsAccepted == nil {
+	if o == nil || IsNil(o.ReturnsAccepted) {
 		return nil, false
 	}
 	return o.ReturnsAccepted, true
@@ -262,7 +265,7 @@ func (o *ItemReturnTerms) GetReturnsAcceptedOk() (*bool, bool) {
 
 // HasReturnsAccepted returns a boolean if a field has been set.
 func (o *ItemReturnTerms) HasReturnsAccepted() bool {
-	if o != nil && o.ReturnsAccepted != nil {
+	if o != nil && !IsNil(o.ReturnsAccepted) {
 		return true
 	}
 
@@ -276,7 +279,7 @@ func (o *ItemReturnTerms) SetReturnsAccepted(v bool) {
 
 // GetReturnShippingCostPayer returns the ReturnShippingCostPayer field value if set, zero value otherwise.
 func (o *ItemReturnTerms) GetReturnShippingCostPayer() string {
-	if o == nil || o.ReturnShippingCostPayer == nil {
+	if o == nil || IsNil(o.ReturnShippingCostPayer) {
 		var ret string
 		return ret
 	}
@@ -286,7 +289,7 @@ func (o *ItemReturnTerms) GetReturnShippingCostPayer() string {
 // GetReturnShippingCostPayerOk returns a tuple with the ReturnShippingCostPayer field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemReturnTerms) GetReturnShippingCostPayerOk() (*string, bool) {
-	if o == nil || o.ReturnShippingCostPayer == nil {
+	if o == nil || IsNil(o.ReturnShippingCostPayer) {
 		return nil, false
 	}
 	return o.ReturnShippingCostPayer, true
@@ -294,7 +297,7 @@ func (o *ItemReturnTerms) GetReturnShippingCostPayerOk() (*string, bool) {
 
 // HasReturnShippingCostPayer returns a boolean if a field has been set.
 func (o *ItemReturnTerms) HasReturnShippingCostPayer() bool {
-	if o != nil && o.ReturnShippingCostPayer != nil {
+	if o != nil && !IsNil(o.ReturnShippingCostPayer) {
 		return true
 	}
 
@@ -307,32 +310,40 @@ func (o *ItemReturnTerms) SetReturnShippingCostPayer(v string) {
 }
 
 func (o ItemReturnTerms) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.ExtendedHolidayReturnsOffered != nil {
-		toSerialize["extendedHolidayReturnsOffered"] = o.ExtendedHolidayReturnsOffered
-	}
-	if o.RefundMethod != nil {
-		toSerialize["refundMethod"] = o.RefundMethod
-	}
-	if o.RestockingFeePercentage != nil {
-		toSerialize["restockingFeePercentage"] = o.RestockingFeePercentage
-	}
-	if o.ReturnInstructions != nil {
-		toSerialize["returnInstructions"] = o.ReturnInstructions
-	}
-	if o.ReturnMethod != nil {
-		toSerialize["returnMethod"] = o.ReturnMethod
-	}
-	if o.ReturnPeriod != nil {
-		toSerialize["returnPeriod"] = o.ReturnPeriod
-	}
-	if o.ReturnsAccepted != nil {
-		toSerialize["returnsAccepted"] = o.ReturnsAccepted
-	}
-	if o.ReturnShippingCostPayer != nil {
-		toSerialize["returnShippingCostPayer"] = o.ReturnShippingCostPayer
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o ItemReturnTerms) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.ExtendedHolidayReturnsOffered) {
+		toSerialize["extendedHolidayReturnsOffered"] = o.ExtendedHolidayReturnsOffered
+	}
+	if !IsNil(o.RefundMethod) {
+		toSerialize["refundMethod"] = o.RefundMethod
+	}
+	if !IsNil(o.RestockingFeePercentage) {
+		toSerialize["restockingFeePercentage"] = o.RestockingFeePercentage
+	}
+	if !IsNil(o.ReturnInstructions) {
+		toSerialize["returnInstructions"] = o.ReturnInstructions
+	}
+	if !IsNil(o.ReturnMethod) {
+		toSerialize["returnMethod"] = o.ReturnMethod
+	}
+	if !IsNil(o.ReturnPeriod) {
+		toSerialize["returnPeriod"] = o.ReturnPeriod
+	}
+	if !IsNil(o.ReturnsAccepted) {
+		toSerialize["returnsAccepted"] = o.ReturnsAccepted
+	}
+	if !IsNil(o.ReturnShippingCostPayer) {
+		toSerialize["returnShippingCostPayer"] = o.ReturnShippingCostPayer
+	}
+	return toSerialize, nil
 }
 
 type NullableItemReturnTerms struct {
