@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,79 +14,88 @@ import (
 	"encoding/json"
 )
 
+// checks if the ItemSummary type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &ItemSummary{}
+
 // ItemSummary The type that defines the fields for the details of a specific item.
 type ItemSummary struct {
-	// An array of containers with the URLs for the images that are in addition to the primary image.  The primary image is returned in the <b> image.imageUrl</b> field.
-	AdditionalImages *[]Image `json:"additionalImages,omitempty"`
-	// This indicates if the item is for adults only. For more information about adult-only items on eBay, see <a href=\"https://pages.ebay.com/help/policies/adult-only.html\" target=\"_blank\">Adult items policy</a> for sellers and <a href=\"https://www.ebay.com/help/terms-conditions/default/searching-adult-items?id=4661\" target=\"_blank\">Adult-Only items on eBay</a> for buyers.
+	// An array of containers with the URLs for the images that are in addition to the primary image. The primary image is returned in the <code>image.imageUrl</code> field.
+	AdditionalImages []Image `json:"additionalImages,omitempty"`
+	// This indicates if the item is for adults only. For more information about adult-only items on eBay, refer to the <a href=\"https://www.ebay.com/help/policies/prohibited-restricted-items/adult-items-policy?id=4278 \" target=\"_blank\">Adult items policy</a>.
 	AdultOnly *bool `json:"adultOnly,omitempty"`
 	// This boolean attribute indicates if coupons are available for the item.
 	AvailableCoupons *bool `json:"availableCoupons,omitempty"`
 	// This integer value indicates the total number of bids that have been placed for an auction item. This field is only returned for auction items.
 	BidCount *int32 `json:"bidCount,omitempty"`
-	// A comma separated list of all the purchase options available for the item. <br><br><b> Values Returned: </b>   <ul> <li><b>FIXED_PRICE</b> - Indicates the buyer can purchase the item for a set price using the Buy It Now button. </li>  <li><b> AUCTION</b> - Indicates the buyer can place a bid for the item. After the first bid is placed, this becomes a live auction item and is the only buying option for this item.</li>  <li><b> BEST_OFFER</b> - Items where the buyer can send the seller a price they're willing to pay for the item. The seller can accept, reject, or send a counter offer. For details about Best Offer, see <a href=\"https://www.ebay.com/help/selling/listings/selling-buy-now/adding-best-offer-listing?id=4144\" target=\"_blank\">Best Offer</a>.  </li></ul> Code so that your app gracefully handles any future changes to this list.
-	BuyingOptions *[]string `json:"buyingOptions,omitempty"`
-	// This container returns the primary category ID of the item (as well as the secondary category if the item was listed in two categories). 
-	Categories *[]Category `json:"categories,omitempty"`
-	// This indicates how well the item matches the <b>compatibility_filter</b> product attributes.  <br><br><b> Valid Values: </b> EXACT or POSSIBLE <br /><br />Code so that your app gracefully handles any future changes to this list. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:CompatibilityMatchEnum'>eBay API documentation</a>
+	// A comma separated list of all the purchase options available for the item.<br><br><b>Values Returned:</b><ul><li><code>FIXED_PRICE</code><br>Indicates the buyer can purchase the item for a set price using the <i>Buy It Now</i> button.</li><li><code>AUCTION</code><br>Indicates the buyer can place a bid for the item. After the first bid is placed, this becomes a live auction item and is the only buying option for this item.</li><li><code>BEST_OFFER</code><br>Items where the buyer can send the seller a price they are willing to pay for the item. The seller can accept, reject, or send a counter offer. For additional information about Best Offer, refer to <a href=\"https://www.ebay.com/help/selling/listings/selling-buy-now/adding-best-offer-listing?id=4144 \" target=\"_blank\">Adding Best Offer to your listing and sending offers to buyers</a>.</li><li><code>CLASSIFIED_AD</code><br>Indicates that the final sales transaction is to be completed outside of the eBay environment.</li></ul>
+	BuyingOptions []string `json:"buyingOptions,omitempty"`
+	// This array returns the name and ID of each category associated with the item, including top level, branch, and leaf categories.
+	Categories []Category `json:"categories,omitempty"`
+	// This indicates how well an item matches the <code>compatibility_filter</code> product attributes.<br><br><b>Valid Values:</b><ul><li><code>EXACT</code></li><li><code>POSSIBLE</li></ul> For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:CompatibilityMatchEnum'>eBay API documentation</a>
 	CompatibilityMatch *string `json:"compatibilityMatch,omitempty"`
-	// This container returns only the product attributes that are compatible with the item. These attributes were specified in the <b>compatibility_filter</b> in the request. This means that if you passed in 5 attributes and only 4 are compatible, only those 4 are returned. If none of the attributes are compatible, this container is not returned.
-	CompatibilityProperties *[]CompatibilityProperty `json:"compatibilityProperties,omitempty"`
-	// The text describing the condition of the item, such as New or Used. For a list of condition names, see <a href=\"https://developer.ebay.com/devzone/finding/callref/enums/conditionIdList.html\" target=\"_blank\">Item Condition IDs and Names</a>.  <br /><br />Code so that your app gracefully handles any future changes to this list. <br /><br /><span class=\"tablenote\"><b> Note: </b>As of September 1, 2021, condition ID 2500 ('Seller Refurbished') is no longer a valid item condition in the <b>Cell Phones & Smartphones</b> category (category ID 9355) for the following marketplaces: US, Canada, UK, Germany, and Australia. This refurbished item condition has been replaced by three new refurbished values, which include 'Excellent - Refurbished' (condition ID 2010), 'Very Good - Refurbished' (condition ID 2020), and 'Good - Refurbished' (condition ID 2030).</span>
+	// This container returns only the product attributes that are compatible with the item. These attributes were specified in the <code>compatibility_filter</code> in the request. This means that if you passed in 5 attributes and only 4 are compatible, only those 4 are returned. If none of the attributes are compatible, this container is not returned.
+	CompatibilityProperties []CompatibilityProperty `json:"compatibilityProperties,omitempty"`
+	// The text describing the condition of the item, such as <b>New</b> or <b>Used</b>. For a list of condition names, refer to <a href=\"/devzone/finding/callref/enums/conditionIdList.html \" target=\"_blank\">Item Condition IDs and Names</a>.
 	Condition *string `json:"condition,omitempty"`
-	// The identifier of the condition of the item. For example, 1000 is the identifier for NEW. For a list of condition names and IDs, see <a https://developer.ebay.com/devzone/finding/callref/enums/conditionIdList.html\" target=\"_blank\">Item Condition IDs and Names</a>. <br /><br />Code so that your app gracefully handles any future changes to this list.<br /><br /><span class=\"tablenote\"><b> Note: </b>As of September 1, 2021, condition ID 2500 ('Seller Refurbished') is no longer a valid item condition in the <b>Cell Phones & Smartphones</b> category (category ID 9355) for the following marketplaces: US, Canada, UK, Germany, and Australia. This refurbished item condition has been replaced by three new refurbished values, which include 'Excellent - Refurbished' (condition ID 2010), 'Very Good - Refurbished' (condition ID 2020), and 'Good - Refurbished' (condition ID 2030).</span>
+	// The identifier of the condition of the item. For example, <code>1000</code> is the identifier for <code>NEW</code>. For a list of condition names and IDs, refer to <a href=\"/devzone/finding/callref/enums/conditionIdList.html \" target=\"_blank\">Item Condition IDs and Names</a>.
 	ConditionId *string `json:"conditionId,omitempty"`
 	CurrentBidPrice *ConvertedAmount `json:"currentBidPrice,omitempty"`
 	DistanceFromPickupLocation *TargetLocation `json:"distanceFromPickupLocation,omitempty"`
-	// This indicates the <a href=\"https://en.wikipedia.org/wiki/European_Union_energy_label\">European energy efficiency</a> rating (EEK) of the item.  Energy efficiency ratings apply to products listed by commercial vendors in electronics categories only. <br /><br />Currently, this field is only applicable for the Germany site, and  is only returned if the seller specified the energy efficiency rating through item specifics at listing time. Rating values include <code>A+++</code>, <code>A++</code>, <code>A+</code>, <code>A</code>, <code>B</code>, <code>C</code>, <code>D</code>, <code>E</code>, <code>F</code>, and <code>G</code>.
+	// This indicates the <a href=\"https://en.wikipedia.org/wiki/European_Union_energy_label \" target=\"_blank\">European energy efficiency</a> rating (EEK) of the item. Energy efficiency ratings apply to products listed by commercial vendors in electronics categories only. <br><br>Currently, this field is only applicable for the Germany site, and is returned only if the seller specifies the energy efficiency rating through item specifics at listing time. Rating values include <code>A+++</code>, <code>A++</code>, <code>A+</code>, <code>A</code>, <code>B</code>, <code>C</code>, <code>D</code>, <code>E</code>, <code>F</code>, and <code>G</code>.
 	EnergyEfficiencyClass *string `json:"energyEfficiencyClass,omitempty"`
 	// An ePID is the eBay product identifier of a product from the eBay product catalog.  This indicates the product in which the item belongs.
 	Epid *string `json:"epid,omitempty"`
 	Image *Image `json:"image,omitempty"`
-	// The URL to the View Item page of the item, which includes the affiliate tracking ID. This field is only returned if the seller enables affiliate tracking for the item by including the <code><a href=\"/api-docs/buy/static/api-browse.html#Headers\">X-EBAY-C-ENDUSERCTX</a></code> request header in the method.  <br /> <br /><span class=\"tablenote\"><b>Note: </b> eBay Partner Network, in order to receive a commission for your sales, you must use this URL to forward your buyer to the ebay.com site. </span>
+	// The URL to the View Item page of the item which includes the affiliate tracking ID.<br><br><span class=\"tablenote\"><b>Note:</b> In order to receive commissions on sales, eBay Partner Network affiliates must use this URL to forward buyers to the listing on the eBay marketplace.</span><br>The <code>itemAffiliateWebUrl</code> is returned only if:<ul><li>The marketplace through which the item is being viewed is part of the eBay Partner Network. Currently Singapore (<code>EBAY_SG</code>) is <b>not</b> supported.<br><br>For additional information, refer to <a href=\"https://partnerhelp.ebay.com/helpcenter/s/article/countries-available-as-a-program-in-EPN?language=en_US \" target=\"_blank\">eBay Partner Network</a>.</li><li>The seller enables affiliate tracking for the item by including the <code><a href=\"/api-docs/buy/static/api-browse.html#Headers\">X-EBAY-C-ENDUSERCTX</a></code> request header in the method.</li></ul>
 	ItemAffiliateWebUrl *string `json:"itemAffiliateWebUrl,omitempty"`
-	// The date and time up to which the item can be purchased.  This value is returned in UTC format (yyyy-MM-ddThh:mm:ss.sssZ), which you can convert into the local time of the buyer.<br /><br /><span class=\"tablenote\"><b> Note: </b>This field is not returned for Good 'Til Cancelled (GTC) listings.</span>
+	// The date and time when the item listing was created. This value is returned in UTC format (yyyy-MM-ddThh:mm:ss.sssZ), which you can convert into the local time of the buyer.
+	ItemCreationDate *string `json:"itemCreationDate,omitempty"`
+	// The date and time up to which the item can be purchased. This value is returned in UTC format (yyyy-MM-ddThh:mm:ss.sssZ), which you can convert into the local time of the buyer.<br><br><span class=\"tablenote\"><b>Note:</b> This field is not returned for Good 'Til Cancelled (GTC) listings.</span>
 	ItemEndDate *string `json:"itemEndDate,omitempty"`
-	// The HATEOAS reference of the parent page of the item group. An item group is an item that has various aspect differences, such as color, size, storage capacity, etc. <br /> <br /><span class=\"tablenote\"> <b>  Note: </b>This field is returned only for item groups.</span>
+	// The HATEOAS reference of the parent page of the item group. An item group is an item that has various aspect differences, such as color, size, storage capacity, etc.<br><br><span class=\"tablenote\"><b>Note:</b> This field is returned only for item groups.</span>
 	ItemGroupHref *string `json:"itemGroupHref,omitempty"`
-	// The indicates the item group type. An item group is an item that has various aspect differences, such as color, size, storage capacity, etc. <br /><br />Currently only the <code>SELLER_DEFINED_VARIATIONS</code> is supported and indicates this is an item group created by the seller. <br /> <br /><span class=\"tablenote\"> <b> Note: </b>This field is returned only for item groups.</span><br /><br />Code so that your app gracefully handles any future changes to this list.
+	// The indicates the item group type. An item group is an item that has various aspect differences, such as color, size, storage capacity, etc. <br><br>Currently only the <code>SELLER_DEFINED_VARIATIONS</code> is supported and indicates this is an item group created by the seller.<br><br><span class=\"tablenote\"><b>Note:</b> This field is returned only for item groups.</span>
 	ItemGroupType *string `json:"itemGroupType,omitempty"`
-	// The URI for the Browse API <a href=\"/api-docs/buy/browse/resources/item/methods/getItem\">getItem</a> method, which can be used to retrieve more details about items in the search results. 
+	// The URI for the Browse API <a href=\"/api-docs/buy/browse/resources/item/methods/getItem\" target=\"_blank\">getItem</a> method, which can be used to retrieve more details about items in the search results.
 	ItemHref *string `json:"itemHref,omitempty"`
 	// The unique RESTful identifier of the item.
 	ItemId *string `json:"itemId,omitempty"`
 	ItemLocation *ItemLocationImpl `json:"itemLocation,omitempty"`
-	// The URL to the View Item page of the item.  This enables you to include a \"Report Item on eBay\" hyperlink that takes the buyer to the View Item page on eBay. From there they can report any issues regarding this item to eBay.
+	// The URL to the View Item page of the item. This enables you to include a \"Report Item on eBay\" hyperlink that takes the buyer to the View Item page on eBay. From there they can report any issues regarding this item to eBay.
 	ItemWebUrl *string `json:"itemWebUrl,omitempty"`
+	// The leaf category IDs of the item. When the item belongs to two leaf categories, the ID values are returned in the order primary, secondary.
+	LeafCategoryIds []string `json:"leafCategoryIds,omitempty"`
 	// The unique identifier of the eBay listing that contains the item. This is the traditional/legacy ID that is often seen in the URL of the listing View Item page.
 	LegacyItemId *string `json:"legacyItemId,omitempty"`
+	// The ID of the eBay marketplace on which the seller listed the item. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/ba:MarketplaceIdEnum'>eBay API documentation</a>
+	ListingMarketplaceId *string `json:"listingMarketplaceId,omitempty"`
 	MarketingPrice *MarketingPrice `json:"marketingPrice,omitempty"`
-	// This container returns the local pickup options available to the buyer. This container is only returned if the user is searching for local pickup items and set the local pickup filters in the method request.
-	PickupOptions *[]PickupOptionSummary `json:"pickupOptions,omitempty"`
+	// This container returns the local pickup options available to the buyer. This container is returned only if the user is searching for local pickup items and set the local pickup filters in the method request.
+	PickupOptions []PickupOptionSummary `json:"pickupOptions,omitempty"`
 	Price *ConvertedAmount `json:"price,omitempty"`
 	// Indicates when in the buying flow the item's price can appear for minimum advertised price (MAP) items, which is the lowest price a retailer can advertise/show for this item. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:PriceDisplayConditionEnum'>eBay API documentation</a>
 	PriceDisplayCondition *string `json:"priceDisplayCondition,omitempty"`
-	// This field is returned as <code>true</code> if the listing is part of a Promoted Listing campaign. Promoted Listings are available to Above Standard and Top Rated sellers with recent sales activity.<br /><br /><span class=\"tablenote\"><b>Note:</b> Priority Listing is returned only with a Best Match sort and will not be returned for other sort options.</span>
+	// This field is returned as <code>true</code> if the listing is part of a Promoted Listing campaign. Promoted Listings are available to <b>Above Standard</b> and <b>Top Rated</b> sellers with recent sales activity.<br><br><span class=\"tablenote\"><b>Note:</b> Priority Listing is returned only with a Best Match sort and will not be returned for other sort options.</span>
 	PriorityListing *bool `json:"priorityListing,omitempty"`
-	// An array of the qualified programs available for the item, such as EBAY_PLUS, AUTHENTICITY_GUARANTEE, and AUTHENTICITY_VERIFICATION.<br /><br />eBay Plus is a premium account option for buyers, which provides benefits such as fast free domestic shipping and free returns on selected items. Top-Rated eBay sellers must opt in to eBay Plus to be able to offer the program on qualifying listings. Sellers must commit to next-day delivery of those items.<br /><br /><span class=\"tablenote\"><b>Note: </b> eBay Plus is available only to buyers in Germany, Austria, and Australia marketplaces.</span><br /><br />The eBay <a href=\"https://pages.ebay.com/authenticity-guarantee/\" target=\"_blank\">Authenticity Guarantee</a> program enables third-party authenticators to perform authentication verification inspections on items such as watches and sneakers.
-	QualifiedPrograms *[]string `json:"qualifiedPrograms,omitempty"`
+	// An array of the qualified programs available for the item, such as <code>EBAY_PLUS</code>, <code>AUTHENTICITY_GUARANTEE</code>, and <code>AUTHENTICITY_VERIFICATION</code>.<br><br>eBay Plus is a premium account option for buyers, which provides benefits such as fast, free domestic shipping and free returns on selected items. Top-Rated eBay sellers must opt in to eBay Plus to be able to offer the program on qualifying listings. Sellers must commit to next-day delivery of those items.<br><br><span class=\"tablenote\"><b>Note: </b> eBay Plus is available only to buyers in the Germany, Austria, and Australia marketplaces.</span><br><br>The eBay <a href=\"https://pages.ebay.com/authenticity-guarantee/ \" target=\"_blank\">Authenticity Guarantee</a> program enables third-party authenticators to perform authentication verification inspections on items such as watches and sneakers.
+	QualifiedPrograms []string `json:"qualifiedPrograms,omitempty"`
 	Seller *Seller `json:"seller,omitempty"`
 	// This container returns the shipping options available to ship the item.
-	ShippingOptions *[]ShippingOptionSummary `json:"shippingOptions,omitempty"`
-	// This text string is derived from the item condition and the item aspects (such as size, color, capacity, model, brand, etc.). Sometimes the title doesn't give enough information but the description is too big. Surfacing the <b>shortDescription</b> can often provide buyers with the additional information that could help them make a buying decision.  <br /><br />For example: <br /><br />    <code>   \"<b> title</b>\": \"Petrel U42W FPV Drone RC Quadcopter w/HD Camera Live Video One Key Off / Landing\", <br />\"<b>shortDescription</b>\": \"1 U42W Quadcopter. Syma X5SW-V3 Wifi FPV RC Drone Quadcopter 2.4Ghz 6-Axis Gyro with Headless Mode. Syma X20 Pocket Drone 2.4Ghz Mini RC Quadcopter Headless Mode Altitude Hold. One Key Take Off / Landing function: allow beginner to easy to fly the drone without any skill.\",</code>       <br /><br /><b>Restriction: </b> This field is returned by the <b> search</b> method only when <b> fieldgroups</b> = <code>EXTENDED</code>.
+	ShippingOptions []ShippingOptionSummary `json:"shippingOptions,omitempty"`
+	// This text string is derived from the item condition and the item aspects (such as size, color, capacity, model, brand, etc.) Sometimes the title does not provide enough information but the description is too big. Surfacing the <code>shortDescription</code> can often provide buyers with the additional information that could help them make a buying decision.<br><br>For example:<pre>\"<b>title</b>\": \"Petrel U42W FPV Drone RC Quadcopter w/HD Camera Live Video One Key Off / Landing\",<br>\"<b>shortDescription</b>\": \"1 U42W Quadcopter. Syma X5SW-V3 Wifi FPV RC Drone Quadcopter 2.4Ghz 6-Axis Gyro with Headless Mode. Syma X20 Pocket Drone 2.4Ghz Mini RC Quadcopter Headless Mode Altitude Hold. One Key Take Off / Landing function: allow beginner to easy to fly the drone without any skill.\",</pre><br><b>Restriction:</b> This field is returned by the <b>search</b> method only when <code>fieldgroups</code> = <code>EXTENDED</code>.
 	ShortDescription *string `json:"shortDescription,omitempty"`
-	// An array of thumbnail images for the item. 
-	ThumbnailImages *[]Image `json:"thumbnailImages,omitempty"`
-	// The seller-created title of the item. <br><br><b>Maximum Length: </b> 80 characters
+	// An array of thumbnail images for the item.
+	ThumbnailImages []Image `json:"thumbnailImages,omitempty"`
+	// The seller-created title of the item.<br><br><b>Maximum Length:</b> 80 characters
 	Title *string `json:"title,omitempty"`
-	// This indicates if the item is a top-rated plus item. There are three benefits of a top-rated plus item: a  minimum 30-day money-back return policy, shipping the item in 1 business day with tracking provided, and the added comfort of knowing that this item is from an experienced seller with the highest buyer ratings. See the <a href=\"https://pages.ebay.com/topratedplus/index.html\" target=\"_blank\">Top Rated Plus Items </a> and  <a href=\"https://pages.ebay.com/help/sell/top-rated.html\" target=\"_blank\">Becoming a Top Rated Seller and qualifying for Top Rated Plus</a> help topics for more information.
+	// This indicates if the item is a top-rated plus item. There are three benefits of a top-rated plus item: a  minimum 30-day money-back return policy; shipping the item in 1 business day with tracking provided; and the added comfort of knowing that this item is from an experienced seller with the highest buyer ratings. For more information, refer to <a href=\"https://pages.ebay.com/topratedplus/index.html \" target=\"_blank\">Look for Top Rated Plus Items</a> and <a href=\"https://www.ebay.com/help/selling/seller-levels-performance-standards/seller-levels-performance-standards?id=4080 \" target=\"_blank\">Seller performance overview</a>.
 	TopRatedBuyingExperience *bool `json:"topRatedBuyingExperience,omitempty"`
 	// The URL to the image that shows the information on the tyre label.
 	TyreLabelImageUrl *string `json:"tyreLabelImageUrl,omitempty"`
 	UnitPrice *ConvertedAmount `json:"unitPrice,omitempty"`
-	// The designation, such as size, weight, volume, count, etc., that was used to specify the quantity of the item. This helps buyers compare prices. <br /><br />For example, the following tells the buyer that the item is 7.99 per 100 grams. <br /><br /><code>\"unitPricingMeasure\": \"100g\",<br /> \"unitPrice\": {<br />&nbsp;&nbsp;\"value\": \"7.99\",<br />&nbsp;&nbsp;\"currency\": \"GBP\"</code>
+	// The designation, such as size, weight, volume, count, etc., that was used to specify the quantity of the item. This helps buyers compare prices.<br><br>For example, the following tells the buyer that the item is 7.99 per 100 grams.<pre>\"unitPricingMeasure\": \"100g\",<br> \"unitPrice\": {<br>&nbsp;&nbsp;\"value\": \"7.99\",<br>&nbsp;&nbsp;\"currency\": \"GBP\"</pre>
 	UnitPricingMeasure *string `json:"unitPricingMeasure,omitempty"`
-	// The number of users that have added the item to their watch list.<br /><br /><span class=\"tablenote\"> <strong>Note:</strong> This field is restricted to applications that have been granted permission to access this feature. You must submit an <a href=\"https://developer.ebay.com/my/support/tickets?tab=app-check\">App Check ticket</a> to request this access. In the App Check form, add a note to the <b>Application Title/Summary</b> and/or <b>Application Details</b> fields that you want access to Watch Count data in the Browse API.</span>
+	// The number of users that have added the item to their watch list.<br><br><span class=\"tablenote\"><b>Note:</b> This field is restricted to applications that have been granted permission to access this feature. You must submit an <a href=\"/my/support/tickets?tab=app-check \" target=\"_blank\">App Check ticket</a> to request this access. In the App Check form, add a note to the <b>Application Title/Summary</b> and/or <b>Application Details</b> fields indicating that you want access to Watch Count data in the Browse API.</span>
 	WatchCount *int32 `json:"watchCount,omitempty"`
 }
 
@@ -109,17 +118,17 @@ func NewItemSummaryWithDefaults() *ItemSummary {
 
 // GetAdditionalImages returns the AdditionalImages field value if set, zero value otherwise.
 func (o *ItemSummary) GetAdditionalImages() []Image {
-	if o == nil || o.AdditionalImages == nil {
+	if o == nil || IsNil(o.AdditionalImages) {
 		var ret []Image
 		return ret
 	}
-	return *o.AdditionalImages
+	return o.AdditionalImages
 }
 
 // GetAdditionalImagesOk returns a tuple with the AdditionalImages field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *ItemSummary) GetAdditionalImagesOk() (*[]Image, bool) {
-	if o == nil || o.AdditionalImages == nil {
+func (o *ItemSummary) GetAdditionalImagesOk() ([]Image, bool) {
+	if o == nil || IsNil(o.AdditionalImages) {
 		return nil, false
 	}
 	return o.AdditionalImages, true
@@ -127,7 +136,7 @@ func (o *ItemSummary) GetAdditionalImagesOk() (*[]Image, bool) {
 
 // HasAdditionalImages returns a boolean if a field has been set.
 func (o *ItemSummary) HasAdditionalImages() bool {
-	if o != nil && o.AdditionalImages != nil {
+	if o != nil && !IsNil(o.AdditionalImages) {
 		return true
 	}
 
@@ -136,12 +145,12 @@ func (o *ItemSummary) HasAdditionalImages() bool {
 
 // SetAdditionalImages gets a reference to the given []Image and assigns it to the AdditionalImages field.
 func (o *ItemSummary) SetAdditionalImages(v []Image) {
-	o.AdditionalImages = &v
+	o.AdditionalImages = v
 }
 
 // GetAdultOnly returns the AdultOnly field value if set, zero value otherwise.
 func (o *ItemSummary) GetAdultOnly() bool {
-	if o == nil || o.AdultOnly == nil {
+	if o == nil || IsNil(o.AdultOnly) {
 		var ret bool
 		return ret
 	}
@@ -151,7 +160,7 @@ func (o *ItemSummary) GetAdultOnly() bool {
 // GetAdultOnlyOk returns a tuple with the AdultOnly field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetAdultOnlyOk() (*bool, bool) {
-	if o == nil || o.AdultOnly == nil {
+	if o == nil || IsNil(o.AdultOnly) {
 		return nil, false
 	}
 	return o.AdultOnly, true
@@ -159,7 +168,7 @@ func (o *ItemSummary) GetAdultOnlyOk() (*bool, bool) {
 
 // HasAdultOnly returns a boolean if a field has been set.
 func (o *ItemSummary) HasAdultOnly() bool {
-	if o != nil && o.AdultOnly != nil {
+	if o != nil && !IsNil(o.AdultOnly) {
 		return true
 	}
 
@@ -173,7 +182,7 @@ func (o *ItemSummary) SetAdultOnly(v bool) {
 
 // GetAvailableCoupons returns the AvailableCoupons field value if set, zero value otherwise.
 func (o *ItemSummary) GetAvailableCoupons() bool {
-	if o == nil || o.AvailableCoupons == nil {
+	if o == nil || IsNil(o.AvailableCoupons) {
 		var ret bool
 		return ret
 	}
@@ -183,7 +192,7 @@ func (o *ItemSummary) GetAvailableCoupons() bool {
 // GetAvailableCouponsOk returns a tuple with the AvailableCoupons field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetAvailableCouponsOk() (*bool, bool) {
-	if o == nil || o.AvailableCoupons == nil {
+	if o == nil || IsNil(o.AvailableCoupons) {
 		return nil, false
 	}
 	return o.AvailableCoupons, true
@@ -191,7 +200,7 @@ func (o *ItemSummary) GetAvailableCouponsOk() (*bool, bool) {
 
 // HasAvailableCoupons returns a boolean if a field has been set.
 func (o *ItemSummary) HasAvailableCoupons() bool {
-	if o != nil && o.AvailableCoupons != nil {
+	if o != nil && !IsNil(o.AvailableCoupons) {
 		return true
 	}
 
@@ -205,7 +214,7 @@ func (o *ItemSummary) SetAvailableCoupons(v bool) {
 
 // GetBidCount returns the BidCount field value if set, zero value otherwise.
 func (o *ItemSummary) GetBidCount() int32 {
-	if o == nil || o.BidCount == nil {
+	if o == nil || IsNil(o.BidCount) {
 		var ret int32
 		return ret
 	}
@@ -215,7 +224,7 @@ func (o *ItemSummary) GetBidCount() int32 {
 // GetBidCountOk returns a tuple with the BidCount field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetBidCountOk() (*int32, bool) {
-	if o == nil || o.BidCount == nil {
+	if o == nil || IsNil(o.BidCount) {
 		return nil, false
 	}
 	return o.BidCount, true
@@ -223,7 +232,7 @@ func (o *ItemSummary) GetBidCountOk() (*int32, bool) {
 
 // HasBidCount returns a boolean if a field has been set.
 func (o *ItemSummary) HasBidCount() bool {
-	if o != nil && o.BidCount != nil {
+	if o != nil && !IsNil(o.BidCount) {
 		return true
 	}
 
@@ -237,17 +246,17 @@ func (o *ItemSummary) SetBidCount(v int32) {
 
 // GetBuyingOptions returns the BuyingOptions field value if set, zero value otherwise.
 func (o *ItemSummary) GetBuyingOptions() []string {
-	if o == nil || o.BuyingOptions == nil {
+	if o == nil || IsNil(o.BuyingOptions) {
 		var ret []string
 		return ret
 	}
-	return *o.BuyingOptions
+	return o.BuyingOptions
 }
 
 // GetBuyingOptionsOk returns a tuple with the BuyingOptions field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *ItemSummary) GetBuyingOptionsOk() (*[]string, bool) {
-	if o == nil || o.BuyingOptions == nil {
+func (o *ItemSummary) GetBuyingOptionsOk() ([]string, bool) {
+	if o == nil || IsNil(o.BuyingOptions) {
 		return nil, false
 	}
 	return o.BuyingOptions, true
@@ -255,7 +264,7 @@ func (o *ItemSummary) GetBuyingOptionsOk() (*[]string, bool) {
 
 // HasBuyingOptions returns a boolean if a field has been set.
 func (o *ItemSummary) HasBuyingOptions() bool {
-	if o != nil && o.BuyingOptions != nil {
+	if o != nil && !IsNil(o.BuyingOptions) {
 		return true
 	}
 
@@ -264,22 +273,22 @@ func (o *ItemSummary) HasBuyingOptions() bool {
 
 // SetBuyingOptions gets a reference to the given []string and assigns it to the BuyingOptions field.
 func (o *ItemSummary) SetBuyingOptions(v []string) {
-	o.BuyingOptions = &v
+	o.BuyingOptions = v
 }
 
 // GetCategories returns the Categories field value if set, zero value otherwise.
 func (o *ItemSummary) GetCategories() []Category {
-	if o == nil || o.Categories == nil {
+	if o == nil || IsNil(o.Categories) {
 		var ret []Category
 		return ret
 	}
-	return *o.Categories
+	return o.Categories
 }
 
 // GetCategoriesOk returns a tuple with the Categories field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *ItemSummary) GetCategoriesOk() (*[]Category, bool) {
-	if o == nil || o.Categories == nil {
+func (o *ItemSummary) GetCategoriesOk() ([]Category, bool) {
+	if o == nil || IsNil(o.Categories) {
 		return nil, false
 	}
 	return o.Categories, true
@@ -287,7 +296,7 @@ func (o *ItemSummary) GetCategoriesOk() (*[]Category, bool) {
 
 // HasCategories returns a boolean if a field has been set.
 func (o *ItemSummary) HasCategories() bool {
-	if o != nil && o.Categories != nil {
+	if o != nil && !IsNil(o.Categories) {
 		return true
 	}
 
@@ -296,12 +305,12 @@ func (o *ItemSummary) HasCategories() bool {
 
 // SetCategories gets a reference to the given []Category and assigns it to the Categories field.
 func (o *ItemSummary) SetCategories(v []Category) {
-	o.Categories = &v
+	o.Categories = v
 }
 
 // GetCompatibilityMatch returns the CompatibilityMatch field value if set, zero value otherwise.
 func (o *ItemSummary) GetCompatibilityMatch() string {
-	if o == nil || o.CompatibilityMatch == nil {
+	if o == nil || IsNil(o.CompatibilityMatch) {
 		var ret string
 		return ret
 	}
@@ -311,7 +320,7 @@ func (o *ItemSummary) GetCompatibilityMatch() string {
 // GetCompatibilityMatchOk returns a tuple with the CompatibilityMatch field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetCompatibilityMatchOk() (*string, bool) {
-	if o == nil || o.CompatibilityMatch == nil {
+	if o == nil || IsNil(o.CompatibilityMatch) {
 		return nil, false
 	}
 	return o.CompatibilityMatch, true
@@ -319,7 +328,7 @@ func (o *ItemSummary) GetCompatibilityMatchOk() (*string, bool) {
 
 // HasCompatibilityMatch returns a boolean if a field has been set.
 func (o *ItemSummary) HasCompatibilityMatch() bool {
-	if o != nil && o.CompatibilityMatch != nil {
+	if o != nil && !IsNil(o.CompatibilityMatch) {
 		return true
 	}
 
@@ -333,17 +342,17 @@ func (o *ItemSummary) SetCompatibilityMatch(v string) {
 
 // GetCompatibilityProperties returns the CompatibilityProperties field value if set, zero value otherwise.
 func (o *ItemSummary) GetCompatibilityProperties() []CompatibilityProperty {
-	if o == nil || o.CompatibilityProperties == nil {
+	if o == nil || IsNil(o.CompatibilityProperties) {
 		var ret []CompatibilityProperty
 		return ret
 	}
-	return *o.CompatibilityProperties
+	return o.CompatibilityProperties
 }
 
 // GetCompatibilityPropertiesOk returns a tuple with the CompatibilityProperties field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *ItemSummary) GetCompatibilityPropertiesOk() (*[]CompatibilityProperty, bool) {
-	if o == nil || o.CompatibilityProperties == nil {
+func (o *ItemSummary) GetCompatibilityPropertiesOk() ([]CompatibilityProperty, bool) {
+	if o == nil || IsNil(o.CompatibilityProperties) {
 		return nil, false
 	}
 	return o.CompatibilityProperties, true
@@ -351,7 +360,7 @@ func (o *ItemSummary) GetCompatibilityPropertiesOk() (*[]CompatibilityProperty, 
 
 // HasCompatibilityProperties returns a boolean if a field has been set.
 func (o *ItemSummary) HasCompatibilityProperties() bool {
-	if o != nil && o.CompatibilityProperties != nil {
+	if o != nil && !IsNil(o.CompatibilityProperties) {
 		return true
 	}
 
@@ -360,12 +369,12 @@ func (o *ItemSummary) HasCompatibilityProperties() bool {
 
 // SetCompatibilityProperties gets a reference to the given []CompatibilityProperty and assigns it to the CompatibilityProperties field.
 func (o *ItemSummary) SetCompatibilityProperties(v []CompatibilityProperty) {
-	o.CompatibilityProperties = &v
+	o.CompatibilityProperties = v
 }
 
 // GetCondition returns the Condition field value if set, zero value otherwise.
 func (o *ItemSummary) GetCondition() string {
-	if o == nil || o.Condition == nil {
+	if o == nil || IsNil(o.Condition) {
 		var ret string
 		return ret
 	}
@@ -375,7 +384,7 @@ func (o *ItemSummary) GetCondition() string {
 // GetConditionOk returns a tuple with the Condition field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetConditionOk() (*string, bool) {
-	if o == nil || o.Condition == nil {
+	if o == nil || IsNil(o.Condition) {
 		return nil, false
 	}
 	return o.Condition, true
@@ -383,7 +392,7 @@ func (o *ItemSummary) GetConditionOk() (*string, bool) {
 
 // HasCondition returns a boolean if a field has been set.
 func (o *ItemSummary) HasCondition() bool {
-	if o != nil && o.Condition != nil {
+	if o != nil && !IsNil(o.Condition) {
 		return true
 	}
 
@@ -397,7 +406,7 @@ func (o *ItemSummary) SetCondition(v string) {
 
 // GetConditionId returns the ConditionId field value if set, zero value otherwise.
 func (o *ItemSummary) GetConditionId() string {
-	if o == nil || o.ConditionId == nil {
+	if o == nil || IsNil(o.ConditionId) {
 		var ret string
 		return ret
 	}
@@ -407,7 +416,7 @@ func (o *ItemSummary) GetConditionId() string {
 // GetConditionIdOk returns a tuple with the ConditionId field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetConditionIdOk() (*string, bool) {
-	if o == nil || o.ConditionId == nil {
+	if o == nil || IsNil(o.ConditionId) {
 		return nil, false
 	}
 	return o.ConditionId, true
@@ -415,7 +424,7 @@ func (o *ItemSummary) GetConditionIdOk() (*string, bool) {
 
 // HasConditionId returns a boolean if a field has been set.
 func (o *ItemSummary) HasConditionId() bool {
-	if o != nil && o.ConditionId != nil {
+	if o != nil && !IsNil(o.ConditionId) {
 		return true
 	}
 
@@ -429,7 +438,7 @@ func (o *ItemSummary) SetConditionId(v string) {
 
 // GetCurrentBidPrice returns the CurrentBidPrice field value if set, zero value otherwise.
 func (o *ItemSummary) GetCurrentBidPrice() ConvertedAmount {
-	if o == nil || o.CurrentBidPrice == nil {
+	if o == nil || IsNil(o.CurrentBidPrice) {
 		var ret ConvertedAmount
 		return ret
 	}
@@ -439,7 +448,7 @@ func (o *ItemSummary) GetCurrentBidPrice() ConvertedAmount {
 // GetCurrentBidPriceOk returns a tuple with the CurrentBidPrice field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetCurrentBidPriceOk() (*ConvertedAmount, bool) {
-	if o == nil || o.CurrentBidPrice == nil {
+	if o == nil || IsNil(o.CurrentBidPrice) {
 		return nil, false
 	}
 	return o.CurrentBidPrice, true
@@ -447,7 +456,7 @@ func (o *ItemSummary) GetCurrentBidPriceOk() (*ConvertedAmount, bool) {
 
 // HasCurrentBidPrice returns a boolean if a field has been set.
 func (o *ItemSummary) HasCurrentBidPrice() bool {
-	if o != nil && o.CurrentBidPrice != nil {
+	if o != nil && !IsNil(o.CurrentBidPrice) {
 		return true
 	}
 
@@ -461,7 +470,7 @@ func (o *ItemSummary) SetCurrentBidPrice(v ConvertedAmount) {
 
 // GetDistanceFromPickupLocation returns the DistanceFromPickupLocation field value if set, zero value otherwise.
 func (o *ItemSummary) GetDistanceFromPickupLocation() TargetLocation {
-	if o == nil || o.DistanceFromPickupLocation == nil {
+	if o == nil || IsNil(o.DistanceFromPickupLocation) {
 		var ret TargetLocation
 		return ret
 	}
@@ -471,7 +480,7 @@ func (o *ItemSummary) GetDistanceFromPickupLocation() TargetLocation {
 // GetDistanceFromPickupLocationOk returns a tuple with the DistanceFromPickupLocation field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetDistanceFromPickupLocationOk() (*TargetLocation, bool) {
-	if o == nil || o.DistanceFromPickupLocation == nil {
+	if o == nil || IsNil(o.DistanceFromPickupLocation) {
 		return nil, false
 	}
 	return o.DistanceFromPickupLocation, true
@@ -479,7 +488,7 @@ func (o *ItemSummary) GetDistanceFromPickupLocationOk() (*TargetLocation, bool) 
 
 // HasDistanceFromPickupLocation returns a boolean if a field has been set.
 func (o *ItemSummary) HasDistanceFromPickupLocation() bool {
-	if o != nil && o.DistanceFromPickupLocation != nil {
+	if o != nil && !IsNil(o.DistanceFromPickupLocation) {
 		return true
 	}
 
@@ -493,7 +502,7 @@ func (o *ItemSummary) SetDistanceFromPickupLocation(v TargetLocation) {
 
 // GetEnergyEfficiencyClass returns the EnergyEfficiencyClass field value if set, zero value otherwise.
 func (o *ItemSummary) GetEnergyEfficiencyClass() string {
-	if o == nil || o.EnergyEfficiencyClass == nil {
+	if o == nil || IsNil(o.EnergyEfficiencyClass) {
 		var ret string
 		return ret
 	}
@@ -503,7 +512,7 @@ func (o *ItemSummary) GetEnergyEfficiencyClass() string {
 // GetEnergyEfficiencyClassOk returns a tuple with the EnergyEfficiencyClass field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetEnergyEfficiencyClassOk() (*string, bool) {
-	if o == nil || o.EnergyEfficiencyClass == nil {
+	if o == nil || IsNil(o.EnergyEfficiencyClass) {
 		return nil, false
 	}
 	return o.EnergyEfficiencyClass, true
@@ -511,7 +520,7 @@ func (o *ItemSummary) GetEnergyEfficiencyClassOk() (*string, bool) {
 
 // HasEnergyEfficiencyClass returns a boolean if a field has been set.
 func (o *ItemSummary) HasEnergyEfficiencyClass() bool {
-	if o != nil && o.EnergyEfficiencyClass != nil {
+	if o != nil && !IsNil(o.EnergyEfficiencyClass) {
 		return true
 	}
 
@@ -525,7 +534,7 @@ func (o *ItemSummary) SetEnergyEfficiencyClass(v string) {
 
 // GetEpid returns the Epid field value if set, zero value otherwise.
 func (o *ItemSummary) GetEpid() string {
-	if o == nil || o.Epid == nil {
+	if o == nil || IsNil(o.Epid) {
 		var ret string
 		return ret
 	}
@@ -535,7 +544,7 @@ func (o *ItemSummary) GetEpid() string {
 // GetEpidOk returns a tuple with the Epid field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetEpidOk() (*string, bool) {
-	if o == nil || o.Epid == nil {
+	if o == nil || IsNil(o.Epid) {
 		return nil, false
 	}
 	return o.Epid, true
@@ -543,7 +552,7 @@ func (o *ItemSummary) GetEpidOk() (*string, bool) {
 
 // HasEpid returns a boolean if a field has been set.
 func (o *ItemSummary) HasEpid() bool {
-	if o != nil && o.Epid != nil {
+	if o != nil && !IsNil(o.Epid) {
 		return true
 	}
 
@@ -557,7 +566,7 @@ func (o *ItemSummary) SetEpid(v string) {
 
 // GetImage returns the Image field value if set, zero value otherwise.
 func (o *ItemSummary) GetImage() Image {
-	if o == nil || o.Image == nil {
+	if o == nil || IsNil(o.Image) {
 		var ret Image
 		return ret
 	}
@@ -567,7 +576,7 @@ func (o *ItemSummary) GetImage() Image {
 // GetImageOk returns a tuple with the Image field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetImageOk() (*Image, bool) {
-	if o == nil || o.Image == nil {
+	if o == nil || IsNil(o.Image) {
 		return nil, false
 	}
 	return o.Image, true
@@ -575,7 +584,7 @@ func (o *ItemSummary) GetImageOk() (*Image, bool) {
 
 // HasImage returns a boolean if a field has been set.
 func (o *ItemSummary) HasImage() bool {
-	if o != nil && o.Image != nil {
+	if o != nil && !IsNil(o.Image) {
 		return true
 	}
 
@@ -589,7 +598,7 @@ func (o *ItemSummary) SetImage(v Image) {
 
 // GetItemAffiliateWebUrl returns the ItemAffiliateWebUrl field value if set, zero value otherwise.
 func (o *ItemSummary) GetItemAffiliateWebUrl() string {
-	if o == nil || o.ItemAffiliateWebUrl == nil {
+	if o == nil || IsNil(o.ItemAffiliateWebUrl) {
 		var ret string
 		return ret
 	}
@@ -599,7 +608,7 @@ func (o *ItemSummary) GetItemAffiliateWebUrl() string {
 // GetItemAffiliateWebUrlOk returns a tuple with the ItemAffiliateWebUrl field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetItemAffiliateWebUrlOk() (*string, bool) {
-	if o == nil || o.ItemAffiliateWebUrl == nil {
+	if o == nil || IsNil(o.ItemAffiliateWebUrl) {
 		return nil, false
 	}
 	return o.ItemAffiliateWebUrl, true
@@ -607,7 +616,7 @@ func (o *ItemSummary) GetItemAffiliateWebUrlOk() (*string, bool) {
 
 // HasItemAffiliateWebUrl returns a boolean if a field has been set.
 func (o *ItemSummary) HasItemAffiliateWebUrl() bool {
-	if o != nil && o.ItemAffiliateWebUrl != nil {
+	if o != nil && !IsNil(o.ItemAffiliateWebUrl) {
 		return true
 	}
 
@@ -619,9 +628,41 @@ func (o *ItemSummary) SetItemAffiliateWebUrl(v string) {
 	o.ItemAffiliateWebUrl = &v
 }
 
+// GetItemCreationDate returns the ItemCreationDate field value if set, zero value otherwise.
+func (o *ItemSummary) GetItemCreationDate() string {
+	if o == nil || IsNil(o.ItemCreationDate) {
+		var ret string
+		return ret
+	}
+	return *o.ItemCreationDate
+}
+
+// GetItemCreationDateOk returns a tuple with the ItemCreationDate field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *ItemSummary) GetItemCreationDateOk() (*string, bool) {
+	if o == nil || IsNil(o.ItemCreationDate) {
+		return nil, false
+	}
+	return o.ItemCreationDate, true
+}
+
+// HasItemCreationDate returns a boolean if a field has been set.
+func (o *ItemSummary) HasItemCreationDate() bool {
+	if o != nil && !IsNil(o.ItemCreationDate) {
+		return true
+	}
+
+	return false
+}
+
+// SetItemCreationDate gets a reference to the given string and assigns it to the ItemCreationDate field.
+func (o *ItemSummary) SetItemCreationDate(v string) {
+	o.ItemCreationDate = &v
+}
+
 // GetItemEndDate returns the ItemEndDate field value if set, zero value otherwise.
 func (o *ItemSummary) GetItemEndDate() string {
-	if o == nil || o.ItemEndDate == nil {
+	if o == nil || IsNil(o.ItemEndDate) {
 		var ret string
 		return ret
 	}
@@ -631,7 +672,7 @@ func (o *ItemSummary) GetItemEndDate() string {
 // GetItemEndDateOk returns a tuple with the ItemEndDate field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetItemEndDateOk() (*string, bool) {
-	if o == nil || o.ItemEndDate == nil {
+	if o == nil || IsNil(o.ItemEndDate) {
 		return nil, false
 	}
 	return o.ItemEndDate, true
@@ -639,7 +680,7 @@ func (o *ItemSummary) GetItemEndDateOk() (*string, bool) {
 
 // HasItemEndDate returns a boolean if a field has been set.
 func (o *ItemSummary) HasItemEndDate() bool {
-	if o != nil && o.ItemEndDate != nil {
+	if o != nil && !IsNil(o.ItemEndDate) {
 		return true
 	}
 
@@ -653,7 +694,7 @@ func (o *ItemSummary) SetItemEndDate(v string) {
 
 // GetItemGroupHref returns the ItemGroupHref field value if set, zero value otherwise.
 func (o *ItemSummary) GetItemGroupHref() string {
-	if o == nil || o.ItemGroupHref == nil {
+	if o == nil || IsNil(o.ItemGroupHref) {
 		var ret string
 		return ret
 	}
@@ -663,7 +704,7 @@ func (o *ItemSummary) GetItemGroupHref() string {
 // GetItemGroupHrefOk returns a tuple with the ItemGroupHref field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetItemGroupHrefOk() (*string, bool) {
-	if o == nil || o.ItemGroupHref == nil {
+	if o == nil || IsNil(o.ItemGroupHref) {
 		return nil, false
 	}
 	return o.ItemGroupHref, true
@@ -671,7 +712,7 @@ func (o *ItemSummary) GetItemGroupHrefOk() (*string, bool) {
 
 // HasItemGroupHref returns a boolean if a field has been set.
 func (o *ItemSummary) HasItemGroupHref() bool {
-	if o != nil && o.ItemGroupHref != nil {
+	if o != nil && !IsNil(o.ItemGroupHref) {
 		return true
 	}
 
@@ -685,7 +726,7 @@ func (o *ItemSummary) SetItemGroupHref(v string) {
 
 // GetItemGroupType returns the ItemGroupType field value if set, zero value otherwise.
 func (o *ItemSummary) GetItemGroupType() string {
-	if o == nil || o.ItemGroupType == nil {
+	if o == nil || IsNil(o.ItemGroupType) {
 		var ret string
 		return ret
 	}
@@ -695,7 +736,7 @@ func (o *ItemSummary) GetItemGroupType() string {
 // GetItemGroupTypeOk returns a tuple with the ItemGroupType field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetItemGroupTypeOk() (*string, bool) {
-	if o == nil || o.ItemGroupType == nil {
+	if o == nil || IsNil(o.ItemGroupType) {
 		return nil, false
 	}
 	return o.ItemGroupType, true
@@ -703,7 +744,7 @@ func (o *ItemSummary) GetItemGroupTypeOk() (*string, bool) {
 
 // HasItemGroupType returns a boolean if a field has been set.
 func (o *ItemSummary) HasItemGroupType() bool {
-	if o != nil && o.ItemGroupType != nil {
+	if o != nil && !IsNil(o.ItemGroupType) {
 		return true
 	}
 
@@ -717,7 +758,7 @@ func (o *ItemSummary) SetItemGroupType(v string) {
 
 // GetItemHref returns the ItemHref field value if set, zero value otherwise.
 func (o *ItemSummary) GetItemHref() string {
-	if o == nil || o.ItemHref == nil {
+	if o == nil || IsNil(o.ItemHref) {
 		var ret string
 		return ret
 	}
@@ -727,7 +768,7 @@ func (o *ItemSummary) GetItemHref() string {
 // GetItemHrefOk returns a tuple with the ItemHref field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetItemHrefOk() (*string, bool) {
-	if o == nil || o.ItemHref == nil {
+	if o == nil || IsNil(o.ItemHref) {
 		return nil, false
 	}
 	return o.ItemHref, true
@@ -735,7 +776,7 @@ func (o *ItemSummary) GetItemHrefOk() (*string, bool) {
 
 // HasItemHref returns a boolean if a field has been set.
 func (o *ItemSummary) HasItemHref() bool {
-	if o != nil && o.ItemHref != nil {
+	if o != nil && !IsNil(o.ItemHref) {
 		return true
 	}
 
@@ -749,7 +790,7 @@ func (o *ItemSummary) SetItemHref(v string) {
 
 // GetItemId returns the ItemId field value if set, zero value otherwise.
 func (o *ItemSummary) GetItemId() string {
-	if o == nil || o.ItemId == nil {
+	if o == nil || IsNil(o.ItemId) {
 		var ret string
 		return ret
 	}
@@ -759,7 +800,7 @@ func (o *ItemSummary) GetItemId() string {
 // GetItemIdOk returns a tuple with the ItemId field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetItemIdOk() (*string, bool) {
-	if o == nil || o.ItemId == nil {
+	if o == nil || IsNil(o.ItemId) {
 		return nil, false
 	}
 	return o.ItemId, true
@@ -767,7 +808,7 @@ func (o *ItemSummary) GetItemIdOk() (*string, bool) {
 
 // HasItemId returns a boolean if a field has been set.
 func (o *ItemSummary) HasItemId() bool {
-	if o != nil && o.ItemId != nil {
+	if o != nil && !IsNil(o.ItemId) {
 		return true
 	}
 
@@ -781,7 +822,7 @@ func (o *ItemSummary) SetItemId(v string) {
 
 // GetItemLocation returns the ItemLocation field value if set, zero value otherwise.
 func (o *ItemSummary) GetItemLocation() ItemLocationImpl {
-	if o == nil || o.ItemLocation == nil {
+	if o == nil || IsNil(o.ItemLocation) {
 		var ret ItemLocationImpl
 		return ret
 	}
@@ -791,7 +832,7 @@ func (o *ItemSummary) GetItemLocation() ItemLocationImpl {
 // GetItemLocationOk returns a tuple with the ItemLocation field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetItemLocationOk() (*ItemLocationImpl, bool) {
-	if o == nil || o.ItemLocation == nil {
+	if o == nil || IsNil(o.ItemLocation) {
 		return nil, false
 	}
 	return o.ItemLocation, true
@@ -799,7 +840,7 @@ func (o *ItemSummary) GetItemLocationOk() (*ItemLocationImpl, bool) {
 
 // HasItemLocation returns a boolean if a field has been set.
 func (o *ItemSummary) HasItemLocation() bool {
-	if o != nil && o.ItemLocation != nil {
+	if o != nil && !IsNil(o.ItemLocation) {
 		return true
 	}
 
@@ -813,7 +854,7 @@ func (o *ItemSummary) SetItemLocation(v ItemLocationImpl) {
 
 // GetItemWebUrl returns the ItemWebUrl field value if set, zero value otherwise.
 func (o *ItemSummary) GetItemWebUrl() string {
-	if o == nil || o.ItemWebUrl == nil {
+	if o == nil || IsNil(o.ItemWebUrl) {
 		var ret string
 		return ret
 	}
@@ -823,7 +864,7 @@ func (o *ItemSummary) GetItemWebUrl() string {
 // GetItemWebUrlOk returns a tuple with the ItemWebUrl field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetItemWebUrlOk() (*string, bool) {
-	if o == nil || o.ItemWebUrl == nil {
+	if o == nil || IsNil(o.ItemWebUrl) {
 		return nil, false
 	}
 	return o.ItemWebUrl, true
@@ -831,7 +872,7 @@ func (o *ItemSummary) GetItemWebUrlOk() (*string, bool) {
 
 // HasItemWebUrl returns a boolean if a field has been set.
 func (o *ItemSummary) HasItemWebUrl() bool {
-	if o != nil && o.ItemWebUrl != nil {
+	if o != nil && !IsNil(o.ItemWebUrl) {
 		return true
 	}
 
@@ -843,9 +884,41 @@ func (o *ItemSummary) SetItemWebUrl(v string) {
 	o.ItemWebUrl = &v
 }
 
+// GetLeafCategoryIds returns the LeafCategoryIds field value if set, zero value otherwise.
+func (o *ItemSummary) GetLeafCategoryIds() []string {
+	if o == nil || IsNil(o.LeafCategoryIds) {
+		var ret []string
+		return ret
+	}
+	return o.LeafCategoryIds
+}
+
+// GetLeafCategoryIdsOk returns a tuple with the LeafCategoryIds field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *ItemSummary) GetLeafCategoryIdsOk() ([]string, bool) {
+	if o == nil || IsNil(o.LeafCategoryIds) {
+		return nil, false
+	}
+	return o.LeafCategoryIds, true
+}
+
+// HasLeafCategoryIds returns a boolean if a field has been set.
+func (o *ItemSummary) HasLeafCategoryIds() bool {
+	if o != nil && !IsNil(o.LeafCategoryIds) {
+		return true
+	}
+
+	return false
+}
+
+// SetLeafCategoryIds gets a reference to the given []string and assigns it to the LeafCategoryIds field.
+func (o *ItemSummary) SetLeafCategoryIds(v []string) {
+	o.LeafCategoryIds = v
+}
+
 // GetLegacyItemId returns the LegacyItemId field value if set, zero value otherwise.
 func (o *ItemSummary) GetLegacyItemId() string {
-	if o == nil || o.LegacyItemId == nil {
+	if o == nil || IsNil(o.LegacyItemId) {
 		var ret string
 		return ret
 	}
@@ -855,7 +928,7 @@ func (o *ItemSummary) GetLegacyItemId() string {
 // GetLegacyItemIdOk returns a tuple with the LegacyItemId field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetLegacyItemIdOk() (*string, bool) {
-	if o == nil || o.LegacyItemId == nil {
+	if o == nil || IsNil(o.LegacyItemId) {
 		return nil, false
 	}
 	return o.LegacyItemId, true
@@ -863,7 +936,7 @@ func (o *ItemSummary) GetLegacyItemIdOk() (*string, bool) {
 
 // HasLegacyItemId returns a boolean if a field has been set.
 func (o *ItemSummary) HasLegacyItemId() bool {
-	if o != nil && o.LegacyItemId != nil {
+	if o != nil && !IsNil(o.LegacyItemId) {
 		return true
 	}
 
@@ -875,9 +948,41 @@ func (o *ItemSummary) SetLegacyItemId(v string) {
 	o.LegacyItemId = &v
 }
 
+// GetListingMarketplaceId returns the ListingMarketplaceId field value if set, zero value otherwise.
+func (o *ItemSummary) GetListingMarketplaceId() string {
+	if o == nil || IsNil(o.ListingMarketplaceId) {
+		var ret string
+		return ret
+	}
+	return *o.ListingMarketplaceId
+}
+
+// GetListingMarketplaceIdOk returns a tuple with the ListingMarketplaceId field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *ItemSummary) GetListingMarketplaceIdOk() (*string, bool) {
+	if o == nil || IsNil(o.ListingMarketplaceId) {
+		return nil, false
+	}
+	return o.ListingMarketplaceId, true
+}
+
+// HasListingMarketplaceId returns a boolean if a field has been set.
+func (o *ItemSummary) HasListingMarketplaceId() bool {
+	if o != nil && !IsNil(o.ListingMarketplaceId) {
+		return true
+	}
+
+	return false
+}
+
+// SetListingMarketplaceId gets a reference to the given string and assigns it to the ListingMarketplaceId field.
+func (o *ItemSummary) SetListingMarketplaceId(v string) {
+	o.ListingMarketplaceId = &v
+}
+
 // GetMarketingPrice returns the MarketingPrice field value if set, zero value otherwise.
 func (o *ItemSummary) GetMarketingPrice() MarketingPrice {
-	if o == nil || o.MarketingPrice == nil {
+	if o == nil || IsNil(o.MarketingPrice) {
 		var ret MarketingPrice
 		return ret
 	}
@@ -887,7 +992,7 @@ func (o *ItemSummary) GetMarketingPrice() MarketingPrice {
 // GetMarketingPriceOk returns a tuple with the MarketingPrice field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetMarketingPriceOk() (*MarketingPrice, bool) {
-	if o == nil || o.MarketingPrice == nil {
+	if o == nil || IsNil(o.MarketingPrice) {
 		return nil, false
 	}
 	return o.MarketingPrice, true
@@ -895,7 +1000,7 @@ func (o *ItemSummary) GetMarketingPriceOk() (*MarketingPrice, bool) {
 
 // HasMarketingPrice returns a boolean if a field has been set.
 func (o *ItemSummary) HasMarketingPrice() bool {
-	if o != nil && o.MarketingPrice != nil {
+	if o != nil && !IsNil(o.MarketingPrice) {
 		return true
 	}
 
@@ -909,17 +1014,17 @@ func (o *ItemSummary) SetMarketingPrice(v MarketingPrice) {
 
 // GetPickupOptions returns the PickupOptions field value if set, zero value otherwise.
 func (o *ItemSummary) GetPickupOptions() []PickupOptionSummary {
-	if o == nil || o.PickupOptions == nil {
+	if o == nil || IsNil(o.PickupOptions) {
 		var ret []PickupOptionSummary
 		return ret
 	}
-	return *o.PickupOptions
+	return o.PickupOptions
 }
 
 // GetPickupOptionsOk returns a tuple with the PickupOptions field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *ItemSummary) GetPickupOptionsOk() (*[]PickupOptionSummary, bool) {
-	if o == nil || o.PickupOptions == nil {
+func (o *ItemSummary) GetPickupOptionsOk() ([]PickupOptionSummary, bool) {
+	if o == nil || IsNil(o.PickupOptions) {
 		return nil, false
 	}
 	return o.PickupOptions, true
@@ -927,7 +1032,7 @@ func (o *ItemSummary) GetPickupOptionsOk() (*[]PickupOptionSummary, bool) {
 
 // HasPickupOptions returns a boolean if a field has been set.
 func (o *ItemSummary) HasPickupOptions() bool {
-	if o != nil && o.PickupOptions != nil {
+	if o != nil && !IsNil(o.PickupOptions) {
 		return true
 	}
 
@@ -936,12 +1041,12 @@ func (o *ItemSummary) HasPickupOptions() bool {
 
 // SetPickupOptions gets a reference to the given []PickupOptionSummary and assigns it to the PickupOptions field.
 func (o *ItemSummary) SetPickupOptions(v []PickupOptionSummary) {
-	o.PickupOptions = &v
+	o.PickupOptions = v
 }
 
 // GetPrice returns the Price field value if set, zero value otherwise.
 func (o *ItemSummary) GetPrice() ConvertedAmount {
-	if o == nil || o.Price == nil {
+	if o == nil || IsNil(o.Price) {
 		var ret ConvertedAmount
 		return ret
 	}
@@ -951,7 +1056,7 @@ func (o *ItemSummary) GetPrice() ConvertedAmount {
 // GetPriceOk returns a tuple with the Price field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetPriceOk() (*ConvertedAmount, bool) {
-	if o == nil || o.Price == nil {
+	if o == nil || IsNil(o.Price) {
 		return nil, false
 	}
 	return o.Price, true
@@ -959,7 +1064,7 @@ func (o *ItemSummary) GetPriceOk() (*ConvertedAmount, bool) {
 
 // HasPrice returns a boolean if a field has been set.
 func (o *ItemSummary) HasPrice() bool {
-	if o != nil && o.Price != nil {
+	if o != nil && !IsNil(o.Price) {
 		return true
 	}
 
@@ -973,7 +1078,7 @@ func (o *ItemSummary) SetPrice(v ConvertedAmount) {
 
 // GetPriceDisplayCondition returns the PriceDisplayCondition field value if set, zero value otherwise.
 func (o *ItemSummary) GetPriceDisplayCondition() string {
-	if o == nil || o.PriceDisplayCondition == nil {
+	if o == nil || IsNil(o.PriceDisplayCondition) {
 		var ret string
 		return ret
 	}
@@ -983,7 +1088,7 @@ func (o *ItemSummary) GetPriceDisplayCondition() string {
 // GetPriceDisplayConditionOk returns a tuple with the PriceDisplayCondition field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetPriceDisplayConditionOk() (*string, bool) {
-	if o == nil || o.PriceDisplayCondition == nil {
+	if o == nil || IsNil(o.PriceDisplayCondition) {
 		return nil, false
 	}
 	return o.PriceDisplayCondition, true
@@ -991,7 +1096,7 @@ func (o *ItemSummary) GetPriceDisplayConditionOk() (*string, bool) {
 
 // HasPriceDisplayCondition returns a boolean if a field has been set.
 func (o *ItemSummary) HasPriceDisplayCondition() bool {
-	if o != nil && o.PriceDisplayCondition != nil {
+	if o != nil && !IsNil(o.PriceDisplayCondition) {
 		return true
 	}
 
@@ -1005,7 +1110,7 @@ func (o *ItemSummary) SetPriceDisplayCondition(v string) {
 
 // GetPriorityListing returns the PriorityListing field value if set, zero value otherwise.
 func (o *ItemSummary) GetPriorityListing() bool {
-	if o == nil || o.PriorityListing == nil {
+	if o == nil || IsNil(o.PriorityListing) {
 		var ret bool
 		return ret
 	}
@@ -1015,7 +1120,7 @@ func (o *ItemSummary) GetPriorityListing() bool {
 // GetPriorityListingOk returns a tuple with the PriorityListing field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetPriorityListingOk() (*bool, bool) {
-	if o == nil || o.PriorityListing == nil {
+	if o == nil || IsNil(o.PriorityListing) {
 		return nil, false
 	}
 	return o.PriorityListing, true
@@ -1023,7 +1128,7 @@ func (o *ItemSummary) GetPriorityListingOk() (*bool, bool) {
 
 // HasPriorityListing returns a boolean if a field has been set.
 func (o *ItemSummary) HasPriorityListing() bool {
-	if o != nil && o.PriorityListing != nil {
+	if o != nil && !IsNil(o.PriorityListing) {
 		return true
 	}
 
@@ -1037,17 +1142,17 @@ func (o *ItemSummary) SetPriorityListing(v bool) {
 
 // GetQualifiedPrograms returns the QualifiedPrograms field value if set, zero value otherwise.
 func (o *ItemSummary) GetQualifiedPrograms() []string {
-	if o == nil || o.QualifiedPrograms == nil {
+	if o == nil || IsNil(o.QualifiedPrograms) {
 		var ret []string
 		return ret
 	}
-	return *o.QualifiedPrograms
+	return o.QualifiedPrograms
 }
 
 // GetQualifiedProgramsOk returns a tuple with the QualifiedPrograms field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *ItemSummary) GetQualifiedProgramsOk() (*[]string, bool) {
-	if o == nil || o.QualifiedPrograms == nil {
+func (o *ItemSummary) GetQualifiedProgramsOk() ([]string, bool) {
+	if o == nil || IsNil(o.QualifiedPrograms) {
 		return nil, false
 	}
 	return o.QualifiedPrograms, true
@@ -1055,7 +1160,7 @@ func (o *ItemSummary) GetQualifiedProgramsOk() (*[]string, bool) {
 
 // HasQualifiedPrograms returns a boolean if a field has been set.
 func (o *ItemSummary) HasQualifiedPrograms() bool {
-	if o != nil && o.QualifiedPrograms != nil {
+	if o != nil && !IsNil(o.QualifiedPrograms) {
 		return true
 	}
 
@@ -1064,12 +1169,12 @@ func (o *ItemSummary) HasQualifiedPrograms() bool {
 
 // SetQualifiedPrograms gets a reference to the given []string and assigns it to the QualifiedPrograms field.
 func (o *ItemSummary) SetQualifiedPrograms(v []string) {
-	o.QualifiedPrograms = &v
+	o.QualifiedPrograms = v
 }
 
 // GetSeller returns the Seller field value if set, zero value otherwise.
 func (o *ItemSummary) GetSeller() Seller {
-	if o == nil || o.Seller == nil {
+	if o == nil || IsNil(o.Seller) {
 		var ret Seller
 		return ret
 	}
@@ -1079,7 +1184,7 @@ func (o *ItemSummary) GetSeller() Seller {
 // GetSellerOk returns a tuple with the Seller field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetSellerOk() (*Seller, bool) {
-	if o == nil || o.Seller == nil {
+	if o == nil || IsNil(o.Seller) {
 		return nil, false
 	}
 	return o.Seller, true
@@ -1087,7 +1192,7 @@ func (o *ItemSummary) GetSellerOk() (*Seller, bool) {
 
 // HasSeller returns a boolean if a field has been set.
 func (o *ItemSummary) HasSeller() bool {
-	if o != nil && o.Seller != nil {
+	if o != nil && !IsNil(o.Seller) {
 		return true
 	}
 
@@ -1101,17 +1206,17 @@ func (o *ItemSummary) SetSeller(v Seller) {
 
 // GetShippingOptions returns the ShippingOptions field value if set, zero value otherwise.
 func (o *ItemSummary) GetShippingOptions() []ShippingOptionSummary {
-	if o == nil || o.ShippingOptions == nil {
+	if o == nil || IsNil(o.ShippingOptions) {
 		var ret []ShippingOptionSummary
 		return ret
 	}
-	return *o.ShippingOptions
+	return o.ShippingOptions
 }
 
 // GetShippingOptionsOk returns a tuple with the ShippingOptions field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *ItemSummary) GetShippingOptionsOk() (*[]ShippingOptionSummary, bool) {
-	if o == nil || o.ShippingOptions == nil {
+func (o *ItemSummary) GetShippingOptionsOk() ([]ShippingOptionSummary, bool) {
+	if o == nil || IsNil(o.ShippingOptions) {
 		return nil, false
 	}
 	return o.ShippingOptions, true
@@ -1119,7 +1224,7 @@ func (o *ItemSummary) GetShippingOptionsOk() (*[]ShippingOptionSummary, bool) {
 
 // HasShippingOptions returns a boolean if a field has been set.
 func (o *ItemSummary) HasShippingOptions() bool {
-	if o != nil && o.ShippingOptions != nil {
+	if o != nil && !IsNil(o.ShippingOptions) {
 		return true
 	}
 
@@ -1128,12 +1233,12 @@ func (o *ItemSummary) HasShippingOptions() bool {
 
 // SetShippingOptions gets a reference to the given []ShippingOptionSummary and assigns it to the ShippingOptions field.
 func (o *ItemSummary) SetShippingOptions(v []ShippingOptionSummary) {
-	o.ShippingOptions = &v
+	o.ShippingOptions = v
 }
 
 // GetShortDescription returns the ShortDescription field value if set, zero value otherwise.
 func (o *ItemSummary) GetShortDescription() string {
-	if o == nil || o.ShortDescription == nil {
+	if o == nil || IsNil(o.ShortDescription) {
 		var ret string
 		return ret
 	}
@@ -1143,7 +1248,7 @@ func (o *ItemSummary) GetShortDescription() string {
 // GetShortDescriptionOk returns a tuple with the ShortDescription field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetShortDescriptionOk() (*string, bool) {
-	if o == nil || o.ShortDescription == nil {
+	if o == nil || IsNil(o.ShortDescription) {
 		return nil, false
 	}
 	return o.ShortDescription, true
@@ -1151,7 +1256,7 @@ func (o *ItemSummary) GetShortDescriptionOk() (*string, bool) {
 
 // HasShortDescription returns a boolean if a field has been set.
 func (o *ItemSummary) HasShortDescription() bool {
-	if o != nil && o.ShortDescription != nil {
+	if o != nil && !IsNil(o.ShortDescription) {
 		return true
 	}
 
@@ -1165,17 +1270,17 @@ func (o *ItemSummary) SetShortDescription(v string) {
 
 // GetThumbnailImages returns the ThumbnailImages field value if set, zero value otherwise.
 func (o *ItemSummary) GetThumbnailImages() []Image {
-	if o == nil || o.ThumbnailImages == nil {
+	if o == nil || IsNil(o.ThumbnailImages) {
 		var ret []Image
 		return ret
 	}
-	return *o.ThumbnailImages
+	return o.ThumbnailImages
 }
 
 // GetThumbnailImagesOk returns a tuple with the ThumbnailImages field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *ItemSummary) GetThumbnailImagesOk() (*[]Image, bool) {
-	if o == nil || o.ThumbnailImages == nil {
+func (o *ItemSummary) GetThumbnailImagesOk() ([]Image, bool) {
+	if o == nil || IsNil(o.ThumbnailImages) {
 		return nil, false
 	}
 	return o.ThumbnailImages, true
@@ -1183,7 +1288,7 @@ func (o *ItemSummary) GetThumbnailImagesOk() (*[]Image, bool) {
 
 // HasThumbnailImages returns a boolean if a field has been set.
 func (o *ItemSummary) HasThumbnailImages() bool {
-	if o != nil && o.ThumbnailImages != nil {
+	if o != nil && !IsNil(o.ThumbnailImages) {
 		return true
 	}
 
@@ -1192,12 +1297,12 @@ func (o *ItemSummary) HasThumbnailImages() bool {
 
 // SetThumbnailImages gets a reference to the given []Image and assigns it to the ThumbnailImages field.
 func (o *ItemSummary) SetThumbnailImages(v []Image) {
-	o.ThumbnailImages = &v
+	o.ThumbnailImages = v
 }
 
 // GetTitle returns the Title field value if set, zero value otherwise.
 func (o *ItemSummary) GetTitle() string {
-	if o == nil || o.Title == nil {
+	if o == nil || IsNil(o.Title) {
 		var ret string
 		return ret
 	}
@@ -1207,7 +1312,7 @@ func (o *ItemSummary) GetTitle() string {
 // GetTitleOk returns a tuple with the Title field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetTitleOk() (*string, bool) {
-	if o == nil || o.Title == nil {
+	if o == nil || IsNil(o.Title) {
 		return nil, false
 	}
 	return o.Title, true
@@ -1215,7 +1320,7 @@ func (o *ItemSummary) GetTitleOk() (*string, bool) {
 
 // HasTitle returns a boolean if a field has been set.
 func (o *ItemSummary) HasTitle() bool {
-	if o != nil && o.Title != nil {
+	if o != nil && !IsNil(o.Title) {
 		return true
 	}
 
@@ -1229,7 +1334,7 @@ func (o *ItemSummary) SetTitle(v string) {
 
 // GetTopRatedBuyingExperience returns the TopRatedBuyingExperience field value if set, zero value otherwise.
 func (o *ItemSummary) GetTopRatedBuyingExperience() bool {
-	if o == nil || o.TopRatedBuyingExperience == nil {
+	if o == nil || IsNil(o.TopRatedBuyingExperience) {
 		var ret bool
 		return ret
 	}
@@ -1239,7 +1344,7 @@ func (o *ItemSummary) GetTopRatedBuyingExperience() bool {
 // GetTopRatedBuyingExperienceOk returns a tuple with the TopRatedBuyingExperience field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetTopRatedBuyingExperienceOk() (*bool, bool) {
-	if o == nil || o.TopRatedBuyingExperience == nil {
+	if o == nil || IsNil(o.TopRatedBuyingExperience) {
 		return nil, false
 	}
 	return o.TopRatedBuyingExperience, true
@@ -1247,7 +1352,7 @@ func (o *ItemSummary) GetTopRatedBuyingExperienceOk() (*bool, bool) {
 
 // HasTopRatedBuyingExperience returns a boolean if a field has been set.
 func (o *ItemSummary) HasTopRatedBuyingExperience() bool {
-	if o != nil && o.TopRatedBuyingExperience != nil {
+	if o != nil && !IsNil(o.TopRatedBuyingExperience) {
 		return true
 	}
 
@@ -1261,7 +1366,7 @@ func (o *ItemSummary) SetTopRatedBuyingExperience(v bool) {
 
 // GetTyreLabelImageUrl returns the TyreLabelImageUrl field value if set, zero value otherwise.
 func (o *ItemSummary) GetTyreLabelImageUrl() string {
-	if o == nil || o.TyreLabelImageUrl == nil {
+	if o == nil || IsNil(o.TyreLabelImageUrl) {
 		var ret string
 		return ret
 	}
@@ -1271,7 +1376,7 @@ func (o *ItemSummary) GetTyreLabelImageUrl() string {
 // GetTyreLabelImageUrlOk returns a tuple with the TyreLabelImageUrl field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetTyreLabelImageUrlOk() (*string, bool) {
-	if o == nil || o.TyreLabelImageUrl == nil {
+	if o == nil || IsNil(o.TyreLabelImageUrl) {
 		return nil, false
 	}
 	return o.TyreLabelImageUrl, true
@@ -1279,7 +1384,7 @@ func (o *ItemSummary) GetTyreLabelImageUrlOk() (*string, bool) {
 
 // HasTyreLabelImageUrl returns a boolean if a field has been set.
 func (o *ItemSummary) HasTyreLabelImageUrl() bool {
-	if o != nil && o.TyreLabelImageUrl != nil {
+	if o != nil && !IsNil(o.TyreLabelImageUrl) {
 		return true
 	}
 
@@ -1293,7 +1398,7 @@ func (o *ItemSummary) SetTyreLabelImageUrl(v string) {
 
 // GetUnitPrice returns the UnitPrice field value if set, zero value otherwise.
 func (o *ItemSummary) GetUnitPrice() ConvertedAmount {
-	if o == nil || o.UnitPrice == nil {
+	if o == nil || IsNil(o.UnitPrice) {
 		var ret ConvertedAmount
 		return ret
 	}
@@ -1303,7 +1408,7 @@ func (o *ItemSummary) GetUnitPrice() ConvertedAmount {
 // GetUnitPriceOk returns a tuple with the UnitPrice field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetUnitPriceOk() (*ConvertedAmount, bool) {
-	if o == nil || o.UnitPrice == nil {
+	if o == nil || IsNil(o.UnitPrice) {
 		return nil, false
 	}
 	return o.UnitPrice, true
@@ -1311,7 +1416,7 @@ func (o *ItemSummary) GetUnitPriceOk() (*ConvertedAmount, bool) {
 
 // HasUnitPrice returns a boolean if a field has been set.
 func (o *ItemSummary) HasUnitPrice() bool {
-	if o != nil && o.UnitPrice != nil {
+	if o != nil && !IsNil(o.UnitPrice) {
 		return true
 	}
 
@@ -1325,7 +1430,7 @@ func (o *ItemSummary) SetUnitPrice(v ConvertedAmount) {
 
 // GetUnitPricingMeasure returns the UnitPricingMeasure field value if set, zero value otherwise.
 func (o *ItemSummary) GetUnitPricingMeasure() string {
-	if o == nil || o.UnitPricingMeasure == nil {
+	if o == nil || IsNil(o.UnitPricingMeasure) {
 		var ret string
 		return ret
 	}
@@ -1335,7 +1440,7 @@ func (o *ItemSummary) GetUnitPricingMeasure() string {
 // GetUnitPricingMeasureOk returns a tuple with the UnitPricingMeasure field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetUnitPricingMeasureOk() (*string, bool) {
-	if o == nil || o.UnitPricingMeasure == nil {
+	if o == nil || IsNil(o.UnitPricingMeasure) {
 		return nil, false
 	}
 	return o.UnitPricingMeasure, true
@@ -1343,7 +1448,7 @@ func (o *ItemSummary) GetUnitPricingMeasureOk() (*string, bool) {
 
 // HasUnitPricingMeasure returns a boolean if a field has been set.
 func (o *ItemSummary) HasUnitPricingMeasure() bool {
-	if o != nil && o.UnitPricingMeasure != nil {
+	if o != nil && !IsNil(o.UnitPricingMeasure) {
 		return true
 	}
 
@@ -1357,7 +1462,7 @@ func (o *ItemSummary) SetUnitPricingMeasure(v string) {
 
 // GetWatchCount returns the WatchCount field value if set, zero value otherwise.
 func (o *ItemSummary) GetWatchCount() int32 {
-	if o == nil || o.WatchCount == nil {
+	if o == nil || IsNil(o.WatchCount) {
 		var ret int32
 		return ret
 	}
@@ -1367,7 +1472,7 @@ func (o *ItemSummary) GetWatchCount() int32 {
 // GetWatchCountOk returns a tuple with the WatchCount field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ItemSummary) GetWatchCountOk() (*int32, bool) {
-	if o == nil || o.WatchCount == nil {
+	if o == nil || IsNil(o.WatchCount) {
 		return nil, false
 	}
 	return o.WatchCount, true
@@ -1375,7 +1480,7 @@ func (o *ItemSummary) GetWatchCountOk() (*int32, bool) {
 
 // HasWatchCount returns a boolean if a field has been set.
 func (o *ItemSummary) HasWatchCount() bool {
-	if o != nil && o.WatchCount != nil {
+	if o != nil && !IsNil(o.WatchCount) {
 		return true
 	}
 
@@ -1388,128 +1493,145 @@ func (o *ItemSummary) SetWatchCount(v int32) {
 }
 
 func (o ItemSummary) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.AdditionalImages != nil {
-		toSerialize["additionalImages"] = o.AdditionalImages
-	}
-	if o.AdultOnly != nil {
-		toSerialize["adultOnly"] = o.AdultOnly
-	}
-	if o.AvailableCoupons != nil {
-		toSerialize["availableCoupons"] = o.AvailableCoupons
-	}
-	if o.BidCount != nil {
-		toSerialize["bidCount"] = o.BidCount
-	}
-	if o.BuyingOptions != nil {
-		toSerialize["buyingOptions"] = o.BuyingOptions
-	}
-	if o.Categories != nil {
-		toSerialize["categories"] = o.Categories
-	}
-	if o.CompatibilityMatch != nil {
-		toSerialize["compatibilityMatch"] = o.CompatibilityMatch
-	}
-	if o.CompatibilityProperties != nil {
-		toSerialize["compatibilityProperties"] = o.CompatibilityProperties
-	}
-	if o.Condition != nil {
-		toSerialize["condition"] = o.Condition
-	}
-	if o.ConditionId != nil {
-		toSerialize["conditionId"] = o.ConditionId
-	}
-	if o.CurrentBidPrice != nil {
-		toSerialize["currentBidPrice"] = o.CurrentBidPrice
-	}
-	if o.DistanceFromPickupLocation != nil {
-		toSerialize["distanceFromPickupLocation"] = o.DistanceFromPickupLocation
-	}
-	if o.EnergyEfficiencyClass != nil {
-		toSerialize["energyEfficiencyClass"] = o.EnergyEfficiencyClass
-	}
-	if o.Epid != nil {
-		toSerialize["epid"] = o.Epid
-	}
-	if o.Image != nil {
-		toSerialize["image"] = o.Image
-	}
-	if o.ItemAffiliateWebUrl != nil {
-		toSerialize["itemAffiliateWebUrl"] = o.ItemAffiliateWebUrl
-	}
-	if o.ItemEndDate != nil {
-		toSerialize["itemEndDate"] = o.ItemEndDate
-	}
-	if o.ItemGroupHref != nil {
-		toSerialize["itemGroupHref"] = o.ItemGroupHref
-	}
-	if o.ItemGroupType != nil {
-		toSerialize["itemGroupType"] = o.ItemGroupType
-	}
-	if o.ItemHref != nil {
-		toSerialize["itemHref"] = o.ItemHref
-	}
-	if o.ItemId != nil {
-		toSerialize["itemId"] = o.ItemId
-	}
-	if o.ItemLocation != nil {
-		toSerialize["itemLocation"] = o.ItemLocation
-	}
-	if o.ItemWebUrl != nil {
-		toSerialize["itemWebUrl"] = o.ItemWebUrl
-	}
-	if o.LegacyItemId != nil {
-		toSerialize["legacyItemId"] = o.LegacyItemId
-	}
-	if o.MarketingPrice != nil {
-		toSerialize["marketingPrice"] = o.MarketingPrice
-	}
-	if o.PickupOptions != nil {
-		toSerialize["pickupOptions"] = o.PickupOptions
-	}
-	if o.Price != nil {
-		toSerialize["price"] = o.Price
-	}
-	if o.PriceDisplayCondition != nil {
-		toSerialize["priceDisplayCondition"] = o.PriceDisplayCondition
-	}
-	if o.PriorityListing != nil {
-		toSerialize["priorityListing"] = o.PriorityListing
-	}
-	if o.QualifiedPrograms != nil {
-		toSerialize["qualifiedPrograms"] = o.QualifiedPrograms
-	}
-	if o.Seller != nil {
-		toSerialize["seller"] = o.Seller
-	}
-	if o.ShippingOptions != nil {
-		toSerialize["shippingOptions"] = o.ShippingOptions
-	}
-	if o.ShortDescription != nil {
-		toSerialize["shortDescription"] = o.ShortDescription
-	}
-	if o.ThumbnailImages != nil {
-		toSerialize["thumbnailImages"] = o.ThumbnailImages
-	}
-	if o.Title != nil {
-		toSerialize["title"] = o.Title
-	}
-	if o.TopRatedBuyingExperience != nil {
-		toSerialize["topRatedBuyingExperience"] = o.TopRatedBuyingExperience
-	}
-	if o.TyreLabelImageUrl != nil {
-		toSerialize["tyreLabelImageUrl"] = o.TyreLabelImageUrl
-	}
-	if o.UnitPrice != nil {
-		toSerialize["unitPrice"] = o.UnitPrice
-	}
-	if o.UnitPricingMeasure != nil {
-		toSerialize["unitPricingMeasure"] = o.UnitPricingMeasure
-	}
-	if o.WatchCount != nil {
-		toSerialize["watchCount"] = o.WatchCount
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o ItemSummary) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.AdditionalImages) {
+		toSerialize["additionalImages"] = o.AdditionalImages
+	}
+	if !IsNil(o.AdultOnly) {
+		toSerialize["adultOnly"] = o.AdultOnly
+	}
+	if !IsNil(o.AvailableCoupons) {
+		toSerialize["availableCoupons"] = o.AvailableCoupons
+	}
+	if !IsNil(o.BidCount) {
+		toSerialize["bidCount"] = o.BidCount
+	}
+	if !IsNil(o.BuyingOptions) {
+		toSerialize["buyingOptions"] = o.BuyingOptions
+	}
+	if !IsNil(o.Categories) {
+		toSerialize["categories"] = o.Categories
+	}
+	if !IsNil(o.CompatibilityMatch) {
+		toSerialize["compatibilityMatch"] = o.CompatibilityMatch
+	}
+	if !IsNil(o.CompatibilityProperties) {
+		toSerialize["compatibilityProperties"] = o.CompatibilityProperties
+	}
+	if !IsNil(o.Condition) {
+		toSerialize["condition"] = o.Condition
+	}
+	if !IsNil(o.ConditionId) {
+		toSerialize["conditionId"] = o.ConditionId
+	}
+	if !IsNil(o.CurrentBidPrice) {
+		toSerialize["currentBidPrice"] = o.CurrentBidPrice
+	}
+	if !IsNil(o.DistanceFromPickupLocation) {
+		toSerialize["distanceFromPickupLocation"] = o.DistanceFromPickupLocation
+	}
+	if !IsNil(o.EnergyEfficiencyClass) {
+		toSerialize["energyEfficiencyClass"] = o.EnergyEfficiencyClass
+	}
+	if !IsNil(o.Epid) {
+		toSerialize["epid"] = o.Epid
+	}
+	if !IsNil(o.Image) {
+		toSerialize["image"] = o.Image
+	}
+	if !IsNil(o.ItemAffiliateWebUrl) {
+		toSerialize["itemAffiliateWebUrl"] = o.ItemAffiliateWebUrl
+	}
+	if !IsNil(o.ItemCreationDate) {
+		toSerialize["itemCreationDate"] = o.ItemCreationDate
+	}
+	if !IsNil(o.ItemEndDate) {
+		toSerialize["itemEndDate"] = o.ItemEndDate
+	}
+	if !IsNil(o.ItemGroupHref) {
+		toSerialize["itemGroupHref"] = o.ItemGroupHref
+	}
+	if !IsNil(o.ItemGroupType) {
+		toSerialize["itemGroupType"] = o.ItemGroupType
+	}
+	if !IsNil(o.ItemHref) {
+		toSerialize["itemHref"] = o.ItemHref
+	}
+	if !IsNil(o.ItemId) {
+		toSerialize["itemId"] = o.ItemId
+	}
+	if !IsNil(o.ItemLocation) {
+		toSerialize["itemLocation"] = o.ItemLocation
+	}
+	if !IsNil(o.ItemWebUrl) {
+		toSerialize["itemWebUrl"] = o.ItemWebUrl
+	}
+	if !IsNil(o.LeafCategoryIds) {
+		toSerialize["leafCategoryIds"] = o.LeafCategoryIds
+	}
+	if !IsNil(o.LegacyItemId) {
+		toSerialize["legacyItemId"] = o.LegacyItemId
+	}
+	if !IsNil(o.ListingMarketplaceId) {
+		toSerialize["listingMarketplaceId"] = o.ListingMarketplaceId
+	}
+	if !IsNil(o.MarketingPrice) {
+		toSerialize["marketingPrice"] = o.MarketingPrice
+	}
+	if !IsNil(o.PickupOptions) {
+		toSerialize["pickupOptions"] = o.PickupOptions
+	}
+	if !IsNil(o.Price) {
+		toSerialize["price"] = o.Price
+	}
+	if !IsNil(o.PriceDisplayCondition) {
+		toSerialize["priceDisplayCondition"] = o.PriceDisplayCondition
+	}
+	if !IsNil(o.PriorityListing) {
+		toSerialize["priorityListing"] = o.PriorityListing
+	}
+	if !IsNil(o.QualifiedPrograms) {
+		toSerialize["qualifiedPrograms"] = o.QualifiedPrograms
+	}
+	if !IsNil(o.Seller) {
+		toSerialize["seller"] = o.Seller
+	}
+	if !IsNil(o.ShippingOptions) {
+		toSerialize["shippingOptions"] = o.ShippingOptions
+	}
+	if !IsNil(o.ShortDescription) {
+		toSerialize["shortDescription"] = o.ShortDescription
+	}
+	if !IsNil(o.ThumbnailImages) {
+		toSerialize["thumbnailImages"] = o.ThumbnailImages
+	}
+	if !IsNil(o.Title) {
+		toSerialize["title"] = o.Title
+	}
+	if !IsNil(o.TopRatedBuyingExperience) {
+		toSerialize["topRatedBuyingExperience"] = o.TopRatedBuyingExperience
+	}
+	if !IsNil(o.TyreLabelImageUrl) {
+		toSerialize["tyreLabelImageUrl"] = o.TyreLabelImageUrl
+	}
+	if !IsNil(o.UnitPrice) {
+		toSerialize["unitPrice"] = o.UnitPrice
+	}
+	if !IsNil(o.UnitPricingMeasure) {
+		toSerialize["unitPricingMeasure"] = o.UnitPricingMeasure
+	}
+	if !IsNil(o.WatchCount) {
+		toSerialize["watchCount"] = o.WatchCount
+	}
+	return toSerialize, nil
 }
 
 type NullableItemSummary struct {
