@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,14 +14,17 @@ import (
 	"encoding/json"
 )
 
+// checks if the Items type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &Items{}
+
 // Items Container for a list of items.
 type Items struct {
 	// An arraylist of all the items.
-	Items *[]CoreItem `json:"items,omitempty"`
+	Items []CoreItem `json:"items,omitempty"`
 	// The total number of items retrieved.
 	Total *int32 `json:"total,omitempty"`
 	// An array of warning messages. These types of errors do not prevent the method from executing but should be checked.
-	Warnings *[]Error `json:"warnings,omitempty"`
+	Warnings []Error `json:"warnings,omitempty"`
 }
 
 // NewItems instantiates a new Items object
@@ -43,17 +46,17 @@ func NewItemsWithDefaults() *Items {
 
 // GetItems returns the Items field value if set, zero value otherwise.
 func (o *Items) GetItems() []CoreItem {
-	if o == nil || o.Items == nil {
+	if o == nil || IsNil(o.Items) {
 		var ret []CoreItem
 		return ret
 	}
-	return *o.Items
+	return o.Items
 }
 
 // GetItemsOk returns a tuple with the Items field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Items) GetItemsOk() (*[]CoreItem, bool) {
-	if o == nil || o.Items == nil {
+func (o *Items) GetItemsOk() ([]CoreItem, bool) {
+	if o == nil || IsNil(o.Items) {
 		return nil, false
 	}
 	return o.Items, true
@@ -61,7 +64,7 @@ func (o *Items) GetItemsOk() (*[]CoreItem, bool) {
 
 // HasItems returns a boolean if a field has been set.
 func (o *Items) HasItems() bool {
-	if o != nil && o.Items != nil {
+	if o != nil && !IsNil(o.Items) {
 		return true
 	}
 
@@ -70,12 +73,12 @@ func (o *Items) HasItems() bool {
 
 // SetItems gets a reference to the given []CoreItem and assigns it to the Items field.
 func (o *Items) SetItems(v []CoreItem) {
-	o.Items = &v
+	o.Items = v
 }
 
 // GetTotal returns the Total field value if set, zero value otherwise.
 func (o *Items) GetTotal() int32 {
-	if o == nil || o.Total == nil {
+	if o == nil || IsNil(o.Total) {
 		var ret int32
 		return ret
 	}
@@ -85,7 +88,7 @@ func (o *Items) GetTotal() int32 {
 // GetTotalOk returns a tuple with the Total field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Items) GetTotalOk() (*int32, bool) {
-	if o == nil || o.Total == nil {
+	if o == nil || IsNil(o.Total) {
 		return nil, false
 	}
 	return o.Total, true
@@ -93,7 +96,7 @@ func (o *Items) GetTotalOk() (*int32, bool) {
 
 // HasTotal returns a boolean if a field has been set.
 func (o *Items) HasTotal() bool {
-	if o != nil && o.Total != nil {
+	if o != nil && !IsNil(o.Total) {
 		return true
 	}
 
@@ -107,17 +110,17 @@ func (o *Items) SetTotal(v int32) {
 
 // GetWarnings returns the Warnings field value if set, zero value otherwise.
 func (o *Items) GetWarnings() []Error {
-	if o == nil || o.Warnings == nil {
+	if o == nil || IsNil(o.Warnings) {
 		var ret []Error
 		return ret
 	}
-	return *o.Warnings
+	return o.Warnings
 }
 
 // GetWarningsOk returns a tuple with the Warnings field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Items) GetWarningsOk() (*[]Error, bool) {
-	if o == nil || o.Warnings == nil {
+func (o *Items) GetWarningsOk() ([]Error, bool) {
+	if o == nil || IsNil(o.Warnings) {
 		return nil, false
 	}
 	return o.Warnings, true
@@ -125,7 +128,7 @@ func (o *Items) GetWarningsOk() (*[]Error, bool) {
 
 // HasWarnings returns a boolean if a field has been set.
 func (o *Items) HasWarnings() bool {
-	if o != nil && o.Warnings != nil {
+	if o != nil && !IsNil(o.Warnings) {
 		return true
 	}
 
@@ -134,21 +137,29 @@ func (o *Items) HasWarnings() bool {
 
 // SetWarnings gets a reference to the given []Error and assigns it to the Warnings field.
 func (o *Items) SetWarnings(v []Error) {
-	o.Warnings = &v
+	o.Warnings = v
 }
 
 func (o Items) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.Items != nil {
-		toSerialize["items"] = o.Items
-	}
-	if o.Total != nil {
-		toSerialize["total"] = o.Total
-	}
-	if o.Warnings != nil {
-		toSerialize["warnings"] = o.Warnings
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o Items) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.Items) {
+		toSerialize["items"] = o.Items
+	}
+	if !IsNil(o.Total) {
+		toSerialize["total"] = o.Total
+	}
+	if !IsNil(o.Warnings) {
+		toSerialize["warnings"] = o.Warnings
+	}
+	return toSerialize, nil
 }
 
 type NullableItems struct {
