@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,6 +14,9 @@ import (
 	"encoding/json"
 )
 
+// checks if the LegalAddress type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &LegalAddress{}
+
 // LegalAddress Type that defines the fields for the seller's address.
 type LegalAddress struct {
 	// The first line of the street address.
@@ -22,7 +25,7 @@ type LegalAddress struct {
 	AddressLine2 *string `json:"addressLine2,omitempty"`
 	// The city of the address.
 	City *string `json:"city,omitempty"`
-	// The two-letter <a href=\"https://www.iso.org/iso-3166-country-codes.html\" target=\"_blank\">ISO 3166</a> standard of the country of the address. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/ba:CountryCodeEnum'>eBay API documentation</a>
+	// The two-letter <a href=\"https://www.iso.org/iso-3166-country-codes.html \" target=\"_blank\">ISO 3166</a> standard code for the country of the address. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/ba:CountryCodeEnum'>eBay API documentation</a>
 	Country *string `json:"country,omitempty"`
 	// The name of the country of the address.
 	CountryName *string `json:"countryName,omitempty"`
@@ -53,7 +56,7 @@ func NewLegalAddressWithDefaults() *LegalAddress {
 
 // GetAddressLine1 returns the AddressLine1 field value if set, zero value otherwise.
 func (o *LegalAddress) GetAddressLine1() string {
-	if o == nil || o.AddressLine1 == nil {
+	if o == nil || IsNil(o.AddressLine1) {
 		var ret string
 		return ret
 	}
@@ -63,7 +66,7 @@ func (o *LegalAddress) GetAddressLine1() string {
 // GetAddressLine1Ok returns a tuple with the AddressLine1 field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *LegalAddress) GetAddressLine1Ok() (*string, bool) {
-	if o == nil || o.AddressLine1 == nil {
+	if o == nil || IsNil(o.AddressLine1) {
 		return nil, false
 	}
 	return o.AddressLine1, true
@@ -71,7 +74,7 @@ func (o *LegalAddress) GetAddressLine1Ok() (*string, bool) {
 
 // HasAddressLine1 returns a boolean if a field has been set.
 func (o *LegalAddress) HasAddressLine1() bool {
-	if o != nil && o.AddressLine1 != nil {
+	if o != nil && !IsNil(o.AddressLine1) {
 		return true
 	}
 
@@ -85,7 +88,7 @@ func (o *LegalAddress) SetAddressLine1(v string) {
 
 // GetAddressLine2 returns the AddressLine2 field value if set, zero value otherwise.
 func (o *LegalAddress) GetAddressLine2() string {
-	if o == nil || o.AddressLine2 == nil {
+	if o == nil || IsNil(o.AddressLine2) {
 		var ret string
 		return ret
 	}
@@ -95,7 +98,7 @@ func (o *LegalAddress) GetAddressLine2() string {
 // GetAddressLine2Ok returns a tuple with the AddressLine2 field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *LegalAddress) GetAddressLine2Ok() (*string, bool) {
-	if o == nil || o.AddressLine2 == nil {
+	if o == nil || IsNil(o.AddressLine2) {
 		return nil, false
 	}
 	return o.AddressLine2, true
@@ -103,7 +106,7 @@ func (o *LegalAddress) GetAddressLine2Ok() (*string, bool) {
 
 // HasAddressLine2 returns a boolean if a field has been set.
 func (o *LegalAddress) HasAddressLine2() bool {
-	if o != nil && o.AddressLine2 != nil {
+	if o != nil && !IsNil(o.AddressLine2) {
 		return true
 	}
 
@@ -117,7 +120,7 @@ func (o *LegalAddress) SetAddressLine2(v string) {
 
 // GetCity returns the City field value if set, zero value otherwise.
 func (o *LegalAddress) GetCity() string {
-	if o == nil || o.City == nil {
+	if o == nil || IsNil(o.City) {
 		var ret string
 		return ret
 	}
@@ -127,7 +130,7 @@ func (o *LegalAddress) GetCity() string {
 // GetCityOk returns a tuple with the City field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *LegalAddress) GetCityOk() (*string, bool) {
-	if o == nil || o.City == nil {
+	if o == nil || IsNil(o.City) {
 		return nil, false
 	}
 	return o.City, true
@@ -135,7 +138,7 @@ func (o *LegalAddress) GetCityOk() (*string, bool) {
 
 // HasCity returns a boolean if a field has been set.
 func (o *LegalAddress) HasCity() bool {
-	if o != nil && o.City != nil {
+	if o != nil && !IsNil(o.City) {
 		return true
 	}
 
@@ -149,7 +152,7 @@ func (o *LegalAddress) SetCity(v string) {
 
 // GetCountry returns the Country field value if set, zero value otherwise.
 func (o *LegalAddress) GetCountry() string {
-	if o == nil || o.Country == nil {
+	if o == nil || IsNil(o.Country) {
 		var ret string
 		return ret
 	}
@@ -159,7 +162,7 @@ func (o *LegalAddress) GetCountry() string {
 // GetCountryOk returns a tuple with the Country field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *LegalAddress) GetCountryOk() (*string, bool) {
-	if o == nil || o.Country == nil {
+	if o == nil || IsNil(o.Country) {
 		return nil, false
 	}
 	return o.Country, true
@@ -167,7 +170,7 @@ func (o *LegalAddress) GetCountryOk() (*string, bool) {
 
 // HasCountry returns a boolean if a field has been set.
 func (o *LegalAddress) HasCountry() bool {
-	if o != nil && o.Country != nil {
+	if o != nil && !IsNil(o.Country) {
 		return true
 	}
 
@@ -181,7 +184,7 @@ func (o *LegalAddress) SetCountry(v string) {
 
 // GetCountryName returns the CountryName field value if set, zero value otherwise.
 func (o *LegalAddress) GetCountryName() string {
-	if o == nil || o.CountryName == nil {
+	if o == nil || IsNil(o.CountryName) {
 		var ret string
 		return ret
 	}
@@ -191,7 +194,7 @@ func (o *LegalAddress) GetCountryName() string {
 // GetCountryNameOk returns a tuple with the CountryName field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *LegalAddress) GetCountryNameOk() (*string, bool) {
-	if o == nil || o.CountryName == nil {
+	if o == nil || IsNil(o.CountryName) {
 		return nil, false
 	}
 	return o.CountryName, true
@@ -199,7 +202,7 @@ func (o *LegalAddress) GetCountryNameOk() (*string, bool) {
 
 // HasCountryName returns a boolean if a field has been set.
 func (o *LegalAddress) HasCountryName() bool {
-	if o != nil && o.CountryName != nil {
+	if o != nil && !IsNil(o.CountryName) {
 		return true
 	}
 
@@ -213,7 +216,7 @@ func (o *LegalAddress) SetCountryName(v string) {
 
 // GetCounty returns the County field value if set, zero value otherwise.
 func (o *LegalAddress) GetCounty() string {
-	if o == nil || o.County == nil {
+	if o == nil || IsNil(o.County) {
 		var ret string
 		return ret
 	}
@@ -223,7 +226,7 @@ func (o *LegalAddress) GetCounty() string {
 // GetCountyOk returns a tuple with the County field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *LegalAddress) GetCountyOk() (*string, bool) {
-	if o == nil || o.County == nil {
+	if o == nil || IsNil(o.County) {
 		return nil, false
 	}
 	return o.County, true
@@ -231,7 +234,7 @@ func (o *LegalAddress) GetCountyOk() (*string, bool) {
 
 // HasCounty returns a boolean if a field has been set.
 func (o *LegalAddress) HasCounty() bool {
-	if o != nil && o.County != nil {
+	if o != nil && !IsNil(o.County) {
 		return true
 	}
 
@@ -245,7 +248,7 @@ func (o *LegalAddress) SetCounty(v string) {
 
 // GetPostalCode returns the PostalCode field value if set, zero value otherwise.
 func (o *LegalAddress) GetPostalCode() string {
-	if o == nil || o.PostalCode == nil {
+	if o == nil || IsNil(o.PostalCode) {
 		var ret string
 		return ret
 	}
@@ -255,7 +258,7 @@ func (o *LegalAddress) GetPostalCode() string {
 // GetPostalCodeOk returns a tuple with the PostalCode field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *LegalAddress) GetPostalCodeOk() (*string, bool) {
-	if o == nil || o.PostalCode == nil {
+	if o == nil || IsNil(o.PostalCode) {
 		return nil, false
 	}
 	return o.PostalCode, true
@@ -263,7 +266,7 @@ func (o *LegalAddress) GetPostalCodeOk() (*string, bool) {
 
 // HasPostalCode returns a boolean if a field has been set.
 func (o *LegalAddress) HasPostalCode() bool {
-	if o != nil && o.PostalCode != nil {
+	if o != nil && !IsNil(o.PostalCode) {
 		return true
 	}
 
@@ -277,7 +280,7 @@ func (o *LegalAddress) SetPostalCode(v string) {
 
 // GetStateOrProvince returns the StateOrProvince field value if set, zero value otherwise.
 func (o *LegalAddress) GetStateOrProvince() string {
-	if o == nil || o.StateOrProvince == nil {
+	if o == nil || IsNil(o.StateOrProvince) {
 		var ret string
 		return ret
 	}
@@ -287,7 +290,7 @@ func (o *LegalAddress) GetStateOrProvince() string {
 // GetStateOrProvinceOk returns a tuple with the StateOrProvince field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *LegalAddress) GetStateOrProvinceOk() (*string, bool) {
-	if o == nil || o.StateOrProvince == nil {
+	if o == nil || IsNil(o.StateOrProvince) {
 		return nil, false
 	}
 	return o.StateOrProvince, true
@@ -295,7 +298,7 @@ func (o *LegalAddress) GetStateOrProvinceOk() (*string, bool) {
 
 // HasStateOrProvince returns a boolean if a field has been set.
 func (o *LegalAddress) HasStateOrProvince() bool {
-	if o != nil && o.StateOrProvince != nil {
+	if o != nil && !IsNil(o.StateOrProvince) {
 		return true
 	}
 
@@ -308,32 +311,40 @@ func (o *LegalAddress) SetStateOrProvince(v string) {
 }
 
 func (o LegalAddress) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.AddressLine1 != nil {
-		toSerialize["addressLine1"] = o.AddressLine1
-	}
-	if o.AddressLine2 != nil {
-		toSerialize["addressLine2"] = o.AddressLine2
-	}
-	if o.City != nil {
-		toSerialize["city"] = o.City
-	}
-	if o.Country != nil {
-		toSerialize["country"] = o.Country
-	}
-	if o.CountryName != nil {
-		toSerialize["countryName"] = o.CountryName
-	}
-	if o.County != nil {
-		toSerialize["county"] = o.County
-	}
-	if o.PostalCode != nil {
-		toSerialize["postalCode"] = o.PostalCode
-	}
-	if o.StateOrProvince != nil {
-		toSerialize["stateOrProvince"] = o.StateOrProvince
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o LegalAddress) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.AddressLine1) {
+		toSerialize["addressLine1"] = o.AddressLine1
+	}
+	if !IsNil(o.AddressLine2) {
+		toSerialize["addressLine2"] = o.AddressLine2
+	}
+	if !IsNil(o.City) {
+		toSerialize["city"] = o.City
+	}
+	if !IsNil(o.Country) {
+		toSerialize["country"] = o.Country
+	}
+	if !IsNil(o.CountryName) {
+		toSerialize["countryName"] = o.CountryName
+	}
+	if !IsNil(o.County) {
+		toSerialize["county"] = o.County
+	}
+	if !IsNil(o.PostalCode) {
+		toSerialize["postalCode"] = o.PostalCode
+	}
+	if !IsNil(o.StateOrProvince) {
+		toSerialize["stateOrProvince"] = o.StateOrProvince
+	}
+	return toSerialize, nil
 }
 
 type NullableLegalAddress struct {
