@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,13 +14,16 @@ import (
 	"encoding/json"
 )
 
+// checks if the MarketingPrice type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &MarketingPrice{}
+
 // MarketingPrice The type that defines the fields that describe a seller discount.
 type MarketingPrice struct {
 	DiscountAmount *ConvertedAmount `json:"discountAmount,omitempty"`
-	// This field expresses the percentage of the seller discount based on the value in the <b>  originalPrice</b> container.
+	// This field expresses the percentage of the seller discount based on the value in the <code>originalPrice</code> container.
 	DiscountPercentage *string `json:"discountPercentage,omitempty"`
 	OriginalPrice *ConvertedAmount `json:"originalPrice,omitempty"`
-	// Indicates the pricing treatment (discount) that was applied to the price of the item. <br /><br /><span class=\"tablenote\"><b>Note: </b> The pricing treatment affects the way and where the discounted price can be displayed.</span> For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:PriceTreatmentEnum'>eBay API documentation</a>
+	// Indicates the pricing treatment (discount) that was applied to the price of the item.<br><br><span class=\"tablenote\"><b>Note:</b> The pricing treatment affects the way and where the discounted price can be displayed.</span> For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:PriceTreatmentEnum'>eBay API documentation</a>
 	PriceTreatment *string `json:"priceTreatment,omitempty"`
 }
 
@@ -43,7 +46,7 @@ func NewMarketingPriceWithDefaults() *MarketingPrice {
 
 // GetDiscountAmount returns the DiscountAmount field value if set, zero value otherwise.
 func (o *MarketingPrice) GetDiscountAmount() ConvertedAmount {
-	if o == nil || o.DiscountAmount == nil {
+	if o == nil || IsNil(o.DiscountAmount) {
 		var ret ConvertedAmount
 		return ret
 	}
@@ -53,7 +56,7 @@ func (o *MarketingPrice) GetDiscountAmount() ConvertedAmount {
 // GetDiscountAmountOk returns a tuple with the DiscountAmount field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *MarketingPrice) GetDiscountAmountOk() (*ConvertedAmount, bool) {
-	if o == nil || o.DiscountAmount == nil {
+	if o == nil || IsNil(o.DiscountAmount) {
 		return nil, false
 	}
 	return o.DiscountAmount, true
@@ -61,7 +64,7 @@ func (o *MarketingPrice) GetDiscountAmountOk() (*ConvertedAmount, bool) {
 
 // HasDiscountAmount returns a boolean if a field has been set.
 func (o *MarketingPrice) HasDiscountAmount() bool {
-	if o != nil && o.DiscountAmount != nil {
+	if o != nil && !IsNil(o.DiscountAmount) {
 		return true
 	}
 
@@ -75,7 +78,7 @@ func (o *MarketingPrice) SetDiscountAmount(v ConvertedAmount) {
 
 // GetDiscountPercentage returns the DiscountPercentage field value if set, zero value otherwise.
 func (o *MarketingPrice) GetDiscountPercentage() string {
-	if o == nil || o.DiscountPercentage == nil {
+	if o == nil || IsNil(o.DiscountPercentage) {
 		var ret string
 		return ret
 	}
@@ -85,7 +88,7 @@ func (o *MarketingPrice) GetDiscountPercentage() string {
 // GetDiscountPercentageOk returns a tuple with the DiscountPercentage field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *MarketingPrice) GetDiscountPercentageOk() (*string, bool) {
-	if o == nil || o.DiscountPercentage == nil {
+	if o == nil || IsNil(o.DiscountPercentage) {
 		return nil, false
 	}
 	return o.DiscountPercentage, true
@@ -93,7 +96,7 @@ func (o *MarketingPrice) GetDiscountPercentageOk() (*string, bool) {
 
 // HasDiscountPercentage returns a boolean if a field has been set.
 func (o *MarketingPrice) HasDiscountPercentage() bool {
-	if o != nil && o.DiscountPercentage != nil {
+	if o != nil && !IsNil(o.DiscountPercentage) {
 		return true
 	}
 
@@ -107,7 +110,7 @@ func (o *MarketingPrice) SetDiscountPercentage(v string) {
 
 // GetOriginalPrice returns the OriginalPrice field value if set, zero value otherwise.
 func (o *MarketingPrice) GetOriginalPrice() ConvertedAmount {
-	if o == nil || o.OriginalPrice == nil {
+	if o == nil || IsNil(o.OriginalPrice) {
 		var ret ConvertedAmount
 		return ret
 	}
@@ -117,7 +120,7 @@ func (o *MarketingPrice) GetOriginalPrice() ConvertedAmount {
 // GetOriginalPriceOk returns a tuple with the OriginalPrice field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *MarketingPrice) GetOriginalPriceOk() (*ConvertedAmount, bool) {
-	if o == nil || o.OriginalPrice == nil {
+	if o == nil || IsNil(o.OriginalPrice) {
 		return nil, false
 	}
 	return o.OriginalPrice, true
@@ -125,7 +128,7 @@ func (o *MarketingPrice) GetOriginalPriceOk() (*ConvertedAmount, bool) {
 
 // HasOriginalPrice returns a boolean if a field has been set.
 func (o *MarketingPrice) HasOriginalPrice() bool {
-	if o != nil && o.OriginalPrice != nil {
+	if o != nil && !IsNil(o.OriginalPrice) {
 		return true
 	}
 
@@ -139,7 +142,7 @@ func (o *MarketingPrice) SetOriginalPrice(v ConvertedAmount) {
 
 // GetPriceTreatment returns the PriceTreatment field value if set, zero value otherwise.
 func (o *MarketingPrice) GetPriceTreatment() string {
-	if o == nil || o.PriceTreatment == nil {
+	if o == nil || IsNil(o.PriceTreatment) {
 		var ret string
 		return ret
 	}
@@ -149,7 +152,7 @@ func (o *MarketingPrice) GetPriceTreatment() string {
 // GetPriceTreatmentOk returns a tuple with the PriceTreatment field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *MarketingPrice) GetPriceTreatmentOk() (*string, bool) {
-	if o == nil || o.PriceTreatment == nil {
+	if o == nil || IsNil(o.PriceTreatment) {
 		return nil, false
 	}
 	return o.PriceTreatment, true
@@ -157,7 +160,7 @@ func (o *MarketingPrice) GetPriceTreatmentOk() (*string, bool) {
 
 // HasPriceTreatment returns a boolean if a field has been set.
 func (o *MarketingPrice) HasPriceTreatment() bool {
-	if o != nil && o.PriceTreatment != nil {
+	if o != nil && !IsNil(o.PriceTreatment) {
 		return true
 	}
 
@@ -170,20 +173,28 @@ func (o *MarketingPrice) SetPriceTreatment(v string) {
 }
 
 func (o MarketingPrice) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.DiscountAmount != nil {
-		toSerialize["discountAmount"] = o.DiscountAmount
-	}
-	if o.DiscountPercentage != nil {
-		toSerialize["discountPercentage"] = o.DiscountPercentage
-	}
-	if o.OriginalPrice != nil {
-		toSerialize["originalPrice"] = o.OriginalPrice
-	}
-	if o.PriceTreatment != nil {
-		toSerialize["priceTreatment"] = o.PriceTreatment
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o MarketingPrice) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.DiscountAmount) {
+		toSerialize["discountAmount"] = o.DiscountAmount
+	}
+	if !IsNil(o.DiscountPercentage) {
+		toSerialize["discountPercentage"] = o.DiscountPercentage
+	}
+	if !IsNil(o.OriginalPrice) {
+		toSerialize["originalPrice"] = o.OriginalPrice
+	}
+	if !IsNil(o.PriceTreatment) {
+		toSerialize["priceTreatment"] = o.PriceTreatment
+	}
+	return toSerialize, nil
 }
 
 type NullableMarketingPrice struct {
