@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,16 +14,19 @@ import (
 	"encoding/json"
 )
 
+// checks if the PaymentMethod type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &PaymentMethod{}
+
 // PaymentMethod struct for PaymentMethod
 type PaymentMethod struct {
 	// The payment method type, such as credit card or cash. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:PaymentMethodTypeEnum'>eBay API documentation</a>
 	PaymentMethodType *string `json:"paymentMethodType,omitempty"`
 	// The payment method brands, including the payment method brand type and logo image.
-	PaymentMethodBrands *[]PaymentMethodBrand `json:"paymentMethodBrands,omitempty"`
+	PaymentMethodBrands []PaymentMethodBrand `json:"paymentMethodBrands,omitempty"`
 	// The payment instructions for the buyer, such as <i>cash in person</i> or <i>contact seller</i>.
-	PaymentInstructions *[]string `json:"paymentInstructions,omitempty"`
+	PaymentInstructions []string `json:"paymentInstructions,omitempty"`
 	// The seller instructions to the buyer, such as <i>accepts credit cards</i> or <i>see description</i>.
-	SellerInstructions *[]string `json:"sellerInstructions,omitempty"`
+	SellerInstructions []string `json:"sellerInstructions,omitempty"`
 }
 
 // NewPaymentMethod instantiates a new PaymentMethod object
@@ -45,7 +48,7 @@ func NewPaymentMethodWithDefaults() *PaymentMethod {
 
 // GetPaymentMethodType returns the PaymentMethodType field value if set, zero value otherwise.
 func (o *PaymentMethod) GetPaymentMethodType() string {
-	if o == nil || o.PaymentMethodType == nil {
+	if o == nil || IsNil(o.PaymentMethodType) {
 		var ret string
 		return ret
 	}
@@ -55,7 +58,7 @@ func (o *PaymentMethod) GetPaymentMethodType() string {
 // GetPaymentMethodTypeOk returns a tuple with the PaymentMethodType field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *PaymentMethod) GetPaymentMethodTypeOk() (*string, bool) {
-	if o == nil || o.PaymentMethodType == nil {
+	if o == nil || IsNil(o.PaymentMethodType) {
 		return nil, false
 	}
 	return o.PaymentMethodType, true
@@ -63,7 +66,7 @@ func (o *PaymentMethod) GetPaymentMethodTypeOk() (*string, bool) {
 
 // HasPaymentMethodType returns a boolean if a field has been set.
 func (o *PaymentMethod) HasPaymentMethodType() bool {
-	if o != nil && o.PaymentMethodType != nil {
+	if o != nil && !IsNil(o.PaymentMethodType) {
 		return true
 	}
 
@@ -77,17 +80,17 @@ func (o *PaymentMethod) SetPaymentMethodType(v string) {
 
 // GetPaymentMethodBrands returns the PaymentMethodBrands field value if set, zero value otherwise.
 func (o *PaymentMethod) GetPaymentMethodBrands() []PaymentMethodBrand {
-	if o == nil || o.PaymentMethodBrands == nil {
+	if o == nil || IsNil(o.PaymentMethodBrands) {
 		var ret []PaymentMethodBrand
 		return ret
 	}
-	return *o.PaymentMethodBrands
+	return o.PaymentMethodBrands
 }
 
 // GetPaymentMethodBrandsOk returns a tuple with the PaymentMethodBrands field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *PaymentMethod) GetPaymentMethodBrandsOk() (*[]PaymentMethodBrand, bool) {
-	if o == nil || o.PaymentMethodBrands == nil {
+func (o *PaymentMethod) GetPaymentMethodBrandsOk() ([]PaymentMethodBrand, bool) {
+	if o == nil || IsNil(o.PaymentMethodBrands) {
 		return nil, false
 	}
 	return o.PaymentMethodBrands, true
@@ -95,7 +98,7 @@ func (o *PaymentMethod) GetPaymentMethodBrandsOk() (*[]PaymentMethodBrand, bool)
 
 // HasPaymentMethodBrands returns a boolean if a field has been set.
 func (o *PaymentMethod) HasPaymentMethodBrands() bool {
-	if o != nil && o.PaymentMethodBrands != nil {
+	if o != nil && !IsNil(o.PaymentMethodBrands) {
 		return true
 	}
 
@@ -104,22 +107,22 @@ func (o *PaymentMethod) HasPaymentMethodBrands() bool {
 
 // SetPaymentMethodBrands gets a reference to the given []PaymentMethodBrand and assigns it to the PaymentMethodBrands field.
 func (o *PaymentMethod) SetPaymentMethodBrands(v []PaymentMethodBrand) {
-	o.PaymentMethodBrands = &v
+	o.PaymentMethodBrands = v
 }
 
 // GetPaymentInstructions returns the PaymentInstructions field value if set, zero value otherwise.
 func (o *PaymentMethod) GetPaymentInstructions() []string {
-	if o == nil || o.PaymentInstructions == nil {
+	if o == nil || IsNil(o.PaymentInstructions) {
 		var ret []string
 		return ret
 	}
-	return *o.PaymentInstructions
+	return o.PaymentInstructions
 }
 
 // GetPaymentInstructionsOk returns a tuple with the PaymentInstructions field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *PaymentMethod) GetPaymentInstructionsOk() (*[]string, bool) {
-	if o == nil || o.PaymentInstructions == nil {
+func (o *PaymentMethod) GetPaymentInstructionsOk() ([]string, bool) {
+	if o == nil || IsNil(o.PaymentInstructions) {
 		return nil, false
 	}
 	return o.PaymentInstructions, true
@@ -127,7 +130,7 @@ func (o *PaymentMethod) GetPaymentInstructionsOk() (*[]string, bool) {
 
 // HasPaymentInstructions returns a boolean if a field has been set.
 func (o *PaymentMethod) HasPaymentInstructions() bool {
-	if o != nil && o.PaymentInstructions != nil {
+	if o != nil && !IsNil(o.PaymentInstructions) {
 		return true
 	}
 
@@ -136,22 +139,22 @@ func (o *PaymentMethod) HasPaymentInstructions() bool {
 
 // SetPaymentInstructions gets a reference to the given []string and assigns it to the PaymentInstructions field.
 func (o *PaymentMethod) SetPaymentInstructions(v []string) {
-	o.PaymentInstructions = &v
+	o.PaymentInstructions = v
 }
 
 // GetSellerInstructions returns the SellerInstructions field value if set, zero value otherwise.
 func (o *PaymentMethod) GetSellerInstructions() []string {
-	if o == nil || o.SellerInstructions == nil {
+	if o == nil || IsNil(o.SellerInstructions) {
 		var ret []string
 		return ret
 	}
-	return *o.SellerInstructions
+	return o.SellerInstructions
 }
 
 // GetSellerInstructionsOk returns a tuple with the SellerInstructions field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *PaymentMethod) GetSellerInstructionsOk() (*[]string, bool) {
-	if o == nil || o.SellerInstructions == nil {
+func (o *PaymentMethod) GetSellerInstructionsOk() ([]string, bool) {
+	if o == nil || IsNil(o.SellerInstructions) {
 		return nil, false
 	}
 	return o.SellerInstructions, true
@@ -159,7 +162,7 @@ func (o *PaymentMethod) GetSellerInstructionsOk() (*[]string, bool) {
 
 // HasSellerInstructions returns a boolean if a field has been set.
 func (o *PaymentMethod) HasSellerInstructions() bool {
-	if o != nil && o.SellerInstructions != nil {
+	if o != nil && !IsNil(o.SellerInstructions) {
 		return true
 	}
 
@@ -168,24 +171,32 @@ func (o *PaymentMethod) HasSellerInstructions() bool {
 
 // SetSellerInstructions gets a reference to the given []string and assigns it to the SellerInstructions field.
 func (o *PaymentMethod) SetSellerInstructions(v []string) {
-	o.SellerInstructions = &v
+	o.SellerInstructions = v
 }
 
 func (o PaymentMethod) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.PaymentMethodType != nil {
-		toSerialize["paymentMethodType"] = o.PaymentMethodType
-	}
-	if o.PaymentMethodBrands != nil {
-		toSerialize["paymentMethodBrands"] = o.PaymentMethodBrands
-	}
-	if o.PaymentInstructions != nil {
-		toSerialize["paymentInstructions"] = o.PaymentInstructions
-	}
-	if o.SellerInstructions != nil {
-		toSerialize["sellerInstructions"] = o.SellerInstructions
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o PaymentMethod) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.PaymentMethodType) {
+		toSerialize["paymentMethodType"] = o.PaymentMethodType
+	}
+	if !IsNil(o.PaymentMethodBrands) {
+		toSerialize["paymentMethodBrands"] = o.PaymentMethodBrands
+	}
+	if !IsNil(o.PaymentInstructions) {
+		toSerialize["paymentInstructions"] = o.PaymentInstructions
+	}
+	if !IsNil(o.SellerInstructions) {
+		toSerialize["sellerInstructions"] = o.SellerInstructions
+	}
+	return toSerialize, nil
 }
 
 type NullablePaymentMethod struct {
