@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -13,6 +13,9 @@ package buybrowse
 import (
 	"encoding/json"
 )
+
+// checks if the PaymentMethodBrand type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &PaymentMethodBrand{}
 
 // PaymentMethodBrand struct for PaymentMethodBrand
 type PaymentMethodBrand struct {
@@ -40,7 +43,7 @@ func NewPaymentMethodBrandWithDefaults() *PaymentMethodBrand {
 
 // GetPaymentMethodBrandType returns the PaymentMethodBrandType field value if set, zero value otherwise.
 func (o *PaymentMethodBrand) GetPaymentMethodBrandType() string {
-	if o == nil || o.PaymentMethodBrandType == nil {
+	if o == nil || IsNil(o.PaymentMethodBrandType) {
 		var ret string
 		return ret
 	}
@@ -50,7 +53,7 @@ func (o *PaymentMethodBrand) GetPaymentMethodBrandType() string {
 // GetPaymentMethodBrandTypeOk returns a tuple with the PaymentMethodBrandType field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *PaymentMethodBrand) GetPaymentMethodBrandTypeOk() (*string, bool) {
-	if o == nil || o.PaymentMethodBrandType == nil {
+	if o == nil || IsNil(o.PaymentMethodBrandType) {
 		return nil, false
 	}
 	return o.PaymentMethodBrandType, true
@@ -58,7 +61,7 @@ func (o *PaymentMethodBrand) GetPaymentMethodBrandTypeOk() (*string, bool) {
 
 // HasPaymentMethodBrandType returns a boolean if a field has been set.
 func (o *PaymentMethodBrand) HasPaymentMethodBrandType() bool {
-	if o != nil && o.PaymentMethodBrandType != nil {
+	if o != nil && !IsNil(o.PaymentMethodBrandType) {
 		return true
 	}
 
@@ -72,7 +75,7 @@ func (o *PaymentMethodBrand) SetPaymentMethodBrandType(v string) {
 
 // GetLogoImage returns the LogoImage field value if set, zero value otherwise.
 func (o *PaymentMethodBrand) GetLogoImage() Image {
-	if o == nil || o.LogoImage == nil {
+	if o == nil || IsNil(o.LogoImage) {
 		var ret Image
 		return ret
 	}
@@ -82,7 +85,7 @@ func (o *PaymentMethodBrand) GetLogoImage() Image {
 // GetLogoImageOk returns a tuple with the LogoImage field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *PaymentMethodBrand) GetLogoImageOk() (*Image, bool) {
-	if o == nil || o.LogoImage == nil {
+	if o == nil || IsNil(o.LogoImage) {
 		return nil, false
 	}
 	return o.LogoImage, true
@@ -90,7 +93,7 @@ func (o *PaymentMethodBrand) GetLogoImageOk() (*Image, bool) {
 
 // HasLogoImage returns a boolean if a field has been set.
 func (o *PaymentMethodBrand) HasLogoImage() bool {
-	if o != nil && o.LogoImage != nil {
+	if o != nil && !IsNil(o.LogoImage) {
 		return true
 	}
 
@@ -103,14 +106,22 @@ func (o *PaymentMethodBrand) SetLogoImage(v Image) {
 }
 
 func (o PaymentMethodBrand) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.PaymentMethodBrandType != nil {
-		toSerialize["paymentMethodBrandType"] = o.PaymentMethodBrandType
-	}
-	if o.LogoImage != nil {
-		toSerialize["logoImage"] = o.LogoImage
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o PaymentMethodBrand) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.PaymentMethodBrandType) {
+		toSerialize["paymentMethodBrandType"] = o.PaymentMethodBrandType
+	}
+	if !IsNil(o.LogoImage) {
+		toSerialize["logoImage"] = o.LogoImage
+	}
+	return toSerialize, nil
 }
 
 type NullablePaymentMethodBrand struct {
