@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,7 +14,10 @@ import (
 	"encoding/json"
 )
 
-// PickupOptionSummary The type that defines the fields for the local pickup options that are available for the item. It is used by the <b>  pickupOptions</b>  container.
+// checks if the PickupOptionSummary type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &PickupOptionSummary{}
+
+// PickupOptionSummary The type that defines the fields for the local pickup options that are available for the item. It is used by the <code>pickupOptions</code> container.
 type PickupOptionSummary struct {
 	// This container returns the local pickup options available to the buyer. Possible values are <code>ARRANGED_LOCATION</code> and <code>STORE</code>.
 	PickupLocationType *string `json:"pickupLocationType,omitempty"`
@@ -39,7 +42,7 @@ func NewPickupOptionSummaryWithDefaults() *PickupOptionSummary {
 
 // GetPickupLocationType returns the PickupLocationType field value if set, zero value otherwise.
 func (o *PickupOptionSummary) GetPickupLocationType() string {
-	if o == nil || o.PickupLocationType == nil {
+	if o == nil || IsNil(o.PickupLocationType) {
 		var ret string
 		return ret
 	}
@@ -49,7 +52,7 @@ func (o *PickupOptionSummary) GetPickupLocationType() string {
 // GetPickupLocationTypeOk returns a tuple with the PickupLocationType field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *PickupOptionSummary) GetPickupLocationTypeOk() (*string, bool) {
-	if o == nil || o.PickupLocationType == nil {
+	if o == nil || IsNil(o.PickupLocationType) {
 		return nil, false
 	}
 	return o.PickupLocationType, true
@@ -57,7 +60,7 @@ func (o *PickupOptionSummary) GetPickupLocationTypeOk() (*string, bool) {
 
 // HasPickupLocationType returns a boolean if a field has been set.
 func (o *PickupOptionSummary) HasPickupLocationType() bool {
-	if o != nil && o.PickupLocationType != nil {
+	if o != nil && !IsNil(o.PickupLocationType) {
 		return true
 	}
 
@@ -70,11 +73,19 @@ func (o *PickupOptionSummary) SetPickupLocationType(v string) {
 }
 
 func (o PickupOptionSummary) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.PickupLocationType != nil {
-		toSerialize["pickupLocationType"] = o.PickupLocationType
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o PickupOptionSummary) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.PickupLocationType) {
+		toSerialize["pickupLocationType"] = o.PickupLocationType
+	}
+	return toSerialize, nil
 }
 
 type NullablePickupOptionSummary struct {
