@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,23 +14,26 @@ import (
 	"encoding/json"
 )
 
+// checks if the Product type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &Product{}
+
 // Product The type that defines the fields for the product information of the item.
 type Product struct {
 	// An array of containers with the URLs for the product images that are in addition to the primary image. 
-	AdditionalImages *[]Image `json:"additionalImages,omitempty"`
+	AdditionalImages []Image `json:"additionalImages,omitempty"`
 	// An array of product identifiers associated with the item. This container is returned if the seller has associated the eBay Product Identifier (ePID) with the item and in the request <b> fieldgroups</b> is set to <code>PRODUCT</code>.
-	AdditionalProductIdentities *[]AdditionalProductIdentity `json:"additionalProductIdentities,omitempty"`
+	AdditionalProductIdentities []AdditionalProductIdentity `json:"additionalProductIdentities,omitempty"`
 	// An array of containers for the product aspects. Each group contains the aspect group name and the aspect name/value pairs.
-	AspectGroups *[]AspectGroup `json:"aspectGroups,omitempty"`
+	AspectGroups []AspectGroup `json:"aspectGroups,omitempty"`
 	// The brand associated with product. To identify the product, this is always used along with MPN (manufacturer part number).
 	Brand *string `json:"brand,omitempty"`
 	// The rich description of an eBay product, which might contain HTML.
 	Description *string `json:"description,omitempty"`
-	// An array of all the possible GTINs values associated with the product. A GTIN is a unique Global Trade Item number of the item as defined by <a href=\"https://www.gtin.info\" target=\"_blank\">https://www.gtin.info</a>. This can be a UPC (Universal Product Code), EAN (European Article Number), or an ISBN (International Standard Book Number) value.
-	Gtins *[]string `json:"gtins,omitempty"`
+	// An array of all the possible GTINs values associated with the product. A GTIN is a unique Global Trade Item number of the item as defined by <a href=\"https://www.gtin.info \" target=\"_blank\">https://www.gtin.info</a>. This can be a UPC (Universal Product Code), EAN (European Article Number), or an ISBN (International Standard Book Number) value.
+	Gtins []string `json:"gtins,omitempty"`
 	Image *Image `json:"image,omitempty"`
 	// An array of all possible MPN values associated with the product. A MPNs is manufacturer part number of the product. To identify the product, this is always used along with brand.
-	Mpns *[]string `json:"mpns,omitempty"`
+	Mpns []string `json:"mpns,omitempty"`
 	// The title of the product.
 	Title *string `json:"title,omitempty"`
 }
@@ -54,17 +57,17 @@ func NewProductWithDefaults() *Product {
 
 // GetAdditionalImages returns the AdditionalImages field value if set, zero value otherwise.
 func (o *Product) GetAdditionalImages() []Image {
-	if o == nil || o.AdditionalImages == nil {
+	if o == nil || IsNil(o.AdditionalImages) {
 		var ret []Image
 		return ret
 	}
-	return *o.AdditionalImages
+	return o.AdditionalImages
 }
 
 // GetAdditionalImagesOk returns a tuple with the AdditionalImages field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Product) GetAdditionalImagesOk() (*[]Image, bool) {
-	if o == nil || o.AdditionalImages == nil {
+func (o *Product) GetAdditionalImagesOk() ([]Image, bool) {
+	if o == nil || IsNil(o.AdditionalImages) {
 		return nil, false
 	}
 	return o.AdditionalImages, true
@@ -72,7 +75,7 @@ func (o *Product) GetAdditionalImagesOk() (*[]Image, bool) {
 
 // HasAdditionalImages returns a boolean if a field has been set.
 func (o *Product) HasAdditionalImages() bool {
-	if o != nil && o.AdditionalImages != nil {
+	if o != nil && !IsNil(o.AdditionalImages) {
 		return true
 	}
 
@@ -81,22 +84,22 @@ func (o *Product) HasAdditionalImages() bool {
 
 // SetAdditionalImages gets a reference to the given []Image and assigns it to the AdditionalImages field.
 func (o *Product) SetAdditionalImages(v []Image) {
-	o.AdditionalImages = &v
+	o.AdditionalImages = v
 }
 
 // GetAdditionalProductIdentities returns the AdditionalProductIdentities field value if set, zero value otherwise.
 func (o *Product) GetAdditionalProductIdentities() []AdditionalProductIdentity {
-	if o == nil || o.AdditionalProductIdentities == nil {
+	if o == nil || IsNil(o.AdditionalProductIdentities) {
 		var ret []AdditionalProductIdentity
 		return ret
 	}
-	return *o.AdditionalProductIdentities
+	return o.AdditionalProductIdentities
 }
 
 // GetAdditionalProductIdentitiesOk returns a tuple with the AdditionalProductIdentities field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Product) GetAdditionalProductIdentitiesOk() (*[]AdditionalProductIdentity, bool) {
-	if o == nil || o.AdditionalProductIdentities == nil {
+func (o *Product) GetAdditionalProductIdentitiesOk() ([]AdditionalProductIdentity, bool) {
+	if o == nil || IsNil(o.AdditionalProductIdentities) {
 		return nil, false
 	}
 	return o.AdditionalProductIdentities, true
@@ -104,7 +107,7 @@ func (o *Product) GetAdditionalProductIdentitiesOk() (*[]AdditionalProductIdenti
 
 // HasAdditionalProductIdentities returns a boolean if a field has been set.
 func (o *Product) HasAdditionalProductIdentities() bool {
-	if o != nil && o.AdditionalProductIdentities != nil {
+	if o != nil && !IsNil(o.AdditionalProductIdentities) {
 		return true
 	}
 
@@ -113,22 +116,22 @@ func (o *Product) HasAdditionalProductIdentities() bool {
 
 // SetAdditionalProductIdentities gets a reference to the given []AdditionalProductIdentity and assigns it to the AdditionalProductIdentities field.
 func (o *Product) SetAdditionalProductIdentities(v []AdditionalProductIdentity) {
-	o.AdditionalProductIdentities = &v
+	o.AdditionalProductIdentities = v
 }
 
 // GetAspectGroups returns the AspectGroups field value if set, zero value otherwise.
 func (o *Product) GetAspectGroups() []AspectGroup {
-	if o == nil || o.AspectGroups == nil {
+	if o == nil || IsNil(o.AspectGroups) {
 		var ret []AspectGroup
 		return ret
 	}
-	return *o.AspectGroups
+	return o.AspectGroups
 }
 
 // GetAspectGroupsOk returns a tuple with the AspectGroups field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Product) GetAspectGroupsOk() (*[]AspectGroup, bool) {
-	if o == nil || o.AspectGroups == nil {
+func (o *Product) GetAspectGroupsOk() ([]AspectGroup, bool) {
+	if o == nil || IsNil(o.AspectGroups) {
 		return nil, false
 	}
 	return o.AspectGroups, true
@@ -136,7 +139,7 @@ func (o *Product) GetAspectGroupsOk() (*[]AspectGroup, bool) {
 
 // HasAspectGroups returns a boolean if a field has been set.
 func (o *Product) HasAspectGroups() bool {
-	if o != nil && o.AspectGroups != nil {
+	if o != nil && !IsNil(o.AspectGroups) {
 		return true
 	}
 
@@ -145,12 +148,12 @@ func (o *Product) HasAspectGroups() bool {
 
 // SetAspectGroups gets a reference to the given []AspectGroup and assigns it to the AspectGroups field.
 func (o *Product) SetAspectGroups(v []AspectGroup) {
-	o.AspectGroups = &v
+	o.AspectGroups = v
 }
 
 // GetBrand returns the Brand field value if set, zero value otherwise.
 func (o *Product) GetBrand() string {
-	if o == nil || o.Brand == nil {
+	if o == nil || IsNil(o.Brand) {
 		var ret string
 		return ret
 	}
@@ -160,7 +163,7 @@ func (o *Product) GetBrand() string {
 // GetBrandOk returns a tuple with the Brand field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Product) GetBrandOk() (*string, bool) {
-	if o == nil || o.Brand == nil {
+	if o == nil || IsNil(o.Brand) {
 		return nil, false
 	}
 	return o.Brand, true
@@ -168,7 +171,7 @@ func (o *Product) GetBrandOk() (*string, bool) {
 
 // HasBrand returns a boolean if a field has been set.
 func (o *Product) HasBrand() bool {
-	if o != nil && o.Brand != nil {
+	if o != nil && !IsNil(o.Brand) {
 		return true
 	}
 
@@ -182,7 +185,7 @@ func (o *Product) SetBrand(v string) {
 
 // GetDescription returns the Description field value if set, zero value otherwise.
 func (o *Product) GetDescription() string {
-	if o == nil || o.Description == nil {
+	if o == nil || IsNil(o.Description) {
 		var ret string
 		return ret
 	}
@@ -192,7 +195,7 @@ func (o *Product) GetDescription() string {
 // GetDescriptionOk returns a tuple with the Description field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Product) GetDescriptionOk() (*string, bool) {
-	if o == nil || o.Description == nil {
+	if o == nil || IsNil(o.Description) {
 		return nil, false
 	}
 	return o.Description, true
@@ -200,7 +203,7 @@ func (o *Product) GetDescriptionOk() (*string, bool) {
 
 // HasDescription returns a boolean if a field has been set.
 func (o *Product) HasDescription() bool {
-	if o != nil && o.Description != nil {
+	if o != nil && !IsNil(o.Description) {
 		return true
 	}
 
@@ -214,17 +217,17 @@ func (o *Product) SetDescription(v string) {
 
 // GetGtins returns the Gtins field value if set, zero value otherwise.
 func (o *Product) GetGtins() []string {
-	if o == nil || o.Gtins == nil {
+	if o == nil || IsNil(o.Gtins) {
 		var ret []string
 		return ret
 	}
-	return *o.Gtins
+	return o.Gtins
 }
 
 // GetGtinsOk returns a tuple with the Gtins field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Product) GetGtinsOk() (*[]string, bool) {
-	if o == nil || o.Gtins == nil {
+func (o *Product) GetGtinsOk() ([]string, bool) {
+	if o == nil || IsNil(o.Gtins) {
 		return nil, false
 	}
 	return o.Gtins, true
@@ -232,7 +235,7 @@ func (o *Product) GetGtinsOk() (*[]string, bool) {
 
 // HasGtins returns a boolean if a field has been set.
 func (o *Product) HasGtins() bool {
-	if o != nil && o.Gtins != nil {
+	if o != nil && !IsNil(o.Gtins) {
 		return true
 	}
 
@@ -241,12 +244,12 @@ func (o *Product) HasGtins() bool {
 
 // SetGtins gets a reference to the given []string and assigns it to the Gtins field.
 func (o *Product) SetGtins(v []string) {
-	o.Gtins = &v
+	o.Gtins = v
 }
 
 // GetImage returns the Image field value if set, zero value otherwise.
 func (o *Product) GetImage() Image {
-	if o == nil || o.Image == nil {
+	if o == nil || IsNil(o.Image) {
 		var ret Image
 		return ret
 	}
@@ -256,7 +259,7 @@ func (o *Product) GetImage() Image {
 // GetImageOk returns a tuple with the Image field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Product) GetImageOk() (*Image, bool) {
-	if o == nil || o.Image == nil {
+	if o == nil || IsNil(o.Image) {
 		return nil, false
 	}
 	return o.Image, true
@@ -264,7 +267,7 @@ func (o *Product) GetImageOk() (*Image, bool) {
 
 // HasImage returns a boolean if a field has been set.
 func (o *Product) HasImage() bool {
-	if o != nil && o.Image != nil {
+	if o != nil && !IsNil(o.Image) {
 		return true
 	}
 
@@ -278,17 +281,17 @@ func (o *Product) SetImage(v Image) {
 
 // GetMpns returns the Mpns field value if set, zero value otherwise.
 func (o *Product) GetMpns() []string {
-	if o == nil || o.Mpns == nil {
+	if o == nil || IsNil(o.Mpns) {
 		var ret []string
 		return ret
 	}
-	return *o.Mpns
+	return o.Mpns
 }
 
 // GetMpnsOk returns a tuple with the Mpns field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Product) GetMpnsOk() (*[]string, bool) {
-	if o == nil || o.Mpns == nil {
+func (o *Product) GetMpnsOk() ([]string, bool) {
+	if o == nil || IsNil(o.Mpns) {
 		return nil, false
 	}
 	return o.Mpns, true
@@ -296,7 +299,7 @@ func (o *Product) GetMpnsOk() (*[]string, bool) {
 
 // HasMpns returns a boolean if a field has been set.
 func (o *Product) HasMpns() bool {
-	if o != nil && o.Mpns != nil {
+	if o != nil && !IsNil(o.Mpns) {
 		return true
 	}
 
@@ -305,12 +308,12 @@ func (o *Product) HasMpns() bool {
 
 // SetMpns gets a reference to the given []string and assigns it to the Mpns field.
 func (o *Product) SetMpns(v []string) {
-	o.Mpns = &v
+	o.Mpns = v
 }
 
 // GetTitle returns the Title field value if set, zero value otherwise.
 func (o *Product) GetTitle() string {
-	if o == nil || o.Title == nil {
+	if o == nil || IsNil(o.Title) {
 		var ret string
 		return ret
 	}
@@ -320,7 +323,7 @@ func (o *Product) GetTitle() string {
 // GetTitleOk returns a tuple with the Title field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Product) GetTitleOk() (*string, bool) {
-	if o == nil || o.Title == nil {
+	if o == nil || IsNil(o.Title) {
 		return nil, false
 	}
 	return o.Title, true
@@ -328,7 +331,7 @@ func (o *Product) GetTitleOk() (*string, bool) {
 
 // HasTitle returns a boolean if a field has been set.
 func (o *Product) HasTitle() bool {
-	if o != nil && o.Title != nil {
+	if o != nil && !IsNil(o.Title) {
 		return true
 	}
 
@@ -341,35 +344,43 @@ func (o *Product) SetTitle(v string) {
 }
 
 func (o Product) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.AdditionalImages != nil {
-		toSerialize["additionalImages"] = o.AdditionalImages
-	}
-	if o.AdditionalProductIdentities != nil {
-		toSerialize["additionalProductIdentities"] = o.AdditionalProductIdentities
-	}
-	if o.AspectGroups != nil {
-		toSerialize["aspectGroups"] = o.AspectGroups
-	}
-	if o.Brand != nil {
-		toSerialize["brand"] = o.Brand
-	}
-	if o.Description != nil {
-		toSerialize["description"] = o.Description
-	}
-	if o.Gtins != nil {
-		toSerialize["gtins"] = o.Gtins
-	}
-	if o.Image != nil {
-		toSerialize["image"] = o.Image
-	}
-	if o.Mpns != nil {
-		toSerialize["mpns"] = o.Mpns
-	}
-	if o.Title != nil {
-		toSerialize["title"] = o.Title
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o Product) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.AdditionalImages) {
+		toSerialize["additionalImages"] = o.AdditionalImages
+	}
+	if !IsNil(o.AdditionalProductIdentities) {
+		toSerialize["additionalProductIdentities"] = o.AdditionalProductIdentities
+	}
+	if !IsNil(o.AspectGroups) {
+		toSerialize["aspectGroups"] = o.AspectGroups
+	}
+	if !IsNil(o.Brand) {
+		toSerialize["brand"] = o.Brand
+	}
+	if !IsNil(o.Description) {
+		toSerialize["description"] = o.Description
+	}
+	if !IsNil(o.Gtins) {
+		toSerialize["gtins"] = o.Gtins
+	}
+	if !IsNil(o.Image) {
+		toSerialize["image"] = o.Image
+	}
+	if !IsNil(o.Mpns) {
+		toSerialize["mpns"] = o.Mpns
+	}
+	if !IsNil(o.Title) {
+		toSerialize["title"] = o.Title
+	}
+	return toSerialize, nil
 }
 
 type NullableProduct struct {
