@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,16 +14,19 @@ import (
 	"encoding/json"
 )
 
+// checks if the Refinement type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &Refinement{}
+
 // Refinement This type defines the fields for the various refinements of an item. You can use the information in this container to create histograms, which help shoppers choose exactly what they want.
 type Refinement struct {
 	// An array of containers for the all the aspect refinements.
-	AspectDistributions *[]AspectDistribution `json:"aspectDistributions,omitempty"`
+	AspectDistributions []AspectDistribution `json:"aspectDistributions,omitempty"`
 	// An array of containers for the all the buying option refinements.
-	BuyingOptionDistributions *[]BuyingOptionDistribution `json:"buyingOptionDistributions,omitempty"`
+	BuyingOptionDistributions []BuyingOptionDistribution `json:"buyingOptionDistributions,omitempty"`
 	// An array of containers for the all the category refinements.
-	CategoryDistributions *[]CategoryDistribution `json:"categoryDistributions,omitempty"`
+	CategoryDistributions []CategoryDistribution `json:"categoryDistributions,omitempty"`
 	// An array of containers for the all the condition refinements.
-	ConditionDistributions *[]ConditionDistribution `json:"conditionDistributions,omitempty"`
+	ConditionDistributions []ConditionDistribution `json:"conditionDistributions,omitempty"`
 	// The identifier of the category that most of the items are part of. 
 	DominantCategoryId *string `json:"dominantCategoryId,omitempty"`
 }
@@ -47,17 +50,17 @@ func NewRefinementWithDefaults() *Refinement {
 
 // GetAspectDistributions returns the AspectDistributions field value if set, zero value otherwise.
 func (o *Refinement) GetAspectDistributions() []AspectDistribution {
-	if o == nil || o.AspectDistributions == nil {
+	if o == nil || IsNil(o.AspectDistributions) {
 		var ret []AspectDistribution
 		return ret
 	}
-	return *o.AspectDistributions
+	return o.AspectDistributions
 }
 
 // GetAspectDistributionsOk returns a tuple with the AspectDistributions field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Refinement) GetAspectDistributionsOk() (*[]AspectDistribution, bool) {
-	if o == nil || o.AspectDistributions == nil {
+func (o *Refinement) GetAspectDistributionsOk() ([]AspectDistribution, bool) {
+	if o == nil || IsNil(o.AspectDistributions) {
 		return nil, false
 	}
 	return o.AspectDistributions, true
@@ -65,7 +68,7 @@ func (o *Refinement) GetAspectDistributionsOk() (*[]AspectDistribution, bool) {
 
 // HasAspectDistributions returns a boolean if a field has been set.
 func (o *Refinement) HasAspectDistributions() bool {
-	if o != nil && o.AspectDistributions != nil {
+	if o != nil && !IsNil(o.AspectDistributions) {
 		return true
 	}
 
@@ -74,22 +77,22 @@ func (o *Refinement) HasAspectDistributions() bool {
 
 // SetAspectDistributions gets a reference to the given []AspectDistribution and assigns it to the AspectDistributions field.
 func (o *Refinement) SetAspectDistributions(v []AspectDistribution) {
-	o.AspectDistributions = &v
+	o.AspectDistributions = v
 }
 
 // GetBuyingOptionDistributions returns the BuyingOptionDistributions field value if set, zero value otherwise.
 func (o *Refinement) GetBuyingOptionDistributions() []BuyingOptionDistribution {
-	if o == nil || o.BuyingOptionDistributions == nil {
+	if o == nil || IsNil(o.BuyingOptionDistributions) {
 		var ret []BuyingOptionDistribution
 		return ret
 	}
-	return *o.BuyingOptionDistributions
+	return o.BuyingOptionDistributions
 }
 
 // GetBuyingOptionDistributionsOk returns a tuple with the BuyingOptionDistributions field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Refinement) GetBuyingOptionDistributionsOk() (*[]BuyingOptionDistribution, bool) {
-	if o == nil || o.BuyingOptionDistributions == nil {
+func (o *Refinement) GetBuyingOptionDistributionsOk() ([]BuyingOptionDistribution, bool) {
+	if o == nil || IsNil(o.BuyingOptionDistributions) {
 		return nil, false
 	}
 	return o.BuyingOptionDistributions, true
@@ -97,7 +100,7 @@ func (o *Refinement) GetBuyingOptionDistributionsOk() (*[]BuyingOptionDistributi
 
 // HasBuyingOptionDistributions returns a boolean if a field has been set.
 func (o *Refinement) HasBuyingOptionDistributions() bool {
-	if o != nil && o.BuyingOptionDistributions != nil {
+	if o != nil && !IsNil(o.BuyingOptionDistributions) {
 		return true
 	}
 
@@ -106,22 +109,22 @@ func (o *Refinement) HasBuyingOptionDistributions() bool {
 
 // SetBuyingOptionDistributions gets a reference to the given []BuyingOptionDistribution and assigns it to the BuyingOptionDistributions field.
 func (o *Refinement) SetBuyingOptionDistributions(v []BuyingOptionDistribution) {
-	o.BuyingOptionDistributions = &v
+	o.BuyingOptionDistributions = v
 }
 
 // GetCategoryDistributions returns the CategoryDistributions field value if set, zero value otherwise.
 func (o *Refinement) GetCategoryDistributions() []CategoryDistribution {
-	if o == nil || o.CategoryDistributions == nil {
+	if o == nil || IsNil(o.CategoryDistributions) {
 		var ret []CategoryDistribution
 		return ret
 	}
-	return *o.CategoryDistributions
+	return o.CategoryDistributions
 }
 
 // GetCategoryDistributionsOk returns a tuple with the CategoryDistributions field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Refinement) GetCategoryDistributionsOk() (*[]CategoryDistribution, bool) {
-	if o == nil || o.CategoryDistributions == nil {
+func (o *Refinement) GetCategoryDistributionsOk() ([]CategoryDistribution, bool) {
+	if o == nil || IsNil(o.CategoryDistributions) {
 		return nil, false
 	}
 	return o.CategoryDistributions, true
@@ -129,7 +132,7 @@ func (o *Refinement) GetCategoryDistributionsOk() (*[]CategoryDistribution, bool
 
 // HasCategoryDistributions returns a boolean if a field has been set.
 func (o *Refinement) HasCategoryDistributions() bool {
-	if o != nil && o.CategoryDistributions != nil {
+	if o != nil && !IsNil(o.CategoryDistributions) {
 		return true
 	}
 
@@ -138,22 +141,22 @@ func (o *Refinement) HasCategoryDistributions() bool {
 
 // SetCategoryDistributions gets a reference to the given []CategoryDistribution and assigns it to the CategoryDistributions field.
 func (o *Refinement) SetCategoryDistributions(v []CategoryDistribution) {
-	o.CategoryDistributions = &v
+	o.CategoryDistributions = v
 }
 
 // GetConditionDistributions returns the ConditionDistributions field value if set, zero value otherwise.
 func (o *Refinement) GetConditionDistributions() []ConditionDistribution {
-	if o == nil || o.ConditionDistributions == nil {
+	if o == nil || IsNil(o.ConditionDistributions) {
 		var ret []ConditionDistribution
 		return ret
 	}
-	return *o.ConditionDistributions
+	return o.ConditionDistributions
 }
 
 // GetConditionDistributionsOk returns a tuple with the ConditionDistributions field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *Refinement) GetConditionDistributionsOk() (*[]ConditionDistribution, bool) {
-	if o == nil || o.ConditionDistributions == nil {
+func (o *Refinement) GetConditionDistributionsOk() ([]ConditionDistribution, bool) {
+	if o == nil || IsNil(o.ConditionDistributions) {
 		return nil, false
 	}
 	return o.ConditionDistributions, true
@@ -161,7 +164,7 @@ func (o *Refinement) GetConditionDistributionsOk() (*[]ConditionDistribution, bo
 
 // HasConditionDistributions returns a boolean if a field has been set.
 func (o *Refinement) HasConditionDistributions() bool {
-	if o != nil && o.ConditionDistributions != nil {
+	if o != nil && !IsNil(o.ConditionDistributions) {
 		return true
 	}
 
@@ -170,12 +173,12 @@ func (o *Refinement) HasConditionDistributions() bool {
 
 // SetConditionDistributions gets a reference to the given []ConditionDistribution and assigns it to the ConditionDistributions field.
 func (o *Refinement) SetConditionDistributions(v []ConditionDistribution) {
-	o.ConditionDistributions = &v
+	o.ConditionDistributions = v
 }
 
 // GetDominantCategoryId returns the DominantCategoryId field value if set, zero value otherwise.
 func (o *Refinement) GetDominantCategoryId() string {
-	if o == nil || o.DominantCategoryId == nil {
+	if o == nil || IsNil(o.DominantCategoryId) {
 		var ret string
 		return ret
 	}
@@ -185,7 +188,7 @@ func (o *Refinement) GetDominantCategoryId() string {
 // GetDominantCategoryIdOk returns a tuple with the DominantCategoryId field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Refinement) GetDominantCategoryIdOk() (*string, bool) {
-	if o == nil || o.DominantCategoryId == nil {
+	if o == nil || IsNil(o.DominantCategoryId) {
 		return nil, false
 	}
 	return o.DominantCategoryId, true
@@ -193,7 +196,7 @@ func (o *Refinement) GetDominantCategoryIdOk() (*string, bool) {
 
 // HasDominantCategoryId returns a boolean if a field has been set.
 func (o *Refinement) HasDominantCategoryId() bool {
-	if o != nil && o.DominantCategoryId != nil {
+	if o != nil && !IsNil(o.DominantCategoryId) {
 		return true
 	}
 
@@ -206,23 +209,31 @@ func (o *Refinement) SetDominantCategoryId(v string) {
 }
 
 func (o Refinement) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.AspectDistributions != nil {
-		toSerialize["aspectDistributions"] = o.AspectDistributions
-	}
-	if o.BuyingOptionDistributions != nil {
-		toSerialize["buyingOptionDistributions"] = o.BuyingOptionDistributions
-	}
-	if o.CategoryDistributions != nil {
-		toSerialize["categoryDistributions"] = o.CategoryDistributions
-	}
-	if o.ConditionDistributions != nil {
-		toSerialize["conditionDistributions"] = o.ConditionDistributions
-	}
-	if o.DominantCategoryId != nil {
-		toSerialize["dominantCategoryId"] = o.DominantCategoryId
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o Refinement) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.AspectDistributions) {
+		toSerialize["aspectDistributions"] = o.AspectDistributions
+	}
+	if !IsNil(o.BuyingOptionDistributions) {
+		toSerialize["buyingOptionDistributions"] = o.BuyingOptionDistributions
+	}
+	if !IsNil(o.CategoryDistributions) {
+		toSerialize["categoryDistributions"] = o.CategoryDistributions
+	}
+	if !IsNil(o.ConditionDistributions) {
+		toSerialize["conditionDistributions"] = o.ConditionDistributions
+	}
+	if !IsNil(o.DominantCategoryId) {
+		toSerialize["dominantCategoryId"] = o.DominantCategoryId
+	}
+	return toSerialize, nil
 }
 
 type NullableRefinement struct {
