@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,12 +14,15 @@ import (
 	"encoding/json"
 )
 
+// checks if the ReviewRating type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &ReviewRating{}
+
 // ReviewRating The type that defines the fields for the rating of a product review.
 type ReviewRating struct {
 	// The average rating given to a product based on customer reviews.
 	AverageRating *string `json:"averageRating,omitempty"`
 	// An array of containers for the product rating histograms that shows the review counts and the product rating.
-	RatingHistograms *[]RatingHistogram `json:"ratingHistograms,omitempty"`
+	RatingHistograms []RatingHistogram `json:"ratingHistograms,omitempty"`
 	// The total number of reviews for the item.
 	ReviewCount *int32 `json:"reviewCount,omitempty"`
 }
@@ -43,7 +46,7 @@ func NewReviewRatingWithDefaults() *ReviewRating {
 
 // GetAverageRating returns the AverageRating field value if set, zero value otherwise.
 func (o *ReviewRating) GetAverageRating() string {
-	if o == nil || o.AverageRating == nil {
+	if o == nil || IsNil(o.AverageRating) {
 		var ret string
 		return ret
 	}
@@ -53,7 +56,7 @@ func (o *ReviewRating) GetAverageRating() string {
 // GetAverageRatingOk returns a tuple with the AverageRating field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ReviewRating) GetAverageRatingOk() (*string, bool) {
-	if o == nil || o.AverageRating == nil {
+	if o == nil || IsNil(o.AverageRating) {
 		return nil, false
 	}
 	return o.AverageRating, true
@@ -61,7 +64,7 @@ func (o *ReviewRating) GetAverageRatingOk() (*string, bool) {
 
 // HasAverageRating returns a boolean if a field has been set.
 func (o *ReviewRating) HasAverageRating() bool {
-	if o != nil && o.AverageRating != nil {
+	if o != nil && !IsNil(o.AverageRating) {
 		return true
 	}
 
@@ -75,17 +78,17 @@ func (o *ReviewRating) SetAverageRating(v string) {
 
 // GetRatingHistograms returns the RatingHistograms field value if set, zero value otherwise.
 func (o *ReviewRating) GetRatingHistograms() []RatingHistogram {
-	if o == nil || o.RatingHistograms == nil {
+	if o == nil || IsNil(o.RatingHistograms) {
 		var ret []RatingHistogram
 		return ret
 	}
-	return *o.RatingHistograms
+	return o.RatingHistograms
 }
 
 // GetRatingHistogramsOk returns a tuple with the RatingHistograms field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *ReviewRating) GetRatingHistogramsOk() (*[]RatingHistogram, bool) {
-	if o == nil || o.RatingHistograms == nil {
+func (o *ReviewRating) GetRatingHistogramsOk() ([]RatingHistogram, bool) {
+	if o == nil || IsNil(o.RatingHistograms) {
 		return nil, false
 	}
 	return o.RatingHistograms, true
@@ -93,7 +96,7 @@ func (o *ReviewRating) GetRatingHistogramsOk() (*[]RatingHistogram, bool) {
 
 // HasRatingHistograms returns a boolean if a field has been set.
 func (o *ReviewRating) HasRatingHistograms() bool {
-	if o != nil && o.RatingHistograms != nil {
+	if o != nil && !IsNil(o.RatingHistograms) {
 		return true
 	}
 
@@ -102,12 +105,12 @@ func (o *ReviewRating) HasRatingHistograms() bool {
 
 // SetRatingHistograms gets a reference to the given []RatingHistogram and assigns it to the RatingHistograms field.
 func (o *ReviewRating) SetRatingHistograms(v []RatingHistogram) {
-	o.RatingHistograms = &v
+	o.RatingHistograms = v
 }
 
 // GetReviewCount returns the ReviewCount field value if set, zero value otherwise.
 func (o *ReviewRating) GetReviewCount() int32 {
-	if o == nil || o.ReviewCount == nil {
+	if o == nil || IsNil(o.ReviewCount) {
 		var ret int32
 		return ret
 	}
@@ -117,7 +120,7 @@ func (o *ReviewRating) GetReviewCount() int32 {
 // GetReviewCountOk returns a tuple with the ReviewCount field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ReviewRating) GetReviewCountOk() (*int32, bool) {
-	if o == nil || o.ReviewCount == nil {
+	if o == nil || IsNil(o.ReviewCount) {
 		return nil, false
 	}
 	return o.ReviewCount, true
@@ -125,7 +128,7 @@ func (o *ReviewRating) GetReviewCountOk() (*int32, bool) {
 
 // HasReviewCount returns a boolean if a field has been set.
 func (o *ReviewRating) HasReviewCount() bool {
-	if o != nil && o.ReviewCount != nil {
+	if o != nil && !IsNil(o.ReviewCount) {
 		return true
 	}
 
@@ -138,17 +141,25 @@ func (o *ReviewRating) SetReviewCount(v int32) {
 }
 
 func (o ReviewRating) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.AverageRating != nil {
-		toSerialize["averageRating"] = o.AverageRating
-	}
-	if o.RatingHistograms != nil {
-		toSerialize["ratingHistograms"] = o.RatingHistograms
-	}
-	if o.ReviewCount != nil {
-		toSerialize["reviewCount"] = o.ReviewCount
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o ReviewRating) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.AverageRating) {
+		toSerialize["averageRating"] = o.AverageRating
+	}
+	if !IsNil(o.RatingHistograms) {
+		toSerialize["ratingHistograms"] = o.RatingHistograms
+	}
+	if !IsNil(o.ReviewCount) {
+		toSerialize["reviewCount"] = o.ReviewCount
+	}
+	return toSerialize, nil
 }
 
 type NullableReviewRating struct {
