@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,26 +14,29 @@ import (
 	"encoding/json"
 )
 
-// SearchPagedCollection The type that defines the fields for a paginated result set. The response consists of 0 or more sequenced <em> pages</em> where each page has 0 or more items. 
+// checks if the SearchPagedCollection type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &SearchPagedCollection{}
+
+// SearchPagedCollection The type that defines the fields for a paginated result set. The response consists of 0 or more sequenced <i>pages</i> where each page has 0 or more items.
 type SearchPagedCollection struct {
 	AutoCorrections *AutoCorrections `json:"autoCorrections,omitempty"`
-	// The URI of the current page of results. <br /><br />The following example of the <b> search</b> method returns items 1 thru 5 from the list of items found. <br /><br /><code>https://api.ebay.com/buy/v1/item_summary/search?q=shirt&limit=5&offset=0</code>.
+	// The URI of the current page of results.<br><br>The following example of the <b>search</b> method returns items 1 thru 5 from the list of items found.<pre>https://api.ebay.com/buy/v1/item_summary/search?q=shirt&limit=5&offset=0</pre>.
 	Href *string `json:"href,omitempty"`
 	// An array of the items on this page. The items are sorted according to the sorting method specified in the request.
-	ItemSummaries *[]ItemSummary `json:"itemSummaries,omitempty"`
-	// The value of the <b>limit</b> parameter submitted in the request, which is the maximum number of items to return on a page, from the result set. A result set is the complete set of items returned by the method.
+	ItemSummaries []ItemSummary `json:"itemSummaries,omitempty"`
+	// The value of the <code>limit</code> parameter submitted in the request, which is the maximum number of items to return on a page, from the result set. A result set is the complete set of items returned by the method.
 	Limit *int32 `json:"limit,omitempty"`
-	// The URI for the next page of results. This value is returned if there is an additional page of results to return from the result set. <br /><br />The following example of the <b> search</b> method returns items 5 thru 10 from the list of items found.<br /> <br /><code>https://api.ebay.com/buy/v1/item_summary/search?query=t-shirts&limit=5&offset=10 </code>
+	// The URI for the next page of results. This value is returned if there is an additional page of results to return from the result set.<br><br>The following example of the <b>search</b> method returns items 5 thru 10 from the list of items found.<pre>https://api.ebay.com/buy/v1/item_summary/search?query=t-shirts&limit=5&offset=10</pre>
 	Next *string `json:"next,omitempty"`
-	// This value indicates the <b>offset</b> used for current page of items being returned. Assume the initial request used an <b>offset</b> of <code>0</code> and a <b>limit</b> of <code>3</code>. Then in the first page of results, this value would be <code>0</code>, and items 1-3 are returned. For the second page, this value is <code>3</code> and so on.
+	// This value indicates the <code>offset</code> used for current page of items being returned. Assume the initial request used an <code>offset</code> of <code>0</code> and a <code>limit</code> of <code>3</code>. Then in the first page of results, this value would be <code>0</code>, and items 1-3 are returned. For the second page, this value is <code>3</code> and so on.
 	Offset *int32 `json:"offset,omitempty"`
-	// The URI for the previous page of results. This is returned if there is a previous page of results from the result set. <br /><br />The following example of the <b> search</b> method returns items 1 thru 5 from the list of items found, which would be the first set of items returned.<br /> <br /><code>https://api.ebay.com/buy/v1/item_summary/search?query=t-shirts&limit=5&offset=0</code>
+	// The URI for the previous page of results. This is returned if there is a previous page of results from the result set.<br><br>The following example of the <b>search</b> method returns items 1 thru 5 from the list of items found, which would be the first set of items returned.<pre>https://api.ebay.com/buy/v1/item_summary/search?query=t-shirts&limit=5&offset=0</pre>
 	Prev *string `json:"prev,omitempty"`
 	Refinement *Refinement `json:"refinement,omitempty"`
-	// The total number of items that match the input criteria.
+	// The total number of items that match the input criteria.<br><br><span class=\"tablenote\"><b>Note:</b> <code>total</code> is just an indicator of the number of listings for a given query. It could vary based on the number of listings with variations included in the result. It is strongly recommended that <code>total</code> not be used in pagination use cases. Instead, use <a href=\"/api-docs/buy/browse/resources/item_summary/methods/search#response.next \">next</a> to determine the results on the next page.</span>
 	Total *int32 `json:"total,omitempty"`
 	// The container with all the warnings for the request.
-	Warnings *[]Error `json:"warnings,omitempty"`
+	Warnings []Error `json:"warnings,omitempty"`
 }
 
 // NewSearchPagedCollection instantiates a new SearchPagedCollection object
@@ -55,7 +58,7 @@ func NewSearchPagedCollectionWithDefaults() *SearchPagedCollection {
 
 // GetAutoCorrections returns the AutoCorrections field value if set, zero value otherwise.
 func (o *SearchPagedCollection) GetAutoCorrections() AutoCorrections {
-	if o == nil || o.AutoCorrections == nil {
+	if o == nil || IsNil(o.AutoCorrections) {
 		var ret AutoCorrections
 		return ret
 	}
@@ -65,7 +68,7 @@ func (o *SearchPagedCollection) GetAutoCorrections() AutoCorrections {
 // GetAutoCorrectionsOk returns a tuple with the AutoCorrections field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SearchPagedCollection) GetAutoCorrectionsOk() (*AutoCorrections, bool) {
-	if o == nil || o.AutoCorrections == nil {
+	if o == nil || IsNil(o.AutoCorrections) {
 		return nil, false
 	}
 	return o.AutoCorrections, true
@@ -73,7 +76,7 @@ func (o *SearchPagedCollection) GetAutoCorrectionsOk() (*AutoCorrections, bool) 
 
 // HasAutoCorrections returns a boolean if a field has been set.
 func (o *SearchPagedCollection) HasAutoCorrections() bool {
-	if o != nil && o.AutoCorrections != nil {
+	if o != nil && !IsNil(o.AutoCorrections) {
 		return true
 	}
 
@@ -87,7 +90,7 @@ func (o *SearchPagedCollection) SetAutoCorrections(v AutoCorrections) {
 
 // GetHref returns the Href field value if set, zero value otherwise.
 func (o *SearchPagedCollection) GetHref() string {
-	if o == nil || o.Href == nil {
+	if o == nil || IsNil(o.Href) {
 		var ret string
 		return ret
 	}
@@ -97,7 +100,7 @@ func (o *SearchPagedCollection) GetHref() string {
 // GetHrefOk returns a tuple with the Href field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SearchPagedCollection) GetHrefOk() (*string, bool) {
-	if o == nil || o.Href == nil {
+	if o == nil || IsNil(o.Href) {
 		return nil, false
 	}
 	return o.Href, true
@@ -105,7 +108,7 @@ func (o *SearchPagedCollection) GetHrefOk() (*string, bool) {
 
 // HasHref returns a boolean if a field has been set.
 func (o *SearchPagedCollection) HasHref() bool {
-	if o != nil && o.Href != nil {
+	if o != nil && !IsNil(o.Href) {
 		return true
 	}
 
@@ -119,17 +122,17 @@ func (o *SearchPagedCollection) SetHref(v string) {
 
 // GetItemSummaries returns the ItemSummaries field value if set, zero value otherwise.
 func (o *SearchPagedCollection) GetItemSummaries() []ItemSummary {
-	if o == nil || o.ItemSummaries == nil {
+	if o == nil || IsNil(o.ItemSummaries) {
 		var ret []ItemSummary
 		return ret
 	}
-	return *o.ItemSummaries
+	return o.ItemSummaries
 }
 
 // GetItemSummariesOk returns a tuple with the ItemSummaries field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *SearchPagedCollection) GetItemSummariesOk() (*[]ItemSummary, bool) {
-	if o == nil || o.ItemSummaries == nil {
+func (o *SearchPagedCollection) GetItemSummariesOk() ([]ItemSummary, bool) {
+	if o == nil || IsNil(o.ItemSummaries) {
 		return nil, false
 	}
 	return o.ItemSummaries, true
@@ -137,7 +140,7 @@ func (o *SearchPagedCollection) GetItemSummariesOk() (*[]ItemSummary, bool) {
 
 // HasItemSummaries returns a boolean if a field has been set.
 func (o *SearchPagedCollection) HasItemSummaries() bool {
-	if o != nil && o.ItemSummaries != nil {
+	if o != nil && !IsNil(o.ItemSummaries) {
 		return true
 	}
 
@@ -146,12 +149,12 @@ func (o *SearchPagedCollection) HasItemSummaries() bool {
 
 // SetItemSummaries gets a reference to the given []ItemSummary and assigns it to the ItemSummaries field.
 func (o *SearchPagedCollection) SetItemSummaries(v []ItemSummary) {
-	o.ItemSummaries = &v
+	o.ItemSummaries = v
 }
 
 // GetLimit returns the Limit field value if set, zero value otherwise.
 func (o *SearchPagedCollection) GetLimit() int32 {
-	if o == nil || o.Limit == nil {
+	if o == nil || IsNil(o.Limit) {
 		var ret int32
 		return ret
 	}
@@ -161,7 +164,7 @@ func (o *SearchPagedCollection) GetLimit() int32 {
 // GetLimitOk returns a tuple with the Limit field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SearchPagedCollection) GetLimitOk() (*int32, bool) {
-	if o == nil || o.Limit == nil {
+	if o == nil || IsNil(o.Limit) {
 		return nil, false
 	}
 	return o.Limit, true
@@ -169,7 +172,7 @@ func (o *SearchPagedCollection) GetLimitOk() (*int32, bool) {
 
 // HasLimit returns a boolean if a field has been set.
 func (o *SearchPagedCollection) HasLimit() bool {
-	if o != nil && o.Limit != nil {
+	if o != nil && !IsNil(o.Limit) {
 		return true
 	}
 
@@ -183,7 +186,7 @@ func (o *SearchPagedCollection) SetLimit(v int32) {
 
 // GetNext returns the Next field value if set, zero value otherwise.
 func (o *SearchPagedCollection) GetNext() string {
-	if o == nil || o.Next == nil {
+	if o == nil || IsNil(o.Next) {
 		var ret string
 		return ret
 	}
@@ -193,7 +196,7 @@ func (o *SearchPagedCollection) GetNext() string {
 // GetNextOk returns a tuple with the Next field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SearchPagedCollection) GetNextOk() (*string, bool) {
-	if o == nil || o.Next == nil {
+	if o == nil || IsNil(o.Next) {
 		return nil, false
 	}
 	return o.Next, true
@@ -201,7 +204,7 @@ func (o *SearchPagedCollection) GetNextOk() (*string, bool) {
 
 // HasNext returns a boolean if a field has been set.
 func (o *SearchPagedCollection) HasNext() bool {
-	if o != nil && o.Next != nil {
+	if o != nil && !IsNil(o.Next) {
 		return true
 	}
 
@@ -215,7 +218,7 @@ func (o *SearchPagedCollection) SetNext(v string) {
 
 // GetOffset returns the Offset field value if set, zero value otherwise.
 func (o *SearchPagedCollection) GetOffset() int32 {
-	if o == nil || o.Offset == nil {
+	if o == nil || IsNil(o.Offset) {
 		var ret int32
 		return ret
 	}
@@ -225,7 +228,7 @@ func (o *SearchPagedCollection) GetOffset() int32 {
 // GetOffsetOk returns a tuple with the Offset field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SearchPagedCollection) GetOffsetOk() (*int32, bool) {
-	if o == nil || o.Offset == nil {
+	if o == nil || IsNil(o.Offset) {
 		return nil, false
 	}
 	return o.Offset, true
@@ -233,7 +236,7 @@ func (o *SearchPagedCollection) GetOffsetOk() (*int32, bool) {
 
 // HasOffset returns a boolean if a field has been set.
 func (o *SearchPagedCollection) HasOffset() bool {
-	if o != nil && o.Offset != nil {
+	if o != nil && !IsNil(o.Offset) {
 		return true
 	}
 
@@ -247,7 +250,7 @@ func (o *SearchPagedCollection) SetOffset(v int32) {
 
 // GetPrev returns the Prev field value if set, zero value otherwise.
 func (o *SearchPagedCollection) GetPrev() string {
-	if o == nil || o.Prev == nil {
+	if o == nil || IsNil(o.Prev) {
 		var ret string
 		return ret
 	}
@@ -257,7 +260,7 @@ func (o *SearchPagedCollection) GetPrev() string {
 // GetPrevOk returns a tuple with the Prev field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SearchPagedCollection) GetPrevOk() (*string, bool) {
-	if o == nil || o.Prev == nil {
+	if o == nil || IsNil(o.Prev) {
 		return nil, false
 	}
 	return o.Prev, true
@@ -265,7 +268,7 @@ func (o *SearchPagedCollection) GetPrevOk() (*string, bool) {
 
 // HasPrev returns a boolean if a field has been set.
 func (o *SearchPagedCollection) HasPrev() bool {
-	if o != nil && o.Prev != nil {
+	if o != nil && !IsNil(o.Prev) {
 		return true
 	}
 
@@ -279,7 +282,7 @@ func (o *SearchPagedCollection) SetPrev(v string) {
 
 // GetRefinement returns the Refinement field value if set, zero value otherwise.
 func (o *SearchPagedCollection) GetRefinement() Refinement {
-	if o == nil || o.Refinement == nil {
+	if o == nil || IsNil(o.Refinement) {
 		var ret Refinement
 		return ret
 	}
@@ -289,7 +292,7 @@ func (o *SearchPagedCollection) GetRefinement() Refinement {
 // GetRefinementOk returns a tuple with the Refinement field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SearchPagedCollection) GetRefinementOk() (*Refinement, bool) {
-	if o == nil || o.Refinement == nil {
+	if o == nil || IsNil(o.Refinement) {
 		return nil, false
 	}
 	return o.Refinement, true
@@ -297,7 +300,7 @@ func (o *SearchPagedCollection) GetRefinementOk() (*Refinement, bool) {
 
 // HasRefinement returns a boolean if a field has been set.
 func (o *SearchPagedCollection) HasRefinement() bool {
-	if o != nil && o.Refinement != nil {
+	if o != nil && !IsNil(o.Refinement) {
 		return true
 	}
 
@@ -311,7 +314,7 @@ func (o *SearchPagedCollection) SetRefinement(v Refinement) {
 
 // GetTotal returns the Total field value if set, zero value otherwise.
 func (o *SearchPagedCollection) GetTotal() int32 {
-	if o == nil || o.Total == nil {
+	if o == nil || IsNil(o.Total) {
 		var ret int32
 		return ret
 	}
@@ -321,7 +324,7 @@ func (o *SearchPagedCollection) GetTotal() int32 {
 // GetTotalOk returns a tuple with the Total field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SearchPagedCollection) GetTotalOk() (*int32, bool) {
-	if o == nil || o.Total == nil {
+	if o == nil || IsNil(o.Total) {
 		return nil, false
 	}
 	return o.Total, true
@@ -329,7 +332,7 @@ func (o *SearchPagedCollection) GetTotalOk() (*int32, bool) {
 
 // HasTotal returns a boolean if a field has been set.
 func (o *SearchPagedCollection) HasTotal() bool {
-	if o != nil && o.Total != nil {
+	if o != nil && !IsNil(o.Total) {
 		return true
 	}
 
@@ -343,17 +346,17 @@ func (o *SearchPagedCollection) SetTotal(v int32) {
 
 // GetWarnings returns the Warnings field value if set, zero value otherwise.
 func (o *SearchPagedCollection) GetWarnings() []Error {
-	if o == nil || o.Warnings == nil {
+	if o == nil || IsNil(o.Warnings) {
 		var ret []Error
 		return ret
 	}
-	return *o.Warnings
+	return o.Warnings
 }
 
 // GetWarningsOk returns a tuple with the Warnings field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *SearchPagedCollection) GetWarningsOk() (*[]Error, bool) {
-	if o == nil || o.Warnings == nil {
+func (o *SearchPagedCollection) GetWarningsOk() ([]Error, bool) {
+	if o == nil || IsNil(o.Warnings) {
 		return nil, false
 	}
 	return o.Warnings, true
@@ -361,7 +364,7 @@ func (o *SearchPagedCollection) GetWarningsOk() (*[]Error, bool) {
 
 // HasWarnings returns a boolean if a field has been set.
 func (o *SearchPagedCollection) HasWarnings() bool {
-	if o != nil && o.Warnings != nil {
+	if o != nil && !IsNil(o.Warnings) {
 		return true
 	}
 
@@ -370,42 +373,50 @@ func (o *SearchPagedCollection) HasWarnings() bool {
 
 // SetWarnings gets a reference to the given []Error and assigns it to the Warnings field.
 func (o *SearchPagedCollection) SetWarnings(v []Error) {
-	o.Warnings = &v
+	o.Warnings = v
 }
 
 func (o SearchPagedCollection) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.AutoCorrections != nil {
-		toSerialize["autoCorrections"] = o.AutoCorrections
-	}
-	if o.Href != nil {
-		toSerialize["href"] = o.Href
-	}
-	if o.ItemSummaries != nil {
-		toSerialize["itemSummaries"] = o.ItemSummaries
-	}
-	if o.Limit != nil {
-		toSerialize["limit"] = o.Limit
-	}
-	if o.Next != nil {
-		toSerialize["next"] = o.Next
-	}
-	if o.Offset != nil {
-		toSerialize["offset"] = o.Offset
-	}
-	if o.Prev != nil {
-		toSerialize["prev"] = o.Prev
-	}
-	if o.Refinement != nil {
-		toSerialize["refinement"] = o.Refinement
-	}
-	if o.Total != nil {
-		toSerialize["total"] = o.Total
-	}
-	if o.Warnings != nil {
-		toSerialize["warnings"] = o.Warnings
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o SearchPagedCollection) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.AutoCorrections) {
+		toSerialize["autoCorrections"] = o.AutoCorrections
+	}
+	if !IsNil(o.Href) {
+		toSerialize["href"] = o.Href
+	}
+	if !IsNil(o.ItemSummaries) {
+		toSerialize["itemSummaries"] = o.ItemSummaries
+	}
+	if !IsNil(o.Limit) {
+		toSerialize["limit"] = o.Limit
+	}
+	if !IsNil(o.Next) {
+		toSerialize["next"] = o.Next
+	}
+	if !IsNil(o.Offset) {
+		toSerialize["offset"] = o.Offset
+	}
+	if !IsNil(o.Prev) {
+		toSerialize["prev"] = o.Prev
+	}
+	if !IsNil(o.Refinement) {
+		toSerialize["refinement"] = o.Refinement
+	}
+	if !IsNil(o.Total) {
+		toSerialize["total"] = o.Total
+	}
+	if !IsNil(o.Warnings) {
+		toSerialize["warnings"] = o.Warnings
+	}
+	return toSerialize, nil
 }
 
 type NullableSearchPagedCollection struct {
