@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,13 +14,16 @@ import (
 	"encoding/json"
 )
 
-// Seller The type that defines the fields for basic information about the seller of the item returned by the <b> item_summary</b> resource.
+// checks if the Seller type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &Seller{}
+
+// Seller The type that defines the fields for basic information about the seller of the item returned by the <code>item_summary</code> resource.
 type Seller struct {
 	// The percentage of the total positive feedback.
 	FeedbackPercentage *string `json:"feedbackPercentage,omitempty"`
 	// The feedback score of the seller. This value is based on the ratings from eBay members that bought items from this seller.
 	FeedbackScore *int32 `json:"feedbackScore,omitempty"`
-	// Indicates if the seller is a business or an individual. This is determined when the seller registers with eBay. If they register for a business account, this value will be BUSINESS. If they register for a private account, this value will be INDIVIDUAL. This designation is required by the tax laws in some countries.   <br /><br />This field is returned only on the following sites. <br /><br />EBAY_AT, EBAY_BE, EBAY_CH, EBAY_DE, EBAY_ES, EBAY_FR, EBAY_GB, EBAY_IE, EBAY_IT, EBAY_PL <br /><br /><b> Valid Values:</b> BUSINESS or INDIVIDUAL <br /><br />Code so that your app gracefully handles any future changes to this list. 
+	// Indicates if the seller is a business or an individual. This is determined when the seller registers with eBay:<ul><li>If they register for a business account, this value will be <code>BUSINESS<code>.</li><li>If they register for a private account, this value will be <code>INDIVIDUAL</code>.</li></ul>This designation is required by the tax laws in some countries.<br><br>This field is returned only on the following sites:<br><br>EBAY_AT, EBAY_BE, EBAY_CH, EBAY_DE, EBAY_ES, EBAY_FR, EBAY_GB, EBAY_IE, EBAY_IT, EBAY_PL<br><br><b>Valid Values:</b> <code>BUSINESS</code> or <code>INDIVIDUAL</code>
 	SellerAccountType *string `json:"sellerAccountType,omitempty"`
 	// The user name created by the seller for use on eBay.
 	Username *string `json:"username,omitempty"`
@@ -45,7 +48,7 @@ func NewSellerWithDefaults() *Seller {
 
 // GetFeedbackPercentage returns the FeedbackPercentage field value if set, zero value otherwise.
 func (o *Seller) GetFeedbackPercentage() string {
-	if o == nil || o.FeedbackPercentage == nil {
+	if o == nil || IsNil(o.FeedbackPercentage) {
 		var ret string
 		return ret
 	}
@@ -55,7 +58,7 @@ func (o *Seller) GetFeedbackPercentage() string {
 // GetFeedbackPercentageOk returns a tuple with the FeedbackPercentage field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Seller) GetFeedbackPercentageOk() (*string, bool) {
-	if o == nil || o.FeedbackPercentage == nil {
+	if o == nil || IsNil(o.FeedbackPercentage) {
 		return nil, false
 	}
 	return o.FeedbackPercentage, true
@@ -63,7 +66,7 @@ func (o *Seller) GetFeedbackPercentageOk() (*string, bool) {
 
 // HasFeedbackPercentage returns a boolean if a field has been set.
 func (o *Seller) HasFeedbackPercentage() bool {
-	if o != nil && o.FeedbackPercentage != nil {
+	if o != nil && !IsNil(o.FeedbackPercentage) {
 		return true
 	}
 
@@ -77,7 +80,7 @@ func (o *Seller) SetFeedbackPercentage(v string) {
 
 // GetFeedbackScore returns the FeedbackScore field value if set, zero value otherwise.
 func (o *Seller) GetFeedbackScore() int32 {
-	if o == nil || o.FeedbackScore == nil {
+	if o == nil || IsNil(o.FeedbackScore) {
 		var ret int32
 		return ret
 	}
@@ -87,7 +90,7 @@ func (o *Seller) GetFeedbackScore() int32 {
 // GetFeedbackScoreOk returns a tuple with the FeedbackScore field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Seller) GetFeedbackScoreOk() (*int32, bool) {
-	if o == nil || o.FeedbackScore == nil {
+	if o == nil || IsNil(o.FeedbackScore) {
 		return nil, false
 	}
 	return o.FeedbackScore, true
@@ -95,7 +98,7 @@ func (o *Seller) GetFeedbackScoreOk() (*int32, bool) {
 
 // HasFeedbackScore returns a boolean if a field has been set.
 func (o *Seller) HasFeedbackScore() bool {
-	if o != nil && o.FeedbackScore != nil {
+	if o != nil && !IsNil(o.FeedbackScore) {
 		return true
 	}
 
@@ -109,7 +112,7 @@ func (o *Seller) SetFeedbackScore(v int32) {
 
 // GetSellerAccountType returns the SellerAccountType field value if set, zero value otherwise.
 func (o *Seller) GetSellerAccountType() string {
-	if o == nil || o.SellerAccountType == nil {
+	if o == nil || IsNil(o.SellerAccountType) {
 		var ret string
 		return ret
 	}
@@ -119,7 +122,7 @@ func (o *Seller) GetSellerAccountType() string {
 // GetSellerAccountTypeOk returns a tuple with the SellerAccountType field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Seller) GetSellerAccountTypeOk() (*string, bool) {
-	if o == nil || o.SellerAccountType == nil {
+	if o == nil || IsNil(o.SellerAccountType) {
 		return nil, false
 	}
 	return o.SellerAccountType, true
@@ -127,7 +130,7 @@ func (o *Seller) GetSellerAccountTypeOk() (*string, bool) {
 
 // HasSellerAccountType returns a boolean if a field has been set.
 func (o *Seller) HasSellerAccountType() bool {
-	if o != nil && o.SellerAccountType != nil {
+	if o != nil && !IsNil(o.SellerAccountType) {
 		return true
 	}
 
@@ -141,7 +144,7 @@ func (o *Seller) SetSellerAccountType(v string) {
 
 // GetUsername returns the Username field value if set, zero value otherwise.
 func (o *Seller) GetUsername() string {
-	if o == nil || o.Username == nil {
+	if o == nil || IsNil(o.Username) {
 		var ret string
 		return ret
 	}
@@ -151,7 +154,7 @@ func (o *Seller) GetUsername() string {
 // GetUsernameOk returns a tuple with the Username field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Seller) GetUsernameOk() (*string, bool) {
-	if o == nil || o.Username == nil {
+	if o == nil || IsNil(o.Username) {
 		return nil, false
 	}
 	return o.Username, true
@@ -159,7 +162,7 @@ func (o *Seller) GetUsernameOk() (*string, bool) {
 
 // HasUsername returns a boolean if a field has been set.
 func (o *Seller) HasUsername() bool {
-	if o != nil && o.Username != nil {
+	if o != nil && !IsNil(o.Username) {
 		return true
 	}
 
@@ -172,20 +175,28 @@ func (o *Seller) SetUsername(v string) {
 }
 
 func (o Seller) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.FeedbackPercentage != nil {
-		toSerialize["feedbackPercentage"] = o.FeedbackPercentage
-	}
-	if o.FeedbackScore != nil {
-		toSerialize["feedbackScore"] = o.FeedbackScore
-	}
-	if o.SellerAccountType != nil {
-		toSerialize["sellerAccountType"] = o.SellerAccountType
-	}
-	if o.Username != nil {
-		toSerialize["username"] = o.Username
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o Seller) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.FeedbackPercentage) {
+		toSerialize["feedbackPercentage"] = o.FeedbackPercentage
+	}
+	if !IsNil(o.FeedbackScore) {
+		toSerialize["feedbackScore"] = o.FeedbackScore
+	}
+	if !IsNil(o.SellerAccountType) {
+		toSerialize["sellerAccountType"] = o.SellerAccountType
+	}
+	if !IsNil(o.Username) {
+		toSerialize["username"] = o.Username
+	}
+	return toSerialize, nil
 }
 
 type NullableSeller struct {
