@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -13,6 +13,9 @@ package buybrowse
 import (
 	"encoding/json"
 )
+
+// checks if the SellerLegalInfo type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &SellerLegalInfo{}
 
 // SellerLegalInfo The type that defines the fields for the contact information for a seller.
 type SellerLegalInfo struct {
@@ -36,7 +39,10 @@ type SellerLegalInfo struct {
 	// This is a free-form string created by the seller. This is the seller's terms or condition, which is in addition to the seller's return policies.
 	TermsOfService *string `json:"termsOfService,omitempty"`
 	// An array of the seller's VAT (value added tax) IDs and the issuing country. VAT is a tax added by some European countries.
-	VatDetails *[]VatDetail `json:"vatDetails,omitempty"`
+	VatDetails []VatDetail `json:"vatDetails,omitempty"`
+	EconomicOperator *EconomicOperator `json:"economicOperator,omitempty"`
+	// The Waste Electrical and Electronic Equipment (WEEE) registration number required for any seller to place electrical and electronic equipment on the market in Germany. This manufacturer number is assigned to the first distributors of electrical and electronic equipment and comprises a country code and an 8-digit sequence of digits (e.g. “WEEE Reg. No. DE 12345678”).
+	WeeeNumber *string `json:"weeeNumber,omitempty"`
 }
 
 // NewSellerLegalInfo instantiates a new SellerLegalInfo object
@@ -58,7 +64,7 @@ func NewSellerLegalInfoWithDefaults() *SellerLegalInfo {
 
 // GetEmail returns the Email field value if set, zero value otherwise.
 func (o *SellerLegalInfo) GetEmail() string {
-	if o == nil || o.Email == nil {
+	if o == nil || IsNil(o.Email) {
 		var ret string
 		return ret
 	}
@@ -68,7 +74,7 @@ func (o *SellerLegalInfo) GetEmail() string {
 // GetEmailOk returns a tuple with the Email field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SellerLegalInfo) GetEmailOk() (*string, bool) {
-	if o == nil || o.Email == nil {
+	if o == nil || IsNil(o.Email) {
 		return nil, false
 	}
 	return o.Email, true
@@ -76,7 +82,7 @@ func (o *SellerLegalInfo) GetEmailOk() (*string, bool) {
 
 // HasEmail returns a boolean if a field has been set.
 func (o *SellerLegalInfo) HasEmail() bool {
-	if o != nil && o.Email != nil {
+	if o != nil && !IsNil(o.Email) {
 		return true
 	}
 
@@ -90,7 +96,7 @@ func (o *SellerLegalInfo) SetEmail(v string) {
 
 // GetFax returns the Fax field value if set, zero value otherwise.
 func (o *SellerLegalInfo) GetFax() string {
-	if o == nil || o.Fax == nil {
+	if o == nil || IsNil(o.Fax) {
 		var ret string
 		return ret
 	}
@@ -100,7 +106,7 @@ func (o *SellerLegalInfo) GetFax() string {
 // GetFaxOk returns a tuple with the Fax field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SellerLegalInfo) GetFaxOk() (*string, bool) {
-	if o == nil || o.Fax == nil {
+	if o == nil || IsNil(o.Fax) {
 		return nil, false
 	}
 	return o.Fax, true
@@ -108,7 +114,7 @@ func (o *SellerLegalInfo) GetFaxOk() (*string, bool) {
 
 // HasFax returns a boolean if a field has been set.
 func (o *SellerLegalInfo) HasFax() bool {
-	if o != nil && o.Fax != nil {
+	if o != nil && !IsNil(o.Fax) {
 		return true
 	}
 
@@ -122,7 +128,7 @@ func (o *SellerLegalInfo) SetFax(v string) {
 
 // GetImprint returns the Imprint field value if set, zero value otherwise.
 func (o *SellerLegalInfo) GetImprint() string {
-	if o == nil || o.Imprint == nil {
+	if o == nil || IsNil(o.Imprint) {
 		var ret string
 		return ret
 	}
@@ -132,7 +138,7 @@ func (o *SellerLegalInfo) GetImprint() string {
 // GetImprintOk returns a tuple with the Imprint field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SellerLegalInfo) GetImprintOk() (*string, bool) {
-	if o == nil || o.Imprint == nil {
+	if o == nil || IsNil(o.Imprint) {
 		return nil, false
 	}
 	return o.Imprint, true
@@ -140,7 +146,7 @@ func (o *SellerLegalInfo) GetImprintOk() (*string, bool) {
 
 // HasImprint returns a boolean if a field has been set.
 func (o *SellerLegalInfo) HasImprint() bool {
-	if o != nil && o.Imprint != nil {
+	if o != nil && !IsNil(o.Imprint) {
 		return true
 	}
 
@@ -154,7 +160,7 @@ func (o *SellerLegalInfo) SetImprint(v string) {
 
 // GetLegalContactFirstName returns the LegalContactFirstName field value if set, zero value otherwise.
 func (o *SellerLegalInfo) GetLegalContactFirstName() string {
-	if o == nil || o.LegalContactFirstName == nil {
+	if o == nil || IsNil(o.LegalContactFirstName) {
 		var ret string
 		return ret
 	}
@@ -164,7 +170,7 @@ func (o *SellerLegalInfo) GetLegalContactFirstName() string {
 // GetLegalContactFirstNameOk returns a tuple with the LegalContactFirstName field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SellerLegalInfo) GetLegalContactFirstNameOk() (*string, bool) {
-	if o == nil || o.LegalContactFirstName == nil {
+	if o == nil || IsNil(o.LegalContactFirstName) {
 		return nil, false
 	}
 	return o.LegalContactFirstName, true
@@ -172,7 +178,7 @@ func (o *SellerLegalInfo) GetLegalContactFirstNameOk() (*string, bool) {
 
 // HasLegalContactFirstName returns a boolean if a field has been set.
 func (o *SellerLegalInfo) HasLegalContactFirstName() bool {
-	if o != nil && o.LegalContactFirstName != nil {
+	if o != nil && !IsNil(o.LegalContactFirstName) {
 		return true
 	}
 
@@ -186,7 +192,7 @@ func (o *SellerLegalInfo) SetLegalContactFirstName(v string) {
 
 // GetLegalContactLastName returns the LegalContactLastName field value if set, zero value otherwise.
 func (o *SellerLegalInfo) GetLegalContactLastName() string {
-	if o == nil || o.LegalContactLastName == nil {
+	if o == nil || IsNil(o.LegalContactLastName) {
 		var ret string
 		return ret
 	}
@@ -196,7 +202,7 @@ func (o *SellerLegalInfo) GetLegalContactLastName() string {
 // GetLegalContactLastNameOk returns a tuple with the LegalContactLastName field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SellerLegalInfo) GetLegalContactLastNameOk() (*string, bool) {
-	if o == nil || o.LegalContactLastName == nil {
+	if o == nil || IsNil(o.LegalContactLastName) {
 		return nil, false
 	}
 	return o.LegalContactLastName, true
@@ -204,7 +210,7 @@ func (o *SellerLegalInfo) GetLegalContactLastNameOk() (*string, bool) {
 
 // HasLegalContactLastName returns a boolean if a field has been set.
 func (o *SellerLegalInfo) HasLegalContactLastName() bool {
-	if o != nil && o.LegalContactLastName != nil {
+	if o != nil && !IsNil(o.LegalContactLastName) {
 		return true
 	}
 
@@ -218,7 +224,7 @@ func (o *SellerLegalInfo) SetLegalContactLastName(v string) {
 
 // GetName returns the Name field value if set, zero value otherwise.
 func (o *SellerLegalInfo) GetName() string {
-	if o == nil || o.Name == nil {
+	if o == nil || IsNil(o.Name) {
 		var ret string
 		return ret
 	}
@@ -228,7 +234,7 @@ func (o *SellerLegalInfo) GetName() string {
 // GetNameOk returns a tuple with the Name field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SellerLegalInfo) GetNameOk() (*string, bool) {
-	if o == nil || o.Name == nil {
+	if o == nil || IsNil(o.Name) {
 		return nil, false
 	}
 	return o.Name, true
@@ -236,7 +242,7 @@ func (o *SellerLegalInfo) GetNameOk() (*string, bool) {
 
 // HasName returns a boolean if a field has been set.
 func (o *SellerLegalInfo) HasName() bool {
-	if o != nil && o.Name != nil {
+	if o != nil && !IsNil(o.Name) {
 		return true
 	}
 
@@ -250,7 +256,7 @@ func (o *SellerLegalInfo) SetName(v string) {
 
 // GetPhone returns the Phone field value if set, zero value otherwise.
 func (o *SellerLegalInfo) GetPhone() string {
-	if o == nil || o.Phone == nil {
+	if o == nil || IsNil(o.Phone) {
 		var ret string
 		return ret
 	}
@@ -260,7 +266,7 @@ func (o *SellerLegalInfo) GetPhone() string {
 // GetPhoneOk returns a tuple with the Phone field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SellerLegalInfo) GetPhoneOk() (*string, bool) {
-	if o == nil || o.Phone == nil {
+	if o == nil || IsNil(o.Phone) {
 		return nil, false
 	}
 	return o.Phone, true
@@ -268,7 +274,7 @@ func (o *SellerLegalInfo) GetPhoneOk() (*string, bool) {
 
 // HasPhone returns a boolean if a field has been set.
 func (o *SellerLegalInfo) HasPhone() bool {
-	if o != nil && o.Phone != nil {
+	if o != nil && !IsNil(o.Phone) {
 		return true
 	}
 
@@ -282,7 +288,7 @@ func (o *SellerLegalInfo) SetPhone(v string) {
 
 // GetRegistrationNumber returns the RegistrationNumber field value if set, zero value otherwise.
 func (o *SellerLegalInfo) GetRegistrationNumber() string {
-	if o == nil || o.RegistrationNumber == nil {
+	if o == nil || IsNil(o.RegistrationNumber) {
 		var ret string
 		return ret
 	}
@@ -292,7 +298,7 @@ func (o *SellerLegalInfo) GetRegistrationNumber() string {
 // GetRegistrationNumberOk returns a tuple with the RegistrationNumber field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SellerLegalInfo) GetRegistrationNumberOk() (*string, bool) {
-	if o == nil || o.RegistrationNumber == nil {
+	if o == nil || IsNil(o.RegistrationNumber) {
 		return nil, false
 	}
 	return o.RegistrationNumber, true
@@ -300,7 +306,7 @@ func (o *SellerLegalInfo) GetRegistrationNumberOk() (*string, bool) {
 
 // HasRegistrationNumber returns a boolean if a field has been set.
 func (o *SellerLegalInfo) HasRegistrationNumber() bool {
-	if o != nil && o.RegistrationNumber != nil {
+	if o != nil && !IsNil(o.RegistrationNumber) {
 		return true
 	}
 
@@ -314,7 +320,7 @@ func (o *SellerLegalInfo) SetRegistrationNumber(v string) {
 
 // GetSellerProvidedLegalAddress returns the SellerProvidedLegalAddress field value if set, zero value otherwise.
 func (o *SellerLegalInfo) GetSellerProvidedLegalAddress() LegalAddress {
-	if o == nil || o.SellerProvidedLegalAddress == nil {
+	if o == nil || IsNil(o.SellerProvidedLegalAddress) {
 		var ret LegalAddress
 		return ret
 	}
@@ -324,7 +330,7 @@ func (o *SellerLegalInfo) GetSellerProvidedLegalAddress() LegalAddress {
 // GetSellerProvidedLegalAddressOk returns a tuple with the SellerProvidedLegalAddress field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SellerLegalInfo) GetSellerProvidedLegalAddressOk() (*LegalAddress, bool) {
-	if o == nil || o.SellerProvidedLegalAddress == nil {
+	if o == nil || IsNil(o.SellerProvidedLegalAddress) {
 		return nil, false
 	}
 	return o.SellerProvidedLegalAddress, true
@@ -332,7 +338,7 @@ func (o *SellerLegalInfo) GetSellerProvidedLegalAddressOk() (*LegalAddress, bool
 
 // HasSellerProvidedLegalAddress returns a boolean if a field has been set.
 func (o *SellerLegalInfo) HasSellerProvidedLegalAddress() bool {
-	if o != nil && o.SellerProvidedLegalAddress != nil {
+	if o != nil && !IsNil(o.SellerProvidedLegalAddress) {
 		return true
 	}
 
@@ -346,7 +352,7 @@ func (o *SellerLegalInfo) SetSellerProvidedLegalAddress(v LegalAddress) {
 
 // GetTermsOfService returns the TermsOfService field value if set, zero value otherwise.
 func (o *SellerLegalInfo) GetTermsOfService() string {
-	if o == nil || o.TermsOfService == nil {
+	if o == nil || IsNil(o.TermsOfService) {
 		var ret string
 		return ret
 	}
@@ -356,7 +362,7 @@ func (o *SellerLegalInfo) GetTermsOfService() string {
 // GetTermsOfServiceOk returns a tuple with the TermsOfService field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *SellerLegalInfo) GetTermsOfServiceOk() (*string, bool) {
-	if o == nil || o.TermsOfService == nil {
+	if o == nil || IsNil(o.TermsOfService) {
 		return nil, false
 	}
 	return o.TermsOfService, true
@@ -364,7 +370,7 @@ func (o *SellerLegalInfo) GetTermsOfServiceOk() (*string, bool) {
 
 // HasTermsOfService returns a boolean if a field has been set.
 func (o *SellerLegalInfo) HasTermsOfService() bool {
-	if o != nil && o.TermsOfService != nil {
+	if o != nil && !IsNil(o.TermsOfService) {
 		return true
 	}
 
@@ -378,17 +384,17 @@ func (o *SellerLegalInfo) SetTermsOfService(v string) {
 
 // GetVatDetails returns the VatDetails field value if set, zero value otherwise.
 func (o *SellerLegalInfo) GetVatDetails() []VatDetail {
-	if o == nil || o.VatDetails == nil {
+	if o == nil || IsNil(o.VatDetails) {
 		var ret []VatDetail
 		return ret
 	}
-	return *o.VatDetails
+	return o.VatDetails
 }
 
 // GetVatDetailsOk returns a tuple with the VatDetails field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *SellerLegalInfo) GetVatDetailsOk() (*[]VatDetail, bool) {
-	if o == nil || o.VatDetails == nil {
+func (o *SellerLegalInfo) GetVatDetailsOk() ([]VatDetail, bool) {
+	if o == nil || IsNil(o.VatDetails) {
 		return nil, false
 	}
 	return o.VatDetails, true
@@ -396,7 +402,7 @@ func (o *SellerLegalInfo) GetVatDetailsOk() (*[]VatDetail, bool) {
 
 // HasVatDetails returns a boolean if a field has been set.
 func (o *SellerLegalInfo) HasVatDetails() bool {
-	if o != nil && o.VatDetails != nil {
+	if o != nil && !IsNil(o.VatDetails) {
 		return true
 	}
 
@@ -405,45 +411,123 @@ func (o *SellerLegalInfo) HasVatDetails() bool {
 
 // SetVatDetails gets a reference to the given []VatDetail and assigns it to the VatDetails field.
 func (o *SellerLegalInfo) SetVatDetails(v []VatDetail) {
-	o.VatDetails = &v
+	o.VatDetails = v
+}
+
+// GetEconomicOperator returns the EconomicOperator field value if set, zero value otherwise.
+func (o *SellerLegalInfo) GetEconomicOperator() EconomicOperator {
+	if o == nil || IsNil(o.EconomicOperator) {
+		var ret EconomicOperator
+		return ret
+	}
+	return *o.EconomicOperator
+}
+
+// GetEconomicOperatorOk returns a tuple with the EconomicOperator field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *SellerLegalInfo) GetEconomicOperatorOk() (*EconomicOperator, bool) {
+	if o == nil || IsNil(o.EconomicOperator) {
+		return nil, false
+	}
+	return o.EconomicOperator, true
+}
+
+// HasEconomicOperator returns a boolean if a field has been set.
+func (o *SellerLegalInfo) HasEconomicOperator() bool {
+	if o != nil && !IsNil(o.EconomicOperator) {
+		return true
+	}
+
+	return false
+}
+
+// SetEconomicOperator gets a reference to the given EconomicOperator and assigns it to the EconomicOperator field.
+func (o *SellerLegalInfo) SetEconomicOperator(v EconomicOperator) {
+	o.EconomicOperator = &v
+}
+
+// GetWeeeNumber returns the WeeeNumber field value if set, zero value otherwise.
+func (o *SellerLegalInfo) GetWeeeNumber() string {
+	if o == nil || IsNil(o.WeeeNumber) {
+		var ret string
+		return ret
+	}
+	return *o.WeeeNumber
+}
+
+// GetWeeeNumberOk returns a tuple with the WeeeNumber field value if set, nil otherwise
+// and a boolean to check if the value has been set.
+func (o *SellerLegalInfo) GetWeeeNumberOk() (*string, bool) {
+	if o == nil || IsNil(o.WeeeNumber) {
+		return nil, false
+	}
+	return o.WeeeNumber, true
+}
+
+// HasWeeeNumber returns a boolean if a field has been set.
+func (o *SellerLegalInfo) HasWeeeNumber() bool {
+	if o != nil && !IsNil(o.WeeeNumber) {
+		return true
+	}
+
+	return false
+}
+
+// SetWeeeNumber gets a reference to the given string and assigns it to the WeeeNumber field.
+func (o *SellerLegalInfo) SetWeeeNumber(v string) {
+	o.WeeeNumber = &v
 }
 
 func (o SellerLegalInfo) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.Email != nil {
-		toSerialize["email"] = o.Email
-	}
-	if o.Fax != nil {
-		toSerialize["fax"] = o.Fax
-	}
-	if o.Imprint != nil {
-		toSerialize["imprint"] = o.Imprint
-	}
-	if o.LegalContactFirstName != nil {
-		toSerialize["legalContactFirstName"] = o.LegalContactFirstName
-	}
-	if o.LegalContactLastName != nil {
-		toSerialize["legalContactLastName"] = o.LegalContactLastName
-	}
-	if o.Name != nil {
-		toSerialize["name"] = o.Name
-	}
-	if o.Phone != nil {
-		toSerialize["phone"] = o.Phone
-	}
-	if o.RegistrationNumber != nil {
-		toSerialize["registrationNumber"] = o.RegistrationNumber
-	}
-	if o.SellerProvidedLegalAddress != nil {
-		toSerialize["sellerProvidedLegalAddress"] = o.SellerProvidedLegalAddress
-	}
-	if o.TermsOfService != nil {
-		toSerialize["termsOfService"] = o.TermsOfService
-	}
-	if o.VatDetails != nil {
-		toSerialize["vatDetails"] = o.VatDetails
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o SellerLegalInfo) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.Email) {
+		toSerialize["email"] = o.Email
+	}
+	if !IsNil(o.Fax) {
+		toSerialize["fax"] = o.Fax
+	}
+	if !IsNil(o.Imprint) {
+		toSerialize["imprint"] = o.Imprint
+	}
+	if !IsNil(o.LegalContactFirstName) {
+		toSerialize["legalContactFirstName"] = o.LegalContactFirstName
+	}
+	if !IsNil(o.LegalContactLastName) {
+		toSerialize["legalContactLastName"] = o.LegalContactLastName
+	}
+	if !IsNil(o.Name) {
+		toSerialize["name"] = o.Name
+	}
+	if !IsNil(o.Phone) {
+		toSerialize["phone"] = o.Phone
+	}
+	if !IsNil(o.RegistrationNumber) {
+		toSerialize["registrationNumber"] = o.RegistrationNumber
+	}
+	if !IsNil(o.SellerProvidedLegalAddress) {
+		toSerialize["sellerProvidedLegalAddress"] = o.SellerProvidedLegalAddress
+	}
+	if !IsNil(o.TermsOfService) {
+		toSerialize["termsOfService"] = o.TermsOfService
+	}
+	if !IsNil(o.VatDetails) {
+		toSerialize["vatDetails"] = o.VatDetails
+	}
+	if !IsNil(o.EconomicOperator) {
+		toSerialize["economicOperator"] = o.EconomicOperator
+	}
+	if !IsNil(o.WeeeNumber) {
+		toSerialize["weeeNumber"] = o.WeeeNumber
+	}
+	return toSerialize, nil
 }
 
 type NullableSellerLegalInfo struct {
