@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,12 +14,15 @@ import (
 	"encoding/json"
 )
 
+// checks if the ShipToLocations type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &ShipToLocations{}
+
 // ShipToLocations The type that defines the fields that include and exclude geographic regions affecting where the item can be shipped. The seller defines these regions when listing the item.
 type ShipToLocations struct {
 	// An array of containers that express the large geographical regions, countries, state/provinces, or special locations within a country where the seller is not willing to ship to.
-	RegionExcluded *[]ShipToRegion `json:"regionExcluded,omitempty"`
+	RegionExcluded []ShipToRegion `json:"regionExcluded,omitempty"`
 	// An array of containers that express the large geographical regions, countries, or state/provinces within a country where the seller is willing to ship to. Prospective buyers must look at the shipping regions under this container, as well as the shipping regions that are under the <b>regionExcluded</b> to see where the seller is willing to ship items. Sellers can specify that they ship 'Worldwide', but then add several large geographical regions (e.g. Asia, Oceania, Middle East) to the exclusion list, or sellers can specify that they ship to Europe and Africa, but then add several individual countries to the exclusion list.
-	RegionIncluded *[]ShipToRegion `json:"regionIncluded,omitempty"`
+	RegionIncluded []ShipToRegion `json:"regionIncluded,omitempty"`
 }
 
 // NewShipToLocations instantiates a new ShipToLocations object
@@ -41,17 +44,17 @@ func NewShipToLocationsWithDefaults() *ShipToLocations {
 
 // GetRegionExcluded returns the RegionExcluded field value if set, zero value otherwise.
 func (o *ShipToLocations) GetRegionExcluded() []ShipToRegion {
-	if o == nil || o.RegionExcluded == nil {
+	if o == nil || IsNil(o.RegionExcluded) {
 		var ret []ShipToRegion
 		return ret
 	}
-	return *o.RegionExcluded
+	return o.RegionExcluded
 }
 
 // GetRegionExcludedOk returns a tuple with the RegionExcluded field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *ShipToLocations) GetRegionExcludedOk() (*[]ShipToRegion, bool) {
-	if o == nil || o.RegionExcluded == nil {
+func (o *ShipToLocations) GetRegionExcludedOk() ([]ShipToRegion, bool) {
+	if o == nil || IsNil(o.RegionExcluded) {
 		return nil, false
 	}
 	return o.RegionExcluded, true
@@ -59,7 +62,7 @@ func (o *ShipToLocations) GetRegionExcludedOk() (*[]ShipToRegion, bool) {
 
 // HasRegionExcluded returns a boolean if a field has been set.
 func (o *ShipToLocations) HasRegionExcluded() bool {
-	if o != nil && o.RegionExcluded != nil {
+	if o != nil && !IsNil(o.RegionExcluded) {
 		return true
 	}
 
@@ -68,22 +71,22 @@ func (o *ShipToLocations) HasRegionExcluded() bool {
 
 // SetRegionExcluded gets a reference to the given []ShipToRegion and assigns it to the RegionExcluded field.
 func (o *ShipToLocations) SetRegionExcluded(v []ShipToRegion) {
-	o.RegionExcluded = &v
+	o.RegionExcluded = v
 }
 
 // GetRegionIncluded returns the RegionIncluded field value if set, zero value otherwise.
 func (o *ShipToLocations) GetRegionIncluded() []ShipToRegion {
-	if o == nil || o.RegionIncluded == nil {
+	if o == nil || IsNil(o.RegionIncluded) {
 		var ret []ShipToRegion
 		return ret
 	}
-	return *o.RegionIncluded
+	return o.RegionIncluded
 }
 
 // GetRegionIncludedOk returns a tuple with the RegionIncluded field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *ShipToLocations) GetRegionIncludedOk() (*[]ShipToRegion, bool) {
-	if o == nil || o.RegionIncluded == nil {
+func (o *ShipToLocations) GetRegionIncludedOk() ([]ShipToRegion, bool) {
+	if o == nil || IsNil(o.RegionIncluded) {
 		return nil, false
 	}
 	return o.RegionIncluded, true
@@ -91,7 +94,7 @@ func (o *ShipToLocations) GetRegionIncludedOk() (*[]ShipToRegion, bool) {
 
 // HasRegionIncluded returns a boolean if a field has been set.
 func (o *ShipToLocations) HasRegionIncluded() bool {
-	if o != nil && o.RegionIncluded != nil {
+	if o != nil && !IsNil(o.RegionIncluded) {
 		return true
 	}
 
@@ -100,18 +103,26 @@ func (o *ShipToLocations) HasRegionIncluded() bool {
 
 // SetRegionIncluded gets a reference to the given []ShipToRegion and assigns it to the RegionIncluded field.
 func (o *ShipToLocations) SetRegionIncluded(v []ShipToRegion) {
-	o.RegionIncluded = &v
+	o.RegionIncluded = v
 }
 
 func (o ShipToLocations) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.RegionExcluded != nil {
-		toSerialize["regionExcluded"] = o.RegionExcluded
-	}
-	if o.RegionIncluded != nil {
-		toSerialize["regionIncluded"] = o.RegionIncluded
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o ShipToLocations) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.RegionExcluded) {
+		toSerialize["regionExcluded"] = o.RegionExcluded
+	}
+	if !IsNil(o.RegionIncluded) {
+		toSerialize["regionIncluded"] = o.RegionIncluded
+	}
+	return toSerialize, nil
 }
 
 type NullableShipToLocations struct {
