@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,26 +14,29 @@ import (
 	"encoding/json"
 )
 
+// checks if the ShippingOption type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &ShippingOption{}
+
 // ShippingOption The type that defines the fields for the details of a shipping provider.
 type ShippingOption struct {
 	AdditionalShippingCostPerUnit *ConvertedAmount `json:"additionalShippingCostPerUnit,omitempty"`
-	// The deadline date that the item must be purchased by in order to be received by the buyer within the delivery window (<b> maxEstimatedDeliveryDate</b> and  <b> minEstimatedDeliveryDate</b> fields). This field is returned only for items that are eligible for 'Same Day Handling'. For these items, the value of this field is what is displayed in the <b> Delivery</b> line on the View Item page.  <br /><br />This value is returned in UTC format (yyyy-MM-ddThh:mm:ss.sssZ), which you can convert into the local time of the buyer.
+	// The deadline date that the item must be purchased by in order to be received by the buyer within the delivery window (<b> maxEstimatedDeliveryDate</b> and  <b> minEstimatedDeliveryDate</b> fields). This field is returned only for items that are eligible for 'Same Day Handling'. For these items, the value of this field is what is displayed in the <b> Delivery</b> line on the View Item page.  <br><br>This value is returned in UTC format (yyyy-MM-ddThh:mm:ss.sssZ), which you can convert into the local time of the buyer.
 	CutOffDateUsedForEstimate *string `json:"cutOffDateUsedForEstimate,omitempty"`
-	// If the item is being shipped by eBay's <a href=\"https://pages.ebay.com/seller-center/shipping/global-shipping-program.html\">Global Shipping Program</a>, this field returns <code>GLOBAL_SHIPPING</code>. Otherwise this field is null. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:FulfilledThroughEnum'>eBay API documentation</a>
+	// If the item is being shipped by the eBay <a href=\"https://pages.ebay.com/seller-center/shipping/global-shipping-program.html \">Global Shipping program</a>, this field returns <code>GLOBAL_SHIPPING</code>.<br><br>If the item is being shipped using the eBay International Shipping program, this field returns <code>INTERNATIONAL_SHIPPING</code>. <br><br>Otherwise, this field is null. For implementation help, refer to <a href='https://developer.ebay.com/api-docs/buy/browse/types/gct:FulfilledThroughEnum'>eBay API documentation</a>
 	FulfilledThrough *string `json:"fulfilledThrough,omitempty"`
-	// Indicates if the seller has committed to shipping the item with eBay Guaranteed Delivery. With eBay Guaranteed Delivery, the  seller is committed to getting the line item to the buyer within 4 business days or less. See the <a href=\"https://www.ebay.com/help/buying/shipping-delivery/buying-items-ebay-guaranteed-delivery?id=4641\">Buying items with eBay Guaranteed Delivery</a> help topic for more details about eBay Guaranteed Delivery.
+	// Although this field is still returned, it can be ignored since eBay Guaranteed Delivery is no longer a supported feature on any marketplace. This field may get removed from the schema in the future.
 	GuaranteedDelivery *bool `json:"guaranteedDelivery,omitempty"`
 	ImportCharges *ConvertedAmount `json:"importCharges,omitempty"`
-	// The end date of the delivery window (latest projected delivery date).  This value is returned in UTC format (yyyy-MM-ddThh:mm:ss.sssZ), which you can convert into the local time of the buyer. <br /> <br /> <span class=\"tablenote\"> <b> Note: </b> For the best accuracy, always include the location of where the item is be shipped in the <code> contextualLocation</code> values of the <a href=\"/api-docs/buy/static/api-browse.html#Headers\"> <code>X-EBAY-C-ENDUSERCTX</code></a> request header.</span> 
+	// The end date of the delivery window (latest projected delivery date).  This value is returned in UTC format (yyyy-MM-ddThh:mm:ss.sssZ), which you can convert into the local time of the buyer. <br> <br> <span class=\"tablenote\"> <b> Note: </b> For the best accuracy, always include the location of where the item is be shipped in the <code> contextualLocation</code> values of the <a href=\"/api-docs/buy/static/api-browse.html#Headers\"> <code>X-EBAY-C-ENDUSERCTX</code></a> request header.</span> 
 	MaxEstimatedDeliveryDate *string `json:"maxEstimatedDeliveryDate,omitempty"`
-	// The start date of the delivery window (earliest projected delivery date). This value is returned in UTC format (yyyy-MM-ddThh:mm:ss.sssZ), which you can convert into the local time of the buyer. <br /> <br /><span class=\"tablenote\"> <b> Note: </b> For the best accuracy, always include the location of where the item is be shipped in the <code> contextualLocation</code> values of the <a href=\"/api-docs/buy/static/api-browse.html#Headers\"> <code>X-EBAY-C-ENDUSERCTX</code></a> request header.</span>
+	// The start date of the delivery window (earliest projected delivery date). This value is returned in UTC format (yyyy-MM-ddThh:mm:ss.sssZ), which you can convert into the local time of the buyer. <br> <br><span class=\"tablenote\"> <b> Note: </b> For the best accuracy, always include the location of where the item is be shipped in the <code> contextualLocation</code> values of the <a href=\"/api-docs/buy/static/api-browse.html#Headers\"> <code>X-EBAY-C-ENDUSERCTX</code></a> request header.</span>
 	MinEstimatedDeliveryDate *string `json:"minEstimatedDeliveryDate,omitempty"`
 	// The number of items used when calculating the estimation information.
 	QuantityUsedForEstimate *int32 `json:"quantityUsedForEstimate,omitempty"`
 	// The name of the shipping provider, such as FedEx, or USPS.
 	ShippingCarrierCode *string `json:"shippingCarrierCode,omitempty"`
 	ShippingCost *ConvertedAmount `json:"shippingCost,omitempty"`
-	// Indicates the class of the shipping cost. <br /><br /><b> Valid Values: </b> FIXED or CALCULATED <br /><br />Code so that your app gracefully handles any future changes to this list. 
+	// Indicates the class of the shipping cost. <br><br><b> Valid Values: </b> FIXED or CALCULATED <br><br>Code so that your app gracefully handles any future changes to this list. 
 	ShippingCostType *string `json:"shippingCostType,omitempty"`
 	// The type of shipping service. For example, USPS First Class.
 	ShippingServiceCode *string `json:"shippingServiceCode,omitempty"`
@@ -63,7 +66,7 @@ func NewShippingOptionWithDefaults() *ShippingOption {
 
 // GetAdditionalShippingCostPerUnit returns the AdditionalShippingCostPerUnit field value if set, zero value otherwise.
 func (o *ShippingOption) GetAdditionalShippingCostPerUnit() ConvertedAmount {
-	if o == nil || o.AdditionalShippingCostPerUnit == nil {
+	if o == nil || IsNil(o.AdditionalShippingCostPerUnit) {
 		var ret ConvertedAmount
 		return ret
 	}
@@ -73,7 +76,7 @@ func (o *ShippingOption) GetAdditionalShippingCostPerUnit() ConvertedAmount {
 // GetAdditionalShippingCostPerUnitOk returns a tuple with the AdditionalShippingCostPerUnit field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetAdditionalShippingCostPerUnitOk() (*ConvertedAmount, bool) {
-	if o == nil || o.AdditionalShippingCostPerUnit == nil {
+	if o == nil || IsNil(o.AdditionalShippingCostPerUnit) {
 		return nil, false
 	}
 	return o.AdditionalShippingCostPerUnit, true
@@ -81,7 +84,7 @@ func (o *ShippingOption) GetAdditionalShippingCostPerUnitOk() (*ConvertedAmount,
 
 // HasAdditionalShippingCostPerUnit returns a boolean if a field has been set.
 func (o *ShippingOption) HasAdditionalShippingCostPerUnit() bool {
-	if o != nil && o.AdditionalShippingCostPerUnit != nil {
+	if o != nil && !IsNil(o.AdditionalShippingCostPerUnit) {
 		return true
 	}
 
@@ -95,7 +98,7 @@ func (o *ShippingOption) SetAdditionalShippingCostPerUnit(v ConvertedAmount) {
 
 // GetCutOffDateUsedForEstimate returns the CutOffDateUsedForEstimate field value if set, zero value otherwise.
 func (o *ShippingOption) GetCutOffDateUsedForEstimate() string {
-	if o == nil || o.CutOffDateUsedForEstimate == nil {
+	if o == nil || IsNil(o.CutOffDateUsedForEstimate) {
 		var ret string
 		return ret
 	}
@@ -105,7 +108,7 @@ func (o *ShippingOption) GetCutOffDateUsedForEstimate() string {
 // GetCutOffDateUsedForEstimateOk returns a tuple with the CutOffDateUsedForEstimate field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetCutOffDateUsedForEstimateOk() (*string, bool) {
-	if o == nil || o.CutOffDateUsedForEstimate == nil {
+	if o == nil || IsNil(o.CutOffDateUsedForEstimate) {
 		return nil, false
 	}
 	return o.CutOffDateUsedForEstimate, true
@@ -113,7 +116,7 @@ func (o *ShippingOption) GetCutOffDateUsedForEstimateOk() (*string, bool) {
 
 // HasCutOffDateUsedForEstimate returns a boolean if a field has been set.
 func (o *ShippingOption) HasCutOffDateUsedForEstimate() bool {
-	if o != nil && o.CutOffDateUsedForEstimate != nil {
+	if o != nil && !IsNil(o.CutOffDateUsedForEstimate) {
 		return true
 	}
 
@@ -127,7 +130,7 @@ func (o *ShippingOption) SetCutOffDateUsedForEstimate(v string) {
 
 // GetFulfilledThrough returns the FulfilledThrough field value if set, zero value otherwise.
 func (o *ShippingOption) GetFulfilledThrough() string {
-	if o == nil || o.FulfilledThrough == nil {
+	if o == nil || IsNil(o.FulfilledThrough) {
 		var ret string
 		return ret
 	}
@@ -137,7 +140,7 @@ func (o *ShippingOption) GetFulfilledThrough() string {
 // GetFulfilledThroughOk returns a tuple with the FulfilledThrough field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetFulfilledThroughOk() (*string, bool) {
-	if o == nil || o.FulfilledThrough == nil {
+	if o == nil || IsNil(o.FulfilledThrough) {
 		return nil, false
 	}
 	return o.FulfilledThrough, true
@@ -145,7 +148,7 @@ func (o *ShippingOption) GetFulfilledThroughOk() (*string, bool) {
 
 // HasFulfilledThrough returns a boolean if a field has been set.
 func (o *ShippingOption) HasFulfilledThrough() bool {
-	if o != nil && o.FulfilledThrough != nil {
+	if o != nil && !IsNil(o.FulfilledThrough) {
 		return true
 	}
 
@@ -159,7 +162,7 @@ func (o *ShippingOption) SetFulfilledThrough(v string) {
 
 // GetGuaranteedDelivery returns the GuaranteedDelivery field value if set, zero value otherwise.
 func (o *ShippingOption) GetGuaranteedDelivery() bool {
-	if o == nil || o.GuaranteedDelivery == nil {
+	if o == nil || IsNil(o.GuaranteedDelivery) {
 		var ret bool
 		return ret
 	}
@@ -169,7 +172,7 @@ func (o *ShippingOption) GetGuaranteedDelivery() bool {
 // GetGuaranteedDeliveryOk returns a tuple with the GuaranteedDelivery field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetGuaranteedDeliveryOk() (*bool, bool) {
-	if o == nil || o.GuaranteedDelivery == nil {
+	if o == nil || IsNil(o.GuaranteedDelivery) {
 		return nil, false
 	}
 	return o.GuaranteedDelivery, true
@@ -177,7 +180,7 @@ func (o *ShippingOption) GetGuaranteedDeliveryOk() (*bool, bool) {
 
 // HasGuaranteedDelivery returns a boolean if a field has been set.
 func (o *ShippingOption) HasGuaranteedDelivery() bool {
-	if o != nil && o.GuaranteedDelivery != nil {
+	if o != nil && !IsNil(o.GuaranteedDelivery) {
 		return true
 	}
 
@@ -191,7 +194,7 @@ func (o *ShippingOption) SetGuaranteedDelivery(v bool) {
 
 // GetImportCharges returns the ImportCharges field value if set, zero value otherwise.
 func (o *ShippingOption) GetImportCharges() ConvertedAmount {
-	if o == nil || o.ImportCharges == nil {
+	if o == nil || IsNil(o.ImportCharges) {
 		var ret ConvertedAmount
 		return ret
 	}
@@ -201,7 +204,7 @@ func (o *ShippingOption) GetImportCharges() ConvertedAmount {
 // GetImportChargesOk returns a tuple with the ImportCharges field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetImportChargesOk() (*ConvertedAmount, bool) {
-	if o == nil || o.ImportCharges == nil {
+	if o == nil || IsNil(o.ImportCharges) {
 		return nil, false
 	}
 	return o.ImportCharges, true
@@ -209,7 +212,7 @@ func (o *ShippingOption) GetImportChargesOk() (*ConvertedAmount, bool) {
 
 // HasImportCharges returns a boolean if a field has been set.
 func (o *ShippingOption) HasImportCharges() bool {
-	if o != nil && o.ImportCharges != nil {
+	if o != nil && !IsNil(o.ImportCharges) {
 		return true
 	}
 
@@ -223,7 +226,7 @@ func (o *ShippingOption) SetImportCharges(v ConvertedAmount) {
 
 // GetMaxEstimatedDeliveryDate returns the MaxEstimatedDeliveryDate field value if set, zero value otherwise.
 func (o *ShippingOption) GetMaxEstimatedDeliveryDate() string {
-	if o == nil || o.MaxEstimatedDeliveryDate == nil {
+	if o == nil || IsNil(o.MaxEstimatedDeliveryDate) {
 		var ret string
 		return ret
 	}
@@ -233,7 +236,7 @@ func (o *ShippingOption) GetMaxEstimatedDeliveryDate() string {
 // GetMaxEstimatedDeliveryDateOk returns a tuple with the MaxEstimatedDeliveryDate field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetMaxEstimatedDeliveryDateOk() (*string, bool) {
-	if o == nil || o.MaxEstimatedDeliveryDate == nil {
+	if o == nil || IsNil(o.MaxEstimatedDeliveryDate) {
 		return nil, false
 	}
 	return o.MaxEstimatedDeliveryDate, true
@@ -241,7 +244,7 @@ func (o *ShippingOption) GetMaxEstimatedDeliveryDateOk() (*string, bool) {
 
 // HasMaxEstimatedDeliveryDate returns a boolean if a field has been set.
 func (o *ShippingOption) HasMaxEstimatedDeliveryDate() bool {
-	if o != nil && o.MaxEstimatedDeliveryDate != nil {
+	if o != nil && !IsNil(o.MaxEstimatedDeliveryDate) {
 		return true
 	}
 
@@ -255,7 +258,7 @@ func (o *ShippingOption) SetMaxEstimatedDeliveryDate(v string) {
 
 // GetMinEstimatedDeliveryDate returns the MinEstimatedDeliveryDate field value if set, zero value otherwise.
 func (o *ShippingOption) GetMinEstimatedDeliveryDate() string {
-	if o == nil || o.MinEstimatedDeliveryDate == nil {
+	if o == nil || IsNil(o.MinEstimatedDeliveryDate) {
 		var ret string
 		return ret
 	}
@@ -265,7 +268,7 @@ func (o *ShippingOption) GetMinEstimatedDeliveryDate() string {
 // GetMinEstimatedDeliveryDateOk returns a tuple with the MinEstimatedDeliveryDate field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetMinEstimatedDeliveryDateOk() (*string, bool) {
-	if o == nil || o.MinEstimatedDeliveryDate == nil {
+	if o == nil || IsNil(o.MinEstimatedDeliveryDate) {
 		return nil, false
 	}
 	return o.MinEstimatedDeliveryDate, true
@@ -273,7 +276,7 @@ func (o *ShippingOption) GetMinEstimatedDeliveryDateOk() (*string, bool) {
 
 // HasMinEstimatedDeliveryDate returns a boolean if a field has been set.
 func (o *ShippingOption) HasMinEstimatedDeliveryDate() bool {
-	if o != nil && o.MinEstimatedDeliveryDate != nil {
+	if o != nil && !IsNil(o.MinEstimatedDeliveryDate) {
 		return true
 	}
 
@@ -287,7 +290,7 @@ func (o *ShippingOption) SetMinEstimatedDeliveryDate(v string) {
 
 // GetQuantityUsedForEstimate returns the QuantityUsedForEstimate field value if set, zero value otherwise.
 func (o *ShippingOption) GetQuantityUsedForEstimate() int32 {
-	if o == nil || o.QuantityUsedForEstimate == nil {
+	if o == nil || IsNil(o.QuantityUsedForEstimate) {
 		var ret int32
 		return ret
 	}
@@ -297,7 +300,7 @@ func (o *ShippingOption) GetQuantityUsedForEstimate() int32 {
 // GetQuantityUsedForEstimateOk returns a tuple with the QuantityUsedForEstimate field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetQuantityUsedForEstimateOk() (*int32, bool) {
-	if o == nil || o.QuantityUsedForEstimate == nil {
+	if o == nil || IsNil(o.QuantityUsedForEstimate) {
 		return nil, false
 	}
 	return o.QuantityUsedForEstimate, true
@@ -305,7 +308,7 @@ func (o *ShippingOption) GetQuantityUsedForEstimateOk() (*int32, bool) {
 
 // HasQuantityUsedForEstimate returns a boolean if a field has been set.
 func (o *ShippingOption) HasQuantityUsedForEstimate() bool {
-	if o != nil && o.QuantityUsedForEstimate != nil {
+	if o != nil && !IsNil(o.QuantityUsedForEstimate) {
 		return true
 	}
 
@@ -319,7 +322,7 @@ func (o *ShippingOption) SetQuantityUsedForEstimate(v int32) {
 
 // GetShippingCarrierCode returns the ShippingCarrierCode field value if set, zero value otherwise.
 func (o *ShippingOption) GetShippingCarrierCode() string {
-	if o == nil || o.ShippingCarrierCode == nil {
+	if o == nil || IsNil(o.ShippingCarrierCode) {
 		var ret string
 		return ret
 	}
@@ -329,7 +332,7 @@ func (o *ShippingOption) GetShippingCarrierCode() string {
 // GetShippingCarrierCodeOk returns a tuple with the ShippingCarrierCode field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetShippingCarrierCodeOk() (*string, bool) {
-	if o == nil || o.ShippingCarrierCode == nil {
+	if o == nil || IsNil(o.ShippingCarrierCode) {
 		return nil, false
 	}
 	return o.ShippingCarrierCode, true
@@ -337,7 +340,7 @@ func (o *ShippingOption) GetShippingCarrierCodeOk() (*string, bool) {
 
 // HasShippingCarrierCode returns a boolean if a field has been set.
 func (o *ShippingOption) HasShippingCarrierCode() bool {
-	if o != nil && o.ShippingCarrierCode != nil {
+	if o != nil && !IsNil(o.ShippingCarrierCode) {
 		return true
 	}
 
@@ -351,7 +354,7 @@ func (o *ShippingOption) SetShippingCarrierCode(v string) {
 
 // GetShippingCost returns the ShippingCost field value if set, zero value otherwise.
 func (o *ShippingOption) GetShippingCost() ConvertedAmount {
-	if o == nil || o.ShippingCost == nil {
+	if o == nil || IsNil(o.ShippingCost) {
 		var ret ConvertedAmount
 		return ret
 	}
@@ -361,7 +364,7 @@ func (o *ShippingOption) GetShippingCost() ConvertedAmount {
 // GetShippingCostOk returns a tuple with the ShippingCost field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetShippingCostOk() (*ConvertedAmount, bool) {
-	if o == nil || o.ShippingCost == nil {
+	if o == nil || IsNil(o.ShippingCost) {
 		return nil, false
 	}
 	return o.ShippingCost, true
@@ -369,7 +372,7 @@ func (o *ShippingOption) GetShippingCostOk() (*ConvertedAmount, bool) {
 
 // HasShippingCost returns a boolean if a field has been set.
 func (o *ShippingOption) HasShippingCost() bool {
-	if o != nil && o.ShippingCost != nil {
+	if o != nil && !IsNil(o.ShippingCost) {
 		return true
 	}
 
@@ -383,7 +386,7 @@ func (o *ShippingOption) SetShippingCost(v ConvertedAmount) {
 
 // GetShippingCostType returns the ShippingCostType field value if set, zero value otherwise.
 func (o *ShippingOption) GetShippingCostType() string {
-	if o == nil || o.ShippingCostType == nil {
+	if o == nil || IsNil(o.ShippingCostType) {
 		var ret string
 		return ret
 	}
@@ -393,7 +396,7 @@ func (o *ShippingOption) GetShippingCostType() string {
 // GetShippingCostTypeOk returns a tuple with the ShippingCostType field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetShippingCostTypeOk() (*string, bool) {
-	if o == nil || o.ShippingCostType == nil {
+	if o == nil || IsNil(o.ShippingCostType) {
 		return nil, false
 	}
 	return o.ShippingCostType, true
@@ -401,7 +404,7 @@ func (o *ShippingOption) GetShippingCostTypeOk() (*string, bool) {
 
 // HasShippingCostType returns a boolean if a field has been set.
 func (o *ShippingOption) HasShippingCostType() bool {
-	if o != nil && o.ShippingCostType != nil {
+	if o != nil && !IsNil(o.ShippingCostType) {
 		return true
 	}
 
@@ -415,7 +418,7 @@ func (o *ShippingOption) SetShippingCostType(v string) {
 
 // GetShippingServiceCode returns the ShippingServiceCode field value if set, zero value otherwise.
 func (o *ShippingOption) GetShippingServiceCode() string {
-	if o == nil || o.ShippingServiceCode == nil {
+	if o == nil || IsNil(o.ShippingServiceCode) {
 		var ret string
 		return ret
 	}
@@ -425,7 +428,7 @@ func (o *ShippingOption) GetShippingServiceCode() string {
 // GetShippingServiceCodeOk returns a tuple with the ShippingServiceCode field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetShippingServiceCodeOk() (*string, bool) {
-	if o == nil || o.ShippingServiceCode == nil {
+	if o == nil || IsNil(o.ShippingServiceCode) {
 		return nil, false
 	}
 	return o.ShippingServiceCode, true
@@ -433,7 +436,7 @@ func (o *ShippingOption) GetShippingServiceCodeOk() (*string, bool) {
 
 // HasShippingServiceCode returns a boolean if a field has been set.
 func (o *ShippingOption) HasShippingServiceCode() bool {
-	if o != nil && o.ShippingServiceCode != nil {
+	if o != nil && !IsNil(o.ShippingServiceCode) {
 		return true
 	}
 
@@ -447,7 +450,7 @@ func (o *ShippingOption) SetShippingServiceCode(v string) {
 
 // GetShipToLocationUsedForEstimate returns the ShipToLocationUsedForEstimate field value if set, zero value otherwise.
 func (o *ShippingOption) GetShipToLocationUsedForEstimate() ShipToLocation {
-	if o == nil || o.ShipToLocationUsedForEstimate == nil {
+	if o == nil || IsNil(o.ShipToLocationUsedForEstimate) {
 		var ret ShipToLocation
 		return ret
 	}
@@ -457,7 +460,7 @@ func (o *ShippingOption) GetShipToLocationUsedForEstimate() ShipToLocation {
 // GetShipToLocationUsedForEstimateOk returns a tuple with the ShipToLocationUsedForEstimate field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetShipToLocationUsedForEstimateOk() (*ShipToLocation, bool) {
-	if o == nil || o.ShipToLocationUsedForEstimate == nil {
+	if o == nil || IsNil(o.ShipToLocationUsedForEstimate) {
 		return nil, false
 	}
 	return o.ShipToLocationUsedForEstimate, true
@@ -465,7 +468,7 @@ func (o *ShippingOption) GetShipToLocationUsedForEstimateOk() (*ShipToLocation, 
 
 // HasShipToLocationUsedForEstimate returns a boolean if a field has been set.
 func (o *ShippingOption) HasShipToLocationUsedForEstimate() bool {
-	if o != nil && o.ShipToLocationUsedForEstimate != nil {
+	if o != nil && !IsNil(o.ShipToLocationUsedForEstimate) {
 		return true
 	}
 
@@ -479,7 +482,7 @@ func (o *ShippingOption) SetShipToLocationUsedForEstimate(v ShipToLocation) {
 
 // GetTrademarkSymbol returns the TrademarkSymbol field value if set, zero value otherwise.
 func (o *ShippingOption) GetTrademarkSymbol() string {
-	if o == nil || o.TrademarkSymbol == nil {
+	if o == nil || IsNil(o.TrademarkSymbol) {
 		var ret string
 		return ret
 	}
@@ -489,7 +492,7 @@ func (o *ShippingOption) GetTrademarkSymbol() string {
 // GetTrademarkSymbolOk returns a tuple with the TrademarkSymbol field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetTrademarkSymbolOk() (*string, bool) {
-	if o == nil || o.TrademarkSymbol == nil {
+	if o == nil || IsNil(o.TrademarkSymbol) {
 		return nil, false
 	}
 	return o.TrademarkSymbol, true
@@ -497,7 +500,7 @@ func (o *ShippingOption) GetTrademarkSymbolOk() (*string, bool) {
 
 // HasTrademarkSymbol returns a boolean if a field has been set.
 func (o *ShippingOption) HasTrademarkSymbol() bool {
-	if o != nil && o.TrademarkSymbol != nil {
+	if o != nil && !IsNil(o.TrademarkSymbol) {
 		return true
 	}
 
@@ -511,7 +514,7 @@ func (o *ShippingOption) SetTrademarkSymbol(v string) {
 
 // GetType returns the Type field value if set, zero value otherwise.
 func (o *ShippingOption) GetType() string {
-	if o == nil || o.Type == nil {
+	if o == nil || IsNil(o.Type) {
 		var ret string
 		return ret
 	}
@@ -521,7 +524,7 @@ func (o *ShippingOption) GetType() string {
 // GetTypeOk returns a tuple with the Type field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOption) GetTypeOk() (*string, bool) {
-	if o == nil || o.Type == nil {
+	if o == nil || IsNil(o.Type) {
 		return nil, false
 	}
 	return o.Type, true
@@ -529,7 +532,7 @@ func (o *ShippingOption) GetTypeOk() (*string, bool) {
 
 // HasType returns a boolean if a field has been set.
 func (o *ShippingOption) HasType() bool {
-	if o != nil && o.Type != nil {
+	if o != nil && !IsNil(o.Type) {
 		return true
 	}
 
@@ -542,53 +545,61 @@ func (o *ShippingOption) SetType(v string) {
 }
 
 func (o ShippingOption) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.AdditionalShippingCostPerUnit != nil {
-		toSerialize["additionalShippingCostPerUnit"] = o.AdditionalShippingCostPerUnit
-	}
-	if o.CutOffDateUsedForEstimate != nil {
-		toSerialize["cutOffDateUsedForEstimate"] = o.CutOffDateUsedForEstimate
-	}
-	if o.FulfilledThrough != nil {
-		toSerialize["fulfilledThrough"] = o.FulfilledThrough
-	}
-	if o.GuaranteedDelivery != nil {
-		toSerialize["guaranteedDelivery"] = o.GuaranteedDelivery
-	}
-	if o.ImportCharges != nil {
-		toSerialize["importCharges"] = o.ImportCharges
-	}
-	if o.MaxEstimatedDeliveryDate != nil {
-		toSerialize["maxEstimatedDeliveryDate"] = o.MaxEstimatedDeliveryDate
-	}
-	if o.MinEstimatedDeliveryDate != nil {
-		toSerialize["minEstimatedDeliveryDate"] = o.MinEstimatedDeliveryDate
-	}
-	if o.QuantityUsedForEstimate != nil {
-		toSerialize["quantityUsedForEstimate"] = o.QuantityUsedForEstimate
-	}
-	if o.ShippingCarrierCode != nil {
-		toSerialize["shippingCarrierCode"] = o.ShippingCarrierCode
-	}
-	if o.ShippingCost != nil {
-		toSerialize["shippingCost"] = o.ShippingCost
-	}
-	if o.ShippingCostType != nil {
-		toSerialize["shippingCostType"] = o.ShippingCostType
-	}
-	if o.ShippingServiceCode != nil {
-		toSerialize["shippingServiceCode"] = o.ShippingServiceCode
-	}
-	if o.ShipToLocationUsedForEstimate != nil {
-		toSerialize["shipToLocationUsedForEstimate"] = o.ShipToLocationUsedForEstimate
-	}
-	if o.TrademarkSymbol != nil {
-		toSerialize["trademarkSymbol"] = o.TrademarkSymbol
-	}
-	if o.Type != nil {
-		toSerialize["type"] = o.Type
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o ShippingOption) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.AdditionalShippingCostPerUnit) {
+		toSerialize["additionalShippingCostPerUnit"] = o.AdditionalShippingCostPerUnit
+	}
+	if !IsNil(o.CutOffDateUsedForEstimate) {
+		toSerialize["cutOffDateUsedForEstimate"] = o.CutOffDateUsedForEstimate
+	}
+	if !IsNil(o.FulfilledThrough) {
+		toSerialize["fulfilledThrough"] = o.FulfilledThrough
+	}
+	if !IsNil(o.GuaranteedDelivery) {
+		toSerialize["guaranteedDelivery"] = o.GuaranteedDelivery
+	}
+	if !IsNil(o.ImportCharges) {
+		toSerialize["importCharges"] = o.ImportCharges
+	}
+	if !IsNil(o.MaxEstimatedDeliveryDate) {
+		toSerialize["maxEstimatedDeliveryDate"] = o.MaxEstimatedDeliveryDate
+	}
+	if !IsNil(o.MinEstimatedDeliveryDate) {
+		toSerialize["minEstimatedDeliveryDate"] = o.MinEstimatedDeliveryDate
+	}
+	if !IsNil(o.QuantityUsedForEstimate) {
+		toSerialize["quantityUsedForEstimate"] = o.QuantityUsedForEstimate
+	}
+	if !IsNil(o.ShippingCarrierCode) {
+		toSerialize["shippingCarrierCode"] = o.ShippingCarrierCode
+	}
+	if !IsNil(o.ShippingCost) {
+		toSerialize["shippingCost"] = o.ShippingCost
+	}
+	if !IsNil(o.ShippingCostType) {
+		toSerialize["shippingCostType"] = o.ShippingCostType
+	}
+	if !IsNil(o.ShippingServiceCode) {
+		toSerialize["shippingServiceCode"] = o.ShippingServiceCode
+	}
+	if !IsNil(o.ShipToLocationUsedForEstimate) {
+		toSerialize["shipToLocationUsedForEstimate"] = o.ShipToLocationUsedForEstimate
+	}
+	if !IsNil(o.TrademarkSymbol) {
+		toSerialize["trademarkSymbol"] = o.TrademarkSymbol
+	}
+	if !IsNil(o.Type) {
+		toSerialize["type"] = o.Type
+	}
+	return toSerialize, nil
 }
 
 type NullableShippingOption struct {
