@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -14,16 +14,19 @@ import (
 	"encoding/json"
 )
 
+// checks if the ShippingOptionSummary type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &ShippingOptionSummary{}
+
 // ShippingOptionSummary The type that defines the fields for the shipping information.
 type ShippingOptionSummary struct {
-	// Indicates if the seller has committed to shipping the item with eBay Guaranteed Delivery. With eBay Guaranteed Delivery, the  seller is committed to getting the line item to the buyer within 4 business days or less. See the <a href=\"https://www.ebay.com/help/buying/shipping-delivery/buying-items-ebay-guaranteed-delivery?id=4641\">Buying items with eBay Guaranteed Delivery</a> help topic for more details about eBay Guaranteed Delivery.
+	// Although this field is still returned, it can be ignored since eBay Guaranteed Delivery is no longer a supported feature on any marketplace. This field may get removed from the schema in the future.
 	GuaranteedDelivery *bool `json:"guaranteedDelivery,omitempty"`
-	// The end date of the delivery window (latest projected delivery date). This value is returned in UTC format (yyyy-MM-ddThh:mm:ss.sssZ), which you can convert into the local time of the buyer. <br /> <br /> <span class=\"tablenote\"> <b> Note: </b> For the best accuracy, always include the <code> contextualLocation</code> values in the <a href=\"/api-docs/buy/static/api-browse.html#Headers\"> <code>X-EBAY-C-ENDUSERCTX</code></a> request header.</span>
+	// The end date of the delivery window (latest projected delivery date). This value is returned in UTC format (yyyy-MM-ddThh:mm:ss.sssZ), which you can convert into the local time of the buyer.<br><br><span class=\"tablenote\"> <b>Note: </b> For the best accuracy, always include the <code>contextualLocation</code> values in the <a href=\"/api-docs/buy/static/api-browse.html#Headers\" target=\"_blank\"><code>X-EBAY-C-ENDUSERCTX</code></a> request header.</span>
 	MaxEstimatedDeliveryDate *string `json:"maxEstimatedDeliveryDate,omitempty"`
-	// The start date of the delivery window (earliest projected delivery date).  This value is returned in UTC format (yyyy-MM-ddThh:mm:ss.sssZ), which you can convert into the local time of the buyer. <br /> <br /><span class=\"tablenote\"> <b> Note: </b> For the best accuracy, always include the <code> contextualLocation</code> values in the <a href=\"/api-docs/buy/static/api-browse.html#Headers\"> <code>X-EBAY-C-ENDUSERCTX</code></a> request header.</span>
+	// The start date of the delivery window (earliest projected delivery date). This value is returned in UTC format (yyyy-MM-ddThh:mm:ss.sssZ), which you can convert into the local time of the buyer.<br><br><span class=\"tablenote\"><b>Note:</b> For the best accuracy, always include the <code>contextualLocation</code> values in the <a href=\"/api-docs/buy/static/api-browse.html#Headers\" target=\"_blank\"><code>X-EBAY-C-ENDUSERCTX</code></a> request header.</span>
 	MinEstimatedDeliveryDate *string `json:"minEstimatedDeliveryDate,omitempty"`
 	ShippingCost *ConvertedAmount `json:"shippingCost,omitempty"`
-	// Indicates the type of shipping used to ship the item. Possible values are <code> FIXED</code> (flat-rate shipping) and <code> CALCULATED</code> (shipping cost calculated based on item and buyer location).
+	// Indicates the type of shipping used to ship the item. Possible values are <code>FIXED</code> (flat-rate shipping) and <code>CALCULATED</code> (shipping cost calculated based on item and buyer location).
 	ShippingCostType *string `json:"shippingCostType,omitempty"`
 }
 
@@ -46,7 +49,7 @@ func NewShippingOptionSummaryWithDefaults() *ShippingOptionSummary {
 
 // GetGuaranteedDelivery returns the GuaranteedDelivery field value if set, zero value otherwise.
 func (o *ShippingOptionSummary) GetGuaranteedDelivery() bool {
-	if o == nil || o.GuaranteedDelivery == nil {
+	if o == nil || IsNil(o.GuaranteedDelivery) {
 		var ret bool
 		return ret
 	}
@@ -56,7 +59,7 @@ func (o *ShippingOptionSummary) GetGuaranteedDelivery() bool {
 // GetGuaranteedDeliveryOk returns a tuple with the GuaranteedDelivery field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOptionSummary) GetGuaranteedDeliveryOk() (*bool, bool) {
-	if o == nil || o.GuaranteedDelivery == nil {
+	if o == nil || IsNil(o.GuaranteedDelivery) {
 		return nil, false
 	}
 	return o.GuaranteedDelivery, true
@@ -64,7 +67,7 @@ func (o *ShippingOptionSummary) GetGuaranteedDeliveryOk() (*bool, bool) {
 
 // HasGuaranteedDelivery returns a boolean if a field has been set.
 func (o *ShippingOptionSummary) HasGuaranteedDelivery() bool {
-	if o != nil && o.GuaranteedDelivery != nil {
+	if o != nil && !IsNil(o.GuaranteedDelivery) {
 		return true
 	}
 
@@ -78,7 +81,7 @@ func (o *ShippingOptionSummary) SetGuaranteedDelivery(v bool) {
 
 // GetMaxEstimatedDeliveryDate returns the MaxEstimatedDeliveryDate field value if set, zero value otherwise.
 func (o *ShippingOptionSummary) GetMaxEstimatedDeliveryDate() string {
-	if o == nil || o.MaxEstimatedDeliveryDate == nil {
+	if o == nil || IsNil(o.MaxEstimatedDeliveryDate) {
 		var ret string
 		return ret
 	}
@@ -88,7 +91,7 @@ func (o *ShippingOptionSummary) GetMaxEstimatedDeliveryDate() string {
 // GetMaxEstimatedDeliveryDateOk returns a tuple with the MaxEstimatedDeliveryDate field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOptionSummary) GetMaxEstimatedDeliveryDateOk() (*string, bool) {
-	if o == nil || o.MaxEstimatedDeliveryDate == nil {
+	if o == nil || IsNil(o.MaxEstimatedDeliveryDate) {
 		return nil, false
 	}
 	return o.MaxEstimatedDeliveryDate, true
@@ -96,7 +99,7 @@ func (o *ShippingOptionSummary) GetMaxEstimatedDeliveryDateOk() (*string, bool) 
 
 // HasMaxEstimatedDeliveryDate returns a boolean if a field has been set.
 func (o *ShippingOptionSummary) HasMaxEstimatedDeliveryDate() bool {
-	if o != nil && o.MaxEstimatedDeliveryDate != nil {
+	if o != nil && !IsNil(o.MaxEstimatedDeliveryDate) {
 		return true
 	}
 
@@ -110,7 +113,7 @@ func (o *ShippingOptionSummary) SetMaxEstimatedDeliveryDate(v string) {
 
 // GetMinEstimatedDeliveryDate returns the MinEstimatedDeliveryDate field value if set, zero value otherwise.
 func (o *ShippingOptionSummary) GetMinEstimatedDeliveryDate() string {
-	if o == nil || o.MinEstimatedDeliveryDate == nil {
+	if o == nil || IsNil(o.MinEstimatedDeliveryDate) {
 		var ret string
 		return ret
 	}
@@ -120,7 +123,7 @@ func (o *ShippingOptionSummary) GetMinEstimatedDeliveryDate() string {
 // GetMinEstimatedDeliveryDateOk returns a tuple with the MinEstimatedDeliveryDate field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOptionSummary) GetMinEstimatedDeliveryDateOk() (*string, bool) {
-	if o == nil || o.MinEstimatedDeliveryDate == nil {
+	if o == nil || IsNil(o.MinEstimatedDeliveryDate) {
 		return nil, false
 	}
 	return o.MinEstimatedDeliveryDate, true
@@ -128,7 +131,7 @@ func (o *ShippingOptionSummary) GetMinEstimatedDeliveryDateOk() (*string, bool) 
 
 // HasMinEstimatedDeliveryDate returns a boolean if a field has been set.
 func (o *ShippingOptionSummary) HasMinEstimatedDeliveryDate() bool {
-	if o != nil && o.MinEstimatedDeliveryDate != nil {
+	if o != nil && !IsNil(o.MinEstimatedDeliveryDate) {
 		return true
 	}
 
@@ -142,7 +145,7 @@ func (o *ShippingOptionSummary) SetMinEstimatedDeliveryDate(v string) {
 
 // GetShippingCost returns the ShippingCost field value if set, zero value otherwise.
 func (o *ShippingOptionSummary) GetShippingCost() ConvertedAmount {
-	if o == nil || o.ShippingCost == nil {
+	if o == nil || IsNil(o.ShippingCost) {
 		var ret ConvertedAmount
 		return ret
 	}
@@ -152,7 +155,7 @@ func (o *ShippingOptionSummary) GetShippingCost() ConvertedAmount {
 // GetShippingCostOk returns a tuple with the ShippingCost field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOptionSummary) GetShippingCostOk() (*ConvertedAmount, bool) {
-	if o == nil || o.ShippingCost == nil {
+	if o == nil || IsNil(o.ShippingCost) {
 		return nil, false
 	}
 	return o.ShippingCost, true
@@ -160,7 +163,7 @@ func (o *ShippingOptionSummary) GetShippingCostOk() (*ConvertedAmount, bool) {
 
 // HasShippingCost returns a boolean if a field has been set.
 func (o *ShippingOptionSummary) HasShippingCost() bool {
-	if o != nil && o.ShippingCost != nil {
+	if o != nil && !IsNil(o.ShippingCost) {
 		return true
 	}
 
@@ -174,7 +177,7 @@ func (o *ShippingOptionSummary) SetShippingCost(v ConvertedAmount) {
 
 // GetShippingCostType returns the ShippingCostType field value if set, zero value otherwise.
 func (o *ShippingOptionSummary) GetShippingCostType() string {
-	if o == nil || o.ShippingCostType == nil {
+	if o == nil || IsNil(o.ShippingCostType) {
 		var ret string
 		return ret
 	}
@@ -184,7 +187,7 @@ func (o *ShippingOptionSummary) GetShippingCostType() string {
 // GetShippingCostTypeOk returns a tuple with the ShippingCostType field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *ShippingOptionSummary) GetShippingCostTypeOk() (*string, bool) {
-	if o == nil || o.ShippingCostType == nil {
+	if o == nil || IsNil(o.ShippingCostType) {
 		return nil, false
 	}
 	return o.ShippingCostType, true
@@ -192,7 +195,7 @@ func (o *ShippingOptionSummary) GetShippingCostTypeOk() (*string, bool) {
 
 // HasShippingCostType returns a boolean if a field has been set.
 func (o *ShippingOptionSummary) HasShippingCostType() bool {
-	if o != nil && o.ShippingCostType != nil {
+	if o != nil && !IsNil(o.ShippingCostType) {
 		return true
 	}
 
@@ -205,23 +208,31 @@ func (o *ShippingOptionSummary) SetShippingCostType(v string) {
 }
 
 func (o ShippingOptionSummary) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.GuaranteedDelivery != nil {
-		toSerialize["guaranteedDelivery"] = o.GuaranteedDelivery
-	}
-	if o.MaxEstimatedDeliveryDate != nil {
-		toSerialize["maxEstimatedDeliveryDate"] = o.MaxEstimatedDeliveryDate
-	}
-	if o.MinEstimatedDeliveryDate != nil {
-		toSerialize["minEstimatedDeliveryDate"] = o.MinEstimatedDeliveryDate
-	}
-	if o.ShippingCost != nil {
-		toSerialize["shippingCost"] = o.ShippingCost
-	}
-	if o.ShippingCostType != nil {
-		toSerialize["shippingCostType"] = o.ShippingCostType
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o ShippingOptionSummary) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.GuaranteedDelivery) {
+		toSerialize["guaranteedDelivery"] = o.GuaranteedDelivery
+	}
+	if !IsNil(o.MaxEstimatedDeliveryDate) {
+		toSerialize["maxEstimatedDeliveryDate"] = o.MaxEstimatedDeliveryDate
+	}
+	if !IsNil(o.MinEstimatedDeliveryDate) {
+		toSerialize["minEstimatedDeliveryDate"] = o.MinEstimatedDeliveryDate
+	}
+	if !IsNil(o.ShippingCost) {
+		toSerialize["shippingCost"] = o.ShippingCost
+	}
+	if !IsNil(o.ShippingCostType) {
+		toSerialize["shippingCostType"] = o.ShippingCostType
+	}
+	return toSerialize, nil
 }
 
 type NullableShippingOptionSummary struct {
