@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -13,6 +13,9 @@ package buybrowse
 import (
 	"encoding/json"
 )
+
+// checks if the TargetLocation type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &TargetLocation{}
 
 // TargetLocation The type that defines the fields for the distance between the item location and the buyer's location. 
 type TargetLocation struct {
@@ -41,7 +44,7 @@ func NewTargetLocationWithDefaults() *TargetLocation {
 
 // GetUnitOfMeasure returns the UnitOfMeasure field value if set, zero value otherwise.
 func (o *TargetLocation) GetUnitOfMeasure() string {
-	if o == nil || o.UnitOfMeasure == nil {
+	if o == nil || IsNil(o.UnitOfMeasure) {
 		var ret string
 		return ret
 	}
@@ -51,7 +54,7 @@ func (o *TargetLocation) GetUnitOfMeasure() string {
 // GetUnitOfMeasureOk returns a tuple with the UnitOfMeasure field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *TargetLocation) GetUnitOfMeasureOk() (*string, bool) {
-	if o == nil || o.UnitOfMeasure == nil {
+	if o == nil || IsNil(o.UnitOfMeasure) {
 		return nil, false
 	}
 	return o.UnitOfMeasure, true
@@ -59,7 +62,7 @@ func (o *TargetLocation) GetUnitOfMeasureOk() (*string, bool) {
 
 // HasUnitOfMeasure returns a boolean if a field has been set.
 func (o *TargetLocation) HasUnitOfMeasure() bool {
-	if o != nil && o.UnitOfMeasure != nil {
+	if o != nil && !IsNil(o.UnitOfMeasure) {
 		return true
 	}
 
@@ -73,7 +76,7 @@ func (o *TargetLocation) SetUnitOfMeasure(v string) {
 
 // GetValue returns the Value field value if set, zero value otherwise.
 func (o *TargetLocation) GetValue() string {
-	if o == nil || o.Value == nil {
+	if o == nil || IsNil(o.Value) {
 		var ret string
 		return ret
 	}
@@ -83,7 +86,7 @@ func (o *TargetLocation) GetValue() string {
 // GetValueOk returns a tuple with the Value field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *TargetLocation) GetValueOk() (*string, bool) {
-	if o == nil || o.Value == nil {
+	if o == nil || IsNil(o.Value) {
 		return nil, false
 	}
 	return o.Value, true
@@ -91,7 +94,7 @@ func (o *TargetLocation) GetValueOk() (*string, bool) {
 
 // HasValue returns a boolean if a field has been set.
 func (o *TargetLocation) HasValue() bool {
-	if o != nil && o.Value != nil {
+	if o != nil && !IsNil(o.Value) {
 		return true
 	}
 
@@ -104,14 +107,22 @@ func (o *TargetLocation) SetValue(v string) {
 }
 
 func (o TargetLocation) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.UnitOfMeasure != nil {
-		toSerialize["unitOfMeasure"] = o.UnitOfMeasure
-	}
-	if o.Value != nil {
-		toSerialize["value"] = o.Value
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o TargetLocation) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.UnitOfMeasure) {
+		toSerialize["unitOfMeasure"] = o.UnitOfMeasure
+	}
+	if !IsNil(o.Value) {
+		toSerialize["value"] = o.Value
+	}
+	return toSerialize, nil
 }
 
 type NullableTargetLocation struct {
