@@ -1,9 +1,9 @@
 /*
 Browse API
 
-<p>The Browse API has the following resources:</p>   <ul> <li><b> item_summary: </b> Lets shoppers search for specific items by keyword, GTIN, category, charity, product, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values.</li>  <li><b> search_by_image: </b><a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> Lets shoppers search for specific items by image. You can refine the results by using URI parameters and filters.</li>   <li><b> item: </b> <ul><li>Lets you retrieve the details of a specific item or all the items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.</li> <li>Provides a bridge between the eBay legacy APIs, such as <b> Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li>  </ul> </li>  <li> <b> shopping_cart: </b> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#experimental\" target=\"_blank\"><img src=\"/cms/img/docs/experimental-icon.svg\" class=\"legend-icon experimental-icon\" alt=\"Experimental Release\" title=\"Experimental Release\" />&nbsp;(Experimental)</a> <a href=\"https://developer.ebay.com/api-docs/static/versioning.html#limited\" target=\"_blank\"> <img src=\"/cms/img/docs/partners-api.svg\" class=\"legend-icon partners-icon\" title=\"Limited Release\"  alt=\"Limited Release\" />(Limited Release)</a> Provides the ability for eBay members to see the contents of their eBay cart, and add, remove, and change the quantity of items in their eBay cart.&nbsp;&nbsp;<b> Note: </b> This resource is not available in the eBay API Explorer.</li></ul>       <p>The <b> item_summary</b>, <b> search_by_image</b>, and <b> item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\">Application access token</a>. The <b> shopping_cart</b> resource calls require a <a href=\"/api-docs/static/oauth-authorization-code-grant.html\">User access token</a>.</p>
+The Browse API has the following resources:<ul><li><b>item_summary:</b><br>Allows shoppers to search for specific items by keyword, GTIN, category, charity, product, image, or item aspects and refine the results by using filters, such as aspects, compatibility, and fields values, or UI parameters.</li><li><b>item:</b><br>Allows shoppers to retrieve the details of a specific item or all items in an item group, which is an item with variations such as color and size and check if a product is compatible with the specified item, such as if a specific car is compatible with a specific part.<br><br>This resource also provides a bridge between the eBay legacy APIs, such as the <a href=\"/api-docs/user-guides/static/finding-user-guide-landing.html\" target=\"_blank\">Finding</b>, and the RESTful APIs, which use different formats for the item IDs.</li></ul>The <b>item_summary</b>, <b>search_by_image</b>, and <b>item</b> resource calls require an <a href=\"/api-docs/static/oauth-client-credentials-grant.html\" target=\"_blank\">Application access token</a>.
 
-API version: v1.11.0
+API version: v1.19.9
 */
 
 // Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.
@@ -13,6 +13,9 @@ package buybrowse
 import (
 	"encoding/json"
 )
+
+// checks if the Taxes type satisfies the MappedNullable interface at compile time
+var _ MappedNullable = &Taxes{}
 
 // Taxes The type that defines the tax fields.
 type Taxes struct {
@@ -48,7 +51,7 @@ func NewTaxesWithDefaults() *Taxes {
 
 // GetEbayCollectAndRemitTax returns the EbayCollectAndRemitTax field value if set, zero value otherwise.
 func (o *Taxes) GetEbayCollectAndRemitTax() bool {
-	if o == nil || o.EbayCollectAndRemitTax == nil {
+	if o == nil || IsNil(o.EbayCollectAndRemitTax) {
 		var ret bool
 		return ret
 	}
@@ -58,7 +61,7 @@ func (o *Taxes) GetEbayCollectAndRemitTax() bool {
 // GetEbayCollectAndRemitTaxOk returns a tuple with the EbayCollectAndRemitTax field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Taxes) GetEbayCollectAndRemitTaxOk() (*bool, bool) {
-	if o == nil || o.EbayCollectAndRemitTax == nil {
+	if o == nil || IsNil(o.EbayCollectAndRemitTax) {
 		return nil, false
 	}
 	return o.EbayCollectAndRemitTax, true
@@ -66,7 +69,7 @@ func (o *Taxes) GetEbayCollectAndRemitTaxOk() (*bool, bool) {
 
 // HasEbayCollectAndRemitTax returns a boolean if a field has been set.
 func (o *Taxes) HasEbayCollectAndRemitTax() bool {
-	if o != nil && o.EbayCollectAndRemitTax != nil {
+	if o != nil && !IsNil(o.EbayCollectAndRemitTax) {
 		return true
 	}
 
@@ -80,7 +83,7 @@ func (o *Taxes) SetEbayCollectAndRemitTax(v bool) {
 
 // GetIncludedInPrice returns the IncludedInPrice field value if set, zero value otherwise.
 func (o *Taxes) GetIncludedInPrice() bool {
-	if o == nil || o.IncludedInPrice == nil {
+	if o == nil || IsNil(o.IncludedInPrice) {
 		var ret bool
 		return ret
 	}
@@ -90,7 +93,7 @@ func (o *Taxes) GetIncludedInPrice() bool {
 // GetIncludedInPriceOk returns a tuple with the IncludedInPrice field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Taxes) GetIncludedInPriceOk() (*bool, bool) {
-	if o == nil || o.IncludedInPrice == nil {
+	if o == nil || IsNil(o.IncludedInPrice) {
 		return nil, false
 	}
 	return o.IncludedInPrice, true
@@ -98,7 +101,7 @@ func (o *Taxes) GetIncludedInPriceOk() (*bool, bool) {
 
 // HasIncludedInPrice returns a boolean if a field has been set.
 func (o *Taxes) HasIncludedInPrice() bool {
-	if o != nil && o.IncludedInPrice != nil {
+	if o != nil && !IsNil(o.IncludedInPrice) {
 		return true
 	}
 
@@ -112,7 +115,7 @@ func (o *Taxes) SetIncludedInPrice(v bool) {
 
 // GetShippingAndHandlingTaxed returns the ShippingAndHandlingTaxed field value if set, zero value otherwise.
 func (o *Taxes) GetShippingAndHandlingTaxed() bool {
-	if o == nil || o.ShippingAndHandlingTaxed == nil {
+	if o == nil || IsNil(o.ShippingAndHandlingTaxed) {
 		var ret bool
 		return ret
 	}
@@ -122,7 +125,7 @@ func (o *Taxes) GetShippingAndHandlingTaxed() bool {
 // GetShippingAndHandlingTaxedOk returns a tuple with the ShippingAndHandlingTaxed field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Taxes) GetShippingAndHandlingTaxedOk() (*bool, bool) {
-	if o == nil || o.ShippingAndHandlingTaxed == nil {
+	if o == nil || IsNil(o.ShippingAndHandlingTaxed) {
 		return nil, false
 	}
 	return o.ShippingAndHandlingTaxed, true
@@ -130,7 +133,7 @@ func (o *Taxes) GetShippingAndHandlingTaxedOk() (*bool, bool) {
 
 // HasShippingAndHandlingTaxed returns a boolean if a field has been set.
 func (o *Taxes) HasShippingAndHandlingTaxed() bool {
-	if o != nil && o.ShippingAndHandlingTaxed != nil {
+	if o != nil && !IsNil(o.ShippingAndHandlingTaxed) {
 		return true
 	}
 
@@ -144,7 +147,7 @@ func (o *Taxes) SetShippingAndHandlingTaxed(v bool) {
 
 // GetTaxJurisdiction returns the TaxJurisdiction field value if set, zero value otherwise.
 func (o *Taxes) GetTaxJurisdiction() TaxJurisdiction {
-	if o == nil || o.TaxJurisdiction == nil {
+	if o == nil || IsNil(o.TaxJurisdiction) {
 		var ret TaxJurisdiction
 		return ret
 	}
@@ -154,7 +157,7 @@ func (o *Taxes) GetTaxJurisdiction() TaxJurisdiction {
 // GetTaxJurisdictionOk returns a tuple with the TaxJurisdiction field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Taxes) GetTaxJurisdictionOk() (*TaxJurisdiction, bool) {
-	if o == nil || o.TaxJurisdiction == nil {
+	if o == nil || IsNil(o.TaxJurisdiction) {
 		return nil, false
 	}
 	return o.TaxJurisdiction, true
@@ -162,7 +165,7 @@ func (o *Taxes) GetTaxJurisdictionOk() (*TaxJurisdiction, bool) {
 
 // HasTaxJurisdiction returns a boolean if a field has been set.
 func (o *Taxes) HasTaxJurisdiction() bool {
-	if o != nil && o.TaxJurisdiction != nil {
+	if o != nil && !IsNil(o.TaxJurisdiction) {
 		return true
 	}
 
@@ -176,7 +179,7 @@ func (o *Taxes) SetTaxJurisdiction(v TaxJurisdiction) {
 
 // GetTaxPercentage returns the TaxPercentage field value if set, zero value otherwise.
 func (o *Taxes) GetTaxPercentage() string {
-	if o == nil || o.TaxPercentage == nil {
+	if o == nil || IsNil(o.TaxPercentage) {
 		var ret string
 		return ret
 	}
@@ -186,7 +189,7 @@ func (o *Taxes) GetTaxPercentage() string {
 // GetTaxPercentageOk returns a tuple with the TaxPercentage field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Taxes) GetTaxPercentageOk() (*string, bool) {
-	if o == nil || o.TaxPercentage == nil {
+	if o == nil || IsNil(o.TaxPercentage) {
 		return nil, false
 	}
 	return o.TaxPercentage, true
@@ -194,7 +197,7 @@ func (o *Taxes) GetTaxPercentageOk() (*string, bool) {
 
 // HasTaxPercentage returns a boolean if a field has been set.
 func (o *Taxes) HasTaxPercentage() bool {
-	if o != nil && o.TaxPercentage != nil {
+	if o != nil && !IsNil(o.TaxPercentage) {
 		return true
 	}
 
@@ -208,7 +211,7 @@ func (o *Taxes) SetTaxPercentage(v string) {
 
 // GetTaxType returns the TaxType field value if set, zero value otherwise.
 func (o *Taxes) GetTaxType() string {
-	if o == nil || o.TaxType == nil {
+	if o == nil || IsNil(o.TaxType) {
 		var ret string
 		return ret
 	}
@@ -218,7 +221,7 @@ func (o *Taxes) GetTaxType() string {
 // GetTaxTypeOk returns a tuple with the TaxType field value if set, nil otherwise
 // and a boolean to check if the value has been set.
 func (o *Taxes) GetTaxTypeOk() (*string, bool) {
-	if o == nil || o.TaxType == nil {
+	if o == nil || IsNil(o.TaxType) {
 		return nil, false
 	}
 	return o.TaxType, true
@@ -226,7 +229,7 @@ func (o *Taxes) GetTaxTypeOk() (*string, bool) {
 
 // HasTaxType returns a boolean if a field has been set.
 func (o *Taxes) HasTaxType() bool {
-	if o != nil && o.TaxType != nil {
+	if o != nil && !IsNil(o.TaxType) {
 		return true
 	}
 
@@ -239,26 +242,34 @@ func (o *Taxes) SetTaxType(v string) {
 }
 
 func (o Taxes) MarshalJSON() ([]byte, error) {
-	toSerialize := map[string]interface{}{}
-	if o.EbayCollectAndRemitTax != nil {
-		toSerialize["ebayCollectAndRemitTax"] = o.EbayCollectAndRemitTax
-	}
-	if o.IncludedInPrice != nil {
-		toSerialize["includedInPrice"] = o.IncludedInPrice
-	}
-	if o.ShippingAndHandlingTaxed != nil {
-		toSerialize["shippingAndHandlingTaxed"] = o.ShippingAndHandlingTaxed
-	}
-	if o.TaxJurisdiction != nil {
-		toSerialize["taxJurisdiction"] = o.TaxJurisdiction
-	}
-	if o.TaxPercentage != nil {
-		toSerialize["taxPercentage"] = o.TaxPercentage
-	}
-	if o.TaxType != nil {
-		toSerialize["taxType"] = o.TaxType
+	toSerialize,err := o.ToMap()
+	if err != nil {
+		return []byte{}, err
 	}
 	return json.Marshal(toSerialize)
+}
+
+func (o Taxes) ToMap() (map[string]interface{}, error) {
+	toSerialize := map[string]interface{}{}
+	if !IsNil(o.EbayCollectAndRemitTax) {
+		toSerialize["ebayCollectAndRemitTax"] = o.EbayCollectAndRemitTax
+	}
+	if !IsNil(o.IncludedInPrice) {
+		toSerialize["includedInPrice"] = o.IncludedInPrice
+	}
+	if !IsNil(o.ShippingAndHandlingTaxed) {
+		toSerialize["shippingAndHandlingTaxed"] = o.ShippingAndHandlingTaxed
+	}
+	if !IsNil(o.TaxJurisdiction) {
+		toSerialize["taxJurisdiction"] = o.TaxJurisdiction
+	}
+	if !IsNil(o.TaxPercentage) {
+		toSerialize["taxPercentage"] = o.TaxPercentage
+	}
+	if !IsNil(o.TaxType) {
+		toSerialize["taxType"] = o.TaxType
+	}
+	return toSerialize, nil
 }
 
 type NullableTaxes struct {
